% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{abs}
\alias{abs}
\alias{@}
\title{DuckDB function abs}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
Absolute value.
}
\section{Overloads}{

\itemize{
\item \code{abs(x = TINYINT)}: Absolute value.
\item \code{abs(x = SMALLINT)}: Absolute value.
\item \code{abs(x = INTEGER)}: Absolute value.
\item \code{abs(x = BIGINT)}: Absolute value.
\item \code{abs(x = HUGEINT)}: Absolute value.
\item \code{abs(x = FLOAT)}: Absolute value.
\item \code{abs(x = DOUBLE)}: Absolute value.
\item \code{abs(x = DECIMAL)}: Absolute value.
\item \code{abs(x = UTINYINT)}: Absolute value.
\item \code{abs(x = USMALLINT)}: Absolute value.
\item \code{abs(x = UINTEGER)}: Absolute value.
\item \code{abs(x = UBIGINT)}: Absolute value.
\item \code{abs(x = UHUGEINT)}: Absolute value.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{abs(-17.4)
}\if{html}{\out{</div>}}
}

