% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{acos}
\alias{acos}
\title{DuckDB function acos}
\usage{
acos(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the arccosine of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{acos(0.5)
}\if{html}{\out{</div>}}
}

