% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{acosh}
\alias{acosh}
\alias{duckdb_fun_acosh}
\title{DuckDB function acosh}
\arguments{
\item{x}{\code{DOUBLE}}
}
\description{
Computes the inverse hyperbolic cos of x.
}
\examples{
\dontrun{
acosh(2.3)
}
}
