% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{add_parquet_key}
\alias{add_parquet_key}
\title{DuckDB function add_parquet_key}
\usage{
add_parquet_key(col0 = VARCHAR, col1 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{add_parquet_key()}.
}
