% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{age}
\alias{age}
\title{DuckDB function age}
\usage{
age(timestamp, timestamp.1)
}
\arguments{
\item{timestamp, timestamp.1}{\verb{TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{INTERVAL}
}
\description{
Subtract arguments, resulting in the time difference between the two timestamps.
}
\section{Overloads}{

\itemize{
\item \code{age(timestamp = TIMESTAMP)}
\item \code{age(timestamp = TIMESTAMP, timestamp.1 = TIMESTAMP)}
\item \code{age(timestamp = `TIMESTAMP WITH TIME ZONE`)}
\item \code{age(timestamp = `TIMESTAMP WITH TIME ZONE`, timestamp.1 = `TIMESTAMP WITH TIME ZONE`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{age(TIMESTAMP '2001-04-10', TIMESTAMP '1992-09-20')
}\if{html}{\out{</div>}}
}

