% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{alias}
\alias{alias}
\alias{duckdb_fun_alias}
\title{DuckDB function alias}
\arguments{
\item{expr}{\code{ANY}}
}
\description{
Returns the name of a given expression.
}
\examples{
\dontrun{
alias(42 + 1)
}
}
