% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{&}
\alias{&}
\title{DuckDB function &}
\arguments{
\item{left}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}

\item{right}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise AND.
}
\section{Overloads}{

\itemize{
\item \code{`&`(left = TINYINT, right = TINYINT)}: Bitwise AND.
\item \code{`&`(left = SMALLINT, right = SMALLINT)}: Bitwise AND.
\item \code{`&`(left = INTEGER, right = INTEGER)}: Bitwise AND.
\item \code{`&`(left = BIGINT, right = BIGINT)}: Bitwise AND.
\item \code{`&`(left = HUGEINT, right = HUGEINT)}: Bitwise AND.
\item \code{`&`(left = UTINYINT, right = UTINYINT)}: Bitwise AND.
\item \code{`&`(left = USMALLINT, right = USMALLINT)}: Bitwise AND.
\item \code{`&`(left = UINTEGER, right = UINTEGER)}: Bitwise AND.
\item \code{`&`(left = UBIGINT, right = UBIGINT)}: Bitwise AND.
\item \code{`&`(left = UHUGEINT, right = UHUGEINT)}: Bitwise AND.
\item \code{`&`(left = BIT, right = BIT)}: Bitwise AND.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{91 & 15
}\if{html}{\out{</div>}}
}

