% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{any_value}
\alias{any_value}
\title{DuckDB function any_value}
\arguments{
\item{arg}{\code{DECIMAL | ANY}}
}
\value{
\code{DECIMAL | ANY}
}
\description{
Returns the first non-NULL value from arg. This function is affected by ordering.
}
\section{Overloads}{

\itemize{
\item \code{any_value(arg = DECIMAL)}
\item \code{any_value(arg = ANY)}
}
}

