% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{approx_count_distinct}
\alias{approx_count_distinct}
\title{DuckDB function approx_count_distinct}
\usage{
approx_count_distinct(any = ANY)
}
\arguments{
\item{any}{\code{ANY}}
}
\value{
\code{BIGINT}
}
\description{
Computes the approximate count of distinct elements using HyperLogLog.
}
\examples{
\dontrun{
approx_count_distinct(A)
}
}
