% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arbitrary}
\alias{arbitrary}
\title{DuckDB function arbitrary}
\arguments{
\item{arg}{\code{DECIMAL | ANY}}
}
\value{
\code{DECIMAL | ANY}
}
\description{
Returns the first value (NULL or non-NULL) from arg. This function is affected by ordering.
}
\section{Overloads}{

\itemize{
\item \code{arbitrary(arg = DECIMAL)}
\item \code{arbitrary(arg = ANY)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{arbitrary(A)
}\if{html}{\out{</div>}}
}

