% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arg_max}
\alias{arg_max}
\alias{argmax}
\alias{max_by}
\title{DuckDB function arg_max}
\arguments{
\item{arg}{\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}}

\item{val}{\verb{INTEGER | BIGINT | HUGEINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | ANY}}

\item{col2}{\code{BIGINT}}
}
\value{
\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY | ANY[]}
}
\description{
Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
}
\section{Overloads}{

\itemize{
\item \code{arg_max(arg = INTEGER, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = INTEGER, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = INTEGER, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = INTEGER, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = INTEGER, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = INTEGER, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = INTEGER, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = INTEGER, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = INTEGER, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BIGINT, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DOUBLE, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = VARCHAR, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DATE, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = TIMESTAMP, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = BLOB, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = DECIMAL, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = INTEGER)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = HUGEINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = DOUBLE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = VARCHAR)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = DATE)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = BLOB)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = ANY)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
\item \code{arg_max(arg = ANY, val = ANY, col2 = BIGINT)}: Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{arg_max(A, B)
}\if{html}{\out{</div>}}
}

