% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arg_max_null}
\alias{arg_max_null}
\title{DuckDB function arg_max_null}
\arguments{
\item{arg}{\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}}

\item{val}{\verb{INTEGER | BIGINT | HUGEINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | ANY}}
}
\value{
\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}
}
\description{
Finds the row with the maximum val. Calculates the arg expression at that row.
}
\section{Overloads}{

\itemize{
\item \code{arg_max_null(arg = INTEGER, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = INTEGER, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = INTEGER, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = INTEGER, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = INTEGER, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = INTEGER, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = INTEGER, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = INTEGER, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = INTEGER, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BIGINT, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DOUBLE, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = VARCHAR, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DATE, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = TIMESTAMP, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = BLOB, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = DECIMAL, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = INTEGER)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = BIGINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = HUGEINT)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = DOUBLE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = VARCHAR)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = DATE)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = TIMESTAMP)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = `TIMESTAMP WITH TIME ZONE`)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = BLOB)}: Finds the row with the maximum val. Calculates the arg expression at that row.
\item \code{arg_max_null(arg = ANY, val = ANY)}: Finds the row with the maximum val. Calculates the arg expression at that row.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{arg_max_null(A, B)
}\if{html}{\out{</div>}}
}

