% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arg_max_nulls_last}
\alias{arg_max_nulls_last}
\title{DuckDB function arg_max_nulls_last}
\arguments{
\item{arg}{\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}}

\item{val}{\verb{INTEGER | BIGINT | HUGEINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | ANY}}

\item{N}{\code{BIGINT}}
}
\value{
\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY | ANY[]}
}
\description{
Finds the rows with N maximum vals, including nulls. Calculates the arg expression at that row.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{arg_min_null_val(A, B, N)
}\if{html}{\out{</div>}}
}

