% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_aggregate}
\alias{array_aggregate}
\title{DuckDB function array_aggregate}
\usage{
array_aggregate(list = `ANY[]`, function_name = VARCHAR)
}
\arguments{
\item{list}{\code{ANY[]}}

\item{function_name}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
Executes the aggregate function \code{function_name} on the elements of \code{list}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_aggregate([1, 2, NULL], 'min')
}\if{html}{\out{</div>}}
}

