% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_append}
\alias{array_append}
\title{DuckDB function array_append}
\usage{
array_append(arr, el)
}
\arguments{
\item{arr}{Unspecified.}

\item{el}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{array_append()}.
}
