% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_cosine_distance}
\alias{array_cosine_distance}
\title{DuckDB function array_cosine_distance}
\usage{
array_cosine_distance(array1, array2)
}
\arguments{
\item{array1, array2}{\code{FLOAT[ANY] | DOUBLE[ANY]}}
}
\value{
\code{FLOAT | DOUBLE}
}
\description{
Computes the cosine distance between two arrays of the same size. The array elements can not be \code{NULL}. The arrays can have any size as long as the size is the same for both arguments.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_cosine_distance(array_value(1.0::FLOAT, 2.0::FLOAT, 3.0::FLOAT), array_value(2.0::FLOAT, 3.0::FLOAT, 4.0::FLOAT))
}\if{html}{\out{</div>}}
}

\seealso{
Other array:
\code{\link[=array_cosine_similarity]{array_cosine_similarity()}},
\code{\link[=array_cross_product]{array_cross_product()}},
\code{\link[=array_distance]{array_distance()}},
\code{\link[=array_inner_product]{array_inner_product()}},
\code{\link[=array_negative_inner_product]{array_negative_inner_product()}},
\code{\link[=array_value]{array_value()}}
}
\concept{array}
