% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_cross_product}
\alias{array_cross_product}
\title{DuckDB function array_cross_product}
\arguments{
\item{array}{\code{FLOAT[3] | DOUBLE[3]}}
}
\value{
\code{FLOAT[3] | DOUBLE[3]}
}
\description{
Computes the cross product of two arrays of size 3. The array elements can not be \code{NULL}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_cross_product(array_value(1.0::FLOAT, 2.0::FLOAT, 3.0::FLOAT), array_value(2.0::FLOAT, 3.0::FLOAT, 4.0::FLOAT))
}\if{html}{\out{</div>}}
}

\seealso{
Other array:
\code{\link[=array_cosine_distance]{array_cosine_distance()}},
\code{\link[=array_cosine_similarity]{array_cosine_similarity()}},
\code{\link[=array_distance]{array_distance()}},
\code{\link[=array_inner_product]{array_inner_product()}},
\code{\link[=array_negative_inner_product]{array_negative_inner_product()}},
\code{\link[=array_value]{array_value()}}
}
\concept{array}
