% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_filter}
\alias{array_filter}
\title{DuckDB function array_filter}
\arguments{
\item{list}{\code{ANY[]}}

\item{lambda}{\code{LAMBDA}}
}
\description{
Constructs a list from those elements of the input list for which the lambda function returns true.
}
\examples{
\dontrun{
list_filter([3, 4, 5], x -> x > 4)
}
}
