% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_grade_up}
\alias{array_grade_up}
\title{DuckDB function array_grade_up}
\arguments{
\item{list}{\code{ANY[]}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR}}
}
\value{
\code{ANY[]}
}
\description{
Works like list_sort, but the results are the indexes that correspond to the position in the original list instead of the actual values.
}
\section{Overloads}{

\itemize{
\item \code{array_grade_up(list = `ANY[]`)}
\item \code{array_grade_up(list = `ANY[]`, col1 = VARCHAR)}
\item \code{array_grade_up(list = `ANY[]`, col1 = VARCHAR, col2 = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_grade_up([3, 6, 1, 2])
}\if{html}{\out{</div>}}
}

