% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_has_all}
\alias{array_has_all}
\title{DuckDB function array_has_all}
\arguments{
\item{list1}{\code{T[]}}

\item{list2}{\code{T[]}}
}
\description{
Returns true if all elements of list2 are in list1. NULLs are ignored.
}
\examples{
\dontrun{
array_has_all([1, 2, 3], [2, 3])
}
}
