% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_has_any}
\alias{array_has_any}
\title{DuckDB function array_has_any}
\usage{
array_has_any(list1 = `T[]`, list2 = `T[]`)
}
\arguments{
\item{list1}{\code{T[]}}

\item{list2}{\code{T[]}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the lists have any element in common. NULLs are ignored.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_has_any([1, 2, 3], [2, 3, 4])
}\if{html}{\out{</div>}}
}

