% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_intersect}
\alias{array_intersect}
\title{DuckDB function array_intersect}
\usage{
array_intersect(l1, l2)
}
\arguments{
\item{l1}{Unspecified.}

\item{l2}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{array_intersect()}.
}
