% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_pop_back}
\alias{array_pop_back}
\title{DuckDB function array_pop_back}
\usage{
array_pop_back(arr)
}
\arguments{
\item{arr}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{array_pop_back()}.
}
