% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_position}
\alias{array_position}
\title{DuckDB function array_position}
\usage{
array_position(list = `T[]`, element = T)
}
\arguments{
\item{list}{\code{T[]}}

\item{element}{\code{T}}
}
\value{
\code{INTEGER}
}
\description{
Returns the index of the \code{element} if the \code{list} contains the \code{element}. If the \code{element} is not found, it returns \code{NULL}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_position([1, 2, NULL], 2)
}\if{html}{\out{</div>}}
}

