% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_prepend}
\alias{array_prepend}
\title{DuckDB function array_prepend}
\usage{
array_prepend(el, arr)
}
\arguments{
\item{el}{Unspecified.}

\item{arr}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{array_prepend()}.
}
