% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_push_back}
\alias{array_push_back}
\title{DuckDB function array_push_back}
\usage{
array_push_back(arr, e)
}
\arguments{
\item{arr}{Unspecified.}

\item{e}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{array_push_back()}.
}
