% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_reverse}
\alias{array_reverse}
\title{DuckDB function array_reverse}
\usage{
array_reverse(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{array_reverse()}.
}
