% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_reverse_sort}
\alias{array_reverse_sort}
\title{DuckDB function array_reverse_sort}
\arguments{
\item{list}{\code{ANY[]}}

\item{col1}{\code{VARCHAR}}
}
\value{
\code{ANY[]}
}
\description{
Sorts the elements of the list in reverse order.
}
\section{Overloads}{

\itemize{
\item \code{array_reverse_sort(list = `ANY[]`)}
\item \code{array_reverse_sort(list = `ANY[]`, col1 = VARCHAR)}
}
}

\examples{
\dontrun{
array_reverse_sort([3, 6, 1, 2])
}
}
