% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_slice}
\alias{array_slice}
\title{DuckDB function array_slice}
\arguments{
\item{list}{\code{ANY}}

\item{begin}{\code{ANY}}

\item{end}{\code{ANY}}

\item{step}{\code{BIGINT}}
}
\value{
\code{ANY}
}
\description{
Extracts a sublist or substring using slice conventions. Negative values are accepted.

list_slice with added step feature.
}
\examples{
\dontrun{
c("array_slice('DuckDB', 3, 4)", "array_slice('DuckDB', 3, NULL)", "array_slice('DuckDB', 0, -3)")
array_slice([4, 5, 6], 1, 3, 2)
}
}
