% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_value}
\alias{array_value}
\title{DuckDB function array_value}
\usage{
array_value()
}
\value{
\code{ARRAY}
}
\description{
Creates an \code{ARRAY} containing the argument values.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_value(1.0::FLOAT, 2.0::FLOAT, 3.0::FLOAT)
}\if{html}{\out{</div>}}
}

\seealso{
Other array:
\code{\link[=array_cosine_distance]{array_cosine_distance()}},
\code{\link[=array_cosine_similarity]{array_cosine_similarity()}},
\code{\link[=array_cross_product]{array_cross_product()}},
\code{\link[=array_distance]{array_distance()}},
\code{\link[=array_inner_product]{array_inner_product()}},
\code{\link[=array_negative_inner_product]{array_negative_inner_product()}}
}
\concept{array}
