% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_where}
\alias{array_where}
\title{DuckDB function array_where}
\arguments{
\item{value_list}{\code{ANY[]}}

\item{mask_list}{\code{BOOLEAN[]}}
}
\description{
Returns a list with the BOOLEANs in mask_list applied as a mask to the value_list.
}
\examples{
\dontrun{
list_where([10, 20, 30, 40], [true, false, false, true])
}
}
