% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arrow->>}
\alias{arrow->>}
\alias{->>}
\title{DuckDB function ->>}
\usage{
`->>`(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{BIGINT | VARCHAR | VARCHAR[]}}
}
\value{
\code{VARCHAR | VARCHAR[]}
}
\description{
DuckDB function \verb{->>()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

