% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arrow_scan_dumb}
\alias{arrow_scan_dumb}
\title{DuckDB function arrow_scan_dumb}
\usage{
arrow_scan_dumb(col0, col1, col2)
}
\arguments{
\item{col0, col1, col2}{\code{POINTER}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{arrow_scan_dumb()}.
}
