% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{asin}
\alias{asin}
\title{DuckDB function asin}
\usage{
asin(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the arcsine of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{asin(0.5)
}\if{html}{\out{</div>}}
}

