% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{asinh}
\alias{asinh}
\title{DuckDB function asinh}
\usage{
asinh(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the inverse hyperbolic sin of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{asinh(0.5)
}\if{html}{\out{</div>}}
}

