% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{at-}
\alias{at-}
\alias{@}
\title{DuckDB function @}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
Absolute value.
}
\section{Overloads}{

\itemize{
\item ```@`(x = TINYINT)``
\item ```@`(x = SMALLINT)``
\item ```@`(x = INTEGER)``
\item ```@`(x = BIGINT)``
\item ```@`(x = HUGEINT)``
\item ```@`(x = FLOAT)``
\item ```@`(x = DOUBLE)``
\item ```@`(x = DECIMAL)``
\item ```@`(x = UTINYINT)``
\item ```@`(x = USMALLINT)``
\item ```@`(x = UINTEGER)``
\item ```@`(x = UBIGINT)``
\item ```@`(x = UHUGEINT)``
}
}

\examples{
\dontrun{
abs(-17.4)
}
}
