% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{atan}
\alias{atan}
\title{DuckDB function atan}
\usage{
atan(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the arctangent of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{atan(0.5)
}\if{html}{\out{</div>}}
}

