% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{atanh}
\alias{atanh}
\title{DuckDB function atanh}
\usage{
atanh(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the inverse hyperbolic tan of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{atanh(0.5)
}\if{html}{\out{</div>}}
}

