% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{avg}
\alias{avg}
\alias{mean}
\title{DuckDB function avg}
\arguments{
\item{x}{\verb{DECIMAL | SMALLINT | INTEGER | BIGINT | HUGEINT | INTERVAL | DOUBLE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TIME | TIME WITH TIME ZONE}}
}
\value{
\verb{DECIMAL | DOUBLE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TIME | TIME WITH TIME ZONE}
}
\description{
Calculates the average value for all tuples in x.
}
\section{Overloads}{

\itemize{
\item \code{avg(x = DECIMAL)}: Calculates the average value for all tuples in x.
\item \code{avg(x = SMALLINT)}: Calculates the average value for all tuples in x.
\item \code{avg(x = INTEGER)}: Calculates the average value for all tuples in x.
\item \code{avg(x = BIGINT)}: Calculates the average value for all tuples in x.
\item \code{avg(x = HUGEINT)}: Calculates the average value for all tuples in x.
\item \code{avg(x = INTERVAL)}: Calculates the average value for all tuples in x.
\item \code{avg(x = DOUBLE)}: Calculates the average value for all tuples in x.
\item \code{avg(x = TIMESTAMP)}: Calculates the average value for all tuples in x.
\item \code{avg(x = `TIMESTAMP WITH TIME ZONE`)}: Calculates the average value for all tuples in x.
\item \code{avg(x = TIME)}: Calculates the average value for all tuples in x.
\item \code{avg(x = `TIME WITH TIME ZONE`)}: Calculates the average value for all tuples in x.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SUM(x) / COUNT(*)
}\if{html}{\out{</div>}}
}

