% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_and}
\alias{bit_and}
\title{DuckDB function bit_and}
\arguments{
\item{arg}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Returns the bitwise AND of all bits in a given expression.
}
\section{Overloads}{

\itemize{
\item \code{bit_and(arg = TINYINT)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = SMALLINT)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = INTEGER)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = BIGINT)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = HUGEINT)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = UTINYINT)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = USMALLINT)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = UINTEGER)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = UBIGINT)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = UHUGEINT)}: Returns the bitwise AND of all bits in a given expression.
\item \code{bit_and(arg = BIT)}: Returns the bitwise AND of all bits in a given expression.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bit_and(A)
}\if{html}{\out{</div>}}
}

