% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_count}
\alias{bit_count}
\title{DuckDB function bit_count}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | BIT}}
}
\value{
\code{TINYINT | BIGINT}
}
\description{
Returns the number of bits that are set.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bit_count(31)
}\if{html}{\out{</div>}}
}

