% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_position}
\alias{bit_position}
\title{DuckDB function bit_position}
\usage{
bit_position(substring = BIT, bitstring = BIT)
}
\arguments{
\item{substring}{\code{BIT}}

\item{bitstring}{\code{BIT}}
}
\value{
\code{INTEGER}
}
\description{
Returns first starting index of the specified substring within bits, or zero if it is not present. The first (leftmost) bit is indexed 1.
}
\examples{
\dontrun{
bit_position('010'::BIT, '1110101'::BIT)
}
}
