% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bitstring}
\alias{bitstring}
\title{DuckDB function bitstring}
\arguments{
\item{bitstring}{\code{VARCHAR | BIT}}

\item{length}{\code{INTEGER}}
}
\value{
\code{BIT}
}
\description{
Pads the bitstring until the specified length.
}
\examples{
\dontrun{
bitstring('1010'::BIT, 7)
}
}
