% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bitstring_agg}
\alias{bitstring_agg}
\title{DuckDB function bitstring_agg}
\arguments{
\item{arg}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}

\item{col1}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}

\item{col2}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{BIT}
}
\description{
Returns a bitstring with bits set for each distinct value.
}
\section{Overloads}{

\itemize{
\item \code{bitstring_agg(arg = TINYINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = TINYINT, col1 = TINYINT, col2 = TINYINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = SMALLINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = SMALLINT, col1 = SMALLINT, col2 = SMALLINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = INTEGER)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = INTEGER, col1 = INTEGER, col2 = INTEGER)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = BIGINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = BIGINT, col1 = BIGINT, col2 = BIGINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = HUGEINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = HUGEINT, col1 = HUGEINT, col2 = HUGEINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = UTINYINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = UTINYINT, col1 = UTINYINT, col2 = UTINYINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = USMALLINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = USMALLINT, col1 = USMALLINT, col2 = USMALLINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = UINTEGER)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = UINTEGER, col1 = UINTEGER, col2 = UINTEGER)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = UBIGINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = UBIGINT, col1 = UBIGINT, col2 = UBIGINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = UHUGEINT)}: Returns a bitstring with bits set for each distinct value.
\item \code{bitstring_agg(arg = UHUGEINT, col1 = UHUGEINT, col2 = UHUGEINT)}: Returns a bitstring with bits set for each distinct value.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bitstring_agg(A)
}\if{html}{\out{</div>}}
}

