% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{broadcast}
\alias{broadcast}
\title{DuckDB function broadcast}
\usage{
broadcast(col0)
}
\arguments{
\item{col0}{\code{INET}}
}
\value{
\code{INET}
}
\description{
DuckDB function \code{broadcast()}.
}
\section{Provided by}{

The \code{inet} extension (\code{LOAD inet;}).
}

