% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{can_cast_implicitly}
\alias{can_cast_implicitly}
\alias{duckdb_fun_can_cast_implicitly}
\title{DuckDB function can_cast_implicitly}
\arguments{
\item{source_type}{\code{ANY}}

\item{target_type}{\code{ANY}}
}
\description{
Whether or not we can implicitly cast from the source type to the other type.
}
\examples{
\dontrun{
can_implicitly_cast(NULL::INTEGER, NULL::BIGINT)
}
}
