% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{cardinality}
\alias{cardinality}
\title{DuckDB function cardinality}
\usage{
cardinality(map = ANY)
}
\arguments{
\item{map}{\code{ANY}}
}
\value{
\code{UBIGINT}
}
\description{
Returns the size of the map (or the number of entries in the map).
}
\examples{
\dontrun{
cardinality( map([4, 2], ['a', 'b']) );
}
}
