% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{cast_to_type}
\alias{cast_to_type}
\title{DuckDB function cast_to_type}
\arguments{
\item{param}{\code{ANY}}

\item{type}{\code{ANY}}
}
\description{
Casts the first argument to the type of the second argument.
}
\examples{
\dontrun{
cast_to_type('42', NULL::INTEGER)
}
}
