% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ceil}
\alias{ceil}
\alias{ceiling}
\title{DuckDB function ceil}
\arguments{
\item{x}{\code{FLOAT | DOUBLE | DECIMAL}}
}
\value{
\code{FLOAT | DOUBLE | DECIMAL}
}
\description{
Rounds the number up.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{ceil(17.4)
}\if{html}{\out{</div>}}
}

