% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ceiling}
\alias{ceiling}
\title{DuckDB function ceiling}
\arguments{
\item{x}{\code{FLOAT | DOUBLE | DECIMAL}}
}
\value{
\code{FLOAT | DOUBLE | DECIMAL}
}
\description{
Rounds the number up.
}
\section{Overloads}{

\itemize{
\item \code{ceiling(x = FLOAT)}
\item \code{ceiling(x = DOUBLE)}
\item \code{ceiling(x = DECIMAL)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{ceiling(17.4)
}\if{html}{\out{</div>}}
}

