% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{character_length}
\alias{character_length}
\title{DuckDB function character_length}
\arguments{
\item{string}{\code{VARCHAR}}

\item{bit}{\code{BIT}}

\item{list}{\code{ANY[]}}
}
\value{
\code{BIGINT}
}
\description{
Number of characters in \code{string}.

Returns the bit-length of the \code{bit} argument.

Returns the length of the \code{list}.
}
\section{Overloads}{

\itemize{
\item \code{character_length(string = VARCHAR)}
\item \code{character_length(bit = BIT)}
\item \code{character_length(list = `ANY[]`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{character_length('Hello🦆')
character_length(42::TINYINT::BIT)
character_length([1,2,3])
}\if{html}{\out{</div>}}
}

