% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{check_peg_parser}
\alias{check_peg_parser}
\title{DuckDB function check_peg_parser}
\usage{
check_peg_parser(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{check_peg_parser()}.
}
\section{Provided by}{

The \code{autocomplete} extension (\code{LOAD autocomplete;}).
}

