% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_treattime.R
\name{code_eventtime}
\alias{code_eventtime}
\title{Code time-to-event variable}
\usage{
code_eventtime(
  unit,
  time,
  treat,
  data,
  baseline = NA,
  never = "min",
  always = "max",
  leads = NA,
  lags = NA
)
}
\arguments{
\item{unit}{unquoted variable name identifying the unit of the panel}

\item{time}{unquoted variable name indicating the ordering of the observations}

\item{treat}{unquoted variable name of the binary treatment status variable}

\item{data}{data frame that includes the variables}

\item{baseline}{baseline or reference period for leads/lags}

\item{never}{coding for units for which never \code{treat=1}}

\item{always}{coding for units for which always \code{treat=1}}

\item{leads}{number of leads to include}

\item{lags}{number of lags to include}
}
\value{
A vector measuring time to event in the units of the panel dataset.
}
\description{
The function constructs a variable measuring the time to treatment
given a panel dataset with a time-varying treatment status indicator.
}
\details{
The output of this function is a factor variable. When passed to a
standard fitting function, the factor variable is broken into
dummies encoding leads and lags.

Most fitting functions are using the first level of a factor variable as a reference category.
This means that the first lead serves as the reference period. This can be changed
by setting a value for the parameter \code{baseline}, e.g. -1 to use the last lead before the treatment as reference period.

To exclude units that never receive treatment or that received treatment
before the first period (i.e. for which the treatment status has no
variance), set the parameters \code{always} and \code{never}
to \code{NA}. The estimation sample is then limited to the
switcher population.

To accumulate lags and leads set the parameters \code{leads}
and \code{lags} to a value that is less than the maximum of
feasible leads/lags.
}
\examples{
library(dd)

data(goldendawn)

goldendawn$t <- code_eventtime(
       unit=muni,
       time=year,
       treat=post,
       data=goldendawn)

m <- lm(gd ~ t + factor(muni) + factor(year), data=goldendawn)
summary(m)



}
\seealso{
\link[stats]{lm}
}
