% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{col_description}
\alias{col_description}
\title{DuckDB function col_description}
\usage{
col_description(table_oid, column_number)
}
\arguments{
\item{table_oid}{Unspecified.}

\item{column_number}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{col_description()}.
}
