% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{collations}
\alias{collations}
\title{DuckDB function collations}
\usage{
collations()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{collations()}.
}
\examples{
\dontrun{
character(0)
}
}
