% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{combine}
\alias{combine}
\title{DuckDB function combine}
\usage{
combine(col0, col1)
}
\arguments{
\item{col0}{\verb{AGGREGATE_STATE<?>}}

\item{col1}{\code{ANY}}
}
\value{
\verb{AGGREGATE_STATE<?>}
}
\description{
DuckDB function \code{combine()}.
}
