% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{constant_or_null}
\alias{constant_or_null}
\title{DuckDB function constant_or_null}
\usage{
constant_or_null(arg1 = ANY, arg2 = ANY)
}
\arguments{
\item{arg1}{\code{ANY}}

\item{arg2}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
If arg2 is NULL, return NULL. Otherwise, return arg1.
}
\examples{
\dontrun{
constant_or_null(42, NULL)
}
}
