% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{copy_dir}
\alias{copy_dir}
\title{DuckDB function copy_dir}
\usage{
copy_dir(src_dir, dst_dir)
}
\arguments{
\item{src_dir, dst_dir}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{copy_dir()}.
}
\section{Provided by}{

The \code{delta} extension (\code{LOAD delta;}).
}

