% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{cos}
\alias{cos}
\title{DuckDB function cos}
\usage{
cos(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the cos of x.
}
\examples{
\dontrun{
cos(90)
}
}
