% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{cosh}
\alias{cosh}
\title{DuckDB function cosh}
\usage{
cosh(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the hyperbolic cos of x.
}
\examples{
\dontrun{
cosh(1)
}
}
