% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{cot}
\alias{cot}
\title{DuckDB function cot}
\usage{
cot(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the cotangent of x.
}
\examples{
\dontrun{
cot(0.5)
}
}
