% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{count}
\alias{count}
\title{DuckDB function count}
\arguments{
\item{arg}{\code{ANY}}
}
\value{
\code{BIGINT}
}
\description{
Returns the number of non-NULL values in arg.
}
\section{Overloads}{

\itemize{
\item \code{count(arg = ANY)}
\item \code{count()}
}
}

\examples{
\dontrun{
count(A)
}
}
