% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{count_if}
\alias{count_if}
\title{DuckDB function count_if}
\usage{
count_if(arg = BOOLEAN)
}
\arguments{
\item{arg}{\code{BOOLEAN}}
}
\value{
\code{HUGEINT}
}
\description{
Counts the total number of TRUE values for a boolean column.
}
\examples{
\dontrun{
count_if(A)
}
}
