% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{count_star}
\alias{count_star}
\title{DuckDB function count_star}
\usage{
count_star()
}
\value{
\code{BIGINT}
}
\description{
DuckDB function \code{count_star()}.
}
\examples{
\dontrun{
character(0)
}
}
