% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{countif}
\alias{countif}
\title{DuckDB function countif}
\usage{
countif(arg = BOOLEAN)
}
\arguments{
\item{arg}{\code{BOOLEAN}}
}
\value{
\code{HUGEINT}
}
\description{
Counts the total number of TRUE values for a boolean column.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{countif(A)
}\if{html}{\out{</div>}}
}

