% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{create_fts_index}
\alias{create_fts_index}
\title{DuckDB function create_fts_index}
\usage{
create_fts_index(col0, col1, overwrite, lower, ignore, stopwords, strip_accents, stemmer)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{overwrite}{\code{BOOLEAN}}

\item{lower}{\code{BOOLEAN}}

\item{ignore}{\code{VARCHAR}}

\item{stopwords}{\code{VARCHAR}}

\item{strip_accents}{\code{BOOLEAN}}

\item{stemmer}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{create_fts_index()}.
}
\section{Provided by}{

The \code{fts} extension (\code{LOAD fts;}).
}

