% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_connection_id}
\alias{current_connection_id}
\title{DuckDB function current_connection_id}
\usage{
current_connection_id()
}
\value{
\code{UBIGINT}
}
\description{
Get the current connection_id.
}
\examples{
\dontrun{
current_connection_id()
}
}
