% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_date}
\alias{current_date}
\title{DuckDB function current_date}
\usage{
current_date()
}
\value{
\code{DATE}
}
\description{
DuckDB function \code{current_date()}.
}
\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

