% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_localtime}
\alias{current_localtime}
\title{DuckDB function current_localtime}
\usage{
current_localtime()
}
\value{
\code{TIME}
}
\description{
DuckDB function \code{current_localtime()}.
}
\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

