% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_localtimestamp}
\alias{current_localtimestamp}
\title{DuckDB function current_localtimestamp}
\usage{
current_localtimestamp()
}
\value{
\code{TIMESTAMP}
}
\description{
DuckDB function \code{current_localtimestamp()}.
}
\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

