% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_query}
\alias{current_query}
\title{DuckDB function current_query}
\value{
\code{VARCHAR}
}
\description{
Returns the current query as a string.
}
\section{Overloads}{

\itemize{
\item \code{current_query()}
\item \code{current_query()}
}
}

\examples{
\dontrun{
current_query()
}
}
