% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_schema}
\alias{current_schema}
\title{DuckDB function current_schema}
\value{
\code{VARCHAR}
}
\description{
Returns the name of the currently active schema. Default is main.
}
\section{Overloads}{

\itemize{
\item \code{current_schema()}
\item \code{current_schema()}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{current_schema()
}\if{html}{\out{</div>}}
}

