% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_schemas}
\alias{current_schemas}
\title{DuckDB function current_schemas}
\arguments{
\item{include_implicit}{\code{BOOLEAN}}
}
\value{
\code{VARCHAR[]}
}
\description{
Returns list of schemas. Pass a parameter of True to include implicit schemas.
}
\section{Overloads}{

\itemize{
\item \code{current_schemas(include_implicit = BOOLEAN)}: Returns list of schemas. Pass a parameter of True to include implicit schemas.
\item \code{current_schemas(include_implicit)}: Returns list of schemas. Pass a parameter of True to include implicit schemas.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{current_schemas(true)
}\if{html}{\out{</div>}}
}

