% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_setting}
\alias{current_setting}
\title{DuckDB function current_setting}
\usage{
current_setting(setting_name = VARCHAR)
}
\arguments{
\item{setting_name}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
Returns the current value of the configuration setting.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{current_setting('access_mode')
}\if{html}{\out{</div>}}
}

