% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{currval}
\alias{currval}
\title{DuckDB function currval}
\usage{
currval(`'sequence_name'` = VARCHAR)
}
\arguments{
\item{'sequence_name'}{\code{VARCHAR}}
}
\value{
\code{BIGINT}
}
\description{
Return the current value of the sequence. Note that nextval must be called at least once prior to calling currval.
}
\examples{
\dontrun{
currval('my_sequence_name')
}
}
