% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{date_add}
\alias{date_add}
\title{DuckDB function date_add}
\usage{
date_add(date, interval)
}
\arguments{
\item{date}{Unspecified.}

\item{interval}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{date_add()}.
}
