% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{date_part}
\alias{date_part}
\alias{datepart}
\title{DuckDB function date_part}
\arguments{
\item{ts}{\code{VARCHAR[] | VARCHAR}}

\item{col1}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{STRUCT() | BIGINT}
}
\description{
Get subfield (equivalent to extract).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{date_part('minute', TIMESTAMP '1992-09-20 20:38:40')
}\if{html}{\out{</div>}}
}

