% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{date_sub}
\alias{date_sub}
\alias{datesub}
\title{DuckDB function date_sub}
\usage{
date_sub(part, startdate, enddate)
}
\arguments{
\item{part}{\code{VARCHAR}}

\item{startdate}{\verb{DATE | TIME | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{enddate}{\verb{DATE | TIME | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
The number of complete partitions between the timestamps.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{date_sub('hour', TIMESTAMPTZ '1992-09-30 23:59:59', TIMESTAMPTZ '1992-10-01 01:58:00')
}\if{html}{\out{</div>}}
}

