% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{date_trunc}
\alias{date_trunc}
\alias{datetrunc}
\title{DuckDB function date_trunc}
\usage{
date_trunc(part, timestamp)
}
\arguments{
\item{part}{\code{VARCHAR}}

\item{timestamp}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\verb{TIMESTAMP | INTERVAL | TIMESTAMP WITH TIME ZONE}
}
\description{
Truncate to specified precision.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{date_trunc('hour', TIMESTAMPTZ '1992-09-20 20:38:40')
}\if{html}{\out{</div>}}
}

