% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{datepart}
\alias{datepart}
\title{DuckDB function datepart}
\arguments{
\item{ts}{\code{VARCHAR[] | VARCHAR}}

\item{col1}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{STRUCT() | BIGINT}
}
\description{
Get subfield (equivalent to extract).
}
\section{Overloads}{

\itemize{
\item \code{datepart(ts = `VARCHAR[]`, col1 = DATE)}
\item \code{datepart(ts = `VARCHAR[]`, col1 = INTERVAL)}
\item \code{datepart(ts = `VARCHAR[]`, col1 = TIME)}
\item \code{datepart(ts = `VARCHAR[]`, col1 = TIMESTAMP)}
\item \code{datepart(ts = `VARCHAR[]`, col1 = `TIME WITH TIME ZONE`)}
\item \code{datepart(ts = `VARCHAR[]`, col1 = TIME_NS)}
\item \code{datepart(ts = VARCHAR, col1 = DATE)}
\item \code{datepart(ts = VARCHAR, col1 = INTERVAL)}
\item \code{datepart(ts = VARCHAR, col1 = TIME)}
\item \code{datepart(ts = VARCHAR, col1 = TIMESTAMP)}
\item \code{datepart(ts = VARCHAR, col1 = `TIME WITH TIME ZONE`)}
\item \code{datepart(ts = VARCHAR, col1 = TIME_NS)}
\item \code{datepart(ts = `VARCHAR[]`, col1 = `TIMESTAMP WITH TIME ZONE`)}
\item \code{datepart(ts = VARCHAR, col1 = `TIMESTAMP WITH TIME ZONE`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{datepart('minute', TIMESTAMP '1992-09-20 20:38:40')
}\if{html}{\out{</div>}}
}

