% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{datetrunc}
\alias{datetrunc}
\title{DuckDB function datetrunc}
\arguments{
\item{part}{\code{VARCHAR}}

\item{timestamp}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\verb{TIMESTAMP | INTERVAL | TIMESTAMP WITH TIME ZONE}
}
\description{
Truncate to specified precision.
}
\examples{
\dontrun{
datetrunc('hour', TIMESTAMPTZ '1992-09-20 20:38:40')
}
}
