% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{dayofyear}
\alias{dayofyear}
\title{DuckDB function dayofyear}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the dayofyear component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{dayofyear(ts = DATE)}
\item \code{dayofyear(ts = INTERVAL)}
\item \code{dayofyear(ts = TIMESTAMP)}
\item \code{dayofyear(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
dayofyear(timestamp '2021-08-03 11:59:44.123456')
}
}
