% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{days_in_month}
\alias{days_in_month}
\title{DuckDB function days_in_month}
\usage{
days_in_month(date)
}
\arguments{
\item{date}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{days_in_month()}.
}
