% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{dbgen}
\alias{dbgen}
\title{DuckDB function dbgen}
\usage{
dbgen(step, suffix, schema, catalog, overwrite, children, sf)
}
\arguments{
\item{step}{\code{UINTEGER}}

\item{suffix}{\code{VARCHAR}}

\item{schema}{\code{VARCHAR}}

\item{catalog}{\code{VARCHAR}}

\item{overwrite}{\code{BOOLEAN}}

\item{children}{\code{UINTEGER}}

\item{sf}{\code{DOUBLE}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{dbgen()}.
}
