% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-dd.R
\name{dd}
\alias{dd}
\title{DuckDB functions}
\usage{
dd
}
\description{
A list of known DuckDB functions, as provided by DuckDB
\verb{r sub("^v", "", duckdb::sql_query("PRAGMA version")$library_version)}.
}
\examples{
dd[1:3]
}
