% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{decade}
\alias{decade}
\title{DuckDB function decade}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the decade component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{decade(ts = DATE)}
\item \code{decade(ts = INTERVAL)}
\item \code{decade(ts = TIMESTAMP)}
\item \code{decade(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
decade(timestamp '2021-08-03 11:59:44.123456')
}
}
