% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{decode}
\alias{decode}
\title{DuckDB function decode}
\arguments{
\item{blob}{\code{BLOB}}

\item{varchar}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Converts \code{blob} to \code{VARCHAR}. Invalid UTF-8 is handled based on the error behavior argument. Can be 'strict' (default, fail), 'replace' to replace invalid characters with '?', or 'ignore' to skip invalid characters.
}
\section{Overloads}{

\itemize{
\item \code{decode(blob = BLOB)}
\item \code{decode(blob = BLOB, varchar = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{decode('\\xC3\\xBC'::BLOB)
decode('\\xA0'::BLOB, 'replace')
decode('\\xA0'::BLOB, 'ignore')
}\if{html}{\out{</div>}}
}

\seealso{
Other blob:
\code{\link[=encode]{encode()}},
\code{\link[=from_base64]{from_base64()}},
\code{\link[=hex]{hex()}},
\code{\link[=md5]{md5()}},
\code{\link[=md5_number]{md5_number()}},
\code{\link[=octet_length]{octet_length()}},
\code{\link{or--or}},
\code{\link[=repeat]{repeat()}},
\code{\link[=sha1]{sha1()}},
\code{\link[=sha256]{sha256()}},
\code{\link[=to_base64]{to_base64()}},
\code{\link[=unbin]{unbin()}},
\code{\link[=unhex]{unhex()}}
}
\concept{blob}
