% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{degrees}
\alias{degrees}
\title{DuckDB function degrees}
\usage{
degrees(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Converts radians to degrees.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{degrees(pi())
}\if{html}{\out{</div>}}
}

