% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{delta_domain_metadata}
\alias{delta_domain_metadata}
\title{DuckDB function delta_domain_metadata}
\usage{
delta_domain_metadata(col0, version)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{version}{\code{UBIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{delta_domain_metadata()}.
}
\section{Provided by}{

The \code{delta} extension (\code{LOAD delta;}).
}

