% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{delta_filter_pushdown_log}
\alias{delta_filter_pushdown_log}
\title{DuckDB function delta_filter_pushdown_log}
\usage{
delta_filter_pushdown_log()
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{delta_filter_pushdown_log()}.
}
