% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{delta_filter_pushdown_log_tpcds}
\alias{delta_filter_pushdown_log_tpcds}
\title{DuckDB function delta_filter_pushdown_log_tpcds}
\usage{
delta_filter_pushdown_log_tpcds()
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{delta_filter_pushdown_log_tpcds()}.
}
