% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{delta_get_transaction_version}
\alias{delta_get_transaction_version}
\title{DuckDB function delta_get_transaction_version}
\usage{
delta_get_transaction_version(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{delta_get_transaction_version()}.
}
\section{Provided by}{

The \code{delta} extension (\code{LOAD delta;}).
}

