% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{delta_list_files}
\alias{delta_list_files}
\title{DuckDB function delta_list_files}
\usage{
delta_list_files(col0, version, delete_count, transform_expression)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{version}{\code{UBIGINT}}

\item{delete_count}{\code{BOOLEAN}}

\item{transform_expression}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{delta_list_files()}.
}
\section{Provided by}{

The \code{delta} extension (\code{LOAD delta;}).
}

