% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{delta_scan}
\alias{delta_scan}
\title{DuckDB function delta_scan}
\usage{
delta_scan(
  col0,
  version,
  log_tail,
  pushdown_partition_info,
  can_have_nan,
  filename,
  pushdown_filters,
  union_by_name,
  debug_use_openssl,
  hive_partitioning,
  parquet_version,
  encryption_config,
  hive_types_autocast,
  binary_as_string,
  explicit_cardinality,
  compression,
  file_row_number,
  hive_types
)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{version}{\code{UBIGINT}}

\item{log_tail}{\verb{STRUCT(file_name VARCHAR, "timestamp" BIGINT, file_size UBIGINT)[]}}

\item{pushdown_partition_info}{\code{BOOLEAN}}

\item{can_have_nan}{\code{BOOLEAN}}

\item{filename}{\code{ANY}}

\item{pushdown_filters}{\code{VARCHAR}}

\item{union_by_name}{\code{BOOLEAN}}

\item{debug_use_openssl}{\code{BOOLEAN}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{parquet_version}{\code{VARCHAR}}

\item{encryption_config}{\code{ANY}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{binary_as_string}{\code{BOOLEAN}}

\item{explicit_cardinality}{\code{UBIGINT}}

\item{compression}{\code{VARCHAR}}

\item{file_row_number}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{delta_scan()}.
}
