% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{delta_set_transaction_version}
\alias{delta_set_transaction_version}
\title{DuckDB function delta_set_transaction_version}
\usage{
delta_set_transaction_version(col0, col1, col2, col3)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{UBIGINT}}

\item{col3}{\code{UBIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{delta_set_transaction_version()}.
}
\section{Provided by}{

The \code{delta} extension (\code{LOAD delta;}).
}

