% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_checkpoint_on_shutdown}
\alias{disable_checkpoint_on_shutdown}
\title{DuckDB function disable_checkpoint_on_shutdown}
\usage{
disable_checkpoint_on_shutdown()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_checkpoint_on_shutdown()}.
}
\examples{
\dontrun{
character(0)
}
}
