% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_logging}
\alias{disable_logging}
\title{DuckDB function disable_logging}
\usage{
disable_logging()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_logging()}.
}
\examples{
\dontrun{
character(0)
}
}
