% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_object_cache}
\alias{disable_object_cache}
\title{DuckDB function disable_object_cache}
\usage{
disable_object_cache()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_object_cache()}.
}
\examples{
\dontrun{
character(0)
}
}
