% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_peg_parser}
\alias{disable_peg_parser}
\title{DuckDB function disable_peg_parser}
\usage{
disable_peg_parser()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_peg_parser()}.
}
