% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_progress_bar}
\alias{disable_progress_bar}
\title{DuckDB function disable_progress_bar}
\usage{
disable_progress_bar()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_progress_bar()}.
}
\examples{
\dontrun{
character(0)
}
}
