% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_verification}
\alias{disable_verification}
\title{DuckDB function disable_verification}
\usage{
disable_verification()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_verification()}.
}
