% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_verify_external}
\alias{disable_verify_external}
\title{DuckDB function disable_verify_external}
\usage{
disable_verify_external()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_verify_external()}.
}
