% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_verify_fetch_row}
\alias{disable_verify_fetch_row}
\title{DuckDB function disable_verify_fetch_row}
\usage{
disable_verify_fetch_row()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_verify_fetch_row()}.
}
\examples{
\dontrun{
character(0)
}
}
