% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_verify_parallelism}
\alias{disable_verify_parallelism}
\title{DuckDB function disable_verify_parallelism}
\usage{
disable_verify_parallelism()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_verify_parallelism()}.
}
\examples{
\dontrun{
character(0)
}
}
