% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{disable_verify_serializer}
\alias{disable_verify_serializer}
\title{DuckDB function disable_verify_serializer}
\usage{
disable_verify_serializer()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{disable_verify_serializer()}.
}
