% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{divide}
\alias{divide}
\alias{//}
\title{DuckDB function divide}
\usage{
divide(col0, col1)
}
\arguments{
\item{col0, col1}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
DuckDB function \code{divide()}.
}
