% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{divorce}
\alias{divorce}
\title{Stevenson-Wolfers Divorce Data}
\format{
A data frame with 1617 rows and 7 variables:
\describe{
\item{stfips}{US State (FIPS Code)}
\item{year}{Calendar Year}
\item{post}{Post reform?}
\item{asmrs}{Suicide rate among women}
\item{pcinc}{Per capita income}
\item{asmrh}{Homicide mortality rate}
\item{cases}{Aid to Families with Dependent Children (AFDC) rate (for a family of four)}
}
}
\usage{
divorce
}
\description{
This data comes from Stevenson and Wolfers (2006) who
study how no-fault unilateral divorce reforms affect female
suicide in the United States. The data here are taken from
the documentation of the Stata ado BACONDECOMP.
}
\references{
Stevenson, Betsey and Justin Wolfers. 2006. "Bargaining in the Shadow of the Law: Divorce Laws and Family Distress." \emph{The Quarterly Journal of Economics}, 121(1), 267-288.

Goodman-Bacon, Andrew and Thomas Goldring and Austin Nichols. 2019. "BACONDECOMP: Stata Module to Perform a Bacon Decomposition of Difference-in-differences Estimation." Statistical Software Components S458676, Boston College Department of Economics, revised 15 Sep 2019.
}
\keyword{datasets}
