% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{drop_fts_index}
\alias{drop_fts_index}
\title{DuckDB function drop_fts_index}
\usage{
drop_fts_index(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{drop_fts_index()}.
}
\section{Provided by}{

The \code{fts} extension (\code{LOAD fts;}).
}

