% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{dsdgen}
\alias{dsdgen}
\title{DuckDB function dsdgen}
\usage{
dsdgen(suffix, schema, catalog, keys, overwrite, sf)
}
\arguments{
\item{suffix}{\code{VARCHAR}}

\item{schema}{\code{VARCHAR}}

\item{catalog}{\code{VARCHAR}}

\item{keys}{\code{BOOLEAN}}

\item{overwrite}{\code{BOOLEAN}}

\item{sf}{\code{DOUBLE}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{dsdgen()}.
}
