% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_approx_database_count}
\alias{duckdb_approx_database_count}
\title{DuckDB function duckdb_approx_database_count}
\usage{
duckdb_approx_database_count()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_approx_database_count()}.
}
\examples{
\dontrun{
character(0)
}
}
