% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_columns}
\alias{duckdb_columns}
\title{DuckDB function duckdb_columns}
\usage{
duckdb_columns()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_columns()}.
}
\examples{
\dontrun{
character(0)
}
}
