% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_connection_count}
\alias{duckdb_connection_count}
\title{DuckDB function duckdb_connection_count}
\usage{
duckdb_connection_count()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_connection_count()}.
}
