% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_constraints}
\alias{duckdb_constraints}
\title{DuckDB function duckdb_constraints}
\usage{
duckdb_constraints()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_constraints()}.
}
