% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_coordinate_systems}
\alias{duckdb_coordinate_systems}
\title{DuckDB function duckdb_coordinate_systems}
\usage{
duckdb_coordinate_systems()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_coordinate_systems()}.
}
