% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_external_file_cache}
\alias{duckdb_external_file_cache}
\title{DuckDB function duckdb_external_file_cache}
\usage{
duckdb_external_file_cache()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_external_file_cache()}.
}
\examples{
\dontrun{
character(0)
}
}
