% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_indexes}
\alias{duckdb_indexes}
\title{DuckDB function duckdb_indexes}
\usage{
duckdb_indexes()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_indexes()}.
}
\examples{
\dontrun{
character(0)
}
}
