% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_keywords}
\alias{duckdb_keywords}
\title{DuckDB function duckdb_keywords}
\usage{
duckdb_keywords()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_keywords()}.
}
\examples{
\dontrun{
character(0)
}
}
