% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_log_contexts}
\alias{duckdb_log_contexts}
\title{DuckDB function duckdb_log_contexts}
\usage{
duckdb_log_contexts()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_log_contexts()}.
}
