% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_logs}
\alias{duckdb_logs}
\title{DuckDB function duckdb_logs}
\usage{
duckdb_logs(denormalized_table = BOOLEAN)
}
\arguments{
\item{denormalized_table}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_logs()}.
}
