% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_logs_parsed}
\alias{duckdb_logs_parsed}
\title{DuckDB function duckdb_logs_parsed}
\usage{
duckdb_logs_parsed(log_type)
}
\arguments{
\item{log_type}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{duckdb_logs_parsed()}.
}
