% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_prepared_statements}
\alias{duckdb_prepared_statements}
\title{DuckDB function duckdb_prepared_statements}
\usage{
duckdb_prepared_statements()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_prepared_statements()}.
}
