% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_profiling_settings}
\alias{duckdb_profiling_settings}
\title{DuckDB function duckdb_profiling_settings}
\usage{
duckdb_profiling_settings()
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{duckdb_profiling_settings()}.
}
