% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{DuckDB_PROJ_Compiled_Version}
\alias{DuckDB_PROJ_Compiled_Version}
\title{DuckDB function DuckDB_PROJ_Compiled_Version}
\usage{
DuckDB_PROJ_Compiled_Version()
}
\value{
\code{VARCHAR}
}
\description{
Returns a text description of the PROJ library version that this instance of DuckDB was compiled against.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT duckdb_proj_compiled_version();
-- +--------------------------------+
-- | duckdb_proj_compiled_version() |
-- |            varchar             |
-- +--------------------------------+
-- | Rel. 9.1.1, December 1st, 2022 |
-- +--------------------------------+
}\if{html}{\out{</div>}}
}

