% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{DuckDB_Proj_Version}
\alias{DuckDB_Proj_Version}
\title{DuckDB function DuckDB_Proj_Version}
\usage{
DuckDB_Proj_Version()
}
\value{
\code{VARCHAR}
}
\description{
Returns a text description of the PROJ library version that is being used by this instance of DuckDB.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT duckdb_proj_version();
-- +-----------------------+
-- | duckdb_proj_version() |
-- |        varchar        |
-- +-----------------------+
-- | 9.1.1                 |geometry_always_xy
-- +-----------------------+
}\if{html}{\out{</div>}}
}

