% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_secret_types}
\alias{duckdb_secret_types}
\title{DuckDB function duckdb_secret_types}
\usage{
duckdb_secret_types()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_secret_types()}.
}
