% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_secrets}
\alias{duckdb_secrets}
\title{DuckDB function duckdb_secrets}
\usage{
duckdb_secrets(redact = BOOLEAN)
}
\arguments{
\item{redact}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_secrets()}.
}
