% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_sequences}
\alias{duckdb_sequences}
\title{DuckDB function duckdb_sequences}
\usage{
duckdb_sequences()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_sequences()}.
}
\examples{
\dontrun{
character(0)
}
}
