% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_settings}
\alias{duckdb_settings}
\title{DuckDB function duckdb_settings}
\usage{
duckdb_settings()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_settings()}.
}
\examples{
\dontrun{
character(0)
}
}
