% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_table_sample}
\alias{duckdb_table_sample}
\title{DuckDB function duckdb_table_sample}
\usage{
duckdb_table_sample(col0 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_table_sample()}.
}
\examples{
\dontrun{
character(0)
}
}
