% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_types}
\alias{duckdb_types}
\title{DuckDB function duckdb_types}
\usage{
duckdb_types()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_types()}.
}
\examples{
\dontrun{
character(0)
}
}
