% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{duckdb_variables}
\alias{duckdb_variables}
\title{DuckDB function duckdb_variables}
\usage{
duckdb_variables()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{duckdb_variables()}.
}
\examples{
\dontrun{
character(0)
}
}
