% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_add_data_files}
\alias{ducklake_add_data_files}
\title{DuckDB function ducklake_add_data_files}
\usage{
ducklake_add_data_files(
  col0,
  col1,
  col2,
  schema,
  hive_partitioning,
  ignore_extra_columns,
  allow_missing
)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR | VARCHAR[]}}

\item{schema}{\code{VARCHAR}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{ignore_extra_columns}{\code{BOOLEAN}}

\item{allow_missing}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_add_data_files()}.
}
