% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_cleanup_old_files}
\alias{ducklake_cleanup_old_files}
\title{DuckDB function ducklake_cleanup_old_files}
\usage{
ducklake_cleanup_old_files(col0, dry_run, cleanup_all, older_than)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{dry_run}{\code{BOOLEAN}}

\item{cleanup_all}{\code{BOOLEAN}}

\item{older_than}{\verb{TIMESTAMP WITH TIME ZONE}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_cleanup_old_files()}.
}
