% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_commit}
\alias{ducklake_commit}
\title{DuckDB function ducklake_commit}
\usage{
ducklake_commit(col0, col1, retry_backoff, retry_wait_ms, max_retry_count)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{BIGINT}}

\item{retry_backoff}{\code{DOUBLE}}

\item{retry_wait_ms}{\code{BIGINT}}

\item{max_retry_count}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_commit()}.
}
