% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_current_snapshot}
\alias{ducklake_current_snapshot}
\title{DuckDB function ducklake_current_snapshot}
\usage{
ducklake_current_snapshot(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_current_snapshot()}.
}
