% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_delete_orphaned_files}
\alias{ducklake_delete_orphaned_files}
\title{DuckDB function ducklake_delete_orphaned_files}
\usage{
ducklake_delete_orphaned_files(col0, dry_run, cleanup_all, older_than)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{dry_run}{\code{BOOLEAN}}

\item{cleanup_all}{\code{BOOLEAN}}

\item{older_than}{\verb{TIMESTAMP WITH TIME ZONE}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_delete_orphaned_files()}.
}
\section{Provided by}{

The \code{ducklake} extension (\code{LOAD ducklake;}).
}

