% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_expire_snapshots}
\alias{ducklake_expire_snapshots}
\title{DuckDB function ducklake_expire_snapshots}
\usage{
ducklake_expire_snapshots(col0, dry_run, versions, older_than)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{dry_run}{\code{BOOLEAN}}

\item{versions}{\code{UBIGINT[]}}

\item{older_than}{\verb{TIMESTAMP WITH TIME ZONE}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_expire_snapshots()}.
}
\section{Provided by}{

The \code{ducklake} extension (\code{LOAD ducklake;}).
}

