% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_flush_inlined_data}
\alias{ducklake_flush_inlined_data}
\title{DuckDB function ducklake_flush_inlined_data}
\usage{
ducklake_flush_inlined_data(col0, table_name, schema_name)
}
\arguments{
\item{col0, table_name, schema_name}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_flush_inlined_data()}.
}
