% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_list_files}
\alias{ducklake_list_files}
\title{DuckDB function ducklake_list_files}
\usage{
ducklake_list_files(col0, col1, snapshot_time, snapshot_version, schema)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{snapshot_time}{\verb{TIMESTAMP WITH TIME ZONE}}

\item{snapshot_version}{\code{BIGINT}}

\item{schema}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_list_files()}.
}
