% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_merge_adjacent_files}
\alias{ducklake_merge_adjacent_files}
\title{DuckDB function ducklake_merge_adjacent_files}
\usage{
ducklake_merge_adjacent_files(
  col0,
  col1,
  schema,
  max_compacted_files,
  max_file_size,
  min_file_size
)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{schema}{\code{VARCHAR}}

\item{max_compacted_files}{\code{UBIGINT}}

\item{max_file_size}{\code{UBIGINT}}

\item{min_file_size}{\code{UBIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_merge_adjacent_files()}.
}
\section{Overloads}{

\itemize{
\item \code{ducklake_merge_adjacent_files(col0 = VARCHAR, col1 = VARCHAR, schema = VARCHAR, max_compacted_files = UBIGINT, max_file_size = UBIGINT, min_file_size = UBIGINT)}
\item \code{ducklake_merge_adjacent_files(col0 = VARCHAR, max_compacted_files = UBIGINT, max_file_size = UBIGINT, min_file_size = UBIGINT)}
}
}

\section{Provided by}{

The \code{ducklake} extension (\code{LOAD ducklake;}).
}

