% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_options}
\alias{ducklake_options}
\title{DuckDB function ducklake_options}
\usage{
ducklake_options(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_options()}.
}
\section{Provided by}{

The \code{ducklake} extension (\code{LOAD ducklake;}).
}

