% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_rewrite_data_files}
\alias{ducklake_rewrite_data_files}
\title{DuckDB function ducklake_rewrite_data_files}
\usage{
ducklake_rewrite_data_files(col0, col1, schema, delete_threshold)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{schema}{\code{VARCHAR}}

\item{delete_threshold}{\code{DOUBLE}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_rewrite_data_files()}.
}
\section{Overloads}{

\itemize{
\item \code{ducklake_rewrite_data_files(col0 = VARCHAR, col1 = VARCHAR, schema = VARCHAR, delete_threshold = DOUBLE)}
\item \code{ducklake_rewrite_data_files(col0 = VARCHAR, delete_threshold = DOUBLE)}
}
}

