% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_set_commit_message}
\alias{ducklake_set_commit_message}
\title{DuckDB function ducklake_set_commit_message}
\usage{
ducklake_set_commit_message(col0, col1, col2, extra_info)
}
\arguments{
\item{col0, col1, col2, extra_info}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_set_commit_message()}.
}
