% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_set_option}
\alias{ducklake_set_option}
\title{DuckDB function ducklake_set_option}
\usage{
ducklake_set_option(col0, col1, col2, schema, table_name)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{ANY}}

\item{schema}{\code{VARCHAR}}

\item{table_name}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_set_option()}.
}
\section{Provided by}{

The \code{ducklake} extension (\code{LOAD ducklake;}).
}

