% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_settings}
\alias{ducklake_settings}
\title{DuckDB function ducklake_settings}
\usage{
ducklake_settings(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_settings()}.
}
