% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_snapshots}
\alias{ducklake_snapshots}
\title{DuckDB function ducklake_snapshots}
\usage{
ducklake_snapshots(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_snapshots()}.
}
