% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_table_changes}
\alias{ducklake_table_changes}
\title{DuckDB function ducklake_table_changes}
\usage{
ducklake_table_changes(catalog, schema_name, table_name, start_snapshot, end_snapshot)
}
\arguments{
\item{catalog, schema_name, table_name, start_snapshot, end_snapshot}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{ducklake_table_changes()}.
}
