% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_table_info}
\alias{ducklake_table_info}
\title{DuckDB function ducklake_table_info}
\usage{
ducklake_table_info(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_table_info()}.
}
