% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ducklake_table_insertions}
\alias{ducklake_table_insertions}
\title{DuckDB function ducklake_table_insertions}
\usage{
ducklake_table_insertions(col0, col1, col2, col3, col4)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR}}

\item{col3}{\verb{BIGINT | TIMESTAMP WITH TIME ZONE}}

\item{col4}{\verb{BIGINT | TIMESTAMP WITH TIME ZONE}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ducklake_table_insertions()}.
}
\section{Provided by}{

The \code{ducklake} extension (\code{LOAD ducklake;}).
}

