% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{element_at}
\alias{element_at}
\title{DuckDB function element_at}
\usage{
element_at(map = `MAP(K, V)`, key = K)
}
\arguments{
\item{map}{\code{MAP(K, V)}}

\item{key}{\code{K}}
}
\description{
Returns a list containing the value for a given key or an empty list if the key is not contained in the map. The type of the key provided in the second parameter must match the type of the map’s keys else an error is returned.
}
\examples{
\dontrun{
element_at(map(['key'], ['val']), 'key')
}
}
