% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enable_checkpoint_on_shutdown}
\alias{enable_checkpoint_on_shutdown}
\title{DuckDB function enable_checkpoint_on_shutdown}
\usage{
enable_checkpoint_on_shutdown()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{enable_checkpoint_on_shutdown()}.
}
\examples{
\dontrun{
character(0)
}
}
