% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enable_logging}
\alias{enable_logging}
\title{DuckDB function enable_logging}
\usage{
enable_logging(
  storage_buffer_size,
  storage_normalize,
  storage_path,
  storage_config,
  storage,
  level
)
}
\arguments{
\item{storage_buffer_size}{\code{UBIGINT}}

\item{storage_normalize}{\code{BOOLEAN}}

\item{storage_path}{\code{VARCHAR}}

\item{storage_config}{\code{ANY}}

\item{storage}{\code{VARCHAR}}

\item{level}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{enable_logging()}.
}
