% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enable_object_cache}
\alias{enable_object_cache}
\title{DuckDB function enable_object_cache}
\usage{
enable_object_cache()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{enable_object_cache()}.
}
\examples{
\dontrun{
character(0)
}
}
