% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enable_peg_parser}
\alias{enable_peg_parser}
\title{DuckDB function enable_peg_parser}
\usage{
enable_peg_parser()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{enable_peg_parser()}.
}
\section{Provided by}{

The \code{autocomplete} extension (\code{LOAD autocomplete;}).
}

