% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enable_progress_bar}
\alias{enable_progress_bar}
\title{DuckDB function enable_progress_bar}
\usage{
enable_progress_bar()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{enable_progress_bar()}.
}
