% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enable_verification}
\alias{enable_verification}
\title{DuckDB function enable_verification}
\usage{
enable_verification()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{enable_verification()}.
}
