% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{entropy}
\alias{entropy}
\title{DuckDB function entropy}
\usage{
entropy(x = ANY)
}
\arguments{
\item{x}{\code{ANY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the log-2 entropy of count input-values.
}
