% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enum_code}
\alias{enum_code}
\alias{duckdb_fun_enum_code}
\title{DuckDB function enum_code}
\arguments{
\item{enum}{\code{ANY}}
}
\description{
Returns the numeric value backing the given enum value.
}
\examples{
\dontrun{
enum_code('happy'::mood)
}
}
