% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enum_first}
\alias{enum_first}
\title{DuckDB function enum_first}
\arguments{
\item{enum}{\code{ANY}}
}
\description{
Returns the first value of the input enum type.
}
\examples{
\dontrun{
enum_first(NULL::mood)
}
}
