% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enum_range}
\alias{enum_range}
\title{DuckDB function enum_range}
\usage{
enum_range(enum = ANY)
}
\arguments{
\item{enum}{\code{ANY}}
}
\value{
\code{VARCHAR[]}
}
\description{
Returns all values of the input enum type as an array.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{enum_range(NULL::mood)
}\if{html}{\out{</div>}}
}

