% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enum_range_boundary}
\alias{enum_range_boundary}
\title{DuckDB function enum_range_boundary}
\usage{
enum_range_boundary(start = ANY, end = ANY)
}
\arguments{
\item{start}{\code{ANY}}

\item{end}{\code{ANY}}
}
\description{
Returns the range between the two given enum values as an array. The values must be of the same enum type. When the first parameter is NULL, the result starts with the first value of the enum type. When the second parameter is NULL, the result ends with the last value of the enum type.
}
\examples{
\dontrun{
enum_range_boundary(NULL, 'happy'::mood)
}
}
