% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{epoch_ms}
\alias{epoch_ms}
\title{DuckDB function epoch_ms}
\arguments{
\item{temporal}{\verb{DATE | TIMESTAMP | INTERVAL | TIME | TIME_NS | TIME WITH TIME ZONE | TIMESTAMP WITH TIME ZONE | BIGINT}}
}
\value{
\code{BIGINT | TIMESTAMP}
}
\description{
Extract the epoch component in milliseconds from a temporal type.
}
\section{Overloads}{

\itemize{
\item \code{epoch_ms(temporal = DATE)}: Extract the epoch component in milliseconds from a temporal type.
\item \code{epoch_ms(temporal = TIMESTAMP)}: Extract the epoch component in milliseconds from a temporal type.
\item \code{epoch_ms(temporal = INTERVAL)}: Extract the epoch component in milliseconds from a temporal type.
\item \code{epoch_ms(temporal = TIME)}: Extract the epoch component in milliseconds from a temporal type.
\item \code{epoch_ms(temporal = TIME_NS)}: Extract the epoch component in milliseconds from a temporal type.
\item \code{epoch_ms(temporal = `TIME WITH TIME ZONE`)}: Extract the epoch component in milliseconds from a temporal type.
\item \code{epoch_ms(temporal = `TIMESTAMP WITH TIME ZONE`)}: Extract the epoch component in milliseconds from a temporal type.
\item \code{epoch_ms(temporal = BIGINT)}: Extract the epoch component in milliseconds from a temporal type.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{epoch_ms(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

