% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{epoch_us}
\alias{epoch_us}
\title{DuckDB function epoch_us}
\arguments{
\item{temporal}{\verb{DATE | TIMESTAMP | INTERVAL | TIME | TIME_NS | TIME WITH TIME ZONE | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the epoch component in microseconds from a temporal type.
}
\section{Overloads}{

\itemize{
\item \code{epoch_us(temporal = DATE)}
\item \code{epoch_us(temporal = TIMESTAMP)}
\item \code{epoch_us(temporal = INTERVAL)}
\item \code{epoch_us(temporal = TIME)}
\item \code{epoch_us(temporal = TIME_NS)}
\item \code{epoch_us(temporal = `TIME WITH TIME ZONE`)}
\item \code{epoch_us(temporal = `TIMESTAMP WITH TIME ZONE`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{epoch_us(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

