% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{error}
\alias{error}
\title{DuckDB function error}
\usage{
error(message = VARCHAR)
}
\arguments{
\item{message}{\code{VARCHAR}}
}
\value{
\code{"NULL"}
}
\description{
Throws the given error message.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{error('access_mode')
}\if{html}{\out{</div>}}
}

