% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{even}
\alias{even}
\title{DuckDB function even}
\usage{
even(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Rounds x to next even number by rounding away from zero.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{even(2.9)
}\if{html}{\out{</div>}}
}

