% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{excel_text}
\alias{excel_text}
\title{DuckDB function excel_text}
\usage{
excel_text(col0, col1)
}
\arguments{
\item{col0}{\code{DOUBLE}}

\item{col1}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{excel_text()}.
}
\section{Provided by}{

The \code{excel} extension (\code{LOAD excel;}).
}

