% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{exp}
\alias{exp}
\title{DuckDB function exp}
\usage{
exp(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes e to the power of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{exp(1)
}\if{html}{\out{</div>}}
}

