% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{extension_versions}
\alias{extension_versions}
\title{DuckDB function extension_versions}
\usage{
extension_versions()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{extension_versions()}.
}
\examples{
\dontrun{
character(0)
}
}
