% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{factorial}
\alias{factorial}
\title{DuckDB function factorial}
\usage{
factorial(x = INTEGER)
}
\arguments{
\item{x}{\code{INTEGER}}
}
\value{
\code{HUGEINT}
}
\description{
Factorial of x. Computes the product of the current integer and all integers below it.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{4!
}\if{html}{\out{</div>}}
}

