% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{family}
\alias{family}
\title{DuckDB function family}
\usage{
family(col0)
}
\arguments{
\item{col0}{\code{INET}}
}
\value{
\code{UTINYINT}
}
\description{
DuckDB function \code{family()}.
}
\section{Provided by}{

The \code{inet} extension (\code{LOAD inet;}).
}

