% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{finalize}
\alias{finalize}
\title{DuckDB function finalize}
\usage{
finalize(col0)
}
\arguments{
\item{col0}{\verb{AGGREGATE_STATE<?>}}
}
\value{
\code{INVALID}
}
\description{
DuckDB function \code{finalize()}.
}
