% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{first}
\alias{first}
\alias{arbitrary}
\title{DuckDB function first}
\usage{
first(arg)
}
\arguments{
\item{arg}{\code{DECIMAL | ANY}}
}
\value{
\code{DECIMAL | ANY | T}
}
\description{
Returns the first value (NULL or non-NULL) from arg. This function is affected by ordering.
}
\section{Overloads}{

\itemize{
\item \code{first(arg = DECIMAL)}
\item \code{first(arg = ANY)}
\item \code{first(col0 = T)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{first(A)
}\if{html}{\out{</div>}}
}

