% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{first_value}
\alias{first_value}
\title{DuckDB function first_value}
\usage{
first_value(col0)
}
\arguments{
\item{col0}{\code{T}}
}
\value{
\code{T}
}
\description{
DuckDB function \code{first_value()}.
}
