% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{flatten}
\alias{flatten}
\title{DuckDB function flatten}
\usage{
flatten(nested_list = `T[][]`)
}
\arguments{
\item{nested_list}{\code{T[][]}}
}
\value{
\code{T[]}
}
\description{
Flattens a nested list by one level.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{flatten([[1, 2, 3], [4, 5]])
}\if{html}{\out{</div>}}
}

