% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{format_pg_type}
\alias{format_pg_type}
\title{DuckDB function format_pg_type}
\usage{
format_pg_type(logical_type, type_name)
}
\arguments{
\item{logical_type}{Unspecified.}

\item{type_name}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{format_pg_type()}.
}
