% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{format_type}
\alias{format_type}
\title{DuckDB function format_type}
\usage{
format_type(type_oid, typemod)
}
\arguments{
\item{type_oid}{Unspecified.}

\item{typemod}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{format_type()}.
}
