% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{formatReadableSize}
\alias{formatReadableSize}
\title{DuckDB function formatReadableSize}
\usage{
formatReadableSize(integer = BIGINT)
}
\arguments{
\item{integer}{\code{BIGINT}}
}
\value{
\code{VARCHAR}
}
\description{
Converts \code{integer} to a human-readable representation using units based on powers of 2 (KiB, MiB, GiB, etc.).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{formatReadableSize(16_000)
}\if{html}{\out{</div>}}
}

