% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{from_hex}
\alias{from_hex}
\title{DuckDB function from_hex}
\arguments{
\item{value}{\code{VARCHAR}}
}
\description{
Converts a value from hexadecimal representation to a blob.
}
\examples{
\dontrun{
unhex('2A')
}
}
