% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{from_json}
\alias{from_json}
\title{DuckDB function from_json}
\usage{
from_json(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
DuckDB function \code{from_json()}.
}
