% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{from_json_strict}
\alias{from_json_strict}
\title{DuckDB function from_json_strict}
\usage{
from_json_strict(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
DuckDB function \code{from_json_strict()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

