% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{fsum}
\alias{fsum}
\title{DuckDB function fsum}
\usage{
fsum(arg = DOUBLE)
}
\arguments{
\item{arg}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Calculates the sum using a more accurate floating point summation (Kahan Sum).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{fsum(A)
}\if{html}{\out{</div>}}
}

