% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{gamma}
\alias{gamma}
\title{DuckDB function gamma}
\usage{
gamma(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\description{
Interpolation of (x-1) factorial (so decimal inputs are allowed).
}
\examples{
\dontrun{
gamma(5.5)
}
}
