% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{gcd}
\alias{gcd}
\title{DuckDB function gcd}
\arguments{
\item{x}{\code{BIGINT | HUGEINT}}

\item{y}{\code{BIGINT | HUGEINT}}
}
\value{
\code{BIGINT | HUGEINT}
}
\description{
Computes the greatest common divisor of x and y.
}
\examples{
\dontrun{
gcd(42, 57)
}
}
