% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{gen_random_uuid}
\alias{gen_random_uuid}
\title{DuckDB function gen_random_uuid}
\usage{
gen_random_uuid()
}
\value{
\code{UUID}
}
\description{
Returns a random UUID v4 similar to this: eeccb8c5-9943-b2bb-bb5e-222f4e14b687.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{gen_random_uuid()
}\if{html}{\out{</div>}}
}

