% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{generate_subscripts}
\alias{generate_subscripts}
\title{DuckDB function generate_subscripts}
\usage{
generate_subscripts(arr, dim)
}
\arguments{
\item{arr}{Unspecified.}

\item{dim}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{generate_subscripts()}.
}
