% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{geometric_mean}
\alias{geometric_mean}
\title{DuckDB function geometric_mean}
\usage{
geometric_mean(x)
}
\arguments{
\item{x}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{geometric_mean()}.
}
