% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{get_bit}
\alias{get_bit}
\title{DuckDB function get_bit}
\usage{
get_bit(bitstring = BIT, index = INTEGER)
}
\arguments{
\item{bitstring}{\code{BIT}}

\item{index}{\code{INTEGER}}
}
\value{
\code{INTEGER}
}
\description{
Extracts the nth bit from bitstring; the first (leftmost) bit is indexed 0.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{get_bit('0110010'::BIT, 2)
}\if{html}{\out{</div>}}
}

