% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{get_block_size}
\alias{get_block_size}
\title{DuckDB function get_block_size}
\usage{
get_block_size(db_name)
}
\arguments{
\item{db_name}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{get_block_size()}.
}
