% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{get_current_time}
\alias{get_current_time}
\title{DuckDB function get_current_time}
\usage{
get_current_time()
}
\value{
\verb{TIME WITH TIME ZONE}
}
\description{
DuckDB function \code{get_current_time()}.
}
\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

