% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{get_current_timestamp}
\alias{get_current_timestamp}
\title{DuckDB function get_current_timestamp}
\usage{
get_current_timestamp()
}
\value{
\verb{TIMESTAMP WITH TIME ZONE}
}
\description{
Returns the current timestamp.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{get_current_timestamp()
}\if{html}{\out{</div>}}
}

