% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{get_delta_test_expression}
\alias{get_delta_test_expression}
\title{DuckDB function get_delta_test_expression}
\usage{
get_delta_test_expression()
}
\value{
\code{VARCHAR[]}
}
\description{
DuckDB function \code{get_delta_test_expression()}.
}
\section{Provided by}{

The \code{delta} extension (\code{LOAD delta;}).
}

