% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{get_type}
\alias{get_type}
\title{DuckDB function get_type}
\usage{
get_type(expression)
}
\arguments{
\item{expression}{\code{ANY}}
}
\value{
\code{TYPE}
}
\description{
Returns the type of the result of the expression.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{get_type('abc')
}\if{html}{\out{</div>}}
}

