% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{get_ui_url}
\alias{get_ui_url}
\title{DuckDB function get_ui_url}
\usage{
get_ui_url()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{get_ui_url()}.
}
\section{Provided by}{

The \code{ui} extension (\code{LOAD ui;}).
}

