% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{getvariable}
\alias{getvariable}
\title{DuckDB function getvariable}
\usage{
getvariable(col0 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\description{
DuckDB function \code{getvariable()}.
}
\examples{
\dontrun{
character(0)
}
}
