% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goldendawn}
\alias{goldendawn}
\title{Golden Dawn Vote Share Data}
\format{
A data frame with 380 rows and 4 variables:
\describe{
\item{muni}{Municipality ID}
\item{year}{Year of election}
\item{post}{Post exposure?}
\item{gd}{Vote share of Golden Dawn Party (in \%)}
}
}
\usage{
goldendawn
}
\description{
This data comes from Dinas, Matakos, Xefteris and Hangartner (2019)
who show that the exposure to the European refugee protection crisis in 2015/2016
increased support for the Golden Dawn, a radical right party in Greece.
}
\references{
Dinas Elias and Konstantinos Matakos, Dimitrios Xefteris, and Dominik Hangartner. 2019. "Waking up the Golden Dawn: Does Exposure to the Refugee Crisis Increase Support for Extreme-right Parties?" \emph{Political Analysis}, 27(2), 244–254.
}
\keyword{datasets}
