% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{grade_up}
\alias{grade_up}
\title{DuckDB function grade_up}
\arguments{
\item{list}{\code{ANY[]}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR}}
}
\description{
Works like list_sort, but the results are the indexes that correspond to the position in the original list instead of the actual values.
}
\examples{
\dontrun{
grade_up([3, 6, 1, 2])
}
}
