% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{>>=}
\alias{>>=}
\title{DuckDB function >>=}
\usage{
`>>=`(col0, col1)
}
\arguments{
\item{col0, col1}{\code{INET}}
}
\value{
\code{BOOLEAN}
}
\description{
DuckDB function \verb{>>=()}.
}
\section{Provided by}{

The \code{inet} extension (\code{LOAD inet;}).
}

