% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{greatest_common_divisor}
\alias{greatest_common_divisor}
\title{DuckDB function greatest_common_divisor}
\arguments{
\item{x}{\code{BIGINT | HUGEINT}}

\item{y}{\code{BIGINT | HUGEINT}}
}
\value{
\code{BIGINT | HUGEINT}
}
\description{
Computes the greatest common divisor of x and y.
}
\section{Overloads}{

\itemize{
\item \code{greatest_common_divisor(x = BIGINT, y = BIGINT)}
\item \code{greatest_common_divisor(x = HUGEINT, y = HUGEINT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{greatest_common_divisor(42, 57)
}\if{html}{\out{</div>}}
}

