% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_column_privilege}
\alias{has_column_privilege}
\title{DuckDB function has_column_privilege}
\usage{
has_column_privilege(user, table, column, privilege)
}
\arguments{
\item{user, table, column, privilege}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_column_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_column_privilege(table, column, privilege)}
\item \code{has_column_privilege(user, table, column, privilege)}
}
}

