% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_database_privilege}
\alias{has_database_privilege}
\title{DuckDB function has_database_privilege}
\arguments{
\item{database}{Unspecified.}

\item{privilege}{Unspecified.}

\item{user}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_database_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_database_privilege(database, privilege)}
\item \code{has_database_privilege(user, database, privilege)}
}
}

