% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_foreign_data_wrapper_privilege}
\alias{has_foreign_data_wrapper_privilege}
\title{DuckDB function has_foreign_data_wrapper_privilege}
\arguments{
\item{fdw}{Unspecified.}

\item{privilege}{Unspecified.}

\item{user}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_foreign_data_wrapper_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_foreign_data_wrapper_privilege(fdw, privilege)}
\item \code{has_foreign_data_wrapper_privilege(user, fdw, privilege)}
}
}

