% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_function_privilege}
\alias{has_function_privilege}
\title{DuckDB function has_function_privilege}
\usage{
has_function_privilege(user, function., privilege)
}
\arguments{
\item{user, function., privilege}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_function_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_function_privilege(function., privilege)}
\item \code{has_function_privilege(user, function., privilege)}
}
}

