% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_language_privilege}
\alias{has_language_privilege}
\title{DuckDB function has_language_privilege}
\usage{
has_language_privilege(user, language, privilege)
}
\arguments{
\item{user, language, privilege}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_language_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_language_privilege(language, privilege)}
\item \code{has_language_privilege(user, language, privilege)}
}
}

