% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_schema_privilege}
\alias{has_schema_privilege}
\title{DuckDB function has_schema_privilege}
\usage{
has_schema_privilege(user, schema, privilege)
}
\arguments{
\item{user, schema, privilege}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_schema_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_schema_privilege(schema, privilege)}
\item \code{has_schema_privilege(user, schema, privilege)}
}
}

