% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_server_privilege}
\alias{has_server_privilege}
\title{DuckDB function has_server_privilege}
\arguments{
\item{server}{Unspecified.}

\item{privilege}{Unspecified.}

\item{user}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_server_privilege()}.
}
