% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_table_privilege}
\alias{has_table_privilege}
\title{DuckDB function has_table_privilege}
\usage{
has_table_privilege(user, table, privilege)
}
\arguments{
\item{user, table, privilege}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_table_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_table_privilege(table, privilege)}
\item \code{has_table_privilege(user, table, privilege)}
}
}

