% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_tablespace_privilege}
\alias{has_tablespace_privilege}
\title{DuckDB function has_tablespace_privilege}
\arguments{
\item{tablespace}{Unspecified.}

\item{privilege}{Unspecified.}

\item{user}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_tablespace_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_tablespace_privilege(tablespace, privilege)}
\item \code{has_tablespace_privilege(user, tablespace, privilege)}
}
}

