% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{histogram}
\alias{histogram}
\title{DuckDB function histogram}
\arguments{
\item{arg}{\code{ANY}}

\item{col1}{\code{ANY[]}}

\item{source}{Unspecified.}

\item{col_name}{Unspecified.}

\item{bin_count}{Unspecified.}

\item{technique}{Unspecified.}
}
\value{
\code{MAP}
}
\description{
Returns a LIST of STRUCTs with the fields bucket and count.
}
\section{Overloads}{

\itemize{
\item \code{histogram(arg = ANY, col1 = `ANY[]`)}
\item \code{histogram(arg = ANY)}
\item \code{histogram(source, col_name, bin_count, technique)}
}
}

\examples{
\dontrun{
histogram(A)
}
}
