% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{histogram_exact}
\alias{histogram_exact}
\title{DuckDB function histogram_exact}
\usage{
histogram_exact(arg = ANY, bins = `ANY[]`)
}
\arguments{
\item{arg}{\code{ANY}}

\item{bins}{\code{ANY[]}}
}
\value{
\code{MAP}
}
\description{
Returns a LIST of STRUCTs with the fields bucket and count matching the buckets exactly.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{histogram_exact(A, [0, 1, 2])
}\if{html}{\out{</div>}}
}

