% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{hnsw_compact_index}
\alias{hnsw_compact_index}
\title{DuckDB function hnsw_compact_index}
\usage{
hnsw_compact_index(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{hnsw_compact_index()}.
}
\section{Provided by}{

The \code{vss} extension (\code{LOAD vss;}).
}

