% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{hnsw_index_scan}
\alias{hnsw_index_scan}
\title{DuckDB function hnsw_index_scan}
\usage{
hnsw_index_scan()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{hnsw_index_scan()}.
}
\section{Provided by}{

The \code{vss} extension (\code{LOAD vss;}).
}

