% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{html_escape}
\alias{html_escape}
\title{DuckDB function html_escape}
\usage{
html_escape(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{BOOLEAN}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{html_escape()}.
}
\section{Overloads}{

\itemize{
\item \code{html_escape(col0 = VARCHAR)}
\item \code{html_escape(col0 = VARCHAR, col1 = BOOLEAN)}
}
}

\section{Provided by}{

The \code{inet} extension (\code{LOAD inet;}).
}

