% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{html_unescape}
\alias{html_unescape}
\title{DuckDB function html_unescape}
\usage{
html_unescape(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{html_unescape()}.
}
