% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_bucket}
\alias{iceberg_bucket}
\title{DuckDB function iceberg_bucket}
\usage{
iceberg_bucket(col0, col1)
}
\arguments{
\item{col0}{\code{INTEGER}}

\item{col1}{\verb{INTEGER | BIGINT | VARCHAR | BLOB | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TIMESTAMP_NS | TIME | UUID | DECIMAL}}
}
\value{
\code{INTEGER}
}
\description{
DuckDB function \code{iceberg_bucket()}.
}
\section{Provided by}{

The \code{iceberg} extension (\code{LOAD iceberg;}).
}

