% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_metadata}
\alias{iceberg_metadata}
\title{DuckDB function iceberg_metadata}
\usage{
iceberg_metadata(
  col0,
  snapshot_from_timestamp,
  snapshot_from_id,
  version_name_format,
  version,
  metadata_compression_codec,
  allow_moved_paths
)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{snapshot_from_timestamp}{\code{TIMESTAMP}}

\item{snapshot_from_id}{\code{UBIGINT}}

\item{version_name_format}{\code{VARCHAR}}

\item{version}{\code{VARCHAR}}

\item{metadata_compression_codec}{\code{VARCHAR}}

\item{allow_moved_paths}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{iceberg_metadata()}.
}
\section{Provided by}{

The \code{iceberg} extension (\code{LOAD iceberg;}).
}

