% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_scan}
\alias{iceberg_scan}
\title{DuckDB function iceberg_scan}
\usage{
iceberg_scan(
  col0,
  snapshot_from_timestamp,
  version,
  metadata_compression_codec,
  mode,
  can_have_nan,
  allow_moved_paths,
  filename,
  union_by_name,
  snapshot_from_id,
  debug_use_openssl,
  hive_partitioning,
  parquet_version,
  encryption_config,
  version_name_format,
  hive_types_autocast,
  binary_as_string,
  explicit_cardinality,
  compression,
  file_row_number,
  hive_types
)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{snapshot_from_timestamp}{\code{TIMESTAMP}}

\item{version}{\code{VARCHAR}}

\item{metadata_compression_codec}{\code{VARCHAR}}

\item{mode}{\code{VARCHAR}}

\item{can_have_nan}{\code{BOOLEAN}}

\item{allow_moved_paths}{\code{BOOLEAN}}

\item{filename}{\code{ANY}}

\item{union_by_name}{\code{BOOLEAN}}

\item{snapshot_from_id}{\code{UBIGINT}}

\item{debug_use_openssl}{\code{BOOLEAN}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{parquet_version}{\code{VARCHAR}}

\item{encryption_config}{\code{ANY}}

\item{version_name_format}{\code{VARCHAR}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{binary_as_string}{\code{BOOLEAN}}

\item{explicit_cardinality}{\code{UBIGINT}}

\item{compression}{\code{VARCHAR}}

\item{file_row_number}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{iceberg_scan()}.
}
\section{Provided by}{

The \code{iceberg} extension (\code{LOAD iceberg;}).
}

