% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_schema_properties}
\alias{iceberg_schema_properties}
\title{DuckDB function iceberg_schema_properties}
\usage{
iceberg_schema_properties(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{iceberg_schema_properties()}.
}
