% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_snapshots}
\alias{iceberg_snapshots}
\title{DuckDB function iceberg_snapshots}
\usage{
iceberg_snapshots(col0, version_name_format, version, metadata_compression_codec)
}
\arguments{
\item{col0, version_name_format, version, metadata_compression_codec}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{iceberg_snapshots()}.
}
