% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_table_properties}
\alias{iceberg_table_properties}
\title{DuckDB function iceberg_table_properties}
\usage{
iceberg_table_properties(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{iceberg_table_properties()}.
}
