% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_to_ducklake}
\alias{iceberg_to_ducklake}
\title{DuckDB function iceberg_to_ducklake}
\usage{
iceberg_to_ducklake(col0, col1, skip_tables)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{skip_tables}{\code{VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{iceberg_to_ducklake()}.
}
