% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_truncate}
\alias{iceberg_truncate}
\title{DuckDB function iceberg_truncate}
\usage{
iceberg_truncate(col0, col1)
}
\arguments{
\item{col0}{\code{INTEGER}}

\item{col1}{\code{INTEGER | BIGINT | VARCHAR | BLOB | DECIMAL}}
}
\value{
\code{INTEGER | BIGINT | VARCHAR | BLOB | DECIMAL}
}
\description{
DuckDB function \code{iceberg_truncate()}.
}
