% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{iceberg_verify_equality_deletes}
\alias{iceberg_verify_equality_deletes}
\title{DuckDB function iceberg_verify_equality_deletes}
\usage{
iceberg_verify_equality_deletes()
}
\value{
\code{BOOLEAN}
}
\description{
DuckDB function \code{iceberg_verify_equality_deletes()}.
}
\section{Provided by}{

The \code{iceberg} extension (\code{LOAD iceberg;}).
}

