% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{icu_calendar_names}
\alias{icu_calendar_names}
\title{DuckDB function icu_calendar_names}
\usage{
icu_calendar_names()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{icu_calendar_names()}.
}
\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

