% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{icu_collate_ee}
\alias{icu_collate_ee}
\title{DuckDB function icu_collate_ee}
\usage{
icu_collate_ee(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{icu_collate_ee()}.
}
