% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{icu_collate_nb_no}
\alias{icu_collate_nb_no}
\title{DuckDB function icu_collate_nb_no}
\usage{
icu_collate_nb_no(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{icu_collate_nb_no()}.
}
\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

