% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{icu_collate_st}
\alias{icu_collate_st}
\title{DuckDB function icu_collate_st}
\usage{
icu_collate_st(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{icu_collate_st()}.
}
\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

