% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{icu_sort_key}
\alias{icu_sort_key}
\title{DuckDB function icu_sort_key}
\usage{
icu_sort_key(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{icu_sort_key()}.
}
