% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{import_database}
\alias{import_database}
\title{DuckDB function import_database}
\usage{
import_database(col0 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{import_database()}.
}
