% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{in_search_path}
\alias{in_search_path}
\title{DuckDB function in_search_path}
\usage{
in_search_path(database_name = VARCHAR, schema_name = VARCHAR)
}
\arguments{
\item{database_name}{\code{VARCHAR}}

\item{schema_name}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns whether or not the database/schema are in the search path.
}
\examples{
\dontrun{
in_search_path('memory', 'main')
}
}
