% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{is_histogram_other_bin}
\alias{is_histogram_other_bin}
\title{DuckDB function is_histogram_other_bin}
\usage{
is_histogram_other_bin(val = ANY)
}
\arguments{
\item{val}{\code{ANY}}
}
\value{
\code{BOOLEAN}
}
\description{
Whether or not the provided value is the histogram "other" bin (used for values not belonging to any provided bin).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{is_histogram_other_bin(v)
}\if{html}{\out{</div>}}
}

