% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{isinf}
\alias{isinf}
\title{DuckDB function isinf}
\arguments{
\item{x}{\verb{FLOAT | DOUBLE | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
Returns true if the floating point value is infinite, false otherwise.
}
\examples{
\dontrun{
isinf('Infinity'::float)
}
}
