% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{jaccard}
\alias{jaccard}
\title{DuckDB function jaccard}
\usage{
jaccard(s1 = VARCHAR, s2 = VARCHAR)
}
\arguments{
\item{s1, s2}{\code{VARCHAR}}
}
\value{
\code{DOUBLE}
}
\description{
The Jaccard similarity between two strings. Characters of different cases (e.g., \code{a} and \code{A}) are considered different. Returns a number between 0 and 1.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{jaccard('duck', 'luck')
}\if{html}{\out{</div>}}
}

\seealso{
Other text_similarity:
\code{\link[=damerau_levenshtein]{damerau_levenshtein()}},
\code{\link[=hamming]{hamming()}},
\code{\link[=jaro_similarity]{jaro_similarity()}},
\code{\link[=jaro_winkler_similarity]{jaro_winkler_similarity()}},
\code{\link[=levenshtein]{levenshtein()}}
}
\concept{text_similarity}
