% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json}
\alias{json}
\title{DuckDB function json}
\usage{
json(x)
}
\arguments{
\item{x}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{json()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

