% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_array}
\alias{json_array}
\title{DuckDB function json_array}
\usage{
json_array()
}
\value{
\code{JSON}
}
\description{
DuckDB function \code{json_array()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

