% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_array_length}
\alias{json_array_length}
\title{DuckDB function json_array_length}
\usage{
json_array_length(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{VARCHAR | VARCHAR[]}}
}
\value{
\code{UBIGINT | UBIGINT[]}
}
\description{
DuckDB function \code{json_array_length()}.
}
\section{Overloads}{

\itemize{
\item \code{json_array_length(col0 = VARCHAR)}
\item \code{json_array_length(col0 = VARCHAR, col1 = VARCHAR)}
\item \code{json_array_length(col0 = VARCHAR, col1 = `VARCHAR[]`)}
\item \code{json_array_length(col0 = JSON)}
\item \code{json_array_length(col0 = JSON, col1 = VARCHAR)}
\item \code{json_array_length(col0 = JSON, col1 = `VARCHAR[]`)}
}
}

