% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_contains}
\alias{json_contains}
\title{DuckDB function json_contains}
\usage{
json_contains(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR | JSON}}
}
\value{
\code{BOOLEAN}
}
\description{
DuckDB function \code{json_contains()}.
}
