% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_deserialize_sql}
\alias{json_deserialize_sql}
\title{DuckDB function json_deserialize_sql}
\usage{
json_deserialize_sql(col0)
}
\arguments{
\item{col0}{\code{JSON}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{json_deserialize_sql()}.
}
