% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_execute_serialized_sql}
\alias{json_execute_serialized_sql}
\title{DuckDB function json_execute_serialized_sql}
\usage{
json_execute_serialized_sql(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{json_execute_serialized_sql()}.
}
