% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_exists}
\alias{json_exists}
\title{DuckDB function json_exists}
\usage{
json_exists(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{VARCHAR | VARCHAR[]}}
}
\value{
\code{BOOLEAN | BOOLEAN[]}
}
\description{
DuckDB function \code{json_exists()}.
}
