% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_extract_path}
\alias{json_extract_path}
\title{DuckDB function json_extract_path}
\usage{
json_extract_path(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{BIGINT | VARCHAR | VARCHAR[]}}
}
\value{
\code{JSON | JSON[]}
}
\description{
DuckDB function \code{json_extract_path()}.
}
