% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_extract_string}
\alias{json_extract_string}
\title{DuckDB function json_extract_string}
\usage{
json_extract_string(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{BIGINT | VARCHAR | VARCHAR[]}}
}
\value{
\code{VARCHAR | VARCHAR[]}
}
\description{
DuckDB function \code{json_extract_string()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

