% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_group_array}
\alias{json_group_array}
\title{DuckDB function json_group_array}
\usage{
json_group_array(x)
}
\arguments{
\item{x}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{json_group_array()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

