% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_group_object}
\alias{json_group_object}
\title{DuckDB function json_group_object}
\usage{
json_group_object(n, v)
}
\arguments{
\item{n, v}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{json_group_object()}.
}
