% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_keys}
\alias{json_keys}
\title{DuckDB function json_keys}
\usage{
json_keys(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{VARCHAR | VARCHAR[]}}
}
\value{
\code{VARCHAR[] | VARCHAR[][]}
}
\description{
DuckDB function \code{json_keys()}.
}
\section{Overloads}{

\itemize{
\item \code{json_keys(col0 = VARCHAR)}
\item \code{json_keys(col0 = VARCHAR, col1 = VARCHAR)}
\item \code{json_keys(col0 = VARCHAR, col1 = `VARCHAR[]`)}
\item \code{json_keys(col0 = JSON)}
\item \code{json_keys(col0 = JSON, col1 = VARCHAR)}
\item \code{json_keys(col0 = JSON, col1 = `VARCHAR[]`)}
}
}

