% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_merge_patch}
\alias{json_merge_patch}
\title{DuckDB function json_merge_patch}
\usage{
json_merge_patch(col0, col1)
}
\arguments{
\item{col0, col1}{\code{JSON}}
}
\value{
\code{JSON}
}
\description{
DuckDB function \code{json_merge_patch()}.
}
