% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_object}
\alias{json_object}
\title{DuckDB function json_object}
\usage{
json_object()
}
\value{
\code{JSON}
}
\description{
DuckDB function \code{json_object()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

