% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_pretty}
\alias{json_pretty}
\title{DuckDB function json_pretty}
\usage{
json_pretty(col0)
}
\arguments{
\item{col0}{\code{JSON}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{json_pretty()}.
}
