% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_quote}
\alias{json_quote}
\title{DuckDB function json_quote}
\usage{
json_quote()
}
\value{
\code{JSON}
}
\description{
DuckDB function \code{json_quote()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

