% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_serialize_sql}
\alias{json_serialize_sql}
\title{DuckDB function json_serialize_sql}
\usage{
json_serialize_sql(col0, col1, col2, col3, col4)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{BOOLEAN}}

\item{col2}{\code{BOOLEAN}}

\item{col3}{\code{BOOLEAN}}

\item{col4}{\code{BOOLEAN}}
}
\value{
\code{JSON}
}
\description{
DuckDB function \code{json_serialize_sql()}.
}
\section{Overloads}{

\itemize{
\item \code{json_serialize_sql(col0 = VARCHAR)}
\item \code{json_serialize_sql(col0 = VARCHAR, col1 = BOOLEAN)}
\item \code{json_serialize_sql(col0 = VARCHAR, col1 = BOOLEAN, col2 = BOOLEAN)}
\item \code{json_serialize_sql(col0 = VARCHAR, col1 = BOOLEAN, col2 = BOOLEAN, col3 = BOOLEAN)}
\item \code{json_serialize_sql(col0 = VARCHAR, col1 = BOOLEAN, col2 = BOOLEAN, col3 = BOOLEAN, col4 = BOOLEAN)}
}
}

\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

