% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_structure}
\alias{json_structure}
\title{DuckDB function json_structure}
\usage{
json_structure(col0)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}
}
\value{
\code{JSON}
}
\description{
DuckDB function \code{json_structure()}.
}
