% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_transform}
\alias{json_transform}
\title{DuckDB function json_transform}
\usage{
json_transform(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
DuckDB function \code{json_transform()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

