% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_transform_strict}
\alias{json_transform_strict}
\title{DuckDB function json_transform_strict}
\usage{
json_transform_strict(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
DuckDB function \code{json_transform_strict()}.
}
