% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_valid}
\alias{json_valid}
\title{DuckDB function json_valid}
\usage{
json_valid(col0)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}
}
\value{
\code{BOOLEAN}
}
\description{
DuckDB function \code{json_valid()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

