% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{json_value}
\alias{json_value}
\title{DuckDB function json_value}
\usage{
json_value(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{BIGINT | VARCHAR | VARCHAR[]}}
}
\value{
\code{VARCHAR | VARCHAR[]}
}
\description{
DuckDB function \code{json_value()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

