% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{kahan_sum}
\alias{kahan_sum}
\title{DuckDB function kahan_sum}
\usage{
kahan_sum(arg = DOUBLE)
}
\arguments{
\item{arg}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Calculates the sum using a more accurate floating point summation (Kahan Sum).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{kahan_sum(A)
}\if{html}{\out{</div>}}
}

