% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{kurtosis}
\alias{kurtosis}
\title{DuckDB function kurtosis}
\usage{
kurtosis(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the excess kurtosis (Fisher’s definition) of all input values, with a bias correction according to the sample size.
}
\examples{
\dontrun{
character(0)
}
}
