% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{kurtosis_pop}
\alias{kurtosis_pop}
\title{DuckDB function kurtosis_pop}
\usage{
kurtosis_pop(x)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the excess kurtosis (Fisher’s definition) of all input values, without bias correction.
}
