% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lance_fts}
\alias{lance_fts}
\title{DuckDB function lance_fts}
\usage{
lance_fts(col0, col1, col2, filter, prefilter, k)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR}}

\item{filter}{\code{VARCHAR}}

\item{prefilter}{\code{BOOLEAN}}

\item{k}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{lance_fts()}.
}
