% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lance_hybrid_search}
\alias{lance_hybrid_search}
\title{DuckDB function lance_hybrid_search}
\usage{
lance_hybrid_search(
  col0,
  col1,
  col2,
  col3,
  col4,
  alpha,
  use_index,
  prefilter,
  refine_factor,
  nprobs,
  oversample_factor,
  k
)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{FLOAT[] | DOUBLE[]}}

\item{col3}{\code{VARCHAR}}

\item{col4}{\code{VARCHAR}}

\item{alpha}{\code{FLOAT}}

\item{use_index}{\code{BOOLEAN}}

\item{prefilter}{\code{BOOLEAN}}

\item{refine_factor}{\code{BIGINT}}

\item{nprobs}{\code{BIGINT}}

\item{oversample_factor}{\code{INTEGER}}

\item{k}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{lance_hybrid_search()}.
}
\section{Provided by}{

The \code{lance} extension (\code{LOAD lance;}).
}

