% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lance_vector_search}
\alias{lance_vector_search}
\title{DuckDB function lance_vector_search}
\usage{
lance_vector_search(
  col0,
  col1,
  col2,
  filter,
  explain_verbose,
  use_index,
  prefilter,
  refine_factor,
  nprobs,
  k
)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{FLOAT[] | DOUBLE[]}}

\item{filter}{\code{VARCHAR}}

\item{explain_verbose}{\code{BOOLEAN}}

\item{use_index}{\code{BOOLEAN}}

\item{prefilter}{\code{BOOLEAN}}

\item{refine_factor}{\code{BIGINT}}

\item{nprobs}{\code{BIGINT}}

\item{k}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{lance_vector_search()}.
}
