% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{last_day}
\alias{last_day}
\title{DuckDB function last_day}
\arguments{
\item{ts}{\verb{DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{DATE}
}
\description{
Returns the last day of the month.
}
\section{Overloads}{

\itemize{
\item \code{last_day(ts = DATE)}
\item \code{last_day(ts = TIMESTAMP)}
\item \code{last_day(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
last_day(TIMESTAMP '1992-03-22 01:02:03.1234')
}
}
