% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{last_value}
\alias{last_value}
\title{DuckDB function last_value}
\usage{
last_value(col0 = T)
}
\arguments{
\item{col0}{\code{T}}
}
\value{
\code{T}
}
\description{
DuckDB function \code{last_value()}.
}
