% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lcase}
\alias{lcase}
\title{DuckDB function lcase}
\usage{
lcase(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Converts \code{string} to lower case.
}
\examples{
\dontrun{
lcase('Hello')
}
}
