% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lead}
\alias{lead}
\title{DuckDB function lead}
\usage{
lead(col0, col1, col2)
}
\arguments{
\item{col0}{\code{T}}

\item{col1}{\code{BIGINT}}

\item{col2}{\code{T}}
}
\value{
\code{T}
}
\description{
DuckDB function \code{lead()}.
}
