% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{least_common_multiple}
\alias{least_common_multiple}
\alias{lcm}
\title{DuckDB function least_common_multiple}
\usage{
least_common_multiple(x, y)
}
\arguments{
\item{x, y}{\code{BIGINT | HUGEINT}}
}
\value{
\code{BIGINT | HUGEINT}
}
\description{
Computes the least common multiple of x and y.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{least_common_multiple(42, 57)
}\if{html}{\out{</div>}}
}

