% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{<=>}
\alias{<=>}
\title{DuckDB function <=>}
\arguments{
\item{list1}{\code{FLOAT[] | DOUBLE[]}}

\item{list2}{\code{FLOAT[] | DOUBLE[]}}
}
\value{
\code{FLOAT | DOUBLE}
}
\description{
Computes the cosine distance between two same-sized lists.
}
\section{Overloads}{

\itemize{
\item \code{`<=>`(list1 = `FLOAT[]`, list2 = `FLOAT[]`)}
\item \code{`<=>`(list1 = `DOUBLE[]`, list2 = `DOUBLE[]`)}
}
}

\examples{
\dontrun{
list_cosine_distance([1, 2, 3], [1, 2, 3])
}
}
