% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{levenshtein}
\alias{levenshtein}
\alias{editdist3}
\title{DuckDB function levenshtein}
\usage{
levenshtein(s1 = VARCHAR, s2 = VARCHAR)
}
\arguments{
\item{s1, s2}{\code{VARCHAR}}
}
\value{
\code{BIGINT}
}
\description{
The minimum number of single-character edits (insertions, deletions or substitutions) required to change one string to the other. Characters of different cases (e.g., \code{a} and \code{A}) are considered different.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{levenshtein('duck', 'db')
}\if{html}{\out{</div>}}
}

\seealso{
Other text_similarity:
\code{\link[=damerau_levenshtein]{damerau_levenshtein()}},
\code{\link[=hamming]{hamming()}},
\code{\link[=jaccard]{jaccard()}},
\code{\link[=jaro_similarity]{jaro_similarity()}},
\code{\link[=jaro_winkler_similarity]{jaro_winkler_similarity()}}
}
\concept{text_similarity}
