% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list}
\alias{list}
\title{DuckDB function list}
\usage{
list(arg = T)
}
\arguments{
\item{arg}{\code{T}}
}
\value{
\code{T[]}
}
\description{
Returns a LIST containing all the values of a column.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list(A)
}\if{html}{\out{</div>}}
}

