% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_aggr}
\alias{list_aggr}
\title{DuckDB function list_aggr}
\usage{
list_aggr(list = `ANY[]`, function_name = VARCHAR)
}
\arguments{
\item{list}{\code{ANY[]}}

\item{function_name}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
Executes the aggregate function \code{function_name} on the elements of \code{list}.
}
\examples{
\dontrun{
list_aggregate([1, 2, NULL], 'min')
}
}
