% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_any_value}
\alias{list_any_value}
\title{DuckDB function list_any_value}
\usage{
list_any_value(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_any_value()}.
}
