% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_append}
\alias{list_append}
\title{DuckDB function list_append}
\usage{
list_append(l, e)
}
\arguments{
\item{l}{Unspecified.}

\item{e}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_append()}.
}
