% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_apply}
\alias{list_apply}
\title{DuckDB function list_apply}
\usage{
list_apply(list = `ANY[]`, `lambda(x)` = LAMBDA)
}
\arguments{
\item{list}{\code{ANY[]}}

\item{lambda(x)}{\code{LAMBDA}}
}
\value{
\code{ANY[]}
}
\description{
Returns a list that is the result of applying the \code{lambda} function to each element of the input \code{list}. The return type is defined by the return type of the \code{lambda} function.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list_apply([1, 2, 3], lambda x : x + 1)
}\if{html}{\out{</div>}}
}

