% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_approx_count_distinct}
\alias{list_approx_count_distinct}
\title{DuckDB function list_approx_count_distinct}
\usage{
list_approx_count_distinct(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_approx_count_distinct()}.
}
