% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_bit_or}
\alias{list_bit_or}
\title{DuckDB function list_bit_or}
\usage{
list_bit_or(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_bit_or()}.
}
