% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_bool_or}
\alias{list_bool_or}
\title{DuckDB function list_bool_or}
\usage{
list_bool_or(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_bool_or()}.
}
