% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_cosine_distance}
\alias{list_cosine_distance}
\title{DuckDB function list_cosine_distance}
\arguments{
\item{list1}{\code{FLOAT[] | DOUBLE[]}}

\item{list2}{\code{FLOAT[] | DOUBLE[]}}
}
\value{
\code{FLOAT | DOUBLE}
}
\description{
Computes the cosine distance between two same-sized lists.
}
\section{Overloads}{

\itemize{
\item \code{list_cosine_distance(list1 = `FLOAT[]`, list2 = `FLOAT[]`)}
\item \code{list_cosine_distance(list1 = `DOUBLE[]`, list2 = `DOUBLE[]`)}
}
}

\examples{
\dontrun{
list_cosine_distance([1, 2, 3], [1, 2, 3])
}
}
