% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_distinct}
\alias{list_distinct}
\alias{duckdb_fun_list_distinct}
\title{DuckDB function list_distinct}
\arguments{
\item{list}{\code{ANY[]}}
}
\description{
Removes all duplicates and NULLs from a list. Does not preserve the original order.
}
\examples{
\dontrun{
list_distinct([1, 1, NULL, -3, 1, 5])
}
}
