% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_extract}
\alias{list_extract}
\title{DuckDB function list_extract}
\arguments{
\item{list}{\code{T[] | VARCHAR}}

\item{index}{\code{BIGINT}}
}
\value{
\code{T | VARCHAR}
}
\description{
Extract the \code{index}th (1-based) value from the list.
}
\examples{
\dontrun{
list_extract([4, 5, 6], 3)
}
}
