% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_filter}
\alias{list_filter}
\alias{array_filter}
\alias{filter}
\title{DuckDB function list_filter}
\usage{
list_filter(list, lambda)
}
\arguments{
\item{list}{\code{ANY[]}}

\item{lambda}{\code{LAMBDA}}
}
\value{
\code{ANY[]}
}
\description{
Constructs a list from those elements of the input \code{list} for which the \code{lambda} function returns \code{true}. DuckDB must be able to cast the \code{lambda} function's return type to \code{BOOL}. The return type of \code{list_filter} is the same as the input list's.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list_filter([3, 4, 5], lambda x : x > 4)
}\if{html}{\out{</div>}}
}

\seealso{
Other lambda:
\code{\link[=list_reduce]{list_reduce()}},
\code{\link[=list_transform]{list_transform()}}

Other list:
\code{\link[=array_extract]{array_extract()}},
\code{\link[=array_length]{array_length()}},
\code{\link[=concat]{concat()}},
\code{\link[=contains]{contains()}},
\code{\link[=flatten]{flatten()}},
\code{\link[=generate_series]{generate_series()}},
\code{\link[=len]{len()}},
\code{\link[=list_aggregate]{list_aggregate()}},
\code{\link[=list_concat]{list_concat()}},
\code{\link[=list_contains]{list_contains()}},
\code{\link[=list_cosine_distance]{list_cosine_distance()}},
\code{\link[=list_cosine_similarity]{list_cosine_similarity()}},
\code{\link[=list_distance]{list_distance()}},
\code{\link[=list_distinct]{list_distinct()}},
\code{\link[=list_extract]{list_extract()}},
\code{\link[=list_grade_up]{list_grade_up()}},
\code{\link[=list_has_all]{list_has_all()}},
\code{\link[=list_has_any]{list_has_any()}},
\code{\link[=list_inner_product]{list_inner_product()}},
\code{\link[=list_intersect]{list_intersect()}},
\code{\link[=list_negative_inner_product]{list_negative_inner_product()}},
\code{\link[=list_position]{list_position()}},
\code{\link[=list_reduce]{list_reduce()}},
\code{\link[=list_resize]{list_resize()}},
\code{\link[=list_reverse_sort]{list_reverse_sort()}},
\code{\link[=list_select]{list_select()}},
\code{\link[=list_slice]{list_slice()}},
\code{\link[=list_sort]{list_sort()}},
\code{\link[=list_transform]{list_transform()}},
\code{\link[=list_unique]{list_unique()}},
\code{\link[=list_value]{list_value()}},
\code{\link[=list_where]{list_where()}},
\code{\link[=list_zip]{list_zip()}},
\code{\link{or--or}},
\code{\link[=range]{range()}},
\code{\link[=repeat]{repeat()}},
\code{\link[=unpivot_list]{unpivot_list()}}
}
\concept{lambda}
\concept{list}
