% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_histogram}
\alias{list_histogram}
\title{DuckDB function list_histogram}
\usage{
list_histogram(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_histogram()}.
}
