% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_indexof}
\alias{list_indexof}
\title{DuckDB function list_indexof}
\usage{
list_indexof(list = `T[]`, element = T)
}
\arguments{
\item{list}{\code{T[]}}

\item{element}{\code{T}}
}
\value{
\code{INTEGER}
}
\description{
Returns the index of the \code{element} if the \code{list} contains the \code{element}. If the \code{element} is not found, it returns \code{NULL}.
}
\examples{
\dontrun{
list_indexof([1, 2, NULL], 2)
}
}
