% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_inner_product}
\alias{list_inner_product}
\title{DuckDB function list_inner_product}
\arguments{
\item{list1}{\code{FLOAT[] | DOUBLE[]}}

\item{list2}{\code{FLOAT[] | DOUBLE[]}}
}
\description{
Computes the inner product between two same-sized lists.
}
\examples{
\dontrun{
list_inner_product([1, 2, 3], [1, 2, 3])
}
}
