% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_intersect}
\alias{list_intersect}
\title{DuckDB function list_intersect}
\usage{
list_intersect(list1 = `T[]`, list2 = `T[]`)
}
\arguments{
\item{list1}{\code{T[]}}

\item{list2}{\code{T[]}}
}
\value{
\code{T[]}
}
\description{
Returns a list containing the distinct elements that are present in both \code{list1} and \code{list2}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list_intersect([1, 2, 3], [2, 3, 4])
}\if{html}{\out{</div>}}
}

