% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_kurtosis_pop}
\alias{list_kurtosis_pop}
\title{DuckDB function list_kurtosis_pop}
\usage{
list_kurtosis_pop(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_kurtosis_pop()}.
}
