% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_negative_dot_product}
\alias{list_negative_dot_product}
\title{DuckDB function list_negative_dot_product}
\arguments{
\item{list1}{\code{FLOAT[] | DOUBLE[]}}

\item{list2}{\code{FLOAT[] | DOUBLE[]}}
}
\value{
\code{FLOAT | DOUBLE}
}
\description{
Computes the negative inner product between two same-sized lists.
}
\section{Overloads}{

\itemize{
\item \code{list_negative_dot_product(list1 = `FLOAT[]`, list2 = `FLOAT[]`)}
\item \code{list_negative_dot_product(list1 = `DOUBLE[]`, list2 = `DOUBLE[]`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list_negative_dot_product([1, 2, 3], [1, 2, 3])
}\if{html}{\out{</div>}}
}

