% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_pack}
\alias{list_pack}
\title{DuckDB function list_pack}
\arguments{
\item{any}{\code{T}}
}
\value{
\code{"NULL"[] | T[]}
}
\description{
Creates a LIST containing the argument values.
}
\section{Overloads}{

\itemize{
\item \code{list_pack()}
\item \code{list_pack(any = T)}
}
}

\examples{
\dontrun{
list_pack(4, 5, 6)
}
}
