% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_prepend}
\alias{list_prepend}
\title{DuckDB function list_prepend}
\usage{
list_prepend(e, l)
}
\arguments{
\item{e}{Unspecified.}

\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_prepend()}.
}
