% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_reverse}
\alias{list_reverse}
\title{DuckDB function list_reverse}
\usage{
list_reverse(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_reverse()}.
}
