% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_select}
\alias{list_select}
\alias{duckdb_fun_list_select}
\title{DuckDB function list_select}
\arguments{
\item{value_list}{\code{ANY[]}}

\item{index_list}{\code{BIGINT[]}}
}
\description{
Returns a list based on the elements selected by the index_list.
}
\examples{
\dontrun{
list_select([10, 20, 30, 40], [1, 4])
}
}
