% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_skewness}
\alias{list_skewness}
\title{DuckDB function list_skewness}
\usage{
list_skewness(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_skewness()}.
}
