% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_string_agg}
\alias{list_string_agg}
\title{DuckDB function list_string_agg}
\usage{
list_string_agg(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_string_agg()}.
}
