% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_unique}
\alias{list_unique}
\title{DuckDB function list_unique}
\arguments{
\item{list}{\code{ANY[]}}
}
\description{
Counts the unique elements of a \code{list}.
}
\examples{
\dontrun{
list_unique([1, 1, NULL, -3, 1, 5])
}
}
