% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_value}
\alias{list_value}
\title{DuckDB function list_value}
\arguments{
\item{any}{\code{T}}
}
\value{
\code{"NULL"[] | T[]}
}
\description{
Creates a LIST containing the argument values.
}
\section{Overloads}{

\itemize{
\item \code{list_value()}
\item \code{list_value(any = T)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list_value(4, 5, 6)
}\if{html}{\out{</div>}}
}

