% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_var_samp}
\alias{list_var_samp}
\title{DuckDB function list_var_samp}
\usage{
list_var_samp(l)
}
\arguments{
\item{l}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_var_samp()}.
}
