% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_where}
\alias{list_where}
\title{DuckDB function list_where}
\usage{
list_where(value_list = `T[]`, mask_list = `BOOLEAN[]`)
}
\arguments{
\item{value_list}{\code{T[]}}

\item{mask_list}{\code{BOOLEAN[]}}
}
\value{
\code{T[]}
}
\description{
Returns a list with the \code{BOOLEAN}s in \code{mask_list} applied as a mask to the \code{value_list}.
}
\examples{
\dontrun{
list_where([10, 20, 30, 40], [true, false, false, true])
}
}
