% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ln}
\alias{ln}
\title{DuckDB function ln}
\usage{
ln(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the natural logarithm of x.
}
\examples{
\dontrun{
ln(2)
}
}
