% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{load_aws_credentials}
\alias{load_aws_credentials}
\title{DuckDB function load_aws_credentials}
\usage{
load_aws_credentials(col0, redact_secret, set_region)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{redact_secret}{\code{BOOLEAN}}

\item{set_region}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{load_aws_credentials()}.
}
\section{Overloads}{

\itemize{
\item \code{load_aws_credentials(redact_secret = BOOLEAN, set_region = BOOLEAN)}
\item \code{load_aws_credentials(col0 = VARCHAR, redact_secret = BOOLEAN, set_region = BOOLEAN)}
}
}

\section{Provided by}{

The \code{aws} extension (\code{LOAD aws;}).
}

