% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{log}
\alias{log}
\title{DuckDB function log}
\arguments{
\item{b}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the logarithm of x to base b. b may be omitted, in which case the default 10.
}
\section{Overloads}{

\itemize{
\item \code{log(b = DOUBLE)}
\item \code{log(b = DOUBLE, x = DOUBLE)}
}
}

\examples{
\dontrun{
log(2, 64)
}
}
