% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{log10}
\alias{log10}
\title{DuckDB function log10}
\usage{
log10(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the 10-log of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{log10(1000)
}\if{html}{\out{</div>}}
}

