% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{log2}
\alias{log2}
\title{DuckDB function log2}
\usage{
log2(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the 2-log of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{log2(8)
}\if{html}{\out{</div>}}
}

