% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{mad}
\alias{mad}
\title{DuckDB function mad}
\arguments{
\item{x}{\verb{DECIMAL | FLOAT | DOUBLE | DATE | TIMESTAMP | TIME | TIMESTAMP WITH TIME ZONE | TIME WITH TIME ZONE}}
}
\value{
\code{DECIMAL | FLOAT | DOUBLE | INTERVAL}
}
\description{
Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
}
\section{Overloads}{

\itemize{
\item \code{mad(x = DECIMAL)}: Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
\item \code{mad(x = FLOAT)}: Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
\item \code{mad(x = DOUBLE)}: Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
\item \code{mad(x = DATE)}: Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
\item \code{mad(x = TIMESTAMP)}: Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
\item \code{mad(x = TIME)}: Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
\item \code{mad(x = `TIMESTAMP WITH TIME ZONE`)}: Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
\item \code{mad(x = `TIME WITH TIME ZONE`)}: Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{mad(x)
}\if{html}{\out{</div>}}
}

