% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{make_date}
\alias{make_date}
\title{DuckDB function make_date}
\usage{
make_date(year, month, day)
}
\arguments{
\item{year, month, day}{\code{BIGINT}}
}
\value{
\code{DATE}
}
\description{
\itemize{
\item \code{make_date(year = BIGINT, month = BIGINT, day = BIGINT)}: The date for the given parts.
 \item \code{make_date(date.struct = `STRUCT("year" BIGINT, "month" BIGINT, "day" BIGINT)`)}: The date for the given struct.
 \item \code{make_date(col0 = INTEGER)}: (Description missing.)
 }
}
\section{Overloads}{

\itemize{
\item \code{make_date(col0 = INTEGER)}
\item \code{make_date(year = BIGINT, month = BIGINT, day = BIGINT)}
\item \code{make_date(date.struct = `STRUCT("year" BIGINT, "month" BIGINT, "day" BIGINT)`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{make_date(1992, 9, 20)
make_date(\{'year': 2024, 'month': 11, 'day': 14\})
}\if{html}{\out{</div>}}
}

