% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{make_timestamp}
\alias{make_timestamp}
\title{DuckDB function make_timestamp}
\arguments{
\item{year}{\code{BIGINT}}

\item{month}{\code{BIGINT}}

\item{day}{\code{BIGINT}}

\item{hour}{\code{BIGINT}}

\item{minute}{\code{BIGINT}}

\item{seconds}{\code{DOUBLE}}
}
\value{
\code{TIMESTAMP}
}
\description{
The timestamp for the given parts.
}
\section{Overloads}{

\itemize{
\item \code{make_timestamp(year = BIGINT, month = BIGINT, day = BIGINT, hour = BIGINT, minute = BIGINT, seconds = DOUBLE)}
\item \code{make_timestamp(year = BIGINT)}
}
}

\examples{
\dontrun{
make_timestamp(1992, 9, 20, 13, 34, 27.123456)
}
}
