% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{make_timestamp_ms}
\alias{make_timestamp_ms}
\title{DuckDB function make_timestamp_ms}
\usage{
make_timestamp_ms(nanos = BIGINT)
}
\arguments{
\item{nanos}{\code{BIGINT}}
}
\value{
\code{TIMESTAMP}
}
\description{
The timestamp for the given microseconds since the epoch.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{make_timestamp_ms(1732117793000000)
}\if{html}{\out{</div>}}
}

