% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{make_timestamp_ns}
\alias{make_timestamp_ns}
\title{DuckDB function make_timestamp_ns}
\usage{
make_timestamp_ns(nanos = BIGINT)
}
\arguments{
\item{nanos}{\code{BIGINT}}
}
\description{
The timestamp for the given nanoseconds since epoch.
}
\examples{
\dontrun{
make_timestamp_ns(1732117793000000000)
}
}
