% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{make_timestamptz}
\alias{make_timestamptz}
\title{DuckDB function make_timestamptz}
\usage{
make_timestamptz(col0, col1, col2, col3, col4, col5, col6)
}
\arguments{
\item{col0}{\code{BIGINT}}

\item{col1}{\code{BIGINT}}

\item{col2}{\code{BIGINT}}

\item{col3}{\code{BIGINT}}

\item{col4}{\code{BIGINT}}

\item{col5}{\code{DOUBLE}}

\item{col6}{\code{VARCHAR}}
}
\value{
\verb{TIMESTAMP WITH TIME ZONE}
}
\description{
DuckDB function \code{make_timestamptz()}.
}
\section{Overloads}{

\itemize{
\item \code{make_timestamptz(col0 = BIGINT, col1 = BIGINT, col2 = BIGINT, col3 = BIGINT, col4 = BIGINT, col5 = DOUBLE)}
\item \code{make_timestamptz(col0 = BIGINT, col1 = BIGINT, col2 = BIGINT, col3 = BIGINT, col4 = BIGINT, col5 = DOUBLE, col6 = VARCHAR)}
\item \code{make_timestamptz(col0 = BIGINT)}
}
}

\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

