% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{make_type}
\alias{make_type}
\title{DuckDB function make_type}
\usage{
make_type(name = VARCHAR)
}
\arguments{
\item{name}{\code{VARCHAR}}
}
\value{
\code{TYPE}
}
\description{
Construct a type from its name and optional parameters.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{make_type('DECIMAL', 10, 2)
}\if{html}{\out{</div>}}
}

