% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map}
\alias{map}
\title{DuckDB function map}
\arguments{
\item{keys}{\code{K[]}}

\item{values}{\code{V[]}}
}
\description{
Creates a map from a set of keys and values.
}
\examples{
\dontrun{
map(['key1', 'key2'], ['val1', 'val2'])
}
}
