% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_concat}
\alias{map_concat}
\title{DuckDB function map_concat}
\usage{
map_concat()
}
\value{
\code{LIST}
}
\description{
Returns a map created from merging the input maps, on key collision the value is taken from the last map with that key.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{map_concat(map([1, 2], ['a', 'b']), map([2, 3], ['c', 'd']));
}\if{html}{\out{</div>}}
}

