% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_contains}
\alias{map_contains}
\title{DuckDB function map_contains}
\arguments{
\item{map}{\code{MAP(ANY, ANY)}}

\item{key}{\code{ANY}}
}
\description{
Checks if a map contains a given key.
}
\examples{
\dontrun{
map_contains(MAP {'key1': 10, 'key2': 20, 'key3': 30}, 'key2')
}
}
