% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_entries}
\alias{map_entries}
\title{DuckDB function map_entries}
\usage{
map_entries(map = `MAP(K, V)`)
}
\arguments{
\item{map}{\code{MAP(K, V)}}
}
\value{
\verb{STRUCT("key" K, "value" V)[]}
}
\description{
Returns the map entries as a list of keys/values.
}
\examples{
\dontrun{
map_entries(map(['key'], ['val']))
}
}
