% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_extract_value}
\alias{map_extract_value}
\title{DuckDB function map_extract_value}
\usage{
map_extract_value(map = `MAP(K, V)`, key = K)
}
\arguments{
\item{map}{\code{MAP(K, V)}}

\item{key}{\code{K}}
}
\value{
\code{V}
}
\description{
Returns the value for a given key or NULL if the key is not contained in the map. The type of the key provided in the second parameter must match the type of the map’s keys else an error is returned.
}
\examples{
\dontrun{
map_extract_value(map(['key'], ['val']), 'key')
}
}
