% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_from_entries}
\alias{map_from_entries}
\title{DuckDB function map_from_entries}
\arguments{
\item{map}{\code{STRUCT(K, V)[]}}
}
\description{
Returns a map created from the entries of the array.
}
\examples{
\dontrun{
map_from_entries([{k: 5, v: 'val1'}, {k: 3, v: 'val2'}]);
}
}
