% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_to_pg_oid}
\alias{map_to_pg_oid}
\title{DuckDB function map_to_pg_oid}
\usage{
map_to_pg_oid(type_name)
}
\arguments{
\item{type_name}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{map_to_pg_oid()}.
}
