% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_values}
\alias{map_values}
\title{DuckDB function map_values}
\usage{
map_values(map = `MAP(K, V)`)
}
\arguments{
\item{map}{\code{MAP(K, V)}}
}
\value{
\code{V[]}
}
\description{
Returns the values of a map as a list.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{map_values(map(['key'], ['val']))
}\if{html}{\out{</div>}}
}

