% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{max_by}
\alias{max_by}
\title{DuckDB function max_by}
\arguments{
\item{arg}{\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}}

\item{val}{\verb{INTEGER | BIGINT | HUGEINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | ANY}}

\item{col2}{\code{BIGINT}}
}
\value{
\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY | ANY[]}
}
\description{
Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
}
\examples{
\dontrun{
max_by(A, B)
}
}
