% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{md5_number_upper}
\alias{md5_number_upper}
\title{DuckDB function md5_number_upper}
\usage{
md5_number_upper(param)
}
\arguments{
\item{param}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{md5_number_upper()}.
}
