% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{mean}
\alias{mean}
\title{DuckDB function mean}
\arguments{
\item{x}{\verb{DECIMAL | SMALLINT | INTEGER | BIGINT | HUGEINT | INTERVAL | DOUBLE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TIME | TIME WITH TIME ZONE}}
}
\value{
\verb{DECIMAL | DOUBLE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TIME | TIME WITH TIME ZONE}
}
\description{
Calculates the average value for all tuples in x.
}
\examples{
\dontrun{
SUM(x) / COUNT(*)
}
}
