% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{median}
\alias{median}
\title{DuckDB function median}
\usage{
median(x = ANY)
}
\arguments{
\item{x}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
Returns the middle value of the set. NULL values are ignored. For even value counts, interpolate-able types (numeric, date/time) return the average of the two middle values. Non-interpolate-able types (everything else) return the lower of the two middle values.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{median(x)
}\if{html}{\out{</div>}}
}

