% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{metadata_info}
\alias{metadata_info}
\title{DuckDB function metadata_info}
\usage{
metadata_info()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{metadata_info()}.
}
\examples{
\dontrun{
character(0)
}
}
