% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{microsecond}
\alias{microsecond}
\title{DuckDB function microsecond}
\usage{
microsecond(ts)
}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the microsecond component from a date or timestamp.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{microsecond(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

