% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{millisecond}
\alias{millisecond}
\title{DuckDB function millisecond}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the millisecond component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{millisecond(ts = DATE)}: Extract the millisecond component from a date or timestamp.
\item \code{millisecond(ts = INTERVAL)}: Extract the millisecond component from a date or timestamp.
\item \code{millisecond(ts = TIME)}: Extract the millisecond component from a date or timestamp.
\item \code{millisecond(ts = TIMESTAMP)}: Extract the millisecond component from a date or timestamp.
\item \code{millisecond(ts = `TIME WITH TIME ZONE`)}: Extract the millisecond component from a date or timestamp.
\item \code{millisecond(ts = TIME_NS)}: Extract the millisecond component from a date or timestamp.
\item \code{millisecond(ts = `TIMESTAMP WITH TIME ZONE`)}: Extract the millisecond component from a date or timestamp.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{millisecond(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

