% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{min}
\alias{min}
\title{DuckDB function min}
\arguments{
\item{arg}{\code{ANY}}

\item{col1}{\code{BIGINT}}
}
\value{
\code{ANY | ANY[]}
}
\description{
Returns the minimum value present in arg.
}
\section{Overloads}{

\itemize{
\item \code{min(arg = ANY)}
\item \code{min(arg = ANY, col1 = BIGINT)}
}
}

\examples{
\dontrun{
min(A)
}
}
