% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{minute}
\alias{minute}
\title{DuckDB function minute}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the minute component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{minute(ts = DATE)}: Extract the minute component from a date or timestamp.
\item \code{minute(ts = INTERVAL)}: Extract the minute component from a date or timestamp.
\item \code{minute(ts = TIME)}: Extract the minute component from a date or timestamp.
\item \code{minute(ts = TIMESTAMP)}: Extract the minute component from a date or timestamp.
\item \code{minute(ts = `TIME WITH TIME ZONE`)}: Extract the minute component from a date or timestamp.
\item \code{minute(ts = TIME_NS)}: Extract the minute component from a date or timestamp.
\item \code{minute(ts = `TIMESTAMP WITH TIME ZONE`)}: Extract the minute component from a date or timestamp.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{minute(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

