% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{mode}
\alias{mode}
\title{DuckDB function mode}
\usage{
mode(x = ANY)
}
\arguments{
\item{x}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
Returns the most frequent value for the values within x. NULL values are ignored.
}
