% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{month}
\alias{month}
\title{DuckDB function month}
\usage{
month(ts)
}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the month component from a date or timestamp.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{month(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

