% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{monthname}
\alias{monthname}
\title{DuckDB function monthname}
\arguments{
\item{ts}{\verb{DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{VARCHAR}
}
\description{
The (English) name of the month.
}
\section{Overloads}{

\itemize{
\item \code{monthname(ts = DATE)}
\item \code{monthname(ts = TIMESTAMP)}
\item \code{monthname(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{monthname(TIMESTAMP '1992-09-20')
}\if{html}{\out{</div>}}
}

