% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{multiply}
\alias{multiply}
\alias{*}
\title{DuckDB function multiply}
\usage{
multiply(col0, col1)
}
\arguments{
\item{col0, col1}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | INTERVAL}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | INTERVAL}
}
\description{
DuckDB function \code{multiply()}.
}
