% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{murmur3_32}
\alias{murmur3_32}
\title{DuckDB function murmur3_32}
\usage{
murmur3_32(col0)
}
\arguments{
\item{col0}{\code{ANY}}
}
\value{
\code{INTEGER}
}
\description{
DuckDB function \code{murmur3_32()}.
}
\section{Provided by}{

The \code{ducklake} extension (\code{LOAD ducklake;}).
}

