% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{mysql_clear_cache}
\alias{mysql_clear_cache}
\title{DuckDB function mysql_clear_cache}
\usage{
mysql_clear_cache()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{mysql_clear_cache()}.
}
\section{Provided by}{

The \code{mysql_scanner} extension (\code{LOAD mysql_scanner;}).
}

