% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{mysql_execute}
\alias{mysql_execute}
\title{DuckDB function mysql_execute}
\usage{
mysql_execute(col0, col1, params)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{params}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{mysql_execute()}.
}
