% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{mysql_query}
\alias{mysql_query}
\title{DuckDB function mysql_query}
\usage{
mysql_query(col0, col1, stream_results, params)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{stream_results}{\code{BOOLEAN}}

\item{params}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{mysql_query()}.
}
\section{Provided by}{

The \code{mysql_scanner} extension (\code{LOAD mysql_scanner;}).
}

