% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{nanosecond}
\alias{nanosecond}
\title{DuckDB function nanosecond}
\arguments{
\item{tsns}{\verb{DATE | TIMESTAMP | INTERVAL | TIME | TIME_NS | TIME WITH TIME ZONE | TIMESTAMP_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the nanosecond component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{nanosecond(tsns = DATE)}: Extract the nanosecond component from a date or timestamp.
\item \code{nanosecond(tsns = TIMESTAMP)}: Extract the nanosecond component from a date or timestamp.
\item \code{nanosecond(tsns = INTERVAL)}: Extract the nanosecond component from a date or timestamp.
\item \code{nanosecond(tsns = TIME)}: Extract the nanosecond component from a date or timestamp.
\item \code{nanosecond(tsns = TIME_NS)}: Extract the nanosecond component from a date or timestamp.
\item \code{nanosecond(tsns = `TIME WITH TIME ZONE`)}: Extract the nanosecond component from a date or timestamp.
\item \code{nanosecond(tsns = TIMESTAMP_NS)}: Extract the nanosecond component from a date or timestamp.
\item \code{nanosecond(tsns = `TIMESTAMP WITH TIME ZONE`)}: Extract the nanosecond component from a date or timestamp.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{nanosecond(timestamp_ns '2021-08-03 11:59:44.123456789')
}\if{html}{\out{</div>}}
}

