% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{netmask}
\alias{netmask}
\title{DuckDB function netmask}
\usage{
netmask(col0)
}
\arguments{
\item{col0}{\code{INET}}
}
\value{
\code{INET}
}
\description{
DuckDB function \code{netmask()}.
}
\section{Provided by}{

The \code{inet} extension (\code{LOAD inet;}).
}

