% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{network}
\alias{network}
\title{DuckDB function network}
\usage{
network(col0)
}
\arguments{
\item{col0}{\code{INET}}
}
\value{
\code{INET}
}
\description{
DuckDB function \code{network()}.
}
\section{Provided by}{

The \code{inet} extension (\code{LOAD inet;}).
}

