% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{normalized_interval}
\alias{normalized_interval}
\title{DuckDB function normalized_interval}
\usage{
normalized_interval(interval = INTERVAL)
}
\arguments{
\item{interval}{\code{INTERVAL}}
}
\value{
\code{INTERVAL}
}
\description{
Normalizes an INTERVAL to an equivalent interval.
}
\examples{
\dontrun{
normalized_interval(INTERVAL '30 days')
}
}
