% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{not-__postfix}
\alias{not-__postfix}
\alias{!__postfix}
\title{DuckDB function !__postfix}
\usage{
`!__postfix`(x = INTEGER)
}
\arguments{
\item{x}{\code{INTEGER}}
}
\value{
\code{HUGEINT}
}
\description{
Factorial of x. Computes the product of the current integer and all integers below it.
}
\examples{
\dontrun{
4!
}
}
