% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{not-~~*}
\alias{not-~~*}
\alias{!~~*}
\title{DuckDB function !~~*}
\usage{
`!~~*`(col0 = VARCHAR, col1 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
DuckDB function \verb{!~~*()}.
}
\examples{
\dontrun{
character(0)
}
}
