% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{now}
\alias{now}
\title{DuckDB function now}
\usage{
now()
}
\value{
\verb{TIMESTAMP WITH TIME ZONE}
}
\description{
Returns the current timestamp.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{now()
}\if{html}{\out{</div>}}
}

