% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{nth_value}
\alias{nth_value}
\title{DuckDB function nth_value}
\usage{
nth_value(col0 = T, col1 = BIGINT)
}
\arguments{
\item{col0}{\code{T}}

\item{col1}{\code{BIGINT}}
}
\value{
\code{T}
}
\description{
DuckDB function \code{nth_value()}.
}
