% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{nullif}
\alias{nullif}
\title{DuckDB function nullif}
\usage{
nullif(a, b)
}
\arguments{
\item{a}{Unspecified.}

\item{b}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{nullif()}.
}
