% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{obj_description}
\alias{obj_description}
\title{DuckDB function obj_description}
\usage{
obj_description(object_oid, catalog_name)
}
\arguments{
\item{object_oid}{Unspecified.}

\item{catalog_name}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{obj_description()}.
}
