% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{or-}
\alias{or-}
\alias{|}
\title{DuckDB function |}
\arguments{
\item{left}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}

\item{right}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise OR.
}
\section{Overloads}{

\itemize{
\item \code{`|`(left = TINYINT, right = TINYINT)}: Bitwise OR.
\item \code{`|`(left = SMALLINT, right = SMALLINT)}: Bitwise OR.
\item \code{`|`(left = INTEGER, right = INTEGER)}: Bitwise OR.
\item \code{`|`(left = BIGINT, right = BIGINT)}: Bitwise OR.
\item \code{`|`(left = HUGEINT, right = HUGEINT)}: Bitwise OR.
\item \code{`|`(left = UTINYINT, right = UTINYINT)}: Bitwise OR.
\item \code{`|`(left = USMALLINT, right = USMALLINT)}: Bitwise OR.
\item \code{`|`(left = UINTEGER, right = UINTEGER)}: Bitwise OR.
\item \code{`|`(left = UBIGINT, right = UBIGINT)}: Bitwise OR.
\item \code{`|`(left = UHUGEINT, right = UHUGEINT)}: Bitwise OR.
\item \code{`|`(left = BIT, right = BIT)}: Bitwise OR.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{32 | 3
}\if{html}{\out{</div>}}
}

