% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ord}
\alias{ord}
\title{DuckDB function ord}
\usage{
ord(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{INTEGER}
}
\description{
Returns an \code{INTEGER} representing the \code{unicode} codepoint of the first character in the \code{string}.
}
\examples{
\dontrun{
[unicode('âbcd'), unicode('â'), unicode(''), unicode(NULL)]
}
}
