% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parquet_full_metadata}
\alias{parquet_full_metadata}
\title{DuckDB function parquet_full_metadata}
\usage{
parquet_full_metadata(col0)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{parquet_full_metadata()}.
}
