% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parquet_scan}
\alias{parquet_scan}
\title{DuckDB function parquet_scan}
\usage{
parquet_scan(
  col0,
  can_have_nan,
  filename,
  union_by_name,
  debug_use_openssl,
  hive_partitioning,
  parquet_version,
  encryption_config,
  hive_types_autocast,
  binary_as_string,
  explicit_cardinality,
  compression,
  file_row_number,
  hive_types,
  schema
)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{can_have_nan}{\code{BOOLEAN}}

\item{filename}{\code{ANY}}

\item{union_by_name}{\code{BOOLEAN}}

\item{debug_use_openssl}{\code{BOOLEAN}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{parquet_version}{\code{VARCHAR}}

\item{encryption_config}{\code{ANY}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{binary_as_string}{\code{BOOLEAN}}

\item{explicit_cardinality}{\code{UBIGINT}}

\item{compression}{\code{VARCHAR}}

\item{file_row_number}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}

\item{schema}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{parquet_scan()}.
}
