% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parquet_schema}
\alias{parquet_schema}
\title{DuckDB function parquet_schema}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{parquet_schema()}.
}
\examples{
\dontrun{
character(0)
}
}
