% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parse_delta_filter_logline}
\alias{parse_delta_filter_logline}
\title{DuckDB function parse_delta_filter_logline}
\usage{
parse_delta_filter_logline(x)
}
\arguments{
\item{x}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{parse_delta_filter_logline()}.
}
