% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parse_duckdb_log_message}
\alias{parse_duckdb_log_message}
\title{DuckDB function parse_duckdb_log_message}
\usage{
parse_duckdb_log_message(type = VARCHAR, message = VARCHAR)
}
\arguments{
\item{type}{\code{VARCHAR}}

\item{message}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
Parse the message into the expected logical type.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{parse_duckdb_log_message('FileSystem', log_message)
}\if{html}{\out{</div>}}
}

