% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_clear_cache}
\alias{pg_clear_cache}
\title{DuckDB function pg_clear_cache}
\usage{
pg_clear_cache()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pg_clear_cache()}.
}
\section{Provided by}{

The \code{postgres_scanner} extension (\code{LOAD postgres_scanner;}).
}

