% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_collation_is_visible}
\alias{pg_collation_is_visible}
\title{DuckDB function pg_collation_is_visible}
\usage{
pg_collation_is_visible(collation_oid)
}
\arguments{
\item{collation_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_collation_is_visible()}.
}
