% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_conversion_is_visible}
\alias{pg_conversion_is_visible}
\title{DuckDB function pg_conversion_is_visible}
\usage{
pg_conversion_is_visible(conversion_oid)
}
\arguments{
\item{conversion_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_conversion_is_visible()}.
}
