% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_function_is_visible}
\alias{pg_function_is_visible}
\title{DuckDB function pg_function_is_visible}
\usage{
pg_function_is_visible(function_oid)
}
\arguments{
\item{function_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_function_is_visible()}.
}
