% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_get_constraintdef}
\alias{pg_get_constraintdef}
\title{DuckDB function pg_get_constraintdef}
\arguments{
\item{constraint_oid}{Unspecified.}

\item{pretty_bool}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pg_get_constraintdef()}.
}
