% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_get_expr}
\alias{pg_get_expr}
\title{DuckDB function pg_get_expr}
\usage{
pg_get_expr(pg_node_tree, relation_oid)
}
\arguments{
\item{pg_node_tree}{Unspecified.}

\item{relation_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_get_expr()}.
}
