% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_get_viewdef}
\alias{pg_get_viewdef}
\title{DuckDB function pg_get_viewdef}
\usage{
pg_get_viewdef(oid)
}
\arguments{
\item{oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_get_viewdef()}.
}
