% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_is_other_temp_schema}
\alias{pg_is_other_temp_schema}
\title{DuckDB function pg_is_other_temp_schema}
\usage{
pg_is_other_temp_schema(schema_id)
}
\arguments{
\item{schema_id}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_is_other_temp_schema()}.
}
