% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_opclass_is_visible}
\alias{pg_opclass_is_visible}
\title{DuckDB function pg_opclass_is_visible}
\usage{
pg_opclass_is_visible(opclass_oid)
}
\arguments{
\item{opclass_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_opclass_is_visible()}.
}
