% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_operator_is_visible}
\alias{pg_operator_is_visible}
\title{DuckDB function pg_operator_is_visible}
\usage{
pg_operator_is_visible(operator_oid)
}
\arguments{
\item{operator_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_operator_is_visible()}.
}
