% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_postmaster_start_time}
\alias{pg_postmaster_start_time}
\title{DuckDB function pg_postmaster_start_time}
\usage{
pg_postmaster_start_time()
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_postmaster_start_time()}.
}
