% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_size_pretty}
\alias{pg_size_pretty}
\title{DuckDB function pg_size_pretty}
\usage{
pg_size_pretty(bytes)
}
\arguments{
\item{bytes}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_size_pretty()}.
}
