% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_sleep}
\alias{pg_sleep}
\title{DuckDB function pg_sleep}
\usage{
pg_sleep(seconds)
}
\arguments{
\item{seconds}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_sleep()}.
}
