% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_table_is_visible}
\alias{pg_table_is_visible}
\title{DuckDB function pg_table_is_visible}
\usage{
pg_table_is_visible(table_oid)
}
\arguments{
\item{table_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_table_is_visible()}.
}
