% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_ts_template_is_visible}
\alias{pg_ts_template_is_visible}
\title{DuckDB function pg_ts_template_is_visible}
\usage{
pg_ts_template_is_visible(template_oid)
}
\arguments{
\item{template_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_ts_template_is_visible()}.
}
