% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_type_is_visible}
\alias{pg_type_is_visible}
\title{DuckDB function pg_type_is_visible}
\usage{
pg_type_is_visible(type_oid)
}
\arguments{
\item{type_oid}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_type_is_visible()}.
}
