% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_typeof}
\alias{pg_typeof}
\title{DuckDB function pg_typeof}
\usage{
pg_typeof(expression)
}
\arguments{
\item{expression}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{pg_typeof()}.
}
