% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pi}
\alias{pi}
\title{DuckDB function pi}
\usage{
pi()
}
\value{
\code{DOUBLE}
}
\description{
Returns the value of pi.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{pi()
}\if{html}{\out{</div>}}
}

