% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{position}
\alias{position}
\title{DuckDB function position}
\arguments{
\item{haystack}{\code{VARCHAR}}

\item{needle}{\code{VARCHAR}}
}
\description{
Returns location of first occurrence of needle in haystack, counting from 1. Returns 0 if no match found.
}
\examples{
\dontrun{
instr('test test','es')
}
}
