% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{postgres_attach}
\alias{postgres_attach}
\title{DuckDB function postgres_attach}
\usage{
postgres_attach(col0, suffix, sink_schema, filter_pushdown, source_schema, overwrite)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{suffix}{\code{VARCHAR}}

\item{sink_schema}{\code{VARCHAR}}

\item{filter_pushdown}{\code{BOOLEAN}}

\item{source_schema}{\code{VARCHAR}}

\item{overwrite}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{postgres_attach()}.
}
