% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{postgres_configure_pool}
\alias{postgres_configure_pool}
\title{DuckDB function postgres_configure_pool}
\usage{
postgres_configure_pool(
  health_check_query,
  max_lifetime_millis,
  enable_reaper_thread,
  enable_thread_local_cache,
  wait_timeout_millis,
  max_connections,
  idle_timeout_millis,
  acquire_mode,
  catalog_name
)
}
\arguments{
\item{health_check_query}{\code{VARCHAR}}

\item{max_lifetime_millis}{\code{UBIGINT}}

\item{enable_reaper_thread}{\code{BOOLEAN}}

\item{enable_thread_local_cache}{\code{BOOLEAN}}

\item{wait_timeout_millis}{\code{UBIGINT}}

\item{max_connections}{\code{UBIGINT}}

\item{idle_timeout_millis}{\code{UBIGINT}}

\item{acquire_mode}{\code{VARCHAR}}

\item{catalog_name}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{postgres_configure_pool()}.
}
\section{Provided by}{

The \code{postgres_scanner} extension (\code{LOAD postgres_scanner;}).
}

