% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{postgres_execute}
\alias{postgres_execute}
\title{DuckDB function postgres_execute}
\usage{
postgres_execute(col0, col1, use_transaction)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{use_transaction}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{postgres_execute()}.
}
\section{Provided by}{

The \code{postgres_scanner} extension (\code{LOAD postgres_scanner;}).
}

