% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{postgres_hstore_get}
\alias{postgres_hstore_get}
\title{DuckDB function postgres_hstore_get}
\usage{
postgres_hstore_get(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{postgres_hstore_get()}.
}
\section{Provided by}{

The \code{postgres_scanner} extension (\code{LOAD postgres_scanner;}).
}

