% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{postgres_hstore_to_json}
\alias{postgres_hstore_to_json}
\title{DuckDB function postgres_hstore_to_json}
\usage{
postgres_hstore_to_json(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
\code{JSON}
}
\description{
DuckDB function \code{postgres_hstore_to_json()}.
}
\section{Provided by}{

The \code{postgres_scanner} extension (\code{LOAD postgres_scanner;}).
}

