% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{postgres_query}
\alias{postgres_query}
\title{DuckDB function postgres_query}
\usage{
postgres_query(col0, col1, params, use_transaction)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{params}{\code{ANY}}

\item{use_transaction}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{postgres_query()}.
}
