% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{postgres_scan}
\alias{postgres_scan}
\title{DuckDB function postgres_scan}
\usage{
postgres_scan(col0, col1, col2)
}
\arguments{
\item{col0, col1, col2}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{postgres_scan()}.
}
