% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{postgres_scan_pushdown}
\alias{postgres_scan_pushdown}
\title{DuckDB function postgres_scan_pushdown}
\usage{
postgres_scan_pushdown(col0, col1, col2)
}
\arguments{
\item{col0, col1, col2}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{postgres_scan_pushdown()}.
}
\section{Provided by}{

The \code{postgres_scanner} extension (\code{LOAD postgres_scanner;}).
}

