% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{^-at}
\alias{^-at}
\alias{^@}
\title{DuckDB function ^@}
\usage{
`^@`(string = VARCHAR, search_string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if \code{string} begins with \code{search_string}.
}
\examples{
\dontrun{
starts_with('abc', 'a')
}
}
