% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pow}
\alias{pow}
\alias{**}
\alias{^}
\alias{power}
\title{DuckDB function pow}
\usage{
pow(x = DOUBLE, y = DOUBLE)
}
\arguments{
\item{x, y}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes x to the power of y.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{pow(2, 3)
power(2, 3)
2 ** 3
2 ^ 3
}\if{html}{\out{</div>}}
}

