% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{power}
\alias{power}
\title{DuckDB function power}
\usage{
power(x = DOUBLE, y = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}

\item{y}{\code{DOUBLE}}
}
\description{
Computes x to the power of y.
}
\examples{
\dontrun{
power(2, 3)
}
}
