% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pragma_database_size}
\alias{pragma_database_size}
\title{DuckDB function pragma_database_size}
\usage{
pragma_database_size()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pragma_database_size()}.
}
\examples{
\dontrun{
character(0)
}
}
