% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pragma_platform}
\alias{pragma_platform}
\title{DuckDB function pragma_platform}
\usage{
pragma_platform()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pragma_platform()}.
}
\examples{
\dontrun{
character(0)
}
}
