% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pragma_rtree_index_info}
\alias{pragma_rtree_index_info}
\title{DuckDB function pragma_rtree_index_info}
\usage{
pragma_rtree_index_info()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pragma_rtree_index_info()}.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

