% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pragma_show}
\alias{pragma_show}
\title{DuckDB function pragma_show}
\usage{
pragma_show(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pragma_show()}.
}
