% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pragma_storage_info}
\alias{pragma_storage_info}
\title{DuckDB function pragma_storage_info}
\usage{
pragma_storage_info(col0 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pragma_storage_info()}.
}
\examples{
\dontrun{
character(0)
}
}
