% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pragma_table_info}
\alias{pragma_table_info}
\title{DuckDB function pragma_table_info}
\usage{
pragma_table_info(col0 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pragma_table_info()}.
}
