% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pragma_version}
\alias{pragma_version}
\title{DuckDB function pragma_version}
\usage{
pragma_version()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pragma_version()}.
}
\examples{
\dontrun{
character(0)
}
}
