% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_active_connections}
\alias{quack_active_connections}
\title{DuckDB function quack_active_connections}
\usage{
quack_active_connections()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{quack_active_connections()}.
}
\section{Provided by}{

The \code{quack} extension (\code{LOAD quack;}).
}

