% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_check_token}
\alias{quack_check_token}
\title{DuckDB function quack_check_token}
\usage{
quack_check_token(col0, col1, col2)
}
\arguments{
\item{col0, col1, col2}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
DuckDB function \code{quack_check_token()}.
}
\section{Provided by}{

The \code{quack} extension (\code{LOAD quack;}).
}

