% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_identify}
\alias{quack_identify}
\title{DuckDB function quack_identify}
\usage{
quack_identify(meta, region, hostname, provider, name)
}
\arguments{
\item{meta, region, hostname, provider, name}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{quack_identify()}.
}
\section{Provided by}{

The \code{quack} extension (\code{LOAD quack;}).
}

