% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_nop_authorization}
\alias{quack_nop_authorization}
\title{DuckDB function quack_nop_authorization}
\usage{
quack_nop_authorization(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{quack_nop_authorization()}.
}
