% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_query}
\alias{quack_query}
\title{DuckDB function quack_query}
\usage{
quack_query(col0, col1, token, disable_ssl)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{token}{\code{VARCHAR}}

\item{disable_ssl}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{quack_query()}.
}
