% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_query_by_name}
\alias{quack_query_by_name}
\title{DuckDB function quack_query_by_name}
\usage{
quack_query_by_name(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{quack_query_by_name()}.
}
