% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_serve}
\alias{quack_serve}
\title{DuckDB function quack_serve}
\usage{
quack_serve(col0, token, allow_other_hostname, disable_ssl)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{token}{\code{VARCHAR}}

\item{allow_other_hostname}{\code{BOOLEAN}}

\item{disable_ssl}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{quack_serve()}.
}
\section{Overloads}{

\itemize{
\item \code{quack_serve(col0 = VARCHAR, token = VARCHAR, allow_other_hostname = BOOLEAN, disable_ssl = BOOLEAN)}
\item \code{quack_serve(token = VARCHAR, allow_other_hostname = BOOLEAN, disable_ssl = BOOLEAN)}
}
}

\section{Provided by}{

The \code{quack} extension (\code{LOAD quack;}).
}

