% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_server_list}
\alias{quack_server_list}
\title{DuckDB function quack_server_list}
\usage{
quack_server_list()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{quack_server_list()}.
}
\section{Provided by}{

The \code{quack} extension (\code{LOAD quack;}).
}

