% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_stop}
\alias{quack_stop}
\title{DuckDB function quack_stop}
\usage{
quack_stop(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{quack_stop()}.
}
\section{Provided by}{

The \code{quack} extension (\code{LOAD quack;}).
}

