% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quack_uri_parser}
\alias{quack_uri_parser}
\title{DuckDB function quack_uri_parser}
\usage{
quack_uri_parser(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{BOOLEAN}}
}
\value{
\verb{STRUCT(host VARCHAR, port USMALLINT, ipv6 BOOLEAN, ssl BOOLEAN, url VARCHAR)}
}
\description{
DuckDB function \code{quack_uri_parser()}.
}
