% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quantile}
\alias{quantile}
\title{DuckDB function quantile}
\arguments{
\item{x}{\code{ANY}}

\item{pos}{\code{DOUBLE | DOUBLE[]}}
}
\value{
\code{ANY}
}
\description{
Returns the exact quantile number between 0 and 1 . If pos is a LIST of FLOATs, then the result is a LIST of the corresponding exact quantiles.
}
\section{Overloads}{

\itemize{
\item \code{quantile(x = ANY, pos = DOUBLE)}
\item \code{quantile(x = ANY, pos = `DOUBLE[]`)}
\item \code{quantile(x = ANY)}
}
}

\examples{
\dontrun{
quantile_disc(x, 0.5)
}
}
