% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quantile_disc}
\alias{quantile_disc}
\title{DuckDB function quantile_disc}
\arguments{
\item{x}{\code{ANY}}

\item{pos}{\code{DOUBLE | DOUBLE[]}}
}
\value{
\code{ANY}
}
\description{
Returns the exact quantile number between 0 and 1 . If pos is a LIST of FLOATs, then the result is a LIST of the corresponding exact quantiles.
}
\section{Overloads}{

\itemize{
\item \code{quantile_disc(x = ANY, pos = DOUBLE)}
\item \code{quantile_disc(x = ANY, pos = `DOUBLE[]`)}
\item \code{quantile_disc(x = ANY)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{quantile_disc(x, 0.5)
}\if{html}{\out{</div>}}
}

