% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{query}
\alias{query}
\title{DuckDB function query}
\usage{
query(col0 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{query()}.
}
\examples{
\dontrun{
character(0)
}
}
