% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{random}
\alias{random}
\title{DuckDB function random}
\usage{
random()
}
\value{
\code{DOUBLE}
}
\description{
Returns a random number between 0 and 1.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{random()
}\if{html}{\out{</div>}}
}

