% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{range}
\alias{range}
\title{DuckDB function range}
\usage{
range(start, stop, step)
}
\arguments{
\item{start}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{stop}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{step}{\code{BIGINT | INTERVAL}}
}
\value{
\verb{BIGINT[] | TIMESTAMP[] | TIMESTAMP WITH TIME ZONE[]}
}
\description{
Creates a list of values between \code{start} and \code{stop} - the stop parameter is exclusive.
}
\section{Overloads}{

\itemize{
\item \code{range(col0 = BIGINT)}
\item \code{range(col0 = BIGINT, col1 = BIGINT)}
\item \code{range(col0 = BIGINT, col1 = BIGINT, col2 = BIGINT)}
\item \code{range(col0 = TIMESTAMP, col1 = TIMESTAMP, col2 = INTERVAL)}
\item \code{range(col0 = `TIMESTAMP WITH TIME ZONE`, col1 = `TIMESTAMP WITH TIME ZONE`, col2 = INTERVAL)}
\item \code{range(start = BIGINT)}
\item \code{range(start = BIGINT, stop = BIGINT)}
\item \code{range(start = BIGINT, stop = BIGINT, step = BIGINT)}
\item \code{range(start = TIMESTAMP, stop = TIMESTAMP, step = INTERVAL)}
\item \code{range(start = `TIMESTAMP WITH TIME ZONE`, stop = `TIMESTAMP WITH TIME ZONE`, step = INTERVAL)}
}
}

\section{Provided by}{

\itemize{
\item \code{range(col0 = BIGINT)}: built in
\item \code{range(col0 = BIGINT, col1 = BIGINT)}: built in
\item \code{range(col0 = BIGINT, col1 = BIGINT, col2 = BIGINT)}: built in
\item \code{range(col0 = TIMESTAMP, col1 = TIMESTAMP, col2 = INTERVAL)}: built in
\item \code{range(col0 = `TIMESTAMP WITH TIME ZONE`, col1 = `TIMESTAMP WITH TIME ZONE`, col2 = INTERVAL)}: \code{icu}
\item \code{range(start = BIGINT)}: built in
\item \code{range(start = BIGINT, stop = BIGINT)}: built in
\item \code{range(start = BIGINT, stop = BIGINT, step = BIGINT)}: built in
\item \code{range(start = TIMESTAMP, stop = TIMESTAMP, step = INTERVAL)}: built in
\item \code{range(start = `TIMESTAMP WITH TIME ZONE`, stop = `TIMESTAMP WITH TIME ZONE`, step = INTERVAL)}: built in
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{range(2, 5, 3)
}\if{html}{\out{</div>}}
}

\seealso{
Other list:
\code{\link[=array_extract]{array_extract()}},
\code{\link[=array_length]{array_length()}},
\code{\link[=concat]{concat()}},
\code{\link[=contains]{contains()}},
\code{\link[=flatten]{flatten()}},
\code{\link[=generate_series]{generate_series()}},
\code{\link[=len]{len()}},
\code{\link[=list_aggregate]{list_aggregate()}},
\code{\link[=list_concat]{list_concat()}},
\code{\link[=list_contains]{list_contains()}},
\code{\link[=list_cosine_distance]{list_cosine_distance()}},
\code{\link[=list_cosine_similarity]{list_cosine_similarity()}},
\code{\link[=list_distance]{list_distance()}},
\code{\link[=list_distinct]{list_distinct()}},
\code{\link[=list_extract]{list_extract()}},
\code{\link[=list_filter]{list_filter()}},
\code{\link[=list_grade_up]{list_grade_up()}},
\code{\link[=list_has_all]{list_has_all()}},
\code{\link[=list_has_any]{list_has_any()}},
\code{\link[=list_inner_product]{list_inner_product()}},
\code{\link[=list_intersect]{list_intersect()}},
\code{\link[=list_negative_inner_product]{list_negative_inner_product()}},
\code{\link[=list_position]{list_position()}},
\code{\link[=list_reduce]{list_reduce()}},
\code{\link[=list_resize]{list_resize()}},
\code{\link[=list_reverse_sort]{list_reverse_sort()}},
\code{\link[=list_select]{list_select()}},
\code{\link[=list_slice]{list_slice()}},
\code{\link[=list_sort]{list_sort()}},
\code{\link[=list_transform]{list_transform()}},
\code{\link[=list_unique]{list_unique()}},
\code{\link[=list_value]{list_value()}},
\code{\link[=list_where]{list_where()}},
\code{\link[=list_zip]{list_zip()}},
\code{\link{or--or}},
\code{\link[=repeat]{repeat()}},
\code{\link[=unpivot_list]{unpivot_list()}}
}
\concept{list}
