% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_avro}
\alias{read_avro}
\title{DuckDB function read_avro}
\usage{
read_avro(
  col0,
  hive_types_autocast,
  hive_types,
  hive_partitioning,
  union_by_name,
  filename
)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{union_by_name}{\code{BOOLEAN}}

\item{filename}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_avro()}.
}
\section{Provided by}{

The \code{avro} extension (\code{LOAD avro;}).
}

