% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_json}
\alias{read_json}
\title{DuckDB function read_json}
\usage{
read_json(
  col0,
  convert_strings_to_integers,
  maximum_sample_files,
  timestamp_format,
  field_appearance_threshold,
  timestampformat,
  map_inference_threshold,
  date_format,
  filename,
  union_by_name,
  compression,
  maximum_depth,
  columns,
  sample_size,
  hive_types,
  hive_types_autocast,
  maximum_object_size,
  format,
  ignore_errors,
  hive_partitioning,
  auto_detect,
  records,
  dateformat
)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{convert_strings_to_integers}{\code{BOOLEAN}}

\item{maximum_sample_files}{\code{BIGINT}}

\item{timestamp_format}{\code{VARCHAR}}

\item{field_appearance_threshold}{\code{DOUBLE}}

\item{timestampformat}{\code{VARCHAR}}

\item{map_inference_threshold}{\code{BIGINT}}

\item{date_format}{\code{VARCHAR}}

\item{filename}{\code{ANY}}

\item{union_by_name}{\code{BOOLEAN}}

\item{compression}{\code{VARCHAR}}

\item{maximum_depth}{\code{BIGINT}}

\item{columns}{\code{ANY}}

\item{sample_size}{\code{BIGINT}}

\item{hive_types}{\code{ANY}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{maximum_object_size}{\code{UINTEGER}}

\item{format}{\code{VARCHAR}}

\item{ignore_errors}{\code{BOOLEAN}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{auto_detect}{\code{BOOLEAN}}

\item{records}{\code{VARCHAR}}

\item{dateformat}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_json()}.
}
