% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_json_objects}
\alias{read_json_objects}
\title{DuckDB function read_json_objects}
\usage{
read_json_objects(
  col0,
  ignore_errors,
  format,
  maximum_object_size,
  hive_types_autocast,
  hive_types,
  hive_partitioning,
  compression,
  union_by_name,
  filename
)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{ignore_errors}{\code{BOOLEAN}}

\item{format}{\code{VARCHAR}}

\item{maximum_object_size}{\code{UINTEGER}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{compression}{\code{VARCHAR}}

\item{union_by_name}{\code{BOOLEAN}}

\item{filename}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_json_objects()}.
}
