% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_parquet}
\alias{read_parquet}
\title{DuckDB function read_parquet}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{can_have_nan}{\code{BOOLEAN}}

\item{filename}{\code{ANY}}

\item{union_by_name}{\code{BOOLEAN}}

\item{debug_use_openssl}{\code{BOOLEAN}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{parquet_version}{\code{VARCHAR}}

\item{encryption_config}{\code{ANY}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{binary_as_string}{\code{BOOLEAN}}

\item{explicit_cardinality}{\code{UBIGINT}}

\item{compression}{\code{VARCHAR}}

\item{file_row_number}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}

\item{schema}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_parquet()}.
}
\section{Overloads}{

\itemize{
\item \code{read_parquet(col0 = VARCHAR, can_have_nan = BOOLEAN, filename = ANY, union_by_name = BOOLEAN, debug_use_openssl = BOOLEAN, hive_partitioning = BOOLEAN, parquet_version = VARCHAR, encryption_config = ANY, hive_types_autocast = BOOLEAN, binary_as_string = BOOLEAN, explicit_cardinality = UBIGINT, compression = VARCHAR, file_row_number = BOOLEAN, hive_types = ANY, schema = ANY)}
\item \code{read_parquet(col0 = `VARCHAR[]`, can_have_nan = BOOLEAN, filename = ANY, union_by_name = BOOLEAN, debug_use_openssl = BOOLEAN, hive_partitioning = BOOLEAN, parquet_version = VARCHAR, encryption_config = ANY, hive_types_autocast = BOOLEAN, binary_as_string = BOOLEAN, explicit_cardinality = UBIGINT, compression = VARCHAR, file_row_number = BOOLEAN, hive_types = ANY, schema = ANY)}
}
}

