% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_postgres_binary}
\alias{read_postgres_binary}
\title{DuckDB function read_postgres_binary}
\usage{
read_postgres_binary(col0, buffer_size, columns)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{buffer_size}{\code{UBIGINT}}

\item{columns}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_postgres_binary()}.
}
