% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_text}
\alias{read_text}
\title{DuckDB function read_text}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_text()}.
}
\section{Overloads}{

\itemize{
\item \code{read_text(col0 = VARCHAR)}
\item \code{read_text(col0 = `VARCHAR[]`)}
}
}

