% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_vortex}
\alias{read_vortex}
\title{DuckDB function read_vortex}
\usage{
read_vortex(col0)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_vortex()}.
}
\section{Provided by}{

The \code{vortex} extension (\code{LOAD vortex;}).
}

