% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_xlsx}
\alias{read_xlsx}
\title{DuckDB function read_xlsx}
\usage{
read_xlsx(
  col0,
  normalize_names,
  empty_as_varchar,
  stop_at_empty,
  sheet,
  range,
  ignore_errors,
  all_varchar,
  header
)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{normalize_names}{\code{BOOLEAN}}

\item{empty_as_varchar}{\code{BOOLEAN}}

\item{stop_at_empty}{\code{BOOLEAN}}

\item{sheet}{\code{VARCHAR}}

\item{range}{\code{VARCHAR}}

\item{ignore_errors}{\code{BOOLEAN}}

\item{all_varchar}{\code{BOOLEAN}}

\item{header}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_xlsx()}.
}
\section{Provided by}{

The \code{excel} extension (\code{LOAD excel;}).
}

