% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{reduce}
\alias{reduce}
\title{DuckDB function reduce}
\arguments{
\item{list}{\code{ANY[]}}

\item{lambda}{\code{LAMBDA}}
}
\description{
Returns a single value that is the result of applying the lambda function to each element of the input list, starting with the first element and then repeatedly applying the lambda function to the result of the previous application and the next element of the list.
}
\examples{
\dontrun{
list_reduce([1, 2, 3], (x, y) -> x + y)
}
}
