% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_escape}
\alias{regexp_escape}
\title{DuckDB function regexp_escape}
\usage{
regexp_escape(string)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Escapes special patterns to turn \code{string} into a regular expression similarly to Python's \code{re.escape} function.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{regexp_escape('https://duckdb.org')
}\if{html}{\out{</div>}}
}

\seealso{
Other regex:
\code{\link[=regexp_extract]{regexp_extract()}},
\code{\link[=regexp_extract_all]{regexp_extract_all()}},
\code{\link[=regexp_full_match]{regexp_full_match()}},
\code{\link[=regexp_matches]{regexp_matches()}},
\code{\link[=regexp_replace]{regexp_replace()}},
\code{\link[=string_split_regex]{string_split_regex()}}
}
\concept{regex}
