% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_extract}
\alias{regexp_extract}
\title{DuckDB function regexp_extract}
\usage{
regexp_extract(string, regex, group, options)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{group}{\code{INTEGER}}

\item{options}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
\itemize{
\item \code{regexp_extract(string = VARCHAR, regex = VARCHAR)}, \code{regexp_extract(string = VARCHAR, regex = VARCHAR, group = INTEGER)}, \code{regexp_extract(string = VARCHAR, regex = VARCHAR, group = INTEGER, options = VARCHAR)}: If \code{string} contains the \code{regex} pattern, returns the capturing group specified by optional parameter \code{group}; otherwise, returns the empty string. The \code{group} must be a constant value. If no \code{group} is given, it defaults to 0. A set of optional regex \code{options} can be set.
 \item \code{regexp_extract(string = VARCHAR, regex = VARCHAR, name_list = `VARCHAR[]`)}, \code{regexp_extract(string = VARCHAR, regex = VARCHAR, name_list = `VARCHAR[]`, options = VARCHAR)}: If \code{string} contains the \code{regex} pattern, returns the capturing groups as a struct with corresponding names from \code{name_list}; otherwise, returns a struct with the same keys and empty strings as values. A set of optional regex \code{options} can be set.
 }
}
\section{Overloads}{

\itemize{
\item \code{regexp_extract(string = VARCHAR, regex = VARCHAR)}
\item \code{regexp_extract(string = VARCHAR, regex = VARCHAR, group = INTEGER)}
\item \code{regexp_extract(string = VARCHAR, regex = VARCHAR, group = INTEGER, options = VARCHAR)}
\item \code{regexp_extract(string = VARCHAR, regex = VARCHAR, name_list = `VARCHAR[]`)}
\item \code{regexp_extract(string = VARCHAR, regex = VARCHAR, name_list = `VARCHAR[]`, options = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{regexp_extract('abcde', '[a-z]\{3\}')
regexp_extract('abc', '([a-z])(b)', 1)
regexp_extract('ABC', '([a-z])(b)', 1, 'i')
regexp_extract('2023-04-15', '(\\d+)-(\\d+)-(\\d+)', ['y', 'm', 'd'])
regexp_extract('John Doe', '([a-z]+) ([a-z]+)', ['first_name', 'last_name'], 'i')
}\if{html}{\out{</div>}}
}

\seealso{
Other regex:
\code{\link[=regexp_escape]{regexp_escape()}},
\code{\link[=regexp_extract_all]{regexp_extract_all()}},
\code{\link[=regexp_full_match]{regexp_full_match()}},
\code{\link[=regexp_matches]{regexp_matches()}},
\code{\link[=regexp_replace]{regexp_replace()}},
\code{\link[=string_split_regex]{string_split_regex()}}
}
\concept{regex}
