% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_extract_all}
\alias{regexp_extract_all}
\title{DuckDB function regexp_extract_all}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{group}{\code{INTEGER | VARCHAR[]}}

\item{options}{\code{VARCHAR}}
}
\value{
\code{VARCHAR[]}
}
\description{
Finds non-overlapping occurrences of the \code{regex} in the \code{string} and returns the corresponding values of the capturing \code{group}. A set of optional regex \code{options} can be set.
}
\section{Overloads}{

\itemize{
\item \code{regexp_extract_all(string = VARCHAR, regex = VARCHAR)}
\item \code{regexp_extract_all(string = VARCHAR, regex = VARCHAR, group = INTEGER)}
\item \code{regexp_extract_all(string = VARCHAR, regex = VARCHAR, group = INTEGER, options = VARCHAR)}
\item \code{regexp_extract_all(string = VARCHAR, regex = VARCHAR, group = `VARCHAR[]`)}
\item \code{regexp_extract_all(string = VARCHAR, regex = VARCHAR, group = `VARCHAR[]`, options = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{regexp_extract_all('Peter: 33, Paul:14', '(\\w+):\\s*(\\d+)', 2)
}\if{html}{\out{</div>}}
}

\seealso{
Other regex:
\code{\link[=regexp_escape]{regexp_escape()}},
\code{\link[=regexp_extract]{regexp_extract()}},
\code{\link[=regexp_full_match]{regexp_full_match()}},
\code{\link[=regexp_matches]{regexp_matches()}},
\code{\link[=regexp_replace]{regexp_replace()}},
\code{\link[=string_split_regex]{string_split_regex()}}
}
\concept{regex}
