% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_full_match}
\alias{regexp_full_match}
\title{DuckDB function regexp_full_match}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if the entire \code{string} matches the \code{regex}. A set of optional regex \code{options} can be set.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{regexp_full_match('anabanana', '(an)*')
}\if{html}{\out{</div>}}
}

\seealso{
Other regex:
\code{\link[=regexp_escape]{regexp_escape()}},
\code{\link[=regexp_extract]{regexp_extract()}},
\code{\link[=regexp_extract_all]{regexp_extract_all()}},
\code{\link[=regexp_matches]{regexp_matches()}},
\code{\link[=regexp_replace]{regexp_replace()}},
\code{\link[=string_split_regex]{string_split_regex()}}
}
\concept{regex}
