% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_matches}
\alias{regexp_matches}
\title{DuckDB function regexp_matches}
\arguments{
\item{string, regex, options}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if \code{string} contains the \code{regex}, \code{false} otherwise. A set of optional regex \code{options} can be set.
}
\section{Overloads}{

\itemize{
\item \code{regexp_matches(string = VARCHAR, regex = VARCHAR)}
\item \code{regexp_matches(string = VARCHAR, regex = VARCHAR, options = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{regexp_matches('anabanana', '(an)*')
}\if{html}{\out{</div>}}
}

\seealso{
Other regex:
\code{\link[=regexp_escape]{regexp_escape()}},
\code{\link[=regexp_extract]{regexp_extract()}},
\code{\link[=regexp_extract_all]{regexp_extract_all()}},
\code{\link[=regexp_full_match]{regexp_full_match()}},
\code{\link[=regexp_replace]{regexp_replace()}},
\code{\link[=string_split_regex]{string_split_regex()}}
}
\concept{regex}
