% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_replace}
\alias{regexp_replace}
\title{DuckDB function regexp_replace}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{replacement}{\code{VARCHAR}}

\item{options}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
If \code{string} contains the \code{regex}, replaces the matching part with \code{replacement}. A set of optional regex \code{options} can be set.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{regexp_replace('hello', '[lo]', '-')
}\if{html}{\out{</div>}}
}

\seealso{
Other regex:
\code{\link[=regexp_escape]{regexp_escape()}},
\code{\link[=regexp_extract]{regexp_extract()}},
\code{\link[=regexp_extract_all]{regexp_extract_all()}},
\code{\link[=regexp_full_match]{regexp_full_match()}},
\code{\link[=regexp_matches]{regexp_matches()}},
\code{\link[=string_split_regex]{string_split_regex()}}
}
\concept{regex}
