% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_split_to_array}
\alias{regexp_split_to_array}
\title{DuckDB function regexp_split_to_array}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{options}{\code{VARCHAR}}
}
\description{
Splits the \code{string} along the \code{regex}. A set of optional regex \code{options} can be set.
}
\examples{
\dontrun{
regexp_split_to_array('hello world; 42', ';? ')
}
}
