% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_split_to_table}
\alias{regexp_split_to_table}
\title{DuckDB function regexp_split_to_table}
\usage{
regexp_split_to_table(text, pattern)
}
\arguments{
\item{text}{Unspecified.}

\item{pattern}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{regexp_split_to_table()}.
}
