% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regr_avgx}
\alias{regr_avgx}
\title{DuckDB function regr_avgx}
\usage{
regr_avgx(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the average of the independent variable for non-NULL pairs in a group, where x is the independent variable and y is the dependent variable.
}
\examples{
\dontrun{
character(0)
}
}
