% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regr_count}
\alias{regr_count}
\title{DuckDB function regr_count}
\usage{
regr_count(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{UINTEGER}
}
\description{
Returns the number of non-NULL number pairs in a group.
}
\examples{
\dontrun{
(SUM(x*y) - SUM(x) * SUM(y) / COUNT(*)) / COUNT(*)
}
}
