% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regr_intercept}
\alias{regr_intercept}
\title{DuckDB function regr_intercept}
\usage{
regr_intercept(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the intercept of the univariate linear regression line for non-NULL pairs in a group.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{AVG(y)-REGR_SLOPE(y, x)*AVG(x)
}\if{html}{\out{</div>}}
}

