% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regr_r2}
\alias{regr_r2}
\title{DuckDB function regr_r2}
\usage{
regr_r2(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the coefficient of determination for non-NULL pairs in a group.
}
