% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regr_sxx}
\alias{regr_sxx}
\title{DuckDB function regr_sxx}
\usage{
regr_sxx(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
DuckDB function \code{regr_sxx()}.
}
\examples{
\dontrun{
REGR_COUNT(y, x) * VAR_POP(x)
}
}
