% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regr_sxy}
\alias{regr_sxy}
\title{DuckDB function regr_sxy}
\usage{
regr_sxy(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the population covariance of input values.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{REGR_COUNT(y, x) * COVAR_POP(y, x)
}\if{html}{\out{</div>}}
}

