% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regr_syy}
\alias{regr_syy}
\title{DuckDB function regr_syy}
\usage{
regr_syy(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
DuckDB function \code{regr_syy()}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{REGR_COUNT(y, x) * VAR_POP(y)
}\if{html}{\out{</div>}}
}

