% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{remap_struct}
\alias{remap_struct}
\title{DuckDB function remap_struct}
\usage{
remap_struct(input, target_type, mapping, defaults)
}
\arguments{
\item{input, target_type, mapping, defaults}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
Map a struct to another struct type, potentially re-ordering, renaming and casting members and filling in defaults for missing values.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{remap_struct(\{'i': 1, 'j': 2\}, NULL::ROW(v1 INT, v2 INT, v3 INT), \{'v1': 'j', 'v3': 'i'\}, \{'v2': NULL::INTEGER\})
}\if{html}{\out{</div>}}
}

