% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{remove_iceberg_schema_properties}
\alias{remove_iceberg_schema_properties}
\title{DuckDB function remove_iceberg_schema_properties}
\usage{
remove_iceberg_schema_properties(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{remove_iceberg_schema_properties()}.
}
\section{Provided by}{

The \code{iceberg} extension (\code{LOAD iceberg;}).
}

