% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{repeat_row}
\alias{repeat_row}
\title{DuckDB function repeat_row}
\usage{
repeat_row(num_rows = BIGINT)
}
\arguments{
\item{num_rows}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{repeat_row()}.
}
