% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{reservoir_quantile}
\alias{reservoir_quantile}
\title{DuckDB function reservoir_quantile}
\arguments{
\item{x}{\code{DECIMAL | TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE}}

\item{quantile}{\code{DOUBLE | DOUBLE[]}}

\item{sample_size}{\code{INTEGER}}
}
\value{
\code{DECIMAL | DECIMAL[] | TINYINT | TINYINT[] | SMALLINT | SMALLINT[] | INTEGER | INTEGER[] | BIGINT | BIGINT[] | HUGEINT | HUGEINT[] | FLOAT | FLOAT[] | DOUBLE | DOUBLE[]}
}
\description{
Gives the approximate quantile using reservoir sampling, the sample size is optional and uses 8192 as a default size.
}
\examples{
\dontrun{
reservoir_quantile(A, 0.5, 1024)
}
}
