% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{reverse}
\alias{reverse}
\title{DuckDB function reverse}
\usage{
reverse(string)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Reverses the \code{string}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{reverse('hello')
}\if{html}{\out{</div>}}
}

\seealso{
Other string:
\code{\link[=array_extract]{array_extract()}},
\code{\link[=ascii]{ascii()}},
\code{\link[=bar]{bar()}},
\code{\link[=bin]{bin()}},
\code{\link[=bit_length]{bit_length()}},
\code{\link[=chr]{chr()}},
\code{\link[=concat]{concat()}},
\code{\link[=concat_ws]{concat_ws()}},
\code{\link[=contains]{contains()}},
\code{\link{format}},
\code{\link[=formatReadableDecimalSize]{formatReadableDecimalSize()}},
\code{\link[=format_bytes]{format_bytes()}},
\code{\link[=from_base64]{from_base64()}},
\code{\link[=greatest]{greatest()}},
\code{\link[=hash]{hash()}},
\code{\link[=hex]{hex()}},
\code{\link[=ilike_escape]{ilike_escape()}},
\code{\link[=instr]{instr()}},
\code{\link[=least]{least()}},
\code{\link[=left]{left()}},
\code{\link[=left_grapheme]{left_grapheme()}},
\code{\link[=len]{len()}},
\code{\link[=length_grapheme]{length_grapheme()}},
\code{\link[=like_escape]{like_escape()}},
\code{\link[=list_slice]{list_slice()}},
\code{\link[=lower]{lower()}},
\code{\link[=lpad]{lpad()}},
\code{\link[=ltrim]{ltrim()}},
\code{\link[=md5]{md5()}},
\code{\link[=md5_number]{md5_number()}},
\code{\link[=nfc_normalize]{nfc_normalize()}},
\code{\link[=not_ilike_escape]{not_ilike_escape()}},
\code{\link[=not_like_escape]{not_like_escape()}},
\code{\link{or--or}},
\code{\link[=parse_dirname]{parse_dirname()}},
\code{\link[=parse_dirpath]{parse_dirpath()}},
\code{\link[=parse_filename]{parse_filename()}},
\code{\link[=parse_formatted_bytes]{parse_formatted_bytes()}},
\code{\link[=parse_path]{parse_path()}},
\code{\link[=prefix]{prefix()}},
\code{\link[=printf]{printf()}},
\code{\link[=repeat]{repeat()}},
\code{\link[=replace]{replace()}},
\code{\link[=right]{right()}},
\code{\link[=right_grapheme]{right_grapheme()}},
\code{\link[=rpad]{rpad()}},
\code{\link[=rtrim]{rtrim()}},
\code{\link[=sha1]{sha1()}},
\code{\link[=sha256]{sha256()}},
\code{\link[=starts_with]{starts_with()}},
\code{\link[=string_split]{string_split()}},
\code{\link[=strip_accents]{strip_accents()}},
\code{\link[=strlen]{strlen()}},
\code{\link[=substring]{substring()}},
\code{\link[=substring_grapheme]{substring_grapheme()}},
\code{\link[=suffix]{suffix()}},
\code{\link[=to_base]{to_base()}},
\code{\link[=to_base64]{to_base64()}},
\code{\link[=translate]{translate()}},
\code{\link[=trim]{trim()}},
\code{\link[=unbin]{unbin()}},
\code{\link[=unhex]{unhex()}},
\code{\link[=unicode]{unicode()}},
\code{\link[=upper]{upper()}},
\code{\link[=url_decode]{url_decode()}},
\code{\link[=url_encode]{url_encode()}}
}
\concept{string}
