% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{round}
\alias{round}
\title{DuckDB function round}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL}}

\item{precision}{\code{INTEGER}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL}
}
\description{
Rounds x to s decimal places.
}
\section{Overloads}{

\itemize{
\item \code{round(x = TINYINT)}: Rounds x to s decimal places.
\item \code{round(x = TINYINT, precision = INTEGER)}: Rounds x to s decimal places.
\item \code{round(x = SMALLINT)}: Rounds x to s decimal places.
\item \code{round(x = SMALLINT, precision = INTEGER)}: Rounds x to s decimal places.
\item \code{round(x = INTEGER)}: Rounds x to s decimal places.
\item \code{round(x = INTEGER, precision = INTEGER)}: Rounds x to s decimal places.
\item \code{round(x = BIGINT)}: Rounds x to s decimal places.
\item \code{round(x = BIGINT, precision = INTEGER)}: Rounds x to s decimal places.
\item \code{round(x = HUGEINT)}: Rounds x to s decimal places.
\item \code{round(x = HUGEINT, precision = INTEGER)}: Rounds x to s decimal places.
\item \code{round(x = FLOAT)}: Rounds x to s decimal places.
\item \code{round(x = FLOAT, precision = INTEGER)}: Rounds x to s decimal places.
\item \code{round(x = DOUBLE)}: Rounds x to s decimal places.
\item \code{round(x = DOUBLE, precision = INTEGER)}: Rounds x to s decimal places.
\item \code{round(x = DECIMAL)}: Rounds x to s decimal places.
\item \code{round(x = DECIMAL, precision = INTEGER)}: Rounds x to s decimal places.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{round(42.4332, 2)
}\if{html}{\out{</div>}}
}

