% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{round_even}
\alias{round_even}
\title{DuckDB function round_even}
\usage{
round_even(x, n)
}
\arguments{
\item{x}{Unspecified.}

\item{n}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{round_even()}.
}
