% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{roundbankers}
\alias{roundbankers}
\title{DuckDB function roundbankers}
\usage{
roundbankers(x, n)
}
\arguments{
\item{x, n}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{roundbankers()}.
}
