% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{row}
\alias{row}
\title{DuckDB function row}
\usage{
row()
}
\value{
\code{STRUCT}
}
\description{
Create an unnamed STRUCT (tuple) containing the argument values.
}
\examples{
\dontrun{
row(i, i \% 4, i / 4)
}
}
