% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{row_to_json}
\alias{row_to_json}
\title{DuckDB function row_to_json}
\usage{
row_to_json()
}
\value{
\code{JSON}
}
\description{
DuckDB function \code{row_to_json()}.
}
\section{Provided by}{

The \code{json} extension (\code{LOAD json;}).
}

