% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{rtree_index_dump}
\alias{rtree_index_dump}
\title{DuckDB function rtree_index_dump}
\usage{
rtree_index_dump(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{rtree_index_dump()}.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

