% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{set_iceberg_table_properties}
\alias{set_iceberg_table_properties}
\title{DuckDB function set_iceberg_table_properties}
\usage{
set_iceberg_table_properties(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{set_iceberg_table_properties()}.
}
