% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{shapefile_meta}
\alias{shapefile_meta}
\title{DuckDB function shapefile_meta}
\usage{
shapefile_meta(col0)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{shapefile_meta()}.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

