% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{shobj_description}
\alias{shobj_description}
\title{DuckDB function shobj_description}
\usage{
shobj_description(object_oid, catalog_name)
}
\arguments{
\item{object_oid}{Unspecified.}

\item{catalog_name}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{shobj_description()}.
}
