% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{show_tables_expanded}
\alias{show_tables_expanded}
\title{DuckDB function show_tables_expanded}
\usage{
show_tables_expanded()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{show_tables_expanded()}.
}
\examples{
\dontrun{
character(0)
}
}
