% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{signbit}
\alias{signbit}
\title{DuckDB function signbit}
\arguments{
\item{x}{\code{FLOAT | DOUBLE}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns whether the signbit is set or not.
}
\examples{
\dontrun{
signbit(-0.0)
}
}
