% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sin}
\alias{sin}
\title{DuckDB function sin}
\usage{
sin(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the sin of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sin(90)
}\if{html}{\out{</div>}}
}

