% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sinh}
\alias{sinh}
\title{DuckDB function sinh}
\usage{
sinh(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the hyperbolic sin of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sinh(1)
}\if{html}{\out{</div>}}
}

