% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{/}
\alias{/}
\title{DuckDB function /}
\arguments{
\item{col0}{\code{FLOAT | DOUBLE | INTERVAL}}

\item{col1}{\code{FLOAT | DOUBLE}}
}
\value{
\code{FLOAT | DOUBLE | INTERVAL}
}
\description{
DuckDB function \verb{/()}.
}
\examples{
\dontrun{
character(0)
}
}
