% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sleep_ms}
\alias{sleep_ms}
\title{DuckDB function sleep_ms}
\usage{
sleep_ms(milliseconds = BIGINT)
}
\arguments{
\item{milliseconds}{\code{BIGINT}}
}
\value{
\code{"NULL"}
}
\description{
Sleeps for the specified number of milliseconds and returns NULL.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sleep_ms(100)
}\if{html}{\out{</div>}}
}

