% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{split_part}
\alias{split_part}
\title{DuckDB function split_part}
\usage{
split_part(string, delimiter, position)
}
\arguments{
\item{string}{Unspecified.}

\item{delimiter}{Unspecified.}

\item{position}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{split_part()}.
}
