% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sql_auto_complete}
\alias{sql_auto_complete}
\title{DuckDB function sql_auto_complete}
\usage{
sql_auto_complete(
  col0,
  max_exact_suggestion_count,
  max_file_suggestion_count,
  max_suggestion_count
)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{max_exact_suggestion_count}{\code{UBIGINT}}

\item{max_file_suggestion_count}{\code{UBIGINT}}

\item{max_suggestion_count}{\code{UBIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{sql_auto_complete()}.
}
\section{Provided by}{

The \code{autocomplete} extension (\code{LOAD autocomplete;}).
}

