% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sqlite_attach}
\alias{sqlite_attach}
\title{DuckDB function sqlite_attach}
\usage{
sqlite_attach(col0, overwrite)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{overwrite}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{sqlite_attach()}.
}
\section{Provided by}{

The \code{sqlite_scanner} extension (\code{LOAD sqlite_scanner;}).
}

