% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sqlite_query}
\alias{sqlite_query}
\title{DuckDB function sqlite_query}
\usage{
sqlite_query(col0, col1, params)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}

\item{params}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{sqlite_query()}.
}
