% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sqlite_scan}
\alias{sqlite_scan}
\title{DuckDB function sqlite_scan}
\usage{
sqlite_scan(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{sqlite_scan()}.
}
