% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sqrt}
\alias{sqrt}
\title{DuckDB function sqrt}
\usage{
sqrt(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the square root of x.
}
\examples{
\dontrun{
sqrt(4)
}
}
