% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Affine}
\alias{ST_Affine}
\title{DuckDB function ST_Affine}
\usage{
ST_Affine(geom, a, b, c, d, e, f, g, h, i, xoff, yoff, zoff)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{a}{\code{DOUBLE}}

\item{b}{\code{DOUBLE}}

\item{c}{\code{DOUBLE}}

\item{d}{\code{DOUBLE}}

\item{e}{\code{DOUBLE}}

\item{f}{\code{DOUBLE}}

\item{g}{\code{DOUBLE}}

\item{h}{\code{DOUBLE}}

\item{i}{\code{DOUBLE}}

\item{xoff}{\code{DOUBLE}}

\item{yoff}{\code{DOUBLE}}

\item{zoff}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Applies an affine transformation to a geometry.

For the 2D variant, the transformation matrix is defined as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{| a b xoff |
| d e yoff |
| 0 0 1    |
}\if{html}{\out{</div>}}

For the 3D variant, the transformation matrix is defined as follows:

\if{html}{\out{<div class="sourceCode">}}\preformatted{| a b c xoff |
| d e f yoff |
| g h i zoff |
| 0 0 0 1    |
}\if{html}{\out{</div>}}

The transformation is applied to all vertices of the geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_Affine(geom = GEOMETRY, a = DOUBLE, b = DOUBLE, c = DOUBLE, d = DOUBLE, e = DOUBLE, f = DOUBLE, g = DOUBLE, h = DOUBLE, i = DOUBLE, xoff = DOUBLE, yoff = DOUBLE, zoff = DOUBLE)}
\item \code{ST_Affine(geom = GEOMETRY, a = DOUBLE, b = DOUBLE, d = DOUBLE, e = DOUBLE, xoff = DOUBLE, yoff = DOUBLE)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Translate a point by (2, 3)
SELECT ST_Affine(ST_Point(1, 1),
                 1, 0,   -- a, b
                 0, 1,   -- d, e
                 2, 3);  -- xoff, yoff
-- POINT (3 4)
-- 
-- -- Scale a geometry by factor 2 in X and Y
-- SELECT ST_Affine(ST_Point(1, 1),
--                  2, 0, 0,   -- a, b, c
--                  0, 2, 0,   -- d, e, f
--                  0, 0, 1,   -- g, h, i
--                  0, 0, 0);  -- xoff, yoff, zoff
-- POINT (2 2)
}\if{html}{\out{</div>}}
}

