% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Area}
\alias{ST_Area}
\title{DuckDB function ST_Area}
\usage{
ST_Area(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Compute the area of a geometry.

Returns \code{0.0} for any geometry that is not a \code{POLYGON}, \code{MULTIPOLYGON} or \code{GEOMETRYCOLLECTION} containing polygon
geometries.

The area is in the same units as the spatial reference system of the geometry.

The \code{POINT_2D} and \code{LINESTRING_2D} overloads of this function always return \code{0.0} but are included for completeness.
}
\section{Overloads}{

\itemize{
\item \code{ST_Area(geom = GEOMETRY)}
\item \code{ST_Area(polygon = POLYGON_2D)}
\item \code{ST_Area(linestring = LINESTRING_2D)}
\item \code{ST_Area(point = POINT_2D)}
\item \code{ST_Area(box = BOX_2D)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Area('POLYGON((0 0, 0 1, 1 1, 1 0, 0 0))'::GEOMETRY);
-- 1.0
}\if{html}{\out{</div>}}
}

