% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_AsHEXWKB}
\alias{ST_AsHEXWKB}
\title{DuckDB function ST_AsHEXWKB}
\usage{
ST_AsHEXWKB(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the geometry as a HEXWKB string.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_AsHexWKB('POLYGON((0 0, 0 1, 1 1, 1 0, 0 0))'::GEOMETRY);
-- 01030000000100000005000000000000000000000000000...
}\if{html}{\out{</div>}}
}

