% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_AsMVTGeom}
\alias{ST_AsMVTGeom}
\title{DuckDB function ST_AsMVTGeom}
\usage{
ST_AsMVTGeom(geom, bounds, extent, buffer, clip_geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{bounds}{\code{BOX_2D}}

\item{extent}{\code{BIGINT}}

\item{buffer}{\code{BIGINT}}

\item{clip_geom}{\code{BOOLEAN}}
}
\value{
\code{GEOMETRY}
}
\description{
Transform and clip geometry to a tile boundary

See "ST_AsMVT" for more details.
}
\section{Overloads}{

\itemize{
\item \code{ST_AsMVTGeom(geom = GEOMETRY, bounds = BOX_2D, extent = BIGINT, buffer = BIGINT, clip_geom = BOOLEAN)}
\item \code{ST_AsMVTGeom(geom = GEOMETRY, bounds = BOX_2D, extent = BIGINT, buffer = BIGINT)}
\item \code{ST_AsMVTGeom(geom = GEOMETRY, bounds = BOX_2D, extent = BIGINT)}
\item \code{ST_AsMVTGeom(geom = GEOMETRY, bounds = BOX_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

