% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{st_astext}
\alias{st_astext}
\alias{st_aswkt}
\title{DuckDB function st_astext}
\usage{
st_astext(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{VARCHAR}
}
\description{
\itemize{
\item \code{st_astext(geom = GEOMETRY)}: Returns the Well-Known Text (WKT) representation of the geometry.
 \item \code{st_astext(point = POINT_2D)}, \code{st_astext(linestring = LINESTRING_2D)}, \code{st_astext(polygon = POLYGON_2D)}, \code{st_astext(box = BOX_2D)}: Returns the geometry as a WKT string.
 }
}
\section{Overloads}{

\itemize{
\item \code{st_astext(geom = GEOMETRY)}
\item \code{st_astext(point = POINT_2D)}
\item \code{st_astext(linestring = LINESTRING_2D)}
\item \code{st_astext(polygon = POLYGON_2D)}
\item \code{st_astext(box = BOX_2D)}
}
}

\section{Provided by}{

\itemize{
\item \code{st_astext(geom = GEOMETRY)}: built in
\item \code{st_astext(point = POINT_2D)}: \code{spatial}
\item \code{st_astext(linestring = LINESTRING_2D)}: \code{spatial}
\item \code{st_astext(polygon = POLYGON_2D)}: \code{spatial}
\item \code{st_astext(box = BOX_2D)}: \code{spatial}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{ST_AsText(ST_GeomFromWKB(X'01010000000000000000000000000000000000000000000000'))
SELECT ST_MakeEnvelope(0, 0, 1, 1);
-- POLYGON ((0 0, 0 1, 1 1, 1 0, 0 0))
}\if{html}{\out{</div>}}
}

\seealso{
Other geometry:
\code{\link[=st_aswkb]{st_aswkb()}},
\code{\link[=st_crs]{st_crs()}},
\code{\link[=st_geomfromwkb]{st_geomfromwkb()}}
}
\concept{geometry}
