% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{st_aswkt}
\alias{st_aswkt}
\title{DuckDB function st_aswkt}
\usage{
st_aswkt(geom = GEOMETRY)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the Well-Known Text (WKT) representation of the geometry.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{ST_AsText(ST_GeomFromWKB(X'01010000000000000000000000000000000000000000000000'))
}\if{html}{\out{</div>}}
}

