% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Azimuth}
\alias{ST_Azimuth}
\title{DuckDB function ST_Azimuth}
\usage{
ST_Azimuth(origin, target)
}
\arguments{
\item{origin, target}{\code{GEOMETRY | POINT_2D}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the azimuth (a clockwise angle measured from north) of two points in radian.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT degrees(ST_Azimuth(ST_Point(0, 0), ST_Point(0, 1)));
-- 90.0
}\if{html}{\out{</div>}}
}

