% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Boundary}
\alias{ST_Boundary}
\title{DuckDB function ST_Boundary}
\usage{
ST_Boundary(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the "boundary" of a geometry.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

