% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Buffer}
\alias{ST_Buffer}
\title{DuckDB function ST_Buffer}
\usage{
ST_Buffer(geom, distance, num_triangles, cap_style, join_style, mitre_limit)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{distance}{\code{DOUBLE}}

\item{num_triangles}{\code{INTEGER}}

\item{cap_style}{\code{VARCHAR}}

\item{join_style}{\code{VARCHAR}}

\item{mitre_limit}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a buffer around the input geometry at the target distance

\code{geom} is the input geometry.

\code{distance} is the target distance for the buffer, using the same units as the input geometry.

\code{num_triangles} represents how many triangles that will be produced to approximate a quarter circle. The larger the number, the smoother the resulting geometry. The default value is 8.

\code{cap_style} must be one of "CAP_ROUND", "CAP_FLAT", "CAP_SQUARE". This parameter is case-insensitive.

\code{join_style} must be one of "JOIN_ROUND", "JOIN_MITRE", "JOIN_BEVEL". This parameter is case-insensitive.

\code{mitre_limit} only applies when \code{join_style} is "JOIN_MITRE". It is the ratio of the distance from the corner to the mitre point to the corner radius. The default value is 1.0.

This is a planar operation and will not take into account the curvature of the earth.
}
\section{Overloads}{

\itemize{
\item \code{ST_Buffer(geom = GEOMETRY, distance = DOUBLE)}
\item \code{ST_Buffer(geom = GEOMETRY, distance = DOUBLE, num_triangles = INTEGER)}
\item \code{ST_Buffer(geom = GEOMETRY, distance = DOUBLE, num_triangles = INTEGER, cap_style = VARCHAR, join_style = VARCHAR, mitre_limit = DOUBLE)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

