% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_BuildArea}
\alias{ST_BuildArea}
\title{DuckDB function ST_BuildArea}
\usage{
ST_BuildArea(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Creates a polygonal geometry by attempting to "fill in" the input geometry.

Unlike ST_Polygonize, this function does not fill in holes.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

