% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Centroid}
\alias{ST_Centroid}
\title{DuckDB function ST_Centroid}
\usage{
ST_Centroid(box)
}
\arguments{
\item{box}{\code{BOX_2D | BOX_2DF}}
}
\value{
\code{GEOMETRY | POINT_2D}
}
\description{
Returns the centroid of a geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_Centroid(geom = GEOMETRY)}
\item \code{ST_Centroid(point = POINT_2D)}
\item \code{ST_Centroid(linestring = LINESTRING_2D)}
\item \code{ST_Centroid(polygon = POLYGON_2D)}
\item \code{ST_Centroid(box = BOX_2D)}
\item \code{ST_Centroid(box = BOX_2DF)}
}
}

