% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ClosestPoint}
\alias{ST_ClosestPoint}
\title{DuckDB function ST_ClosestPoint}
\usage{
ST_ClosestPoint(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the closest point on the first geometry to the second geometry.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

