% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Collect}
\alias{ST_Collect}
\title{DuckDB function ST_Collect}
\usage{
ST_Collect(geoms)
}
\arguments{
\item{geoms}{\code{GEOMETRY[]}}
}
\value{
\code{GEOMETRY}
}
\description{
Collects a list of geometries into a collection geometry.
\itemize{
\item If all geometries are \code{POINT}'s, a \code{MULTIPOINT} is returned.
\item If all geometries are \code{LINESTRING}'s, a \code{MULTILINESTRING} is returned.
\item If all geometries are \code{POLYGON}'s, a \code{MULTIPOLYGON} is returned.
\item Otherwise if the input collection contains a mix of geometry types, a \code{GEOMETRYCOLLECTION} is returned.
}

Empty and \code{NULL} geometries are ignored. If all geometries are empty or \code{NULL}, a \verb{GEOMETRYCOLLECTION EMPTY} is returned.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- With all POINT's, a MULTIPOINT is returned
SELECT ST_Collect([ST_Point(1, 2), ST_Point(3, 4)]);
-- MULTIPOINT (1 2, 3 4)
-- 
-- -- With mixed geometry types, a GEOMETRYCOLLECTION is returned
-- SELECT ST_Collect([ST_Point(1, 2), ST_GeomFromText('LINESTRING(3 4, 5 6)')]);
-- GEOMETRYCOLLECTION (POINT (1 2), LINESTRING (3 4, 5 6))
-- 
-- -- Note that the empty geometry is ignored, so the result is a MULTIPOINT
-- SELECT ST_Collect([ST_Point(1, 2), NULL, ST_GeomFromText('GEOMETRYCOLLECTION EMPTY')]);
-- MULTIPOINT (1 2)
-- 
-- -- If all geometries are empty or NULL, a GEOMETRYCOLLECTION EMPTY is returned
-- SELECT ST_Collect([NULL, ST_GeomFromText('GEOMETRYCOLLECTION EMPTY')]);
-- GEOMETRYCOLLECTION EMPTY
-- 
-- -- Tip: You can use the `ST_Collect` function together with the `list()` aggregate function to collect multiple rows of geometries into a single geometry collection:
-- 
-- CREATE TABLE points (geom GEOMETRY);
-- 
-- INSERT INTO points VALUES (ST_Point(1, 2)), (ST_Point(3, 4));
-- 
-- SELECT ST_Collect(list(geom)) FROM points;
-- MULTIPOINT (1 2, 3 4)
}\if{html}{\out{</div>}}
}

