% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CollectionExtract}
\alias{ST_CollectionExtract}
\title{DuckDB function ST_CollectionExtract}
\usage{
ST_CollectionExtract(geom, type)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{type}{\code{INTEGER}}
}
\value{
\code{GEOMETRY}
}
\description{
Extracts geometries from a GeometryCollection into a typed multi geometry.

If the input geometry is a GeometryCollection, the function will return a multi geometry, determined by the \code{type} parameter.
\itemize{
\item if \code{type} = 1, returns a MultiPoint containing all the Points in the collection
\item if \code{type} = 2, returns a MultiLineString containing all the LineStrings in the collection
\item if \code{type} = 3, returns a MultiPolygon containing all the Polygons in the collection
}

If no \code{type} parameters is provided, the function will return a multi geometry matching the highest "surface dimension"
of the contained geometries. E.g. if the collection contains only Points, a MultiPoint will be returned. But if the
collection contains both Points and LineStrings, a MultiLineString will be returned. Similarly, if the collection
contains Polygons, a MultiPolygon will be returned. Contained geometries of a lower surface dimension will be ignored.

If the input geometry contains nested GeometryCollections, their geometries will be extracted recursively and included
into the final multi geometry as well.

If the input geometry is not a GeometryCollection, the function will return the input geometry as is.
}
\section{Overloads}{

\itemize{
\item \code{ST_CollectionExtract(geom = GEOMETRY, type = INTEGER)}
\item \code{ST_CollectionExtract(geom = GEOMETRY)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_CollectionExtract('MULTIPOINT(1 2, 3 4)'::GEOMETRY, 1);
-- MULTIPOINT (1 2, 3 4)
}\if{html}{\out{</div>}}
}

