% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ConcaveHull}
\alias{ST_ConcaveHull}
\title{DuckDB function ST_ConcaveHull}
\usage{
ST_ConcaveHull(geom, ratio, allowHoles)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{ratio}{\code{DOUBLE}}

\item{allowHoles}{\code{BOOLEAN}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the 'concave' hull of the input geometry, containing all of the source input's points, and which can be used to create polygons from points. The ratio parameter dictates the level of concavity; 1.0 returns the convex hull; and 0 indicates to return the most concave hull possible. Set allowHoles to a non-zero value to allow output containing holes.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

