% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Contains}
\alias{ST_Contains}
\title{DuckDB function ST_Contains}
\usage{
ST_Contains(geom1, geom2)
}
\arguments{
\item{geom1}{\code{POLYGON_2D | GEOMETRY}}

\item{geom2}{\code{POINT_2D | GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the first geometry contains the second geometry

In contrast to \code{ST_ContainsProperly}, this function will also return true if \code{geom2} is contained strictly on the boundary of \code{geom1}.
 A geometry always \code{ST_Contains} itself, but does not \code{ST_ContainsProperly} itself.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

