% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ContainsProperly}
\alias{ST_ContainsProperly}
\title{DuckDB function ST_ContainsProperly}
\usage{
ST_ContainsProperly(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the first geometry \"properly\" contains the second geometry

In contrast to \code{ST_Contains}, this function does not return true if \code{geom2} is contained strictly on the boundary of \code{geom1}.
 A geometry always \code{ST_Contains} itself, but does not \code{ST_ContainsProperly} itself.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

