% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ConvexHull}
\alias{ST_ConvexHull}
\title{DuckDB function ST_ConvexHull}
\usage{
ST_ConvexHull(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the convex hull enclosing the geometry.
}
