% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CoverageClean}
\alias{ST_CoverageClean}
\title{DuckDB function ST_CoverageClean}
\usage{
ST_CoverageClean(geoms, snapping_distance, gap_maximum_width)
}
\arguments{
\item{geoms}{\code{GEOMETRY[]}}

\item{snapping_distance}{\code{DOUBLE}}

\item{gap_maximum_width}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Aligns the edges of a list of polygons whose edges are meant to align but are in fact exact matches.

Returns a collection of fixed polygons with the same size and order as the input polygons. EMPTY will be used in place of collapsed polygons.
}
\section{Overloads}{

\itemize{
\item \code{ST_CoverageClean(geoms = `GEOMETRY[]`, snapping_distance = DOUBLE, gap_maximum_width = DOUBLE)}
\item \code{ST_CoverageClean(geoms = `GEOMETRY[]`, snapping_distance = DOUBLE)}
\item \code{ST_CoverageClean(geoms = `GEOMETRY[]`)}
}
}

