% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CoverageInvalidEdges_Agg}
\alias{ST_CoverageInvalidEdges_Agg}
\title{DuckDB function ST_CoverageInvalidEdges_Agg}
\usage{
ST_CoverageInvalidEdges_Agg(col0, col1)
}
\arguments{
\item{col0}{\code{GEOMETRY}}

\item{col1}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the invalid edges of a coverage geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_CoverageInvalidEdges_Agg(col0 = GEOMETRY)}
\item \code{ST_CoverageInvalidEdges_Agg(col0 = GEOMETRY, col1 = DOUBLE)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

