% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CoverageSimplify}
\alias{ST_CoverageSimplify}
\title{DuckDB function ST_CoverageSimplify}
\usage{
ST_CoverageSimplify(geoms, tolerance, simplify_boundary)
}
\arguments{
\item{geoms}{\code{GEOMETRY[]}}

\item{tolerance}{\code{DOUBLE}}

\item{simplify_boundary}{\code{BOOLEAN}}
}
\value{
\code{GEOMETRY}
}
\description{
Simplify the edges in a polygonal coverage, preserving the coverange by ensuring that the there are no seams between the resulting simplified polygons.

By default, the boundary of the coverage is also simplified, but this can be controlled with the optional third 'simplify_boundary' parameter.
}
\section{Overloads}{

\itemize{
\item \code{ST_CoverageSimplify(geoms = `GEOMETRY[]`, tolerance = DOUBLE, simplify_boundary = BOOLEAN)}
\item \code{ST_CoverageSimplify(geoms = `GEOMETRY[]`, tolerance = DOUBLE)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

