% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CoverageSimplify_Agg}
\alias{ST_CoverageSimplify_Agg}
\title{DuckDB function ST_CoverageSimplify_Agg}
\usage{
ST_CoverageSimplify_Agg(col0, col1, col2)
}
\arguments{
\item{col0}{\code{GEOMETRY}}

\item{col1}{\code{DOUBLE}}

\item{col2}{\code{BOOLEAN}}
}
\value{
\code{GEOMETRY}
}
\description{
Simplifies a set of geometries while maintaining coverage.
}
\section{Overloads}{

\itemize{
\item \code{ST_CoverageSimplify_Agg(col0 = GEOMETRY, col1 = DOUBLE)}
\item \code{ST_CoverageSimplify_Agg(col0 = GEOMETRY, col1 = DOUBLE, col2 = BOOLEAN)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

