% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CoverageUnion_Agg}
\alias{ST_CoverageUnion_Agg}
\title{DuckDB function ST_CoverageUnion_Agg}
\usage{
ST_CoverageUnion_Agg(col0)
}
\arguments{
\item{col0}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Unions a set of geometries while maintaining coverage.
}
