% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CoveredBy}
\alias{ST_CoveredBy}
\title{DuckDB function ST_CoveredBy}
\usage{
ST_CoveredBy(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if geom1 is "covered by" geom2.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

