% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Covers}
\alias{ST_Covers}
\title{DuckDB function ST_Covers}
\usage{
ST_Covers(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geom1 "covers" geom2.
}
