% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Crosses}
\alias{ST_Crosses}
\title{DuckDB function ST_Crosses}
\usage{
ST_Crosses(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if geom1 "crosses" geom2.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

