% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Difference}
\alias{ST_Difference}
\title{DuckDB function ST_Difference}
\usage{
ST_Difference(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the "difference" between two geometries.
}
