% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Disjoint}
\alias{ST_Disjoint}
\title{DuckDB function ST_Disjoint}
\usage{
ST_Disjoint(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometries are disjoint.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

