% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Distance}
\alias{ST_Distance}
\title{DuckDB function ST_Distance}
\usage{
ST_Distance(point1, point2)
}
\arguments{
\item{point1, point2}{\code{POINT_2D}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the planar distance between two geometries.
}
\section{Overloads}{

\itemize{
\item \code{ST_Distance(point1 = POINT_2D, point2 = POINT_2D)}
\item \code{ST_Distance(point = POINT_2D, linestring = LINESTRING_2D)}
\item \code{ST_Distance(linestring = LINESTRING_2D, point = POINT_2D)}
\item \code{ST_Distance(geom1 = GEOMETRY, geom2 = GEOMETRY)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Distance('POINT (0 0)'::GEOMETRY, 'POINT (3 4)'::GEOMETRY);
-- 5.0
-- 
-- -- Z coordinates are ignored
-- SELECT ST_Distance('POINT Z (0 0 0)'::GEOMETRY, 'POINT Z (3 4 5)'::GEOMETRY);
-- 5.0
}\if{html}{\out{</div>}}
}

