% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Distance_GEOS}
\alias{ST_Distance_GEOS}
\title{DuckDB function ST_Distance_GEOS}
\usage{
ST_Distance_GEOS(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the planar distance between two geometries.
}
