% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Distance_Sphere}
\alias{ST_Distance_Sphere}
\title{DuckDB function ST_Distance_Sphere}
\usage{
ST_Distance_Sphere(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the haversine (great circle) distance between two geometries.
\itemize{
\item Only supports POINT geometries.
\item Returns the distance in meters.
\item The input is expected to be in WGS84 (EPSG:4326) coordinates, using a [latitude, longitude] axis order.
}
}
\section{Overloads}{

\itemize{
\item \code{ST_Distance_Sphere(geom1 = GEOMETRY, geom2 = GEOMETRY)}
\item \code{ST_Distance_Sphere(point1 = POINT_2D, point2 = POINT_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

