% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Drivers}
\alias{ST_Drivers}
\title{DuckDB function ST_Drivers}
\usage{
ST_Drivers()
}
\value{
Unspecified.
}
\description{
Returns the list of supported GDAL drivers and file formats

Note that far from all of these drivers have been tested properly.
 Some may require additional options to be passed to work as expected.
 If you run into any issues please first consult the \href{https://gdal.org/drivers/vector/index.html}{consult the GDAL docs}.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT * FROM ST_Drivers();
}\if{html}{\out{</div>}}
}

