% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Dump}
\alias{ST_Dump}
\title{DuckDB function ST_Dump}
\usage{
ST_Dump(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\verb{STRUCT(geom GEOMETRY, path INTEGER[])[]}
}
\description{
Dumps a geometry into a list of sub-geometries and their "path" in the original geometry.

You can use the \code{unnest(res, recursive := true)} function to explode the resulting list of structs into multiple rows.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Dump('MULTIPOINT(1 2, 3 4)'::GEOMETRY);
-- [\{'geom': 'POINT(1 2)', 'path': [0]\}, \{'geom': 'POINT(3 4)', 'path': [1]\}]
-- 
-- SELECT unnest(ST_Dump('MULTIPOINT(1 2, 3 4)'::GEOMETRY), recursive := true);
-- -- +-------------+---------+
-- -- |    geom     |  path   |
-- -- |  geometry   | int32[] |
-- -- +-------------+---------+
-- -- | POINT (1 2) | [1]     |
-- -- | POINT (3 4) | [2]     |
-- -- +-------------+---------+
}\if{html}{\out{</div>}}
}

