% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_DWithin_GEOS}
\alias{ST_DWithin_GEOS}
\title{DuckDB function ST_DWithin_GEOS}
\usage{
ST_DWithin_GEOS(geom1, geom2, distance)
}
\arguments{
\item{geom1}{\code{GEOMETRY}}

\item{geom2}{\code{GEOMETRY}}

\item{distance}{\code{DOUBLE}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if two geometries are within a target distance of each-other.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

