% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_EndPoint}
\alias{ST_EndPoint}
\title{DuckDB function ST_EndPoint}
\usage{
ST_EndPoint(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY | POINT_2D}
}
\description{
Returns the end point of a LINESTRING.
}
\section{Overloads}{

\itemize{
\item \code{ST_EndPoint(geom = GEOMETRY)}
\item \code{ST_EndPoint(line = LINESTRING_2D)}
}
}

