% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Envelope}
\alias{ST_Envelope}
\title{DuckDB function ST_Envelope}
\usage{
ST_Envelope(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the minimum bounding rectangle of a geometry as a polygon geometry.
}
