% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Equals}
\alias{ST_Equals}
\title{DuckDB function ST_Equals}
\usage{
ST_Equals(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometries are "equal".
}
