% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Expand}
\alias{ST_Expand}
\title{DuckDB function ST_Expand}
\usage{
ST_Expand(geom, distance)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{distance}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Expand the input geometry by the specified distance, returning a polygon.

\code{geom} is the input geometry.

\code{distance} is the target distance for the expansion, using the same units as the input geometry.

This is a planar operation and will not take into account the curvature of the earth.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_AsText(ST_Expand(ST_GeomFromText('POINT(20 30)'), 0.1));
}\if{html}{\out{</div>}}
}

