% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Extent}
\alias{ST_Extent}
\title{DuckDB function ST_Extent}
\usage{
ST_Extent(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOX_2D}
}
\description{
Returns the minimal bounding box enclosing the input geometry.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

