% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Extent_Agg}
\alias{ST_Extent_Agg}
\title{DuckDB function ST_Extent_Agg}
\usage{
ST_Extent_Agg(col0)
}
\arguments{
\item{col0}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Computes the minimal-bounding-box polygon containing the set of input geometries.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Extent_Agg(geom) FROM UNNEST([ST_Point(1,1), ST_Point(5,5)]) AS _(geom);
-- POLYGON ((1 1, 1 5, 5 5, 5 1, 1 1))
}\if{html}{\out{</div>}}
}

