% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ExteriorRing}
\alias{ST_ExteriorRing}
\title{DuckDB function ST_ExteriorRing}
\usage{
ST_ExteriorRing(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY | LINESTRING_2D}
}
\description{
Returns the exterior ring (shell) of a polygon geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_ExteriorRing(geom = GEOMETRY)}
\item \code{ST_ExteriorRing(polygon = POLYGON_2D)}
}
}

