% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_FlipCoordinates}
\alias{ST_FlipCoordinates}
\title{DuckDB function ST_FlipCoordinates}
\usage{
ST_FlipCoordinates(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY | POINT_2D | LINESTRING_2D | POLYGON_2D | BOX_2D}
}
\description{
Returns a new geometry with the coordinates of the input geometry "flipped" so that x = y and y = x.
}
\section{Overloads}{

\itemize{
\item \code{ST_FlipCoordinates(geom = GEOMETRY)}
\item \code{ST_FlipCoordinates(point = POINT_2D)}
\item \code{ST_FlipCoordinates(linestring = LINESTRING_2D)}
\item \code{ST_FlipCoordinates(polygon = POLYGON_2D)}
\item \code{ST_FlipCoordinates(box = BOX_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

