% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Force2D}
\alias{ST_Force2D}
\title{DuckDB function ST_Force2D}
\usage{
ST_Force2D(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Forces the vertices of a geometry to have X and Y components

This function will drop any Z and M values from the input geometry, if present. If the input geometry is already 2D, it will be returned as is.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

