% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Force3DM}
\alias{ST_Force3DM}
\title{DuckDB function ST_Force3DM}
\usage{
ST_Force3DM(geom, m)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{m}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Forces the vertices of a geometry to have X, Y and M components

The following cases apply:
\itemize{
\item If the input geometry has a Z component but no M component, the Z component will be replaced with the new M value.
\item If the input geometry has a M component but no Z component, it will be returned as is.
\item If the input geometry has both a Z component and a M component, the Z component will be removed.
\item Otherwise, if the input geometry has neither a Z or M component, the new M value will be added to the vertices of the input geometry.
}
}
