% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Force3DZ}
\alias{ST_Force3DZ}
\title{DuckDB function ST_Force3DZ}
\usage{
ST_Force3DZ(geom, z)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{z}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Forces the vertices of a geometry to have X, Y and Z components

The following cases apply:
\itemize{
\item If the input geometry has a M component but no Z component, the M component will be replaced with the new Z value.
\item If the input geometry has a Z component but no M component, it will be returned as is.
\item If the input geometry has both a Z component and a M component, the M component will be removed.
\item Otherwise, if the input geometry has neither a Z or M component, the new Z value will be added to the vertices of the input geometry.
}
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

