% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Force4D}
\alias{ST_Force4D}
\title{DuckDB function ST_Force4D}
\usage{
ST_Force4D(geom, z, m)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{z}{\code{DOUBLE}}

\item{m}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Forces the vertices of a geometry to have X, Y, Z and M components

The following cases apply:
\itemize{
\item If the input geometry has a Z component but no M component, the new M value will be added to the vertices of the input geometry.
\item If the input geometry has a M component but no Z component, the new Z value will be added to the vertices of the input geometry.
\item If the input geometry has both a Z component and a M component, the geometry will be returned as is.
\item Otherwise, if the input geometry has neither a Z or M component, the new Z and M values will be added to the vertices of the input geometry.
}
}
