% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_GeneratePoints}
\alias{ST_GeneratePoints}
\title{DuckDB function ST_GeneratePoints}
\usage{
ST_GeneratePoints(col0, col1, col2)
}
\arguments{
\item{col0}{\code{BOX_2D}}

\item{col1}{\code{BIGINT}}

\item{col2}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
Generates a set of random points within the specified bounding box.

Takes a bounding box (min_x, min_y, max_x, max_y), a count of points to generate, and optionally a seed for the random number generator.
}
\section{Overloads}{

\itemize{
\item \code{ST_GeneratePoints(col0 = BOX_2D, col1 = BIGINT)}
\item \code{ST_GeneratePoints(col0 = BOX_2D, col1 = BIGINT, col2 = BIGINT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT * FROM ST_GeneratePoints(\{min_x: 0, min_y:0, max_x:10, max_y:10\}::BOX_2D, 5, 42);
}\if{html}{\out{</div>}}
}

