% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_GeometryType}
\alias{ST_GeometryType}
\title{DuckDB function ST_GeometryType}
\usage{
ST_GeometryType(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{ANY}
}
\description{
Returns a 'GEOMETRY_TYPE' enum identifying the input geometry type. Possible enum return types are: \code{POINT}, \code{LINESTRING}, \code{POLYGON}, \code{MULTIPOINT}, \code{MULTILINESTRING}, \code{MULTIPOLYGON}, and \code{GEOMETRYCOLLECTION}.
}
\section{Overloads}{

\itemize{
\item \code{ST_GeometryType(geom = GEOMETRY)}
\item \code{ST_GeometryType(point = POINT_2D)}
\item \code{ST_GeometryType(linestring = LINESTRING_2D)}
\item \code{ST_GeometryType(polygon = POLYGON_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT DISTINCT ST_GeometryType(ST_GeomFromText('POINT(1 1)'));
-- POINT
}\if{html}{\out{</div>}}
}

