% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_GeomFromGeoJSON}
\alias{ST_GeomFromGeoJSON}
\title{DuckDB function ST_GeomFromGeoJSON}
\usage{
ST_GeomFromGeoJSON(geojson)
}
\arguments{
\item{geojson}{\code{JSON | VARCHAR}}
}
\value{
\code{GEOMETRY}
}
\description{
Deserializes a GEOMETRY from a GeoJSON fragment.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_GeomFromGeoJSON('\{"type": "Point", "coordinates": [1.0, 2.0]\}');
-- POINT (1 2)
}\if{html}{\out{</div>}}
}

