% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_GeomFromHEXEWKB}
\alias{ST_GeomFromHEXEWKB}
\title{DuckDB function ST_GeomFromHEXEWKB}
\usage{
ST_GeomFromHEXEWKB(hexwkb)
}
\arguments{
\item{hexwkb}{\code{VARCHAR}}
}
\value{
\code{GEOMETRY}
}
\description{
Deserialize a GEOMETRY from a HEX(E)WKB encoded string

DuckDB spatial doesn't currently differentiate between \code{WKB} and \code{EWKB}, so \code{ST_GeomFromHEXWKB} and \code{ST_GeomFromHEXEWKB} are just aliases of each other.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

