% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_GeomFromText}
\alias{ST_GeomFromText}
\title{DuckDB function ST_GeomFromText}
\usage{
ST_GeomFromText(wkt, ignore_invalid)
}
\arguments{
\item{wkt}{\code{VARCHAR}}

\item{ignore_invalid}{\code{BOOLEAN}}
}
\value{
\code{GEOMETRY}
}
\description{
Deserialize a GEOMETRY from a WKT encoded string.
}
\section{Overloads}{

\itemize{
\item \code{ST_GeomFromText(wkt = VARCHAR)}
\item \code{ST_GeomFromText(wkt = VARCHAR, ignore_invalid = BOOLEAN)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

