% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{st_geomfromwkb}
\alias{st_geomfromwkb}
\title{DuckDB function st_geomfromwkb}
\usage{
st_geomfromwkb(wkb = BLOB)
}
\arguments{
\item{wkb}{\code{BLOB}}
}
\value{
\code{GEOMETRY}
}
\description{
Creates a geometry from Well-Known Binary (WKB) representation.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{ST_GeomFromWKB(X'01010000000000000000000000000000000000000000000000')
}\if{html}{\out{</div>}}
}

\seealso{
Other geometry:
\code{\link[=st_astext]{st_astext()}},
\code{\link[=st_aswkb]{st_aswkb()}},
\code{\link[=st_crs]{st_crs()}}
}
\concept{geometry}
