% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_HasM}
\alias{ST_HasM}
\title{DuckDB function ST_HasM}
\usage{
ST_HasM(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Check if the input geometry has M values.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- HasM for a 2D geometry
SELECT ST_HasM(ST_GeomFromText('POINT(1 1)'));
-- false
-- 
-- -- HasM for a 3DZ geometry
-- SELECT ST_HasM(ST_GeomFromText('POINT Z(1 1 1)'));
-- false
-- 
-- -- HasM for a 3DM geometry
-- SELECT ST_HasM(ST_GeomFromText('POINT M(1 1 1)'));
-- true
-- 
-- -- HasM for a 4D geometry
-- SELECT ST_HasM(ST_GeomFromText('POINT ZM(1 1 1 1)'));
-- true
}\if{html}{\out{</div>}}
}

