% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Hilbert}
\alias{ST_Hilbert}
\title{DuckDB function ST_Hilbert}
\usage{
ST_Hilbert(box, bounds)
}
\arguments{
\item{box, bounds}{\code{BOX_2D | BOX_2DF}}
}
\value{
\code{UINTEGER}
}
\description{
Encodes the X and Y values as the hilbert curve index for a curve covering the given bounding box.
 If a geometry is provided, the center of the approximate bounding box is used as the point to encode.
 If no bounding box is provided, the hilbert curve index is mapped to the full range of a single-precision float.
 For the BOX_2D and BOX_2DF variants, the center of the box is used as the point to encode.
}
\section{Overloads}{

\itemize{
\item \code{ST_Hilbert(x = DOUBLE, y = DOUBLE, bounds = BOX_2D)}
\item \code{ST_Hilbert(geom = GEOMETRY, bounds = BOX_2D)}
\item \code{ST_Hilbert(geom = GEOMETRY)}
\item \code{ST_Hilbert(box = BOX_2D, bounds = BOX_2D)}
\item \code{ST_Hilbert(box = BOX_2DF, bounds = BOX_2DF)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

