% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_InteriorRingN}
\alias{ST_InteriorRingN}
\title{DuckDB function ST_InteriorRingN}
\usage{
ST_InteriorRingN(geom, n)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{n}{\code{BIGINT}}
}
\value{
\code{GEOMETRY | LINESTRING_2D}
}
\description{
Returns the N-th interior ring (hole) of a POLYGON as a LINESTRING. Indexing is 1-based  (n = 1 returns the first interior ring). Returns NULL if the polygon is empty or has fewer than N interior rings.
}
\section{Overloads}{

\itemize{
\item \code{ST_InteriorRingN(geom = GEOMETRY, n = BIGINT)}
\item \code{ST_InteriorRingN(polygon = POLYGON_2D, n = BIGINT)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_AsText(ST_InteriorRingN(ST_GeomFromText('POLYGON((0 0,10 0,10 10,0 10,0 0),(2 2,4 2,4 4,2 4,2 2))'), 1));
}\if{html}{\out{</div>}}
}

