% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Intersection}
\alias{ST_Intersection}
\title{DuckDB function ST_Intersection}
\usage{
ST_Intersection(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the intersection of two geometries.
}
