% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Intersection_Agg}
\alias{ST_Intersection_Agg}
\title{DuckDB function ST_Intersection_Agg}
\usage{
ST_Intersection_Agg(col0)
}
\arguments{
\item{col0}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Computes the intersection of a set of geometries.
}
