% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Intersects}
\alias{ST_Intersects}
\title{DuckDB function ST_Intersects}
\usage{
ST_Intersects(box1, box2)
}
\arguments{
\item{box1, box2}{\code{BOX_2D}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometries intersect.
}
\section{Overloads}{

\itemize{
\item \code{ST_Intersects(box1 = BOX_2D, box2 = BOX_2D)}
\item \code{ST_Intersects(geom1 = GEOMETRY, geom2 = GEOMETRY)}
}
}

