% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{st_intersects_extent}
\alias{st_intersects_extent}
\title{DuckDB function st_intersects_extent}
\usage{
st_intersects_extent(geom1 = GEOMETRY, geom2 = GEOMETRY)
}
\arguments{
\item{geom1}{\code{GEOMETRY}}

\item{geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometries bounding boxes intersect.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{'POINT(5 5)'::GEOMETRY && 'LINESTRING(0 0, 10 20)'::GEOMETRY;
}\if{html}{\out{</div>}}
}

