% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_IsClosed}
\alias{ST_IsClosed}
\title{DuckDB function ST_IsClosed}
\usage{
ST_IsClosed(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Check if a geometry is 'closed'.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

