% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_IsEmpty}
\alias{ST_IsEmpty}
\title{DuckDB function ST_IsEmpty}
\usage{
ST_IsEmpty(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometry is "empty".
}
\section{Overloads}{

\itemize{
\item \code{ST_IsEmpty(geom = GEOMETRY)}
\item \code{ST_IsEmpty(linestring = LINESTRING_2D)}
\item \code{ST_IsEmpty(polygon = POLYGON_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

