% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_IsRing}
\alias{ST_IsRing}
\title{DuckDB function ST_IsRing}
\usage{
ST_IsRing(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometry is a ring (both ST_IsClosed and ST_IsSimple).
}
