% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_IsSimple}
\alias{ST_IsSimple}
\title{DuckDB function ST_IsSimple}
\usage{
ST_IsSimple(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometry is simple.
}
