% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_IsValid}
\alias{ST_IsValid}
\title{DuckDB function ST_IsValid}
\usage{
ST_IsValid(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometry is valid.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

