% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Length}
\alias{ST_Length}
\title{DuckDB function ST_Length}
\usage{
ST_Length(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the length of the input line geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_Length(geom = GEOMETRY)}
\item \code{ST_Length(linestring = LINESTRING_2D)}
}
}

