% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Length_Spheroid}
\alias{ST_Length_Spheroid}
\title{DuckDB function ST_Length_Spheroid}
\usage{
ST_Length_Spheroid(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the length of the input geometry in meters, using an ellipsoidal model of the earth

The input geometry is assumed to be in the \href{https://en.wikipedia.org/wiki/World_Geodetic_System}{EPSG:4326} coordinate system (WGS84), with [latitude, longitude] axis order and the length is returned in meters. This function uses the \href{https://geographiclib.sourceforge.io/}{GeographicLib} library, calculating the length using an ellipsoidal model of the earth. This is a highly accurate method for calculating the length of a line geometry taking the curvature of the earth into account, but is also the slowest.

Returns \code{0.0} for any geometry that is not a \code{LINESTRING}, \code{MULTILINESTRING} or \code{GEOMETRYCOLLECTION} containing line geometries.
}
\section{Overloads}{

\itemize{
\item \code{ST_Length_Spheroid(geom = GEOMETRY)}
\item \code{ST_Length_Spheroid(line = LINESTRING_2D)}
}
}

