% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_LineInterpolatePoint}
\alias{ST_LineInterpolatePoint}
\title{DuckDB function ST_LineInterpolatePoint}
\usage{
ST_LineInterpolatePoint(line, fraction)
}
\arguments{
\item{line}{\code{GEOMETRY}}

\item{fraction}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a point interpolated along a line at a fraction of total 2D length.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

