% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_LineInterpolatePoints}
\alias{ST_LineInterpolatePoints}
\title{DuckDB function ST_LineInterpolatePoints}
\usage{
ST_LineInterpolatePoints(line, fraction, repeat.)
}
\arguments{
\item{line}{\code{GEOMETRY}}

\item{fraction}{\code{DOUBLE}}

\item{repeat.}{\code{BOOLEAN}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a multi-point interpolated along a line at a fraction of total 2D length.

if repeat is false, the result is a single point, (and equivalent to ST_LineInterpolatePoint),
otherwise, the result is a multi-point with points repeated at the fraction interval.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

