% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_LineLocatePoint}
\alias{ST_LineLocatePoint}
\title{DuckDB function ST_LineLocatePoint}
\usage{
ST_LineLocatePoint(line, point)
}
\arguments{
\item{line, point}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the location on a line closest to a point as a fraction of the total 2D length of the line.
}
