% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_LineMerge}
\alias{ST_LineMerge}
\title{DuckDB function ST_LineMerge}
\usage{
ST_LineMerge(geom, preserve_direction)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{preserve_direction}{\code{BOOLEAN}}
}
\value{
\code{GEOMETRY}
}
\description{
"Merges" the input line geometry, optionally taking direction into account.
}
\section{Overloads}{

\itemize{
\item \code{ST_LineMerge(geom = GEOMETRY)}
\item \code{ST_LineMerge(geom = GEOMETRY, preserve_direction = BOOLEAN)}
}
}

