% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_LineString2DFromWKB}
\alias{ST_LineString2DFromWKB}
\title{DuckDB function ST_LineString2DFromWKB}
\usage{
ST_LineString2DFromWKB(blob)
}
\arguments{
\item{blob}{\code{BLOB}}
}
\value{
\code{LINESTRING_2D}
}
\description{
Deserialize a LINESTRING_2D from a WKB encoded blob.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

