% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_LineSubstring}
\alias{ST_LineSubstring}
\title{DuckDB function ST_LineSubstring}
\usage{
ST_LineSubstring(line, start_fraction, end_fraction)
}
\arguments{
\item{line}{\code{GEOMETRY}}

\item{start_fraction}{\code{DOUBLE}}

\item{end_fraction}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a substring of a line between two fractions of total 2D length.
}
