% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_LocateAlong}
\alias{ST_LocateAlong}
\title{DuckDB function ST_LocateAlong}
\usage{
ST_LocateAlong(line, measure, offset)
}
\arguments{
\item{line}{\code{GEOMETRY}}

\item{measure}{\code{DOUBLE}}

\item{offset}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a point or multi-point, containing the point(s) at the geometry with the given measure

For a LINESTRING, or MULTILINESTRING, the location is determined by interpolating between M values
For a POINT and MULTIPOINT, the point is returned if the measure matches the M value of the vertex, otherwise an empty geometry is returned
For a POLYGON, only the exterior ring is considered, and treated as a LINESTRING

If offset is provided, the resulting point(s) is offset by the given amount perpendicular to the line direction.
}
\section{Overloads}{

\itemize{
\item \code{ST_LocateAlong(line = GEOMETRY, measure = DOUBLE, offset = DOUBLE)}
\item \code{ST_LocateAlong(line = GEOMETRY, measure = DOUBLE)}
}
}

