% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_M}
\alias{ST_M}
\title{DuckDB function ST_M}
\usage{
ST_M(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the M coordinate of a point geometry.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_M(ST_Point(1, 2, 3, 4))
}\if{html}{\out{</div>}}
}

