% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MakeEnvelope}
\alias{ST_MakeEnvelope}
\title{DuckDB function ST_MakeEnvelope}
\usage{
ST_MakeEnvelope(min_x, min_y, max_x, max_y)
}
\arguments{
\item{min_x, min_y, max_x, max_y}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Create a rectangular polygon from min/max coordinates.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

