% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MakePolygon}
\alias{ST_MakePolygon}
\title{DuckDB function ST_MakePolygon}
\usage{
ST_MakePolygon(shell, holes)
}
\arguments{
\item{shell}{\code{GEOMETRY}}

\item{holes}{\code{GEOMETRY[]}}
}
\value{
\code{GEOMETRY}
}
\description{
\itemize{
\item \code{ST_MakePolygon(shell = GEOMETRY)}: Create a POLYGON from a LINESTRING shell.
 \item \code{ST_MakePolygon(shell = GEOMETRY, holes = `GEOMETRY[]`)}: Create a POLYGON from a LINESTRING shell and a list of LINESTRING holes.
 }
}
\section{Overloads}{

\itemize{
\item \code{ST_MakePolygon(shell = GEOMETRY)}
\item \code{ST_MakePolygon(shell = GEOMETRY, holes = `GEOMETRY[]`)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_MakePolygon(ST_LineString([ST_Point(0, 0), ST_Point(1, 0), ST_Point(1, 1), ST_Point(0, 0)]));
SELECT ST_MakePolygon(ST_LineString([ST_Point(0, 0), ST_Point(1, 0), ST_Point(1, 1), ST_Point(0, 0)]), [ST_LineString([ST_Point(0.25, 0.25), ST_Point(0.75, 0.25), ST_Point(0.75, 0.75), ST_Point(0.25, 0.25)])]);
}\if{html}{\out{</div>}}
}

