% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MakeValid}
\alias{ST_MakeValid}
\title{DuckDB function ST_MakeValid}
\usage{
ST_MakeValid(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a valid representation of the geometry.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

