% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MinimumRotatedRectangle}
\alias{ST_MinimumRotatedRectangle}
\title{DuckDB function ST_MinimumRotatedRectangle}
\usage{
ST_MinimumRotatedRectangle(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the minimum rotated rectangle that bounds the input geometry, finding the surrounding box that has the lowest area by using a rotated rectangle, rather than taking the lowest and highest coordinate values as per ST_Envelope().
}
