% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MMax}
\alias{ST_MMax}
\title{DuckDB function ST_MMax}
\usage{
ST_MMax(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the maximum M coordinate of a geometry.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_MMax(ST_Point(1, 2, 3, 4))
}\if{html}{\out{</div>}}
}

