% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MMin}
\alias{ST_MMin}
\title{DuckDB function ST_MMin}
\usage{
ST_MMin(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the minimum M coordinate of a geometry.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_MMin(ST_Point(1, 2, 3, 4))
}\if{html}{\out{</div>}}
}

