% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_NInteriorRings}
\alias{ST_NInteriorRings}
\title{DuckDB function ST_NInteriorRings}
\usage{
ST_NInteriorRings(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{INTEGER}
}
\description{
Returns the number of interior rings of a polygon.
}
\section{Overloads}{

\itemize{
\item \code{ST_NInteriorRings(geom = GEOMETRY)}
\item \code{ST_NInteriorRings(polygon = POLYGON_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

