% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Normalize}
\alias{ST_Normalize}
\title{DuckDB function ST_Normalize}
\usage{
ST_Normalize(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the "normalized" representation of the geometry.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

