% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_NPoints}
\alias{ST_NPoints}
\title{DuckDB function ST_NPoints}
\usage{
ST_NPoints(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{UINTEGER | UBIGINT}
}
\description{
Returns the number of vertices within a geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_NPoints(geom = GEOMETRY)}
\item \code{ST_NPoints(point = POINT_2D)}
\item \code{ST_NPoints(linestring = LINESTRING_2D)}
\item \code{ST_NPoints(polygon = POLYGON_2D)}
\item \code{ST_NPoints(box = BOX_2D)}
}
}

