% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_NumInteriorRings}
\alias{ST_NumInteriorRings}
\title{DuckDB function ST_NumInteriorRings}
\usage{
ST_NumInteriorRings(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{INTEGER}
}
\description{
Returns the number of interior rings of a polygon.
}
\section{Overloads}{

\itemize{
\item \code{ST_NumInteriorRings(geom = GEOMETRY)}
\item \code{ST_NumInteriorRings(polygon = POLYGON_2D)}
}
}

