% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_NumPoints}
\alias{ST_NumPoints}
\title{DuckDB function ST_NumPoints}
\usage{
ST_NumPoints(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{UINTEGER | UBIGINT}
}
\description{
Returns the number of vertices within a geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_NumPoints(geom = GEOMETRY)}
\item \code{ST_NumPoints(point = POINT_2D)}
\item \code{ST_NumPoints(linestring = LINESTRING_2D)}
\item \code{ST_NumPoints(polygon = POLYGON_2D)}
\item \code{ST_NumPoints(box = BOX_2D)}
}
}

