% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Overlaps}
\alias{ST_Overlaps}
\title{DuckDB function ST_Overlaps}
\usage{
ST_Overlaps(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the geometries overlap.
}
