% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Perimeter}
\alias{ST_Perimeter}
\title{DuckDB function ST_Perimeter}
\usage{
ST_Perimeter(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the length of the perimeter of the geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_Perimeter(geom = GEOMETRY)}
\item \code{ST_Perimeter(polygon = POLYGON_2D)}
\item \code{ST_Perimeter(box = BOX_2D)}
}
}

