% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Perimeter_Spheroid}
\alias{ST_Perimeter_Spheroid}
\title{DuckDB function ST_Perimeter_Spheroid}
\usage{
ST_Perimeter_Spheroid(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the length of the perimeter in meters using an ellipsoidal model of the earths surface

The input geometry is assumed to be in the \href{https://en.wikipedia.org/wiki/World_Geodetic_System}{EPSG:4326} coordinate system (WGS84), with [latitude, longitude] axis order and the length is returned in meters. This function uses the \href{https://geographiclib.sourceforge.io/}{GeographicLib} library, calculating the perimeter using an ellipsoidal model of the earth. This is a highly accurate method for calculating the perimeter of a polygon taking the curvature of the earth into account, but is also the slowest.

Returns \code{0.0} for any geometry that is not a \code{POLYGON}, \code{MULTIPOLYGON} or \code{GEOMETRYCOLLECTION} containing polygon geometries.
}
\section{Overloads}{

\itemize{
\item \code{ST_Perimeter_Spheroid(geom = GEOMETRY)}
\item \code{ST_Perimeter_Spheroid(poly = POLYGON_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

