% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Point2D}
\alias{ST_Point2D}
\title{DuckDB function ST_Point2D}
\usage{
ST_Point2D(x, y)
}
\arguments{
\item{x, y}{\code{DOUBLE}}
}
\value{
\code{POINT_2D}
}
\description{
Creates a POINT_2D.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

