% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Point3D}
\alias{ST_Point3D}
\title{DuckDB function ST_Point3D}
\usage{
ST_Point3D(x, y, z)
}
\arguments{
\item{x, y, z}{\code{DOUBLE}}
}
\value{
\code{POINT_3D}
}
\description{
Creates a POINT_3D.
}
