% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Point4D}
\alias{ST_Point4D}
\title{DuckDB function ST_Point4D}
\usage{
ST_Point4D(x, y, z, m)
}
\arguments{
\item{x, y, z, m}{\code{DOUBLE}}
}
\value{
\code{POINT_4D}
}
\description{
Creates a POINT_4D.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

