% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_PointN}
\alias{ST_PointN}
\title{DuckDB function ST_PointN}
\usage{
ST_PointN(geom, index)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{index}{\code{INTEGER}}
}
\value{
\code{GEOMETRY | POINT_2D}
}
\description{
Returns the n'th vertex from the input geometry as a point geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_PointN(geom = GEOMETRY, index = INTEGER)}
\item \code{ST_PointN(linestring = LINESTRING_2D, index = INTEGER)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

