% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_PointOnSurface}
\alias{ST_PointOnSurface}
\title{DuckDB function ST_PointOnSurface}
\usage{
ST_PointOnSurface(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a point guaranteed to lie on the surface of the geometry.
}
