% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Polygon2DFromWKB}
\alias{ST_Polygon2DFromWKB}
\title{DuckDB function ST_Polygon2DFromWKB}
\usage{
ST_Polygon2DFromWKB(blob)
}
\arguments{
\item{blob}{\code{BLOB}}
}
\value{
\code{POLYGON_2D}
}
\description{
Deserialize a POLYGON_2D from a WKB encoded blob.
}
