% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Polygonize}
\alias{ST_Polygonize}
\title{DuckDB function ST_Polygonize}
\usage{
ST_Polygonize(geometries)
}
\arguments{
\item{geometries}{\code{GEOMETRY[]}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a polygonized representation of the input geometries.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Create a polygon from a closed linestring ring
SELECT ST_Polygonize([
    ST_GeomFromText('LINESTRING(0 0, 0 10, 10 10, 10 0, 0 0)')
]);
---
GEOMETRYCOLLECTION (POLYGON ((0 0, 0 10, 10 10, 10 0, 0 0)))
}\if{html}{\out{</div>}}
}

