% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_QuadKey}
\alias{ST_QuadKey}
\title{DuckDB function ST_QuadKey}
\usage{
ST_QuadKey(longitude, latitude, level)
}
\arguments{
\item{longitude}{\code{DOUBLE}}

\item{latitude}{\code{DOUBLE}}

\item{level}{\code{INTEGER}}
}
\value{
\code{VARCHAR}
}
\description{
Compute the \href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{quadkey} for a given lon/lat point at a given level.
 Note that the parameter order is \strong{longitude}, \strong{latitude}.

\code{level} has to be between 1 and 23, inclusive.

The input coordinates will be clamped to the lon/lat bounds of the earth (longitude between -180 and 180, latitude between -85.05112878 and 85.05112878).

The geometry overload throws an error if the input geometry is not a \code{POINT}.
}
\section{Overloads}{

\itemize{
\item \code{ST_QuadKey(longitude = DOUBLE, latitude = DOUBLE, level = INTEGER)}
\item \code{ST_QuadKey(point = GEOMETRY, level = INTEGER)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_QuadKey(ST_Point(11.08, 49.45), 10);
-- 1333203202
}\if{html}{\out{</div>}}
}

