% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Read}
\alias{ST_Read}
\title{DuckDB function ST_Read}
\usage{
ST_Read(
  col0,
  keep_wkb,
  max_batch_size,
  layer,
  sibling_files,
  allowed_drivers,
  open_options
)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{keep_wkb}{\code{BOOLEAN}}

\item{max_batch_size}{\code{INTEGER}}

\item{layer}{\code{VARCHAR}}

\item{sibling_files}{\code{VARCHAR[]}}

\item{allowed_drivers}{\code{VARCHAR[]}}

\item{open_options}{\code{VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
Read and import a variety of geospatial file formats using the GDAL library.

The \code{ST_Read} table function is based on the \href{https://gdal.org/index.html}{GDAL} translator library and enables reading spatial data from a variety of geospatial vector file formats as if they were DuckDB tables.

See \href{#st_drivers}{ST_Drivers} for a list of supported file formats and drivers.

Except for the \code{path} parameter, all parameters are optional.\tabular{lll}{
   Parameter \tab Type \tab Description \cr
   \code{path} \tab VARCHAR \tab The path to the file to read. Mandatory \cr
   \code{sequential_layer_scan} \tab BOOLEAN \tab If set to true, the table function will scan through all layers sequentially and return the first layer that matches the given layer name. This is required for some drivers to work properly, e.g., the OSM driver. \cr
   \code{spatial_filter} \tab WKB_BLOB \tab If set to a WKB blob, the table function will only return rows that intersect with the given WKB geometry. Some drivers may support efficient spatial filtering natively, in which case it will be pushed down. Otherwise the filtering is done by GDAL which may be much slower. \cr
   \code{open_options} \tab VARCHAR[] \tab A list of key-value pairs that are passed to the GDAL driver to control the opening of the file. E.g., the GeoJSON driver supports a FLATTEN_NESTED_ATTRIBUTES=YES option to flatten nested attributes. \cr
   \code{layer} \tab VARCHAR \tab The name of the layer to read from the file. If NULL, the first layer is returned. Can also be a layer index (starting at 0). \cr
   \code{allowed_drivers} \tab VARCHAR[] \tab A list of GDAL driver names that are allowed to be used to open the file. If empty, all drivers are allowed. \cr
   \code{sibling_files} \tab VARCHAR[] \tab A list of sibling files that are required to open the file. E.g., the ESRI Shapefile driver requires a .shx file to be present. Although most of the time these can be discovered automatically. \cr
   \code{spatial_filter_box} \tab BOX_2D \tab If set to a BOX_2D, the table function will only return rows that intersect with the given bounding box. Similar to spatial_filter. \cr
   \code{keep_wkb} \tab BOOLEAN \tab If set, the table function will return geometries in a wkb_geometry column with the type WKB_BLOB (which can be cast to BLOB) instead of GEOMETRY. This is useful if you want to use DuckDB with more exotic geometry subtypes that DuckDB spatial doesn't support representing in the GEOMETRY type yet. \cr
}


Note that GDAL is single-threaded, so this table function will not be able to make full use of parallelism.

By using \code{ST_Read}, the spatial extension also provides “replacement scans” for common geospatial file formats, allowing you to query files of these formats as if they were tables directly.

\if{html}{\out{<div class="sourceCode sql">}}\preformatted{SELECT * FROM './path/to/some/shapefile/dataset.shp';
}\if{html}{\out{</div>}}

In practice this is just syntax-sugar for calling ST_Read, so there is no difference in performance. If you want to pass additional options, you should use the ST_Read table function directly.

The following formats are currently recognized by their file extension:\tabular{ll}{
   Format \tab Extension \cr
   ESRI ShapeFile \tab .shp \cr
   GeoPackage \tab .gpkg \cr
   FlatGeoBuf \tab .fgb \cr
}
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Read a Shapefile
SELECT * FROM ST_Read('some/file/path/filename.shp');

-- Read a GeoJSON file
CREATE TABLE my_geojson_table AS SELECT * FROM ST_Read('some/file/path/filename.json');
}\if{html}{\out{</div>}}
}

