% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Read_Meta}
\alias{ST_Read_Meta}
\title{DuckDB function ST_Read_Meta}
\usage{
ST_Read_Meta(col0)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
Read the metadata from a variety of geospatial file formats using the GDAL library.

The \code{ST_Read_Meta} table function accompanies the \code{ST_Read} table function, but instead of reading the contents of a file, this function scans the metadata instead.
 Since the data model of the underlying GDAL library is quite flexible, most of the interesting metadata is within the returned \code{layers} column, which is a somewhat complex nested structure of DuckDB \code{STRUCT} and \code{LIST} types.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Find the coordinate reference system authority name and code for the first layers first geometry column in the file
SELECT
    layers[1].geometry_fields[1].crs.auth_name as name,
    layers[1].geometry_fields[1].crs.auth_code as code
FROM st_read_meta('../../tmp/data/amsterdam_roads.fgb');
}\if{html}{\out{</div>}}
}

