% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ReadSHP}
\alias{ST_ReadSHP}
\title{DuckDB function ST_ReadSHP}
\usage{
ST_ReadSHP(col0, encoding)
}
\arguments{
\item{col0, encoding}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
Read a Shapefile without relying on the GDAL library.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

