% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ReducePrecision}
\alias{ST_ReducePrecision}
\title{DuckDB function ST_ReducePrecision}
\usage{
ST_ReducePrecision(geom, precision)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{precision}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the geometry with all vertices reduced to the given precision.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

