% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Reverse}
\alias{ST_Reverse}
\title{DuckDB function ST_Reverse}
\usage{
ST_Reverse(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the geometry with the order of its vertices reversed.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

