% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Rotate}
\alias{ST_Rotate}
\title{DuckDB function ST_Rotate}
\usage{
ST_Rotate(geom, radians)
}
\arguments{
\item{geom, radians}{Unspecified.}
}
\value{
Unspecified.
}
\description{
Alias of ST_RotateZ.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

