% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_RotateX}
\alias{ST_RotateX}
\title{DuckDB function ST_RotateX}
\usage{
ST_RotateX(geom, radians)
}
\arguments{
\item{geom, radians}{Unspecified.}
}
\value{
Unspecified.
}
\description{
Rotates a geometry around the X axis. This is a shorthand macro for calling ST_Affine.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Rotate a 3D point 90 degrees (π/2 radians) around the X-axis
SELECT ST_RotateX(ST_GeomFromText('POINT Z(0 1 0)'), pi()/2);
-- POINT Z (0 0 1)
}\if{html}{\out{</div>}}
}

