% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_RotateZ}
\alias{ST_RotateZ}
\title{DuckDB function ST_RotateZ}
\usage{
ST_RotateZ(geom, radians)
}
\arguments{
\item{geom, radians}{Unspecified.}
}
\value{
Unspecified.
}
\description{
Rotates a geometry around the Z axis. This is a shorthand macro for calling ST_Affine.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Rotate a point 90 degrees (π/2 radians) around the Z-axis
SELECT ST_RotateZ(ST_Point(1, 0), pi()/2);
-- POINT (0 1)
}\if{html}{\out{</div>}}
}

