% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Scale}
\alias{ST_Scale}
\title{DuckDB function ST_Scale}
\usage{
ST_Scale(geom, xs, ys, zs)
}
\arguments{
\item{geom, xs, ys, zs}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ST_Scale()}.
}
\section{Overloads}{

\itemize{
\item \code{ST_Scale(geom, xs, ys, zs)}
\item \code{ST_Scale(geom, xs, ys)}
}
}

