% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{st_setcrs}
\alias{st_setcrs}
\title{DuckDB function st_setcrs}
\usage{
st_setcrs(geom, crs)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{crs}{\code{VARCHAR}}
}
\value{
\code{GEOMETRY}
}
\description{
Sets the Coordinate Reference System (CRS) identifier of the geometry.
}
