% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ShortestLine}
\alias{ST_ShortestLine}
\title{DuckDB function ST_ShortestLine}
\usage{
ST_ShortestLine(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the shortest line between two geometries.
}
