% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Simplify}
\alias{ST_Simplify}
\title{DuckDB function ST_Simplify}
\usage{
ST_Simplify(geom, tolerance)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{tolerance}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns a simplified version of the geometry.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

