% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Snap}
\alias{ST_Snap}
\title{DuckDB function ST_Snap}
\usage{
ST_Snap(geom, snap_to, tolerance)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{snap_to}{\code{GEOMETRY}}

\item{tolerance}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Snaps the vertices and segments of a geometry to another geometry's vertices within the given tolerance.
}
