% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_StartPoint}
\alias{ST_StartPoint}
\title{DuckDB function ST_StartPoint}
\usage{
ST_StartPoint(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY | POINT_2D}
}
\description{
Returns the start point of a LINESTRING.
}
\section{Overloads}{

\itemize{
\item \code{ST_StartPoint(geom = GEOMETRY)}
\item \code{ST_StartPoint(line = LINESTRING_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

