% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_SymDifference}
\alias{ST_SymDifference}
\title{DuckDB function ST_SymDifference}
\usage{
ST_SymDifference(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the symmetric difference of two geometries.
}
