% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Translate}
\alias{ST_Translate}
\title{DuckDB function ST_Translate}
\usage{
ST_Translate(geom, dx, dy, dz)
}
\arguments{
\item{geom, dx, dy, dz}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ST_Translate()}.
}
\section{Overloads}{

\itemize{
\item \code{ST_Translate(geom, dx, dy, dz)}
\item \code{ST_Translate(geom, dx, dy)}
}
}

