% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Union}
\alias{ST_Union}
\title{DuckDB function ST_Union}
\usage{
ST_Union(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the union of two geometries.
}
