% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Union_Agg}
\alias{ST_Union_Agg}
\title{DuckDB function ST_Union_Agg}
\usage{
ST_Union_Agg(col0)
}
\arguments{
\item{col0}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Computes the union of a set of input geometries.
}
