% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_VoronoiDiagram}
\alias{ST_VoronoiDiagram}
\title{DuckDB function ST_VoronoiDiagram}
\usage{
ST_VoronoiDiagram(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the Voronoi diagram of the supplied MultiPoint geometry.
}
