% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Within}
\alias{ST_Within}
\title{DuckDB function ST_Within}
\usage{
ST_Within(geom1, geom2)
}
\arguments{
\item{geom1}{\code{POINT_2D | GEOMETRY}}

\item{geom2}{\code{POLYGON_2D | GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the first geometry is within the second.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

