% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_WithinProperly}
\alias{ST_WithinProperly}
\title{DuckDB function ST_WithinProperly}
\usage{
ST_WithinProperly(geom1, geom2)
}
\arguments{
\item{geom1, geom2}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the first geometry \"properly\" is contained by the second geometry

This function functions the same as \code{ST_ContainsProperly}, but the arguments are swapped.
}
