% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_X}
\alias{ST_X}
\title{DuckDB function ST_X}
\usage{
ST_X(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the X coordinate of a point geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_X(geom = GEOMETRY)}
\item \code{ST_X(point = POINT_2D)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_X(ST_Point(1, 2))
}\if{html}{\out{</div>}}
}

