% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_XMax}
\alias{ST_XMax}
\title{DuckDB function ST_XMax}
\usage{
ST_XMax(box)
}
\arguments{
\item{box}{\code{BOX_2D | BOX_2DF}}
}
\value{
\code{DOUBLE | FLOAT}
}
\description{
Returns the maximum X coordinate of a geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_XMax(geom = GEOMETRY)}
\item \code{ST_XMax(point = POINT_2D)}
\item \code{ST_XMax(line = LINESTRING_2D)}
\item \code{ST_XMax(polygon = POLYGON_2D)}
\item \code{ST_XMax(box = BOX_2D)}
\item \code{ST_XMax(box = BOX_2DF)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_XMax(ST_Point(1, 2))
}\if{html}{\out{</div>}}
}

