% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Y}
\alias{ST_Y}
\title{DuckDB function ST_Y}
\usage{
ST_Y(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the Y coordinate of a point geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_Y(geom = GEOMETRY)}
\item \code{ST_Y(point = POINT_2D)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Y(ST_Point(1, 2))
}\if{html}{\out{</div>}}
}

