% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_YMax}
\alias{ST_YMax}
\title{DuckDB function ST_YMax}
\usage{
ST_YMax(box)
}
\arguments{
\item{box}{\code{BOX_2D | BOX_2DF}}
}
\value{
\code{DOUBLE | FLOAT}
}
\description{
Returns the maximum Y coordinate of a geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_YMax(geom = GEOMETRY)}
\item \code{ST_YMax(point = POINT_2D)}
\item \code{ST_YMax(line = LINESTRING_2D)}
\item \code{ST_YMax(polygon = POLYGON_2D)}
\item \code{ST_YMax(box = BOX_2D)}
\item \code{ST_YMax(box = BOX_2DF)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_YMax(ST_Point(1, 2))
}\if{html}{\out{</div>}}
}

