% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_YMin}
\alias{ST_YMin}
\title{DuckDB function ST_YMin}
\usage{
ST_YMin(box)
}
\arguments{
\item{box}{\code{BOX_2D | BOX_2DF}}
}
\value{
\code{DOUBLE | FLOAT}
}
\description{
Returns the minimum Y coordinate of a geometry.
}
\section{Overloads}{

\itemize{
\item \code{ST_YMin(geom = GEOMETRY)}
\item \code{ST_YMin(point = POINT_2D)}
\item \code{ST_YMin(line = LINESTRING_2D)}
\item \code{ST_YMin(polygon = POLYGON_2D)}
\item \code{ST_YMin(box = BOX_2D)}
\item \code{ST_YMin(box = BOX_2DF)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_YMin(ST_Point(1, 2))
}\if{html}{\out{</div>}}
}

