% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Z}
\alias{ST_Z}
\title{DuckDB function ST_Z}
\usage{
ST_Z(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the Z coordinate of a point geometry.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Z(ST_Point(1, 2, 3))
}\if{html}{\out{</div>}}
}

