% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ZMax}
\alias{ST_ZMax}
\title{DuckDB function ST_ZMax}
\usage{
ST_ZMax(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the maximum Z coordinate of a geometry.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_ZMax(ST_Point(1, 2, 3))
}\if{html}{\out{</div>}}
}

