% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ZMin}
\alias{ST_ZMin}
\title{DuckDB function ST_ZMin}
\usage{
ST_ZMin(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the minimum Z coordinate of a geometry.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_ZMin(ST_Point(1, 2, 3))
}\if{html}{\out{</div>}}
}

