% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{start_ui_server}
\alias{start_ui_server}
\title{DuckDB function start_ui_server}
\usage{
start_ui_server()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{start_ui_server()}.
}
\section{Provided by}{

The \code{ui} extension (\code{LOAD ui;}).
}

