% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{stddev}
\alias{stddev}
\title{DuckDB function stddev}
\usage{
stddev(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the sample standard deviation.
}
\examples{
\dontrun{
sqrt(var_samp(x))
}
}
