% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{stem}
\alias{stem}
\title{DuckDB function stem}
\usage{
stem(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
DuckDB function \code{stem()}.
}
\section{Provided by}{

The \code{fts} extension (\code{LOAD fts;}).
}

