% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{stop_ui_server}
\alias{stop_ui_server}
\title{DuckDB function stop_ui_server}
\usage{
stop_ui_server()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{stop_ui_server()}.
}
\section{Provided by}{

The \code{ui} extension (\code{LOAD ui;}).
}

