% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{str_split}
\alias{str_split}
\title{DuckDB function str_split}
\usage{
str_split(string = VARCHAR, separator = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\value{
\code{VARCHAR[]}
}
\description{
Splits the \code{string} along the \code{separator}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{str_split('hello-world', '-')
}\if{html}{\out{</div>}}
}

