% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{str_split_regex}
\alias{str_split_regex}
\title{DuckDB function str_split_regex}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{options}{\code{VARCHAR}}
}
\value{
\code{VARCHAR[]}
}
\description{
Splits the \code{string} along the \code{regex}. A set of optional regex \code{options} can be set.
}
\section{Overloads}{

\itemize{
\item \code{str_split_regex(string = VARCHAR, regex = VARCHAR)}
\item \code{str_split_regex(string = VARCHAR, regex = VARCHAR, options = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{str_split_regex('hello world; 42', ';? ')
}\if{html}{\out{</div>}}
}

