% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{string_agg}
\alias{string_agg}
\title{DuckDB function string_agg}
\arguments{
\item{str}{\code{ANY}}

\item{arg}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Concatenates the column string values with an optional separator.
}
\examples{
\dontrun{
string_agg(A, '-')
}
}
