% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{string_split_regex}
\alias{string_split_regex}
\alias{regexp_split_to_array}
\alias{str_split_regex}
\title{DuckDB function string_split_regex}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{options}{\code{VARCHAR}}
}
\value{
\code{VARCHAR[]}
}
\description{
Splits the \code{string} along the \code{regex}. A set of optional regex \code{options} can be set.
}
\section{Overloads}{

\itemize{
\item \code{string_split_regex(string = VARCHAR, regex = VARCHAR)}
\item \code{string_split_regex(string = VARCHAR, regex = VARCHAR, options = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{string_split_regex('hello world; 42', ';? ')
}\if{html}{\out{</div>}}
}

\seealso{
Other regex:
\code{\link[=regexp_escape]{regexp_escape()}},
\code{\link[=regexp_extract]{regexp_extract()}},
\code{\link[=regexp_extract_all]{regexp_extract_all()}},
\code{\link[=regexp_full_match]{regexp_full_match()}},
\code{\link[=regexp_matches]{regexp_matches()}},
\code{\link[=regexp_replace]{regexp_replace()}}
}
\concept{regex}
