% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{string_to_array}
\alias{string_to_array}
\title{DuckDB function string_to_array}
\arguments{
\item{string}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\description{
Splits the \code{string} along the \code{separator}.
}
\examples{
\dontrun{
string_split('hello-world', '-')
}
}
