% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strpos}
\alias{strpos}
\title{DuckDB function strpos}
\usage{
strpos(string = VARCHAR, search_string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\value{
\code{BIGINT}
}
\description{
Returns location of first occurrence of \code{search_string} in \code{string}, counting from 1. Returns 0 if no match found.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{strpos('test test', 'es')
}\if{html}{\out{</div>}}
}

