% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_concat}
\alias{struct_concat}
\title{DuckDB function struct_concat}
\usage{
struct_concat()
}
\value{
\code{STRUCT}
}
\description{
Merge the multiple STRUCTs into a single STRUCT.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{struct_concat(struct_pack(i := 4), struct_pack(s := 'string'))
}\if{html}{\out{</div>}}
}

