% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_extract}
\alias{struct_extract}
\title{DuckDB function struct_extract}
\arguments{
\item{struct}{\code{STRUCT}}

\item{'entry'}{\code{VARCHAR | BIGINT}}
}
\description{
Extract the named entry from the STRUCT.
}
\examples{
\dontrun{
struct_extract({'i': 3, 'v2': 3, 'v3': 0}, 'i')
}
}
