% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_extract_at}
\alias{struct_extract_at}
\title{DuckDB function struct_extract_at}
\usage{
struct_extract_at(struct, entry)
}
\arguments{
\item{struct}{\code{STRUCT}}

\item{entry}{\code{BIGINT}}
}
\value{
\code{ANY}
}
\description{
Extract the entry from the STRUCT by position (starts at 1!).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{struct_extract_at(\{'i': 3, 'v2': 3, 'v3': 0\}, 2)
}\if{html}{\out{</div>}}
}

