% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_has}
\alias{struct_has}
\title{DuckDB function struct_has}
\usage{
struct_has(struct = STRUCT, `'entry'` = ANY)
}
\arguments{
\item{struct}{\code{STRUCT}}

\item{'entry'}{\code{ANY}}
}
\value{
\code{BOOLEAN}
}
\description{
Check if an unnamed STRUCT contains the value.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{struct_has(ROW(3, 3, 0), 3)
}\if{html}{\out{</div>}}
}

