% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_keys}
\alias{struct_keys}
\title{DuckDB function struct_keys}
\usage{
struct_keys(struct = ANY)
}
\arguments{
\item{struct}{\code{ANY}}
}
\value{
\code{VARCHAR[]}
}
\description{
Returns the field names of a STRUCT as a list.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{struct_keys(\{'a': 1, 'b': 2\})
}\if{html}{\out{</div>}}
}

