% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_pack}
\alias{struct_pack}
\title{DuckDB function struct_pack}
\usage{
struct_pack()
}
\value{
\code{STRUCT}
}
\description{
Create a STRUCT containing the argument values. The entry name will be the bound variable name.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{struct_pack(i := 4, s := 'string')
}\if{html}{\out{</div>}}
}

