% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_position}
\alias{struct_position}
\title{DuckDB function struct_position}
\arguments{
\item{struct}{\code{STRUCT}}

\item{'entry'}{\code{ANY}}
}
\description{
Get the position of the entry in an unnamed STRUCT, starting at 1.
}
\examples{
\dontrun{
struct_position(ROW(3, 3, 0), 3)
}
}
