% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_update}
\alias{struct_update}
\title{DuckDB function struct_update}
\usage{
struct_update()
}
\value{
\code{STRUCT}
}
\description{
Changes field(s)/value(s) to an existing STRUCT with the argument values. The entry name(s) will be the bound variable name(s).
}
\examples{
\dontrun{
struct_update({'a': 1}, a := 2)
}
}
