% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_values}
\alias{struct_values}
\title{DuckDB function struct_values}
\usage{
struct_values(struct = STRUCT)
}
\arguments{
\item{struct}{\code{STRUCT}}
}
\value{
\code{STRUCT}
}
\description{
Returns the field values of a STRUCT as an UnnamedStruct.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{struct_values(\{'a': 1, 'b': 'world'\})
}\if{html}{\out{</div>}}
}

