% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sum}
\alias{sum}
\title{DuckDB function sum}
\usage{
sum(arg)
}
\arguments{
\item{arg}{\code{DECIMAL | BOOLEAN | SMALLINT | INTEGER | BIGINT | HUGEINT | DOUBLE | BIGNUM}}
}
\value{
\code{DECIMAL | HUGEINT | DOUBLE | BIGNUM}
}
\description{
Calculates the sum value for all tuples in arg.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sum(A)
}\if{html}{\out{</div>}}
}

