% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{summary}
\alias{summary}
\title{DuckDB function summary}
\usage{
summary(col0 = TABLE)
}
\arguments{
\item{col0}{\code{TABLE}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{summary()}.
}
\examples{
\dontrun{
character(0)
}
}
