% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{switch}
\alias{switch}
\title{DuckDB function switch}
\arguments{
\item{key}{\code{K | MAP(K, V)}}

\item{map}{\code{MAP(K, V) | V}}

\item{value}{\code{V}}
}
\value{
\code{V}
}
\description{
Creates a switch statement similar to CASE WHEN/THEN.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{switch(x, map(\{1 : 1\}, default)
}\if{html}{\out{</div>}}
}

