% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{table_info}
\alias{table_info}
\title{DuckDB function table_info}
\usage{
table_info(col0 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{table_info()}.
}
