% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{tan}
\alias{tan}
\title{DuckDB function tan}
\usage{
tan(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the tan of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{tan(90)
}\if{html}{\out{</div>}}
}

