% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{tanh}
\alias{tanh}
\title{DuckDB function tanh}
\usage{
tanh(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the hyperbolic tan of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{tanh(1)
}\if{html}{\out{</div>}}
}

