% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{test_all_types}
\alias{test_all_types}
\title{DuckDB function test_all_types}
\usage{
test_all_types(use_large_bignum, use_large_enum)
}
\arguments{
\item{use_large_bignum, use_large_enum}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{test_all_types()}.
}
