% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{test_vector_types}
\alias{test_vector_types}
\title{DuckDB function test_vector_types}
\usage{
test_vector_types(col0 = ANY, all_flat = BOOLEAN)
}
\arguments{
\item{col0}{\code{ANY}}

\item{all_flat}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{test_vector_types()}.
}
\examples{
\dontrun{
character(0)
}
}
