% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_eventcoef.R
\name{tidy_eventcoef}
\alias{tidy_eventcoef}
\title{Tidy lead/lag coefficients to data frame}
\usage{
tidy_eventcoef(model, varname, baseline = NA, ...)
}
\arguments{
\item{model}{fitted model object}

\item{varname}{quoted (partial) name of the time-to-event variable}

\item{baseline}{optional reference period to add (see details)}

\item{...}{further parameters passed to \code{tidy()}, e.g. \code{conf.int=TRUE}}
}
\value{
A subset of the data frame as returned by \code{\link[broom]{tidy}}
with the estimates for the leads and lags and a variable \code{eventtime}.
}
\description{
Tidy lead/lag coefficients to data frame
}
\details{
The function extracts the coefficients using \code{\link[broom]{tidy}} and
then subsets to the relevant coefficients based on \code{varname}.

The parameter \code{baseline} allows to add an estimate
with value zero and with a confidence interval of length zero to
visually highlight the reference period for the leads and lags.

All additional arguments are passed further to
\code{\link[broom]{tidy}}. This can be a global argument
(e.g., \code{conf.int=TRUE}) or extractor-specific
argument (e.g., \code{se.type='robust'}) when
using \code{\link[lfe]{felm}} for fitting (see \code{\link[broom]{tidy.felm}}).
}
\examples{
library(dd)

data(goldendawn)

goldendawn$ttime <- code_eventtime(
       unit=muni,
       time=year,
       treat=post,
       data=goldendawn)

m <- lm(gd ~ ttime + factor(muni) + factor(year), data=goldendawn)
summary(m)

toplot <- tidy_eventcoef(m, varname='ttime', conf.int=TRUE)

with(toplot, plot(eventtime,estimate, pch=20, ylim=c(-2,3)))
with(toplot, segments(eventtime,conf.low, eventtime, conf.high))


}
\seealso{
\code{\link[broom]{tidy}}
}
