% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{**}
\alias{**}
\title{DuckDB function **}
\usage{
`**`(x = DOUBLE, y = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}

\item{y}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes x to the power of y.
}
\examples{
\dontrun{
2 ** 3
}
}
