% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{timetz_byte_comparable}
\alias{timetz_byte_comparable}
\title{DuckDB function timetz_byte_comparable}
\arguments{
\item{time_tz}{\verb{TIME WITH TIME ZONE}}
}
\description{
Converts a TIME WITH TIME ZONE to an integer sort key.
}
\examples{
\dontrun{
timetz_byte_comparable('18:18:16.21-07:00'::TIME_TZ)
}
}
