% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{timezone}
\alias{timezone}
\title{DuckDB function timezone}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE | VARCHAR}}

\item{col1}{\verb{TIME WITH TIME ZONE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\verb{BIGINT | TIME WITH TIME ZONE | TIMESTAMP WITH TIME ZONE | TIMESTAMP}
}
\description{
Extract the timezone component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{timezone(ts = DATE)}
\item \code{timezone(ts = INTERVAL)}
\item \code{timezone(ts = INTERVAL, col1 = `TIME WITH TIME ZONE`)}
\item \code{timezone(ts = TIMESTAMP)}
\item \code{timezone(ts = `TIMESTAMP WITH TIME ZONE`)}
\item \code{timezone(ts = VARCHAR, col1 = TIMESTAMP)}
\item \code{timezone(ts = VARCHAR, col1 = `TIMESTAMP WITH TIME ZONE`)}
\item \code{timezone(ts = VARCHAR, col1 = `TIME WITH TIME ZONE`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{timezone(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

