% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{timezone_minute}
\alias{timezone_minute}
\title{DuckDB function timezone_minute}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the timezone_minute component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{timezone_minute(ts = DATE)}: Extract the timezone_minute component from a date or timestamp.
\item \code{timezone_minute(ts = INTERVAL)}: Extract the timezone_minute component from a date or timestamp.
\item \code{timezone_minute(ts = TIMESTAMP)}: Extract the timezone_minute component from a date or timestamp.
\item \code{timezone_minute(ts = `TIMESTAMP WITH TIME ZONE`)}: Extract the timezone_minute component from a date or timestamp.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{timezone_minute(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

