% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_binary}
\alias{to_binary}
\title{DuckDB function to_binary}
\arguments{
\item{string}{\code{VARCHAR}}

\item{value}{\code{BIGNUM | UBIGINT | BIGINT | HUGEINT | UHUGEINT}}
}
\value{
\code{VARCHAR}
}
\description{
Converts the \code{string} to binary representation.

Converts the \code{value} to binary representation.
}
\section{Overloads}{

\itemize{
\item \code{to_binary(string = VARCHAR)}
\item \code{to_binary(value = BIGNUM)}
\item \code{to_binary(value = UBIGINT)}
\item \code{to_binary(value = BIGINT)}
\item \code{to_binary(value = HUGEINT)}
\item \code{to_binary(value = UHUGEINT)}
}
}

\examples{
\dontrun{
to_binary('Aa')
to_binary(42)
}
}
