% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_hours}
\alias{to_hours}
\title{DuckDB function to_hours}
\usage{
to_hours(integer = BIGINT)
}
\arguments{
\item{integer}{\code{BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a hour interval.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{to_hours(5)
}\if{html}{\out{</div>}}
}

