% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_millennia}
\alias{to_millennia}
\title{DuckDB function to_millennia}
\arguments{
\item{integer}{\code{INTEGER | BIGINT}}
}
\description{
Construct a millenium interval.
}
\examples{
\dontrun{
to_millennia(1)
}
}
