% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_milliseconds}
\alias{to_milliseconds}
\alias{duckdb_fun_to_milliseconds}
\title{DuckDB function to_milliseconds}
\arguments{
\item{double}{\code{DOUBLE}}
}
\description{
Construct a millisecond interval.
}
\examples{
\dontrun{
to_milliseconds(5.5)
}
}
