% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_minutes}
\alias{to_minutes}
\title{DuckDB function to_minutes}
\usage{
to_minutes(integer = BIGINT)
}
\arguments{
\item{integer}{\code{BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a minute interval.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{to_minutes(5)
}\if{html}{\out{</div>}}
}

