% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_months}
\alias{to_months}
\title{DuckDB function to_months}
\arguments{
\item{integer}{\code{INTEGER | BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a month interval.
}
\section{Overloads}{

\itemize{
\item \code{to_months(integer = INTEGER)}: Construct a month interval.
\item \code{to_months(integer = BIGINT)}: Construct a month interval.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{to_months(5)
}\if{html}{\out{</div>}}
}

