% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_seconds}
\alias{to_seconds}
\title{DuckDB function to_seconds}
\usage{
to_seconds(double = DOUBLE)
}
\arguments{
\item{double}{\code{DOUBLE}}
}
\description{
Construct a second interval.
}
\examples{
\dontrun{
to_seconds(5.5)
}
}
