% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_years}
\alias{to_years}
\title{DuckDB function to_years}
\arguments{
\item{integer}{\code{INTEGER | BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a year interval.
}
\section{Overloads}{

\itemize{
\item \code{to_years(integer = INTEGER)}
\item \code{to_years(integer = BIGINT)}
}
}

\examples{
\dontrun{
to_years(5)
}
}
