% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{today}
\alias{today}
\title{DuckDB function today}
\usage{
today()
}
\value{
\code{DATE}
}
\description{
DuckDB function \code{today()}.
}
\section{Provided by}{

The \code{icu} extension (\code{LOAD icu;}).
}

