% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{tpcds}
\alias{tpcds}
\title{DuckDB function tpcds}
\usage{
tpcds(col0)
}
\arguments{
\item{col0}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{tpcds()}.
}
\section{Provided by}{

The \code{tpcds} extension (\code{LOAD tpcds;}).
}

