% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{tpcds_answers}
\alias{tpcds_answers}
\title{DuckDB function tpcds_answers}
\usage{
tpcds_answers()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{tpcds_answers()}.
}
\section{Provided by}{

The \code{tpcds} extension (\code{LOAD tpcds;}).
}

