% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{truncate_duckdb_logs}
\alias{truncate_duckdb_logs}
\title{DuckDB function truncate_duckdb_logs}
\usage{
truncate_duckdb_logs()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{truncate_duckdb_logs()}.
}
