% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{try_strptime}
\alias{try_strptime}
\title{DuckDB function try_strptime}
\arguments{
\item{text}{\code{VARCHAR}}

\item{format}{\code{VARCHAR | VARCHAR[]}}
}
\description{
Converts the \code{string} text to timestamp according to the format string. Returns \code{NULL} on failure.
}
\examples{
\dontrun{
try_strptime('Wed, 1 January 1992 - 08:38:40 PM', '\%a, \%-d \%B \%Y - \%I:\%M:\%S \%p')
}
}
