% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{~}
\alias{~}
\title{DuckDB function ~}
\arguments{
\item{input}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise NOT.
}
\section{Overloads}{

\itemize{
\item \code{`~`(input = TINYINT)}: Bitwise NOT.
\item \code{`~`(input = SMALLINT)}: Bitwise NOT.
\item \code{`~`(input = INTEGER)}: Bitwise NOT.
\item \code{`~`(input = BIGINT)}: Bitwise NOT.
\item \code{`~`(input = HUGEINT)}: Bitwise NOT.
\item \code{`~`(input = UTINYINT)}: Bitwise NOT.
\item \code{`~`(input = USMALLINT)}: Bitwise NOT.
\item \code{`~`(input = UINTEGER)}: Bitwise NOT.
\item \code{`~`(input = UBIGINT)}: Bitwise NOT.
\item \code{`~`(input = UHUGEINT)}: Bitwise NOT.
\item \code{`~`(input = BIT)}: Bitwise NOT.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{~15
}\if{html}{\out{</div>}}
}

