% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{txid_current}
\alias{txid_current}
\title{DuckDB function txid_current}
\usage{
txid_current()
}
\value{
\code{UBIGINT}
}
\description{
Returns the current transaction’s ID (a BIGINT). It will assign a new one if the current transaction does not have one already.
}
\examples{
\dontrun{
txid_current()
}
}
