% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{typeof}
\alias{typeof}
\title{DuckDB function typeof}
\usage{
typeof(expression = ANY)
}
\arguments{
\item{expression}{\code{ANY}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the name of the data type of the result of the expression.
}
\examples{
\dontrun{
typeof('abc')
}
}
