% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ucase}
\alias{ucase}
\title{DuckDB function ucase}
\usage{
ucase(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Converts \code{string} to upper case.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{ucase('Hello')
}\if{html}{\out{</div>}}
}

