% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ui_is_started}
\alias{ui_is_started}
\title{DuckDB function ui_is_started}
\usage{
ui_is_started()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{ui_is_started()}.
}
\section{Provided by}{

The \code{ui} extension (\code{LOAD ui;}).
}

