% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{union_tag}
\alias{union_tag}
\title{DuckDB function union_tag}
\usage{
union_tag(union = UNION)
}
\arguments{
\item{union}{\code{UNION}}
}
\value{
\code{ANY}
}
\description{
Retrieve the currently selected tag of the union as an ENUM.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{union_tag(union_value(k := 'foo'))
}\if{html}{\out{</div>}}
}

