% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{unity_catalog_checkpoint_table}
\alias{unity_catalog_checkpoint_table}
\title{DuckDB function unity_catalog_checkpoint_table}
\usage{
unity_catalog_checkpoint_table(col0)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{unity_catalog_checkpoint_table()}.
}
