% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{unnest}
\alias{unnest}
\title{DuckDB function unnest}
\usage{
unnest(col0 = ANY)
}
\arguments{
\item{col0}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{unnest()}.
}
\examples{
\dontrun{
character(0)
}
}
