% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{user_agent}
\alias{user_agent}
\title{DuckDB function user_agent}
\usage{
user_agent()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{user_agent()}.
}
\examples{
\dontrun{
character(0)
}
}
