% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{uuid}
\alias{uuid}
\title{DuckDB function uuid}
\usage{
uuid()
}
\value{
\code{UUID}
}
\description{
Returns a random UUID v4 similar to this: eeccb8c5-9943-b2bb-bb5e-222f4e14b687.
}
\examples{
\dontrun{
uuid()
}
}
