% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{uuid_extract_timestamp}
\alias{uuid_extract_timestamp}
\title{DuckDB function uuid_extract_timestamp}
\arguments{
\item{uuid}{\code{UUID}}
}
\description{
Extract the timestamp for the given UUID v7.
}
\examples{
\dontrun{
uuid_extract_timestamp('019482e4-1441-7aad-8127-eec99573b0a0')
}
}
