% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{uuid_extract_version}
\alias{uuid_extract_version}
\title{DuckDB function uuid_extract_version}
\usage{
uuid_extract_version(uuid = UUID)
}
\arguments{
\item{uuid}{\code{UUID}}
}
\value{
\code{UINTEGER}
}
\description{
Extract a version for the given UUID.
}
\examples{
\dontrun{
uuid_extract_version('019482e4-1441-7aad-8127-eec99573b0a0')
}
}
