% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{uuidv4}
\alias{uuidv4}
\title{DuckDB function uuidv4}
\usage{
uuidv4()
}
\value{
\code{UUID}
}
\description{
Returns a random UUIDv4 similar to this: eeccb8c5-9943-b2bb-bb5e-222f4e14b687.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{uuidv4()
}\if{html}{\out{</div>}}
}

