% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{uuidv7}
\alias{uuidv7}
\title{DuckDB function uuidv7}
\usage{
uuidv7()
}
\value{
\code{UUID}
}
\description{
Returns a random UUID v7 similar to this: 019482e4-1441-7aad-8127-eec99573b0a0.
}
\examples{
\dontrun{
uuidv7()
}
}
