% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{var_pop}
\alias{var_pop}
\title{DuckDB function var_pop}
\usage{
var_pop(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the population variance.
}
\examples{
\dontrun{
character(0)
}
}
