% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{var_samp}
\alias{var_samp}
\title{DuckDB function var_samp}
\usage{
var_samp(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the sample variance of all input values.
}
\examples{
\dontrun{
(SUM(x^2) - SUM(x)^2 / COUNT(x)) / (COUNT(x) - 1)
}
}
