% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{variant_bytes_to_variant}
\alias{variant_bytes_to_variant}
\title{DuckDB function variant_bytes_to_variant}
\usage{
variant_bytes_to_variant(col0)
}
\arguments{
\item{col0}{\code{BLOB}}
}
\value{
\code{VARIANT}
}
\description{
DuckDB function \code{variant_bytes_to_variant()}.
}
