% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{variant_extract}
\alias{variant_extract}
\title{DuckDB function variant_extract}
\usage{
variant_extract(input_variant, field)
}
\arguments{
\item{input_variant}{\code{VARIANT}}

\item{field}{\code{VARCHAR}}
}
\value{
\code{VARIANT}
}
\description{
\itemize{
\item \code{variant_extract(input_variant = VARIANT, field = VARCHAR)}: Returns the \code{field} from the \code{input_variant} if it's an OBJECT.
 \item \code{variant_extract(input_variant = VARIANT, index = UINTEGER)}: Returns the entry at \code{index} from the \code{input_variant} if it's an ARRAY.
 }
}
\section{Overloads}{

\itemize{
\item \code{variant_extract(input_variant = VARIANT, field = VARCHAR)}
\item \code{variant_extract(input_variant = VARIANT, index = UINTEGER)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{variant_extract(\{'a': 42, 'b': [1,2,3])::VARIANT, 'b')
variant_extract([1,2,3])::VARIANT, 0)
}\if{html}{\out{</div>}}
}

\seealso{
Other variant:
\code{\link[=variant_normalize]{variant_normalize()}},
\code{\link[=variant_typeof]{variant_typeof()}}
}
\concept{variant}
