% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{variant_normalize}
\alias{variant_normalize}
\title{DuckDB function variant_normalize}
\usage{
variant_normalize(input_variant)
}
\arguments{
\item{input_variant}{\code{VARIANT}}
}
\value{
\code{VARIANT}
}
\description{
Normalizes the \code{input_variant} to a canonical representation.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{variant_normalize(\{'b': [1,2,3], 'a': 42\})::VARIANT)
}\if{html}{\out{</div>}}
}

\seealso{
Other variant:
\code{\link[=variant_extract]{variant_extract()}},
\code{\link[=variant_typeof]{variant_typeof()}}
}
\concept{variant}
