% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{variant_to_parquet_variant}
\alias{variant_to_parquet_variant}
\title{DuckDB function variant_to_parquet_variant}
\usage{
variant_to_parquet_variant(col0)
}
\arguments{
\item{col0}{\code{VARIANT}}
}
\value{
\code{ANY}
}
\description{
DuckDB function \code{variant_to_parquet_variant()}.
}
