% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{variant_typeof}
\alias{variant_typeof}
\title{DuckDB function variant_typeof}
\usage{
variant_typeof(input_variant = VARIANT)
}
\arguments{
\item{input_variant}{\code{VARIANT}}
}
\description{
Returns the internal type of the \code{input_variant}.
}
