% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{vector_type}
\alias{vector_type}
\title{DuckDB function vector_type}
\usage{
vector_type(col = ANY)
}
\arguments{
\item{col}{\code{ANY}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the VectorType of a given column.
}
\examples{
\dontrun{
vector_type(col)
}
}
