% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{verify_external}
\alias{verify_external}
\title{DuckDB function verify_external}
\usage{
verify_external()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{verify_external()}.
}
\examples{
\dontrun{
character(0)
}
}
