% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{verify_parallelism}
\alias{verify_parallelism}
\title{DuckDB function verify_parallelism}
\usage{
verify_parallelism()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{verify_parallelism()}.
}
\examples{
\dontrun{
character(0)
}
}
