% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{verify_serializer}
\alias{verify_serializer}
\title{DuckDB function verify_serializer}
\usage{
verify_serializer()
}
\value{
Unspecified.
}
\description{
DuckDB function \code{verify_serializer()}.
}
\examples{
\dontrun{
character(0)
}
}
