% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{vortex_scan}
\alias{vortex_scan}
\title{DuckDB function vortex_scan}
\usage{
vortex_scan(col0)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{vortex_scan()}.
}
