% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{vss_join}
\alias{vss_join}
\title{DuckDB function vss_join}
\usage{
vss_join(left_table, right_table, left_col, right_col, k, metric)
}
\arguments{
\item{left_table, right_table, left_col, right_col, k, metric}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{vss_join()}.
}
