% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{vss_match}
\alias{vss_match}
\title{DuckDB function vss_match}
\usage{
vss_match(right_table, left_col, right_col, k, metric)
}
\arguments{
\item{right_table, left_col, right_col, k, metric}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{vss_match()}.
}
