% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{wavg}
\alias{wavg}
\title{DuckDB function wavg}
\usage{
wavg(value, weight)
}
\arguments{
\item{value}{Unspecified.}

\item{weight}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{wavg()}.
}
