% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{week}
\alias{week}
\title{DuckDB function week}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the week component from a date or timestamp.
}
\examples{
\dontrun{
week(timestamp '2021-08-03 11:59:44.123456')
}
}
