% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{weekday}
\alias{weekday}
\title{DuckDB function weekday}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the weekday component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{weekday(ts = DATE)}
\item \code{weekday(ts = INTERVAL)}
\item \code{weekday(ts = TIMESTAMP)}
\item \code{weekday(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
weekday(timestamp '2021-08-03 11:59:44.123456')
}
}
