% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{weighted_avg}
\alias{weighted_avg}
\title{DuckDB function weighted_avg}
\usage{
weighted_avg(value, weight)
}
\arguments{
\item{value}{Unspecified.}

\item{weight}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{weighted_avg()}.
}
