% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{which_secret}
\alias{which_secret}
\title{DuckDB function which_secret}
\usage{
which_secret(col0, col1)
}
\arguments{
\item{col0, col1}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{which_secret()}.
}
