% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{whoami}
\alias{whoami}
\title{DuckDB function whoami}
\usage{
whoami()
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{whoami()}.
}
\section{Provided by}{

The \code{quack} extension (\code{LOAD quack;}).
}

