% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{write_blob}
\alias{write_blob}
\title{DuckDB function write_blob}
\usage{
write_blob(col0, col1)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{BLOB}}
}
\value{
\code{BOOLEAN}
}
\description{
DuckDB function \code{write_blob()}.
}
\section{Provided by}{

The \code{delta} extension (\code{LOAD delta;}).
}

