% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{write_log}
\alias{write_log}
\title{DuckDB function write_log}
\usage{
write_log(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
Writes to the logger.
}
\examples{
\dontrun{
write_log('Hello')
}
}
