% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{xor}
\alias{xor}
\title{DuckDB function xor}
\usage{
xor(left, right)
}
\arguments{
\item{left, right}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise XOR.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{xor(17, 5)
}\if{html}{\out{</div>}}
}

