% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{abs}
\alias{abs}
\alias{@}
\title{DuckDB function abs}
\usage{
abs(x)
}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
Absolute value.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{abs(-17.4)
}\if{html}{\out{</div>}}
}

