% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{abs}
\alias{abs}
\title{DuckDB function abs}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
Absolute value.
}
\section{Overloads}{

\itemize{
\item \code{abs(x = TINYINT)}
\item \code{abs(x = SMALLINT)}
\item \code{abs(x = INTEGER)}
\item \code{abs(x = BIGINT)}
\item \code{abs(x = HUGEINT)}
\item \code{abs(x = FLOAT)}
\item \code{abs(x = DOUBLE)}
\item \code{abs(x = DECIMAL)}
\item \code{abs(x = UTINYINT)}
\item \code{abs(x = USMALLINT)}
\item \code{abs(x = UINTEGER)}
\item \code{abs(x = UBIGINT)}
\item \code{abs(x = UHUGEINT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{abs(-17.4)
}\if{html}{\out{</div>}}
}

