% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{acosh}
\alias{acosh}
\title{DuckDB function acosh}
\usage{
acosh(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the inverse hyperbolic cos of x.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{acosh(2.3)
}\if{html}{\out{</div>}}
}

