% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{age}
\alias{age}
\title{DuckDB function age}
\arguments{
\item{timestamp}{\verb{TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{INTERVAL}
}
\description{
Subtract arguments, resulting in the time difference between the two timestamps.
}
\examples{
\dontrun{
age(TIMESTAMP '2001-04-10', TIMESTAMP '1992-09-20')
}
}
