% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{alias}
\alias{alias}
\title{DuckDB function alias}
\usage{
alias(expr = ANY)
}
\arguments{
\item{expr}{\code{ANY}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the name of a given expression.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{alias(42 + 1)
}\if{html}{\out{</div>}}
}

