% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{&}
\alias{&}
\title{DuckDB function &}
\arguments{
\item{left}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}

\item{right}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise AND.
}
\section{Overloads}{

\itemize{
\item ```&`(left = TINYINT, right = TINYINT)``
\item ```&`(left = SMALLINT, right = SMALLINT)``
\item ```&`(left = INTEGER, right = INTEGER)``
\item ```&`(left = BIGINT, right = BIGINT)``
\item ```&`(left = HUGEINT, right = HUGEINT)``
\item ```&`(left = UTINYINT, right = UTINYINT)``
\item ```&`(left = USMALLINT, right = USMALLINT)``
\item ```&`(left = UINTEGER, right = UINTEGER)``
\item ```&`(left = UBIGINT, right = UBIGINT)``
\item ```&`(left = UHUGEINT, right = UHUGEINT)``
\item ```&`(left = BIT, right = BIT)``
}
}

\examples{
\dontrun{
91 & 15
}
}
