% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{approx_quantile}
\alias{approx_quantile}
\title{DuckDB function approx_quantile}
\arguments{
\item{x}{\verb{DECIMAL | SMALLINT | INTEGER | BIGINT | HUGEINT | DOUBLE | DATE | TIME | TIME WITH TIME ZONE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TINYINT | FLOAT}}

\item{pos}{\code{FLOAT | FLOAT[]}}
}
\value{
\verb{DECIMAL | SMALLINT | INTEGER | BIGINT | HUGEINT | DOUBLE | DATE | TIME | TIME WITH TIME ZONE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | DECIMAL[] | TINYINT[] | SMALLINT[] | INTEGER[] | BIGINT[] | HUGEINT[] | FLOAT[] | DOUBLE[] | DATE[] | TIME[] | TIME WITH TIME ZONE[] | TIMESTAMP[] | TIMESTAMP WITH TIME ZONE[]}
}
\description{
Computes the approximate quantile using T-Digest.
}
\examples{
\dontrun{
approx_quantile(x, 0.5)
}
}
