% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{approx_quantile}
\alias{approx_quantile}
\title{DuckDB function approx_quantile}
\arguments{
\item{x}{\verb{DECIMAL | SMALLINT | INTEGER | BIGINT | HUGEINT | DOUBLE | DATE | TIME | TIME WITH TIME ZONE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TINYINT | FLOAT}}

\item{pos}{\code{FLOAT | FLOAT[]}}
}
\value{
\verb{DECIMAL | SMALLINT | INTEGER | BIGINT | HUGEINT | DOUBLE | DATE | TIME | TIME WITH TIME ZONE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | DECIMAL[] | TINYINT[] | SMALLINT[] | INTEGER[] | BIGINT[] | HUGEINT[] | FLOAT[] | DOUBLE[] | DATE[] | TIME[] | TIME WITH TIME ZONE[] | TIMESTAMP[] | TIMESTAMP WITH TIME ZONE[]}
}
\description{
Computes the approximate quantile using T-Digest.
}
\section{Overloads}{

\itemize{
\item \code{approx_quantile(x = DECIMAL, pos = FLOAT)}
\item \code{approx_quantile(x = SMALLINT, pos = FLOAT)}
\item \code{approx_quantile(x = INTEGER, pos = FLOAT)}
\item \code{approx_quantile(x = BIGINT, pos = FLOAT)}
\item \code{approx_quantile(x = HUGEINT, pos = FLOAT)}
\item \code{approx_quantile(x = DOUBLE, pos = FLOAT)}
\item \code{approx_quantile(x = DATE, pos = FLOAT)}
\item \code{approx_quantile(x = TIME, pos = FLOAT)}
\item \code{approx_quantile(x = `TIME WITH TIME ZONE`, pos = FLOAT)}
\item \code{approx_quantile(x = TIMESTAMP, pos = FLOAT)}
\item \code{approx_quantile(x = `TIMESTAMP WITH TIME ZONE`, pos = FLOAT)}
\item \code{approx_quantile(x = DECIMAL, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = TINYINT, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = SMALLINT, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = INTEGER, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = BIGINT, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = HUGEINT, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = FLOAT, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = DOUBLE, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = DATE, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = TIME, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = `TIME WITH TIME ZONE`, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = TIMESTAMP, pos = `FLOAT[]`)}
\item \code{approx_quantile(x = `TIMESTAMP WITH TIME ZONE`, pos = `FLOAT[]`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{approx_quantile(x, 0.5)
}\if{html}{\out{</div>}}
}

