% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{approx_top_k}
\alias{approx_top_k}
\title{DuckDB function approx_top_k}
\usage{
approx_top_k(val = ANY, k = BIGINT)
}
\arguments{
\item{val}{\code{ANY}}

\item{k}{\code{BIGINT}}
}
\value{
\code{ANY[]}
}
\description{
Finds the k approximately most occurring values in the data set.
}
\examples{
\dontrun{
approx_top_k(x, 5)
}
}
