% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arg_max}
\alias{arg_max}
\title{DuckDB function arg_max}
\arguments{
\item{arg}{\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}}

\item{val}{\verb{INTEGER | BIGINT | HUGEINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | ANY}}

\item{col2}{\code{BIGINT}}
}
\value{
\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY | ANY[]}
}
\description{
Finds the row with the maximum val. Calculates the non-NULL arg expression at that row.
}
\section{Overloads}{

\itemize{
\item \code{arg_max(arg = INTEGER, val = INTEGER)}
\item \code{arg_max(arg = INTEGER, val = BIGINT)}
\item \code{arg_max(arg = INTEGER, val = HUGEINT)}
\item \code{arg_max(arg = INTEGER, val = DOUBLE)}
\item \code{arg_max(arg = INTEGER, val = VARCHAR)}
\item \code{arg_max(arg = INTEGER, val = DATE)}
\item \code{arg_max(arg = INTEGER, val = TIMESTAMP)}
\item \code{arg_max(arg = INTEGER, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = INTEGER, val = BLOB)}
\item \code{arg_max(arg = BIGINT, val = INTEGER)}
\item \code{arg_max(arg = BIGINT, val = BIGINT)}
\item \code{arg_max(arg = BIGINT, val = HUGEINT)}
\item \code{arg_max(arg = BIGINT, val = DOUBLE)}
\item \code{arg_max(arg = BIGINT, val = VARCHAR)}
\item \code{arg_max(arg = BIGINT, val = DATE)}
\item \code{arg_max(arg = BIGINT, val = TIMESTAMP)}
\item \code{arg_max(arg = BIGINT, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = BIGINT, val = BLOB)}
\item \code{arg_max(arg = DOUBLE, val = INTEGER)}
\item \code{arg_max(arg = DOUBLE, val = BIGINT)}
\item \code{arg_max(arg = DOUBLE, val = HUGEINT)}
\item \code{arg_max(arg = DOUBLE, val = DOUBLE)}
\item \code{arg_max(arg = DOUBLE, val = VARCHAR)}
\item \code{arg_max(arg = DOUBLE, val = DATE)}
\item \code{arg_max(arg = DOUBLE, val = TIMESTAMP)}
\item \code{arg_max(arg = DOUBLE, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = DOUBLE, val = BLOB)}
\item \code{arg_max(arg = VARCHAR, val = INTEGER)}
\item \code{arg_max(arg = VARCHAR, val = BIGINT)}
\item \code{arg_max(arg = VARCHAR, val = HUGEINT)}
\item \code{arg_max(arg = VARCHAR, val = DOUBLE)}
\item \code{arg_max(arg = VARCHAR, val = VARCHAR)}
\item \code{arg_max(arg = VARCHAR, val = DATE)}
\item \code{arg_max(arg = VARCHAR, val = TIMESTAMP)}
\item \code{arg_max(arg = VARCHAR, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = VARCHAR, val = BLOB)}
\item \code{arg_max(arg = DATE, val = INTEGER)}
\item \code{arg_max(arg = DATE, val = BIGINT)}
\item \code{arg_max(arg = DATE, val = HUGEINT)}
\item \code{arg_max(arg = DATE, val = DOUBLE)}
\item \code{arg_max(arg = DATE, val = VARCHAR)}
\item \code{arg_max(arg = DATE, val = DATE)}
\item \code{arg_max(arg = DATE, val = TIMESTAMP)}
\item \code{arg_max(arg = DATE, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = DATE, val = BLOB)}
\item \code{arg_max(arg = TIMESTAMP, val = INTEGER)}
\item \code{arg_max(arg = TIMESTAMP, val = BIGINT)}
\item \code{arg_max(arg = TIMESTAMP, val = HUGEINT)}
\item \code{arg_max(arg = TIMESTAMP, val = DOUBLE)}
\item \code{arg_max(arg = TIMESTAMP, val = VARCHAR)}
\item \code{arg_max(arg = TIMESTAMP, val = DATE)}
\item \code{arg_max(arg = TIMESTAMP, val = TIMESTAMP)}
\item \code{arg_max(arg = TIMESTAMP, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = TIMESTAMP, val = BLOB)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = INTEGER)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = BIGINT)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = HUGEINT)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = DOUBLE)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = VARCHAR)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = DATE)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = TIMESTAMP)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = `TIMESTAMP WITH TIME ZONE`, val = BLOB)}
\item \code{arg_max(arg = BLOB, val = INTEGER)}
\item \code{arg_max(arg = BLOB, val = BIGINT)}
\item \code{arg_max(arg = BLOB, val = HUGEINT)}
\item \code{arg_max(arg = BLOB, val = DOUBLE)}
\item \code{arg_max(arg = BLOB, val = VARCHAR)}
\item \code{arg_max(arg = BLOB, val = DATE)}
\item \code{arg_max(arg = BLOB, val = TIMESTAMP)}
\item \code{arg_max(arg = BLOB, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = BLOB, val = BLOB)}
\item \code{arg_max(arg = DECIMAL, val = INTEGER)}
\item \code{arg_max(arg = DECIMAL, val = BIGINT)}
\item \code{arg_max(arg = DECIMAL, val = HUGEINT)}
\item \code{arg_max(arg = DECIMAL, val = DOUBLE)}
\item \code{arg_max(arg = DECIMAL, val = VARCHAR)}
\item \code{arg_max(arg = DECIMAL, val = DATE)}
\item \code{arg_max(arg = DECIMAL, val = TIMESTAMP)}
\item \code{arg_max(arg = DECIMAL, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = DECIMAL, val = BLOB)}
\item \code{arg_max(arg = ANY, val = INTEGER)}
\item \code{arg_max(arg = ANY, val = BIGINT)}
\item \code{arg_max(arg = ANY, val = HUGEINT)}
\item \code{arg_max(arg = ANY, val = DOUBLE)}
\item \code{arg_max(arg = ANY, val = VARCHAR)}
\item \code{arg_max(arg = ANY, val = DATE)}
\item \code{arg_max(arg = ANY, val = TIMESTAMP)}
\item \code{arg_max(arg = ANY, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max(arg = ANY, val = BLOB)}
\item \code{arg_max(arg = ANY, val = ANY)}
\item \code{arg_max(arg = ANY, val = ANY, col2 = BIGINT)}
}
}

\examples{
\dontrun{
arg_max(A, B)
}
}
