% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arg_max_null}
\alias{arg_max_null}
\title{DuckDB function arg_max_null}
\arguments{
\item{arg}{\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}}

\item{val}{\verb{INTEGER | BIGINT | HUGEINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | ANY}}
}
\value{
\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}
}
\description{
Finds the row with the maximum val. Calculates the arg expression at that row.
}
\examples{
\dontrun{
arg_max_null(A, B)
}
}
