% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arg_max_null}
\alias{arg_max_null}
\title{DuckDB function arg_max_null}
\arguments{
\item{arg}{\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}}

\item{val}{\verb{INTEGER | BIGINT | HUGEINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | ANY}}
}
\value{
\verb{INTEGER | BIGINT | DOUBLE | VARCHAR | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | BLOB | DECIMAL | ANY}
}
\description{
Finds the row with the maximum val. Calculates the arg expression at that row.
}
\section{Overloads}{

\itemize{
\item \code{arg_max_null(arg = INTEGER, val = INTEGER)}
\item \code{arg_max_null(arg = INTEGER, val = BIGINT)}
\item \code{arg_max_null(arg = INTEGER, val = HUGEINT)}
\item \code{arg_max_null(arg = INTEGER, val = DOUBLE)}
\item \code{arg_max_null(arg = INTEGER, val = VARCHAR)}
\item \code{arg_max_null(arg = INTEGER, val = DATE)}
\item \code{arg_max_null(arg = INTEGER, val = TIMESTAMP)}
\item \code{arg_max_null(arg = INTEGER, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = INTEGER, val = BLOB)}
\item \code{arg_max_null(arg = BIGINT, val = INTEGER)}
\item \code{arg_max_null(arg = BIGINT, val = BIGINT)}
\item \code{arg_max_null(arg = BIGINT, val = HUGEINT)}
\item \code{arg_max_null(arg = BIGINT, val = DOUBLE)}
\item \code{arg_max_null(arg = BIGINT, val = VARCHAR)}
\item \code{arg_max_null(arg = BIGINT, val = DATE)}
\item \code{arg_max_null(arg = BIGINT, val = TIMESTAMP)}
\item \code{arg_max_null(arg = BIGINT, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = BIGINT, val = BLOB)}
\item \code{arg_max_null(arg = DOUBLE, val = INTEGER)}
\item \code{arg_max_null(arg = DOUBLE, val = BIGINT)}
\item \code{arg_max_null(arg = DOUBLE, val = HUGEINT)}
\item \code{arg_max_null(arg = DOUBLE, val = DOUBLE)}
\item \code{arg_max_null(arg = DOUBLE, val = VARCHAR)}
\item \code{arg_max_null(arg = DOUBLE, val = DATE)}
\item \code{arg_max_null(arg = DOUBLE, val = TIMESTAMP)}
\item \code{arg_max_null(arg = DOUBLE, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = DOUBLE, val = BLOB)}
\item \code{arg_max_null(arg = VARCHAR, val = INTEGER)}
\item \code{arg_max_null(arg = VARCHAR, val = BIGINT)}
\item \code{arg_max_null(arg = VARCHAR, val = HUGEINT)}
\item \code{arg_max_null(arg = VARCHAR, val = DOUBLE)}
\item \code{arg_max_null(arg = VARCHAR, val = VARCHAR)}
\item \code{arg_max_null(arg = VARCHAR, val = DATE)}
\item \code{arg_max_null(arg = VARCHAR, val = TIMESTAMP)}
\item \code{arg_max_null(arg = VARCHAR, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = VARCHAR, val = BLOB)}
\item \code{arg_max_null(arg = DATE, val = INTEGER)}
\item \code{arg_max_null(arg = DATE, val = BIGINT)}
\item \code{arg_max_null(arg = DATE, val = HUGEINT)}
\item \code{arg_max_null(arg = DATE, val = DOUBLE)}
\item \code{arg_max_null(arg = DATE, val = VARCHAR)}
\item \code{arg_max_null(arg = DATE, val = DATE)}
\item \code{arg_max_null(arg = DATE, val = TIMESTAMP)}
\item \code{arg_max_null(arg = DATE, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = DATE, val = BLOB)}
\item \code{arg_max_null(arg = TIMESTAMP, val = INTEGER)}
\item \code{arg_max_null(arg = TIMESTAMP, val = BIGINT)}
\item \code{arg_max_null(arg = TIMESTAMP, val = HUGEINT)}
\item \code{arg_max_null(arg = TIMESTAMP, val = DOUBLE)}
\item \code{arg_max_null(arg = TIMESTAMP, val = VARCHAR)}
\item \code{arg_max_null(arg = TIMESTAMP, val = DATE)}
\item \code{arg_max_null(arg = TIMESTAMP, val = TIMESTAMP)}
\item \code{arg_max_null(arg = TIMESTAMP, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = TIMESTAMP, val = BLOB)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = INTEGER)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = BIGINT)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = HUGEINT)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = DOUBLE)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = VARCHAR)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = DATE)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = TIMESTAMP)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = `TIMESTAMP WITH TIME ZONE`, val = BLOB)}
\item \code{arg_max_null(arg = BLOB, val = INTEGER)}
\item \code{arg_max_null(arg = BLOB, val = BIGINT)}
\item \code{arg_max_null(arg = BLOB, val = HUGEINT)}
\item \code{arg_max_null(arg = BLOB, val = DOUBLE)}
\item \code{arg_max_null(arg = BLOB, val = VARCHAR)}
\item \code{arg_max_null(arg = BLOB, val = DATE)}
\item \code{arg_max_null(arg = BLOB, val = TIMESTAMP)}
\item \code{arg_max_null(arg = BLOB, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = BLOB, val = BLOB)}
\item \code{arg_max_null(arg = DECIMAL, val = INTEGER)}
\item \code{arg_max_null(arg = DECIMAL, val = BIGINT)}
\item \code{arg_max_null(arg = DECIMAL, val = HUGEINT)}
\item \code{arg_max_null(arg = DECIMAL, val = DOUBLE)}
\item \code{arg_max_null(arg = DECIMAL, val = VARCHAR)}
\item \code{arg_max_null(arg = DECIMAL, val = DATE)}
\item \code{arg_max_null(arg = DECIMAL, val = TIMESTAMP)}
\item \code{arg_max_null(arg = DECIMAL, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = DECIMAL, val = BLOB)}
\item \code{arg_max_null(arg = ANY, val = INTEGER)}
\item \code{arg_max_null(arg = ANY, val = BIGINT)}
\item \code{arg_max_null(arg = ANY, val = HUGEINT)}
\item \code{arg_max_null(arg = ANY, val = DOUBLE)}
\item \code{arg_max_null(arg = ANY, val = VARCHAR)}
\item \code{arg_max_null(arg = ANY, val = DATE)}
\item \code{arg_max_null(arg = ANY, val = TIMESTAMP)}
\item \code{arg_max_null(arg = ANY, val = `TIMESTAMP WITH TIME ZONE`)}
\item \code{arg_max_null(arg = ANY, val = BLOB)}
\item \code{arg_max_null(arg = ANY, val = ANY)}
}
}

\examples{
\dontrun{
arg_max_null(A, B)
}
}
