% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_agg}
\alias{array_agg}
\title{DuckDB function array_agg}
\usage{
array_agg(arg = T)
}
\arguments{
\item{arg}{\code{T}}
}
\value{
\code{T[]}
}
\description{
Returns a LIST containing all the values of a column.
}
\examples{
\dontrun{
array_agg(A)
}
}
