% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_cat}
\alias{array_cat}
\title{DuckDB function array_cat}
\usage{
array_cat()
}
\value{
\code{ANY[]}
}
\description{
Concatenates lists. \code{NULL} inputs are skipped. See also operator \code{||}.
}
\examples{
\dontrun{
array_cat([2, 3], [4, 5, 6], [7])
}
}
