% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_contains}
\alias{array_contains}
\title{DuckDB function array_contains}
\usage{
array_contains(list = `T[]`, element = T)
}
\arguments{
\item{list}{\code{T[]}}

\item{element}{\code{T}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the list contains the element.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_contains([1, 2, NULL], 1)
}\if{html}{\out{</div>}}
}

