% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_cosine_similarity}
\alias{array_cosine_similarity}
\title{DuckDB function array_cosine_similarity}
\arguments{
\item{array1}{\code{FLOAT[ANY] | DOUBLE[ANY]}}

\item{array2}{\code{FLOAT[ANY] | DOUBLE[ANY]}}
}
\value{
\code{FLOAT | DOUBLE}
}
\description{
Computes the cosine similarity between two arrays of the same size. The array elements can not be \code{NULL}. The arrays can have any size as long as the size is the same for both arguments.
}
\section{Overloads}{

\itemize{
\item \code{array_cosine_similarity(array1 = `FLOAT[ANY]`, array2 = `FLOAT[ANY]`)}
\item \code{array_cosine_similarity(array1 = `DOUBLE[ANY]`, array2 = `DOUBLE[ANY]`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_cosine_similarity(array_value(1.0::FLOAT, 2.0::FLOAT, 3.0::FLOAT), array_value(2.0::FLOAT, 3.0::FLOAT, 4.0::FLOAT))
}\if{html}{\out{</div>}}
}

