% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_cross_product}
\alias{array_cross_product}
\title{DuckDB function array_cross_product}
\arguments{
\item{array}{\code{FLOAT[3] | DOUBLE[3]}}
}
\value{
\code{FLOAT[3] | DOUBLE[3]}
}
\description{
Computes the cross product of two arrays of size 3. The array elements can not be \code{NULL}.
}
\section{Overloads}{

\itemize{
\item \code{array_cross_product(array = `FLOAT[3]`, array = `FLOAT[3]`)}
\item \code{array_cross_product(array = `DOUBLE[3]`, array = `DOUBLE[3]`)}
}
}

\examples{
\dontrun{
array_cross_product(array_value(1.0::FLOAT, 2.0::FLOAT, 3.0::FLOAT), array_value(2.0::FLOAT, 3.0::FLOAT, 4.0::FLOAT))
}
}
