% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_distinct}
\alias{array_distinct}
\title{DuckDB function array_distinct}
\usage{
array_distinct(list = `T[]`)
}
\arguments{
\item{list}{\code{T[]}}
}
\value{
\code{T[]}
}
\description{
Removes all duplicates and \code{NULL} values from a list. Does not preserve the original order.
}
\examples{
\dontrun{
array_distinct([1, 1, NULL, -3, 1, 5])
}
}
