% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_filter}
\alias{array_filter}
\title{DuckDB function array_filter}
\usage{
array_filter(list = `ANY[]`, `lambda(x)` = LAMBDA)
}
\arguments{
\item{list}{\code{ANY[]}}

\item{lambda(x)}{\code{LAMBDA}}
}
\description{
Constructs a list from those elements of the input \code{list} for which the \code{lambda} function returns \code{true}. DuckDB must be able to cast the \code{lambda} function's return type to \code{BOOL}. The return type of \code{list_filter} is the same as the input list's.
}
\examples{
\dontrun{
array_filter([3, 4, 5], lambda x : x > 4)
}
}
