% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_has_all}
\alias{array_has_all}
\alias{duckdb_fun_array_has_all}
\title{DuckDB function array_has_all}
\arguments{
\item{l1}{\code{ANY[]}}

\item{l2}{\code{ANY[]}}
}
\description{
Returns true if all elements of l2 are in l1. NULLs are ignored.
}
\examples{
\dontrun{
list_has_all([1, 2, 3], [2, 3])
}
}
