% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_length}
\alias{array_length}
\title{DuckDB function array_length}
\arguments{
\item{list}{\code{ANY[]}}

\item{dimension}{\code{BIGINT}}
}
\description{
Returns the length of the \code{list}.

\code{array_length} for lists with dimensions other than 1 not implemented.
}
\examples{
\dontrun{
array_length([1, 2, 3])
}
}
