% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_position}
\alias{array_position}
\alias{duckdb_fun_array_position}
\title{DuckDB function array_position}
\arguments{
\item{list}{\code{ANY[]}}

\item{element}{\code{ANY}}
}
\description{
Returns the index of the element if the list contains the element. If the element is not found, it returns NULL.
}
\examples{
\dontrun{
list_position([1, 2, NULL], 2)
}
}
