% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_resize}
\alias{array_resize}
\title{DuckDB function array_resize}
\arguments{
\item{list}{\code{ANY[]}}

\item{size[}{\code{ANY}}

\item{value]}{\code{ANY}}
}
\description{
Resizes the \code{list} to contain \code{size} elements. Initializes new elements with \code{value} or \code{NULL} if \code{value} is not set.
}
\examples{
\dontrun{
array_resize([1, 2, 3], 5, 0)
}
}
