% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_select}
\alias{array_select}
\title{DuckDB function array_select}
\usage{
array_select(value_list = `T[]`, index_list = `BIGINT[]`)
}
\arguments{
\item{value_list}{\code{T[]}}

\item{index_list}{\code{BIGINT[]}}
}
\value{
\code{T[]}
}
\description{
Returns a list based on the elements selected by the \code{index_list}.
}
\examples{
\dontrun{
array_select([10, 20, 30, 40], [1, 4])
}
}
