% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_sort}
\alias{array_sort}
\title{DuckDB function array_sort}
\arguments{
\item{list}{\code{ANY[]}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR}}
}
\value{
\code{ANY[]}
}
\description{
Sorts the elements of the list.
}
\examples{
\dontrun{
array_sort([3, 6, 1, 2])
}
}
