% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_transform}
\alias{array_transform}
\title{DuckDB function array_transform}
\usage{
array_transform(list = `ANY[]`, `lambda(x)` = LAMBDA)
}
\arguments{
\item{list}{\code{ANY[]}}

\item{lambda(x)}{\code{LAMBDA}}
}
\value{
\code{ANY[]}
}
\description{
Returns a list that is the result of applying the \code{lambda} function to each element of the input \code{list}. The return type is defined by the return type of the \code{lambda} function.
}
\examples{
\dontrun{
array_transform([1, 2, 3], lambda x : x + 1)
}
}
