% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_unique}
\alias{array_unique}
\title{DuckDB function array_unique}
\usage{
array_unique(list = `ANY[]`)
}
\arguments{
\item{list}{\code{ANY[]}}
}
\description{
Counts the unique elements of a \code{list}.
}
\examples{
\dontrun{
array_unique([1, 1, NULL, -3, 1, 5])
}
}
