% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_unique}
\alias{array_unique}
\title{DuckDB function array_unique}
\arguments{
\item{list}{\code{ANY[]}}
}
\description{
Counts the unique elements of a list.
}
\examples{
\dontrun{
list_unique([1, 1, NULL, -3, 1, 5])
}
}
