% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_value}
\alias{array_value}
\title{DuckDB function array_value}
\usage{
array_value()
}
\value{
\code{ARRAY}
}
\description{
Creates an \code{ARRAY} containing the argument values.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{array_value(1.0::FLOAT, 2.0::FLOAT, 3.0::FLOAT)
}\if{html}{\out{</div>}}
}

