% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{array_where}
\alias{array_where}
\title{DuckDB function array_where}
\arguments{
\item{value_list}{\code{T[]}}

\item{mask_list}{\code{BOOLEAN[]}}
}
\description{
Returns a list with the \code{BOOLEAN}s in \code{mask_list} applied as a mask to the \code{value_list}.
}
\examples{
\dontrun{
array_where([10, 20, 30, 40], [true, false, false, true])
}
}
