% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{arrow_scan_dumb}
\alias{arrow_scan_dumb}
\title{DuckDB function arrow_scan_dumb}
\usage{
arrow_scan_dumb(col0 = POINTER, col1 = POINTER, col2 = POINTER)
}
\arguments{
\item{col0}{\code{POINTER}}

\item{col1}{\code{POINTER}}

\item{col2}{\code{POINTER}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{arrow_scan_dumb()}.
}
\examples{
\dontrun{
character(0)
}
}
