% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ascii}
\alias{ascii}
\title{DuckDB function ascii}
\usage{
ascii(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Returns an integer that represents the Unicode code point of the first character of the \code{string}.
}
\examples{
\dontrun{
ascii('Ω')
}
}
