% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{asinh}
\alias{asinh}
\alias{duckdb_fun_asinh}
\title{DuckDB function asinh}
\arguments{
\item{x}{\code{DOUBLE}}
}
\description{
Computes the inverse hyperbolic sin of x.
}
\examples{
\dontrun{
asinh(0.5)
}
}
