% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{at-}
\alias{at-}
\alias{@}
\title{DuckDB function @}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
Absolute value.
}
\section{Overloads}{

\itemize{
\item \code{`@`(x = TINYINT)}
\item \code{`@`(x = SMALLINT)}
\item \code{`@`(x = INTEGER)}
\item \code{`@`(x = BIGINT)}
\item \code{`@`(x = HUGEINT)}
\item \code{`@`(x = FLOAT)}
\item \code{`@`(x = DOUBLE)}
\item \code{`@`(x = DECIMAL)}
\item \code{`@`(x = UTINYINT)}
\item \code{`@`(x = USMALLINT)}
\item \code{`@`(x = UINTEGER)}
\item \code{`@`(x = UBIGINT)}
\item \code{`@`(x = UHUGEINT)}
}
}

\examples{
\dontrun{
abs(-17.4)
}
}
