% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{at-}
\alias{at-}
\alias{@}
\title{DuckDB function @}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
Absolute value.
}
\examples{
\dontrun{
abs(-17.4)
}
}
