% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{atan2}
\alias{atan2}
\title{DuckDB function atan2}
\usage{
atan2(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\description{
Computes the arctangent (y, x).
}
\examples{
\dontrun{
atan2(1.0, 0.0)
}
}
