% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{atanh}
\alias{atanh}
\alias{duckdb_fun_atanh}
\title{DuckDB function atanh}
\arguments{
\item{x}{\code{DOUBLE}}
}
\description{
Computes the inverse hyperbolic tan of x.
}
\examples{
\dontrun{
atanh(0.5)
}
}
