% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bar}
\alias{bar}
\title{DuckDB function bar}
\arguments{
\item{x}{\code{DOUBLE}}

\item{min}{\code{DOUBLE}}

\item{max}{\code{DOUBLE}}

\item{width}{\code{DOUBLE}}
}
\description{
Draws a band whose width is proportional to (\code{x - min}) and equal to \code{width} characters when \code{x} = \code{max}. \code{width} defaults to 80.
}
\examples{
\dontrun{
bar(5, 0, 20, 10)
}
}
