% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bar}
\alias{bar}
\title{DuckDB function bar}
\arguments{
\item{x}{\code{DOUBLE}}

\item{min}{\code{DOUBLE}}

\item{max}{\code{DOUBLE}}

\item{width}{\code{DOUBLE}}
}
\value{
\code{VARCHAR}
}
\description{
Draws a band whose width is proportional to (\code{x - min}) and equal to \code{width} characters when \code{x} = \code{max}. \code{width} defaults to 80.
}
\section{Overloads}{

\itemize{
\item \code{bar(x = DOUBLE, min = DOUBLE, max = DOUBLE, width = DOUBLE)}
\item \code{bar(x = DOUBLE, min = DOUBLE, max = DOUBLE)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bar(5, 0, 20, 10)
}\if{html}{\out{</div>}}
}

