% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{base64}
\alias{base64}
\title{DuckDB function base64}
\usage{
base64(blob = BLOB)
}
\arguments{
\item{blob}{\code{BLOB}}
}
\value{
\code{VARCHAR}
}
\description{
Converts a \code{blob} to a base64 encoded string.
}
\examples{
\dontrun{
base64('A'::BLOB)
}
}
