% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bin}
\alias{bin}
\title{DuckDB function bin}
\arguments{
\item{string}{\code{VARCHAR}}

\item{value}{\code{BIGNUM | UBIGINT | BIGINT | HUGEINT | UHUGEINT}}
}
\value{
\code{VARCHAR}
}
\description{
Converts the \code{string} to binary representation.

Converts the \code{value} to binary representation.
}
\section{Overloads}{

\itemize{
\item \code{bin(string = VARCHAR)}
\item \code{bin(value = BIGNUM)}
\item \code{bin(value = UBIGINT)}
\item \code{bin(value = BIGINT)}
\item \code{bin(value = HUGEINT)}
\item \code{bin(value = UHUGEINT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bin('Aa')
bin(42)
}\if{html}{\out{</div>}}
}

