% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bin}
\alias{bin}
\title{DuckDB function bin}
\arguments{
\item{string}{\code{VARCHAR}}

\item{value}{\code{BIGNUM | UBIGINT | BIGINT | HUGEINT | UHUGEINT}}
}
\description{
Converts the \code{string} to binary representation.

Converts the \code{value} to binary representation.
}
\examples{
\dontrun{
bin('Aa')
bin(42)
}
}
