% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_and}
\alias{bit_and}
\title{DuckDB function bit_and}
\arguments{
\item{arg}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Returns the bitwise AND of all bits in a given expression.
}
\examples{
\dontrun{
bit_and(A)
}
}
