% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_count}
\alias{bit_count}
\title{DuckDB function bit_count}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | BIT}}
}
\value{
\code{TINYINT | BIGINT}
}
\description{
Returns the number of bits that are set.
}
\examples{
\dontrun{
bit_count(31)
}
}
