% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_length}
\alias{bit_length}
\title{DuckDB function bit_length}
\arguments{
\item{string}{\code{VARCHAR}}

\item{bit}{\code{BIT}}
}
\value{
\code{BIGINT}
}
\description{
Number of bits in a \code{string}.

Returns the bit-length of the \code{bit} argument.
}
\section{Overloads}{

\itemize{
\item \code{bit_length(string = VARCHAR)}
\item \code{bit_length(bit = BIT)}
}
}

\examples{
\dontrun{
bit_length('abc')
bit_length(42::TINYINT::BIT)
}
}
