% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_or}
\alias{bit_or}
\title{DuckDB function bit_or}
\usage{
bit_or(arg)
}
\arguments{
\item{arg}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Returns the bitwise OR of all bits in a given expression.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bit_or(A)
}\if{html}{\out{</div>}}
}

