% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_or}
\alias{bit_or}
\title{DuckDB function bit_or}
\arguments{
\item{arg}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Returns the bitwise OR of all bits in a given expression.
}
\section{Overloads}{

\itemize{
\item \code{bit_or(arg = TINYINT)}
\item \code{bit_or(arg = SMALLINT)}
\item \code{bit_or(arg = INTEGER)}
\item \code{bit_or(arg = BIGINT)}
\item \code{bit_or(arg = HUGEINT)}
\item \code{bit_or(arg = UTINYINT)}
\item \code{bit_or(arg = USMALLINT)}
\item \code{bit_or(arg = UINTEGER)}
\item \code{bit_or(arg = UBIGINT)}
\item \code{bit_or(arg = UHUGEINT)}
\item \code{bit_or(arg = BIT)}
}
}

\examples{
\dontrun{
bit_or(A)
}
}
