% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bit_xor}
\alias{bit_xor}
\title{DuckDB function bit_xor}
\arguments{
\item{arg}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Returns the bitwise XOR of all bits in a given expression.
}
\section{Overloads}{

\itemize{
\item \code{bit_xor(arg = TINYINT)}
\item \code{bit_xor(arg = SMALLINT)}
\item \code{bit_xor(arg = INTEGER)}
\item \code{bit_xor(arg = BIGINT)}
\item \code{bit_xor(arg = HUGEINT)}
\item \code{bit_xor(arg = UTINYINT)}
\item \code{bit_xor(arg = USMALLINT)}
\item \code{bit_xor(arg = UINTEGER)}
\item \code{bit_xor(arg = UBIGINT)}
\item \code{bit_xor(arg = UHUGEINT)}
\item \code{bit_xor(arg = BIT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bit_xor(A)
}\if{html}{\out{</div>}}
}

