% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bitstring}
\alias{bitstring}
\title{DuckDB function bitstring}
\arguments{
\item{bitstring}{\code{VARCHAR | BIT}}

\item{length}{\code{INTEGER}}
}
\value{
\code{BIT}
}
\description{
Pads the bitstring until the specified length.
}
\section{Overloads}{

\itemize{
\item \code{bitstring(bitstring = VARCHAR, length = INTEGER)}
\item \code{bitstring(bitstring = BIT, length = INTEGER)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bitstring('1010'::BIT, 7)
}\if{html}{\out{</div>}}
}

