% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bool_and}
\alias{bool_and}
\title{DuckDB function bool_and}
\usage{
bool_and(arg = BOOLEAN)
}
\arguments{
\item{arg}{\code{BOOLEAN}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns TRUE if every input value is TRUE, otherwise FALSE.
}
\examples{
\dontrun{
bool_and(A)
}
}
