% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{bool_or}
\alias{bool_or}
\title{DuckDB function bool_or}
\usage{
bool_or(arg = BOOLEAN)
}
\arguments{
\item{arg}{\code{BOOLEAN}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns TRUE if any input value is TRUE, otherwise FALSE.
}
\examples{
\dontrun{
bool_or(A)
}
}
