% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{can_cast_implicitly}
\alias{can_cast_implicitly}
\title{DuckDB function can_cast_implicitly}
\usage{
can_cast_implicitly(source_type = ANY, target_type = ANY)
}
\arguments{
\item{source_type}{\code{ANY}}

\item{target_type}{\code{ANY}}
}
\value{
\code{BOOLEAN}
}
\description{
Whether or not we can implicitly cast from the source type to the other type.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{can_cast_implicitly(NULL::INTEGER, NULL::BIGINT)
}\if{html}{\out{</div>}}
}

