% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{cbrt}
\alias{cbrt}
\title{DuckDB function cbrt}
\usage{
cbrt(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the cube root of x.
}
\examples{
\dontrun{
cbrt(8)
}
}
