% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ceil}
\alias{ceil}
\title{DuckDB function ceil}
\arguments{
\item{x}{\code{FLOAT | DOUBLE | DECIMAL}}
}
\value{
\code{FLOAT | DOUBLE | DECIMAL}
}
\description{
Rounds the number up.
}
\section{Overloads}{

\itemize{
\item \code{ceil(x = FLOAT)}
\item \code{ceil(x = DOUBLE)}
\item \code{ceil(x = DECIMAL)}
}
}

\examples{
\dontrun{
ceil(17.4)
}
}
