% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{char_length}
\alias{char_length}
\title{DuckDB function char_length}
\arguments{
\item{string}{\code{VARCHAR}}

\item{bit}{\code{BIT}}

\item{list}{\code{ANY[]}}
}
\value{
\code{BIGINT}
}
\description{
Number of characters in \code{string}.

Returns the bit-length of the \code{bit} argument.

Returns the length of the \code{list}.
}
\examples{
\dontrun{
char_length('Hello🦆')
char_length(42::TINYINT::BIT)
char_length([1,2,3])
}
}
