% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{character_length}
\alias{character_length}
\title{DuckDB function character_length}
\arguments{
\item{string}{\code{VARCHAR}}

\item{bit}{\code{BIT}}

\item{list}{\code{ANY[]}}
}
\description{
Number of characters in \code{string}.

Returns the bit-length of the \code{bit} argument.

Returns the length of the \code{list}.
}
\examples{
\dontrun{
character_length('Hello🦆')
character_length(42::TINYINT::BIT)
character_length([1,2,3])
}
}
