% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{checkpoint}
\alias{checkpoint}
\title{DuckDB function checkpoint}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{checkpoint()}.
}
\examples{
\dontrun{
character(0)
}
}
