% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{checkpoint}
\alias{checkpoint}
\title{DuckDB function checkpoint}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{checkpoint()}.
}
\section{Overloads}{

\itemize{
\item \code{checkpoint()}
\item \code{checkpoint(col0 = VARCHAR)}
}
}

