% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{chr}
\alias{chr}
\title{DuckDB function chr}
\usage{
chr(code_point = INTEGER)
}
\arguments{
\item{code_point}{\code{INTEGER}}
}
\value{
\code{VARCHAR}
}
\description{
Returns a character which is corresponding the ASCII code value or Unicode code point.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{chr(65)
}\if{html}{\out{</div>}}
}

