% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{chr}
\alias{chr}
\alias{duckdb_fun_chr}
\title{DuckDB function chr}
\arguments{
\item{code_point}{\code{INTEGER}}
}
\description{
Returns a character which is corresponding the ASCII code value or Unicode code point.
}
\examples{
\dontrun{
chr(65)
}
}
