% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{concat}
\alias{concat}
\title{DuckDB function concat}
\usage{
concat(value = ANY)
}
\arguments{
\item{value}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
Concatenates multiple strings or lists. \code{NULL} inputs are skipped. See also operator \code{||}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{concat('Hello', ' ', 'World')
concat([1, 2, 3], NULL, [4, 5, 6])
}\if{html}{\out{</div>}}
}

