% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{concat_ws}
\alias{concat_ws}
\title{DuckDB function concat_ws}
\usage{
concat_ws(separator = VARCHAR, string = ANY)
}
\arguments{
\item{separator}{\code{VARCHAR}}

\item{string}{\code{ANY}}
}
\value{
\code{VARCHAR}
}
\description{
Concatenates many strings, separated by \code{separator}. \code{NULL} inputs are skipped.
}
\examples{
\dontrun{
concat_ws(', ', 'Banana', 'Apple', 'Melon')
}
}
