% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{contains}
\alias{contains}
\title{DuckDB function contains}
\usage{
contains(string = VARCHAR, search_string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\description{
Returns \code{true} if \code{search_string} is found within \code{string}.
}
\examples{
\dontrun{
contains('abc', 'a')
}
}
