% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{contains}
\alias{contains}
\title{DuckDB function contains}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}

\item{col0}{\code{T[] | MAP(K, V) | STRUCT}}

\item{col1}{\code{T | K | ANY}}
}
\description{
Returns \code{true} if \code{search_string} is found within \code{string}.
}
\examples{
\dontrun{
contains('abc', 'a')
character(0)
}
}
