% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{contains}
\alias{contains}
\title{DuckDB function contains}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}

\item{col0}{\code{T[] | MAP(K, V) | STRUCT}}

\item{col1}{\code{T | K | ANY}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if \code{search_string} is found within \code{string}.
}
\section{Overloads}{

\itemize{
\item \code{contains(string = VARCHAR, search_string = VARCHAR)}
\item \code{contains(col0 = `T[]`, col1 = T)}
\item \code{contains(col0 = `MAP(K, V)`, col1 = K)}
\item \code{contains(col0 = STRUCT, col1 = ANY)}
}
}

\examples{
\dontrun{
contains('abc', 'a')
}
}
