% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{copy_database}
\alias{copy_database}
\title{DuckDB function copy_database}
\usage{
copy_database(col0 = VARCHAR, col1 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{col1}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{copy_database()}.
}
