% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{corr}
\alias{corr}
\title{DuckDB function corr}
\usage{
corr(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the correlation coefficient for non-NULL pairs in a group.
}
\examples{
\dontrun{
COVAR_POP(y, x) / (STDDEV_POP(x) * STDDEV_POP(y))
}
}
