% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{covar_pop}
\alias{covar_pop}
\title{DuckDB function covar_pop}
\usage{
covar_pop(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the population covariance of input values.
}
\examples{
\dontrun{
(SUM(x*y) - SUM(x) * SUM(y) / COUNT(*)) / COUNT(*)
}
}
