% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{covar_samp}
\alias{covar_samp}
\title{DuckDB function covar_samp}
\usage{
covar_samp(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the sample covariance for non-NULL pairs in a group.
}
\examples{
\dontrun{
(SUM(x*y) - SUM(x) * SUM(y) / COUNT(*)) / (COUNT(*) - 1)
}
}
