% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{create_sort_key}
\alias{create_sort_key}
\title{DuckDB function create_sort_key}
\usage{
create_sort_key(parameters... = ANY)
}
\arguments{
\item{parameters...}{\code{ANY}}
}
\value{
\code{BLOB}
}
\description{
Constructs a binary-comparable sort key based on a set of input parameters and sort qualifiers.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{create_sort_key('A', 'DESC')
}\if{html}{\out{</div>}}
}

