% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_database}
\alias{current_database}
\title{DuckDB function current_database}
\value{
\code{VARCHAR}
}
\description{
Returns the name of the currently active database.
}
\section{Overloads}{

\itemize{
\item \code{current_database()}
\item \code{current_database()}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{current_database()
}\if{html}{\out{</div>}}
}

