% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_database}
\alias{current_database}
\title{DuckDB function current_database}
\usage{
current_database()
}
\value{
\code{VARCHAR}
}
\description{
Returns the name of the currently active database.
}
\examples{
\dontrun{
current_database()
}
}
