% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_schema}
\alias{current_schema}
\title{DuckDB function current_schema}
\usage{
current_schema()
}
\value{
\code{VARCHAR}
}
\description{
Returns the name of the currently active schema. Default is main.
}
\examples{
\dontrun{
current_schema()
}
}
