% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_schemas}
\alias{current_schemas}
\title{DuckDB function current_schemas}
\arguments{
\item{include_implicit}{\code{BOOLEAN}}
}
\value{
\code{VARCHAR[]}
}
\description{
Returns list of schemas. Pass a parameter of True to include implicit schemas.
}
\section{Overloads}{

\itemize{
\item \code{current_schemas(include_implicit = BOOLEAN)}
\item \code{current_schemas(include_implicit)}
}
}

\examples{
\dontrun{
current_schemas(true)
}
}
