% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_setting}
\alias{current_setting}
\alias{duckdb_fun_current_setting}
\title{DuckDB function current_setting}
\arguments{
\item{setting_name}{\code{VARCHAR}}
}
\description{
Returns the current value of the configuration setting.
}
\examples{
\dontrun{
current_setting('access_mode')
}
}
