% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{current_transaction_id}
\alias{current_transaction_id}
\title{DuckDB function current_transaction_id}
\usage{
current_transaction_id()
}
\value{
\code{UBIGINT}
}
\description{
Get the current global transaction_id.
}
\examples{
\dontrun{
current_transaction_id()
}
}
