% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{damerau_levenshtein}
\alias{damerau_levenshtein}
\title{DuckDB function damerau_levenshtein}
\usage{
damerau_levenshtein(s1 = VARCHAR, s2 = VARCHAR)
}
\arguments{
\item{s1}{\code{VARCHAR}}

\item{s2}{\code{VARCHAR}}
}
\value{
\code{BIGINT}
}
\description{
Extension of Levenshtein distance to also include transposition of adjacent characters as an allowed edit operation. In other words, the minimum number of edit operations (insertions, deletions, substitutions or transpositions) required to change one string to another. Characters of different cases (e.g., \code{a} and \code{A}) are considered different.
}
\examples{
\dontrun{
damerau_levenshtein('duckdb', 'udckbd')
}
}
