% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{date_diff}
\alias{date_diff}
\title{DuckDB function date_diff}
\arguments{
\item{part}{\code{VARCHAR}}

\item{startdate}{\verb{DATE | TIME | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{enddate}{\verb{DATE | TIME | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
The number of partition boundaries between the timestamps.
}
\examples{
\dontrun{
date_diff('hour', TIMESTAMPTZ '1992-09-30 23:59:59', TIMESTAMPTZ '1992-10-01 01:58:00')
}
}
