% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{date_part}
\alias{date_part}
\title{DuckDB function date_part}
\arguments{
\item{ts}{\code{VARCHAR[] | VARCHAR}}

\item{col1}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{STRUCT() | BIGINT}
}
\description{
Get subfield (equivalent to extract).
}
\examples{
\dontrun{
date_part('minute', TIMESTAMP '1992-09-20 20:38:40')
}
}
