% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{date_trunc}
\alias{date_trunc}
\title{DuckDB function date_trunc}
\arguments{
\item{part}{\code{VARCHAR}}

\item{timestamp}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\verb{TIMESTAMP | INTERVAL | TIMESTAMP WITH TIME ZONE}
}
\description{
Truncate to specified precision.
}
\section{Overloads}{

\itemize{
\item \code{date_trunc(part = VARCHAR, timestamp = DATE)}
\item \code{date_trunc(part = VARCHAR, timestamp = INTERVAL)}
\item \code{date_trunc(part = VARCHAR, timestamp = TIMESTAMP)}
\item \code{date_trunc(part = VARCHAR, timestamp = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
date_trunc('hour', TIMESTAMPTZ '1992-09-20 20:38:40')
}
}
