% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{datediff}
\alias{datediff}
\title{DuckDB function datediff}
\arguments{
\item{part}{\code{VARCHAR}}

\item{startdate}{\verb{DATE | TIME | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{enddate}{\verb{DATE | TIME | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
The number of partition boundaries between the timestamps.
}
\section{Overloads}{

\itemize{
\item \code{datediff(part = VARCHAR, startdate = DATE, enddate = DATE)}
\item \code{datediff(part = VARCHAR, startdate = TIME, enddate = TIME)}
\item \code{datediff(part = VARCHAR, startdate = TIMESTAMP, enddate = TIMESTAMP)}
\item \code{datediff(part = VARCHAR, startdate = `TIMESTAMP WITH TIME ZONE`, enddate = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
datediff('hour', TIMESTAMPTZ '1992-09-30 23:59:59', TIMESTAMPTZ '1992-10-01 01:58:00')
}
}
