% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{datesub}
\alias{datesub}
\title{DuckDB function datesub}
\arguments{
\item{part}{\code{VARCHAR}}

\item{startdate}{\verb{DATE | TIME | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{enddate}{\verb{DATE | TIME | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
The number of complete partitions between the timestamps.
}
\examples{
\dontrun{
datesub('hour', TIMESTAMPTZ '1992-09-30 23:59:59', TIMESTAMPTZ '1992-10-01 01:58:00')
}
}
