% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{day}
\alias{day}
\title{DuckDB function day}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the day component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{day(ts = DATE)}
\item \code{day(ts = INTERVAL)}
\item \code{day(ts = TIMESTAMP)}
\item \code{day(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
day(timestamp '2021-08-03 11:59:44.123456')
}
}
