% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{dayname}
\alias{dayname}
\title{DuckDB function dayname}
\arguments{
\item{ts}{\verb{DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
The (English) name of the weekday.
}
\examples{
\dontrun{
dayname(TIMESTAMP '1992-03-22')
}
}
