% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{dayofmonth}
\alias{dayofmonth}
\title{DuckDB function dayofmonth}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the dayofmonth component from a date or timestamp.
}
\examples{
\dontrun{
dayofmonth(timestamp '2021-08-03 11:59:44.123456')
}
}
