% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{dayofmonth}
\alias{dayofmonth}
\title{DuckDB function dayofmonth}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the dayofmonth component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{dayofmonth(ts = DATE)}
\item \code{dayofmonth(ts = INTERVAL)}
\item \code{dayofmonth(ts = TIMESTAMP)}
\item \code{dayofmonth(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{dayofmonth(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

