% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd-package.R
\docType{package}
\name{dd-package}
\alias{dd-package}
\title{dd: DuckDB Functions}
\description{
Lists DuckDB functions for integration in R's help system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/dd}
  \item \url{https://cynkra.github.io/dd}
  \item Report bugs at \url{https://github.com/cynkra/dd/issues}
}

}
\keyword{internal}
