% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd-package.R
\docType{package}
\name{dd-package}
\alias{dd}
\alias{dd-package}
\title{dd: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/kirill}
  \item \url{https://krlmlr.github.io/kirill}
  \item Report bugs at \url{https://github.com/krlmlr/kirill/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
