% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd-package.R
\docType{package}
\name{dd-package}
\alias{dd-package}
\title{dd: Functions Provided by DuckDB}
\description{
Lists DuckDB functions for integration in R's help system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/dd}
  \item \url{https://cynkra.github.io/dd}
  \item Report bugs at \url{https://github.com/cynkra/dd/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Other contributors:
\itemize{
  \item cynkra GmbH \email{mail@cynkra.com} (\href{https://ror.org/0335t7e62}{ROR}) [funder, copyright holder]
}

}
\keyword{internal}
