% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-dd.R
\docType{data}
\name{dd}
\alias{dd}
\title{DuckDB functions}
\format{
An object of class \code{list} of length 713.
}
\usage{
dd
}
\description{
A list of known DuckDB functions.
}
\examples{
dd[1:3]
}
\keyword{datasets}
