% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{decode}
\alias{decode}
\alias{duckdb_fun_decode}
\title{DuckDB function decode}
\arguments{
\item{blob}{\code{BLOB}}
}
\description{
Convert blob to varchar. Fails if blob is not valid utf-8.
}
\examples{
\dontrun{
decode('\xC3\xBC'::BLOB)
}
}
