% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{decode}
\alias{decode}
\title{DuckDB function decode}
\arguments{
\item{blob}{\code{BLOB}}
}
\description{
Converts \code{blob} to \code{VARCHAR}. Fails if \code{blob} is not valid UTF-8.
}
\examples{
\dontrun{
decode('\xC3\xBC'::BLOB)
}
}
