% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{degrees}
\alias{degrees}
\alias{duckdb_fun_degrees}
\title{DuckDB function degrees}
\arguments{
\item{x}{\code{DOUBLE}}
}
\description{
Converts radians to degrees.
}
\examples{
\dontrun{
degrees(pi())
}
}
