% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enable_logging}
\alias{enable_logging}
\title{DuckDB function enable_logging}
\usage{
enable_logging(level = VARCHAR, storage = VARCHAR, storage_config = ANY, storage_path = VARCHAR, storage_normalize = BOOLEAN, storage_buffer_size = UBIGINT)
}
\arguments{
\item{level}{\code{VARCHAR}}

\item{storage}{\code{VARCHAR}}

\item{storage_config}{\code{ANY}}

\item{storage_path}{\code{VARCHAR}}

\item{storage_normalize}{\code{BOOLEAN}}

\item{storage_buffer_size}{\code{UBIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{enable_logging()}.
}
\examples{
\dontrun{
character(0)
}
}
