% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{encode}
\alias{encode}
\title{DuckDB function encode}
\usage{
encode(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{BLOB}
}
\description{
Converts the \code{string} to \code{BLOB}. Converts UTF-8 characters into literal encoding.
}
\examples{
\dontrun{
encode('my_string_with_ü')
}
}
