% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ends_with}
\alias{ends_with}
\title{DuckDB function ends_with}
\usage{
ends_with(string = VARCHAR, search_string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if \code{string} ends with \code{search_string}.
}
\examples{
\dontrun{
ends_with('abc', 'bc')
}
}
