% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enum_code}
\alias{enum_code}
\title{DuckDB function enum_code}
\usage{
enum_code(enum = ANY)
}
\arguments{
\item{enum}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
Returns the numeric value backing the given enum value.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{enum_code('happy'::mood)
}\if{html}{\out{</div>}}
}

