% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enum_first}
\alias{enum_first}
\title{DuckDB function enum_first}
\usage{
enum_first(enum = ANY)
}
\arguments{
\item{enum}{\code{ANY}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the first value of the input enum type.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{enum_first(NULL::mood)
}\if{html}{\out{</div>}}
}

