% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{enum_range}
\alias{enum_range}
\alias{duckdb_fun_enum_range}
\title{DuckDB function enum_range}
\arguments{
\item{enum}{\code{ANY}}
}
\description{
Returns all values of the input enum type as an array.
}
\examples{
\dontrun{
enum_range(NULL::mood)
}
}
