% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{epoch}
\alias{epoch}
\title{DuckDB function epoch}
\arguments{
\item{temporal}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{DOUBLE}
}
\description{
Extract the epoch component from a temporal type.
}
\section{Overloads}{

\itemize{
\item \code{epoch(temporal = DATE)}
\item \code{epoch(temporal = INTERVAL)}
\item \code{epoch(temporal = TIME)}
\item \code{epoch(temporal = TIMESTAMP)}
\item \code{epoch(temporal = `TIME WITH TIME ZONE`)}
\item \code{epoch(temporal = TIME_NS)}
\item \code{epoch(temporal = `TIMESTAMP WITH TIME ZONE`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{epoch(timestamp '2021-08-03 11:59:44.123456')
}\if{html}{\out{</div>}}
}

