% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{epoch}
\alias{epoch}
\title{DuckDB function epoch}
\arguments{
\item{temporal}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the epoch component from a temporal type.
}
\examples{
\dontrun{
epoch(timestamp '2021-08-03 11:59:44.123456')
}
}
