% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{epoch_ms}
\alias{epoch_ms}
\title{DuckDB function epoch_ms}
\arguments{
\item{temporal}{\verb{DATE | TIMESTAMP | INTERVAL | TIME | TIME_NS | TIME WITH TIME ZONE | TIMESTAMP WITH TIME ZONE | BIGINT}}
}
\description{
Extract the epoch component in milliseconds from a temporal type.
}
\examples{
\dontrun{
epoch_ms(timestamp '2021-08-03 11:59:44.123456')
}
}
