% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{epoch_us}
\alias{epoch_us}
\title{DuckDB function epoch_us}
\arguments{
\item{temporal}{\verb{DATE | TIMESTAMP | INTERVAL | TIME | TIME_NS | TIME WITH TIME ZONE | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the epoch component in microseconds from a temporal type.
}
\examples{
\dontrun{
epoch_us(timestamp '2021-08-03 11:59:44.123456')
}
}
