% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{equi_width_bins}
\alias{equi_width_bins}
\title{DuckDB function equi_width_bins}
\arguments{
\item{min}{\code{BIGINT | DOUBLE | TIMESTAMP | ANY}}

\item{max}{\code{BIGINT | DOUBLE | TIMESTAMP | ANY}}

\item{bin_count}{\code{BIGINT}}

\item{nice_rounding}{\code{BOOLEAN}}
}
\value{
\code{ANY[]}
}
\description{
Generates bin_count equi-width bins between the min and max. If enabled nice_rounding makes the numbers more readable/less jagged.
}
\examples{
\dontrun{
equi_width_bins(0, 10, 2, true)
}
}
