% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{era}
\alias{era}
\title{DuckDB function era}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the era component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{era(ts = DATE)}
\item \code{era(ts = INTERVAL)}
\item \code{era(ts = TIMESTAMP)}
\item \code{era(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
era(timestamp '2021-08-03 11:59:44.123456')
}
}
