% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{error}
\alias{error}
\title{DuckDB function error}
\usage{
error(message = VARCHAR)
}
\arguments{
\item{message}{\code{VARCHAR}}
}
\value{
\code{"NULL"}
}
\description{
Throws the given error message.
}
\examples{
\dontrun{
error('access_mode')
}
}
