% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{factorial}
\alias{factorial}
\title{DuckDB function factorial}
\usage{
factorial(x = INTEGER)
}
\arguments{
\item{x}{\code{INTEGER}}
}
\description{
Factorial of x. Computes the product of the current integer and all integers below it.
}
\examples{
\dontrun{
4!
}
}
