% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{favg}
\alias{favg}
\title{DuckDB function favg}
\usage{
favg(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Calculates the average using a more accurate floating point summation (Kahan Sum).
}
\examples{
\dontrun{
favg(A)
}
}
