% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{first}
\alias{first}
\alias{arbitrary}
\title{DuckDB function first}
\arguments{
\item{arg}{\code{DECIMAL | ANY}}

\item{col0}{\code{T}}
}
\value{
\code{DECIMAL | ANY | T}
}
\description{
Returns the first value (NULL or non-NULL) from arg. This function is affected by ordering.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{first(A)
}\if{html}{\out{</div>}}
}

