% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{first}
\alias{first}
\title{DuckDB function first}
\arguments{
\item{arg}{\code{DECIMAL | ANY}}
}
\value{
\code{DECIMAL | ANY}
}
\description{
Returns the first value (NULL or non-NULL) from arg. This function is affected by ordering.
}
\section{Overloads}{

\itemize{
\item \code{first(arg = DECIMAL)}
\item \code{first(arg = ANY)}
}
}

\examples{
\dontrun{
first(A)
}
}
