% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{flatten}
\alias{flatten}
\alias{duckdb_fun_flatten}
\title{DuckDB function flatten}
\arguments{
\item{nested_list}{\code{ANY[][]}}
}
\description{
Flatten a nested list by one level.
}
\examples{
\dontrun{
flatten([[1, 2, 3], [4, 5]])
}
}
