% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{flatten}
\alias{flatten}
\title{DuckDB function flatten}
\usage{
flatten(nested_list = `T[][]`)
}
\arguments{
\item{nested_list}{\code{T[][]}}
}
\description{
Flattens a nested list by one level.
}
\examples{
\dontrun{
flatten([[1, 2, 3], [4, 5]])
}
}
