% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{floor}
\alias{floor}
\title{DuckDB function floor}
\arguments{
\item{x}{\code{FLOAT | DOUBLE | DECIMAL}}
}
\value{
\code{FLOAT | DOUBLE | DECIMAL}
}
\description{
Rounds the number down.
}
\section{Overloads}{

\itemize{
\item \code{floor(x = FLOAT)}
\item \code{floor(x = DOUBLE)}
\item \code{floor(x = DECIMAL)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{floor(17.4)
}\if{html}{\out{</div>}}
}

