% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{format}
\alias{format}
\title{DuckDB function format}
\usage{
format(format = VARCHAR)
}
\arguments{
\item{format}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Formats a string using the fmt syntax.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{format('Benchmark "\{\}" took \{\} seconds', 'CSV', 42)
}\if{html}{\out{</div>}}
}

