% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{format_bytes}
\alias{format_bytes}
\title{DuckDB function format_bytes}
\usage{
format_bytes(integer = BIGINT)
}
\arguments{
\item{integer}{\code{BIGINT}}
}
\description{
Converts \code{integer} to a human-readable representation using units based on powers of 2 (KiB, MiB, GiB, etc.).
}
\examples{
\dontrun{
format_bytes(16_000)
}
}
