% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{format_bytes}
\alias{format_bytes}
\title{DuckDB function format_bytes}
\arguments{
\item{bytes}{\code{BIGINT}}
}
\description{
Converts bytes to a human-readable presentation (e.g. 16000 -> 15.6 KiB).
}
\examples{
\dontrun{
format_bytes(1000 * 16)
}
}
