% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{formatReadableDecimalSize}
\alias{formatReadableDecimalSize}
\title{DuckDB function formatReadableDecimalSize}
\usage{
formatReadableDecimalSize(integer = BIGINT)
}
\arguments{
\item{integer}{\code{BIGINT}}
}
\value{
\code{VARCHAR}
}
\description{
Converts \code{integer} to a human-readable representation using units based on powers of 10 (KB, MB, GB, etc.).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{formatReadableDecimalSize(16_000)
}\if{html}{\out{</div>}}
}

