% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{formatReadableDecimalSize}
\alias{formatReadableDecimalSize}
\title{DuckDB function formatReadableDecimalSize}
\arguments{
\item{bytes}{\code{BIGINT}}
}
\description{
Converts bytes to a human-readable presentation (e.g. 16000 -> 16.0 KB).
}
\examples{
\dontrun{
format_bytes(1000 * 16)
}
}
