% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{formatReadableDecimalSize}
\alias{formatReadableDecimalSize}
\title{DuckDB function formatReadableDecimalSize}
\usage{
formatReadableDecimalSize(integer = BIGINT)
}
\arguments{
\item{integer}{\code{BIGINT}}
}
\description{
Converts \code{integer} to a human-readable representation using units based on powers of 10 (KB, MB, GB, etc.).
}
\examples{
\dontrun{
formatReadableDecimalSize(16_000)
}
}
