% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{formatReadableSize}
\alias{formatReadableSize}
\title{DuckDB function formatReadableSize}
\arguments{
\item{integer}{\code{BIGINT}}
}
\description{
Converts \code{integer} to a human-readable representation using units based on powers of 2 (KiB, MiB, GiB, etc.).
}
\examples{
\dontrun{
formatReadableSize(16_000)
}
}
