% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{formatReadableSize}
\alias{formatReadableSize}
\title{DuckDB function formatReadableSize}
\arguments{
\item{bytes}{\code{BIGINT}}
}
\description{
Converts bytes to a human-readable presentation (e.g., 16000 -> 15.6 KiB).
}
\examples{
\dontrun{
format_bytes(1000 * 16)
}
}
