% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{from_base64}
\alias{from_base64}
\alias{duckdb_fun_from_base64}
\title{DuckDB function from_base64}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Convert a base64 encoded string to a character string.
}
\examples{
\dontrun{
from_base64('QQ==')
}
}
