% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{from_base64}
\alias{from_base64}
\title{DuckDB function from_base64}
\usage{
from_base64(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{BLOB}
}
\description{
Converts a base64 encoded \code{string} to a character string (\code{BLOB}).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{from_base64('QQ==')
}\if{html}{\out{</div>}}
}

