% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{from_binary}
\alias{from_binary}
\title{DuckDB function from_binary}
\usage{
from_binary(value = VARCHAR)
}
\arguments{
\item{value}{\code{VARCHAR}}
}
\value{
\code{BLOB}
}
\description{
Converts a \code{value} from binary representation to a blob.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{from_binary('0110')
}\if{html}{\out{</div>}}
}

