% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{from_binary}
\alias{from_binary}
\title{DuckDB function from_binary}
\arguments{
\item{value}{\code{VARCHAR}}
}
\description{
Converts a \code{value} from binary representation to a blob.
}
\examples{
\dontrun{
from_binary('0110')
}
}
