% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{from_hex}
\alias{from_hex}
\title{DuckDB function from_hex}
\usage{
from_hex(value = VARCHAR)
}
\arguments{
\item{value}{\code{VARCHAR}}
}
\value{
\code{BLOB}
}
\description{
Converts a \code{value} from hexadecimal representation to a blob.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{from_hex('2A')
}\if{html}{\out{</div>}}
}

