% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{generate_series}
\alias{generate_series}
\title{DuckDB function generate_series}
\arguments{
\item{col0}{\code{BIGINT | TIMESTAMP}}

\item{col1}{\code{BIGINT | TIMESTAMP}}

\item{col2}{\code{BIGINT | INTERVAL}}

\item{start}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{stop}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{step}{\code{BIGINT | INTERVAL}}
}
\value{
\verb{BIGINT[] | TIMESTAMP[] | TIMESTAMP WITH TIME ZONE[]}
}
\description{
Creates a list of values between \code{start} and \code{stop} - the stop parameter is inclusive.
}
\section{Overloads}{

\itemize{
\item \code{generate_series(col0 = BIGINT)}
\item \code{generate_series(col0 = BIGINT, col1 = BIGINT)}
\item \code{generate_series(col0 = BIGINT, col1 = BIGINT, col2 = BIGINT)}
\item \code{generate_series(col0 = TIMESTAMP, col1 = TIMESTAMP, col2 = INTERVAL)}
\item \code{generate_series(start = BIGINT)}
\item \code{generate_series(start = BIGINT, stop = BIGINT)}
\item \code{generate_series(start = BIGINT, stop = BIGINT, step = BIGINT)}
\item \code{generate_series(start = TIMESTAMP, stop = TIMESTAMP, step = INTERVAL)}
\item \code{generate_series(start = `TIMESTAMP WITH TIME ZONE`, stop = `TIMESTAMP WITH TIME ZONE`, step = INTERVAL)}
}
}

\examples{
\dontrun{
generate_series(2, 5, 3)
}
}
