% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{generate_series}
\alias{generate_series}
\title{DuckDB function generate_series}
\arguments{
\item{start}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{stop}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{step}{\code{BIGINT | INTERVAL}}
}
\description{
Creates a list of values between \code{start} and \code{stop} - the stop parameter is inclusive.
}
\examples{
\dontrun{
generate_series(2, 5, 3)
}
}
