% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{get_bit}
\alias{get_bit}
\alias{duckdb_fun_get_bit}
\title{DuckDB function get_bit}
\arguments{
\item{bitstring}{\code{BIT}}

\item{index}{\code{INTEGER}}
}
\description{
Extracts the nth bit from bitstring; the first (leftmost) bit is indexed 0.
}
\examples{
\dontrun{
get_bit('0110010'::BIT, 2)
}
}
