% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{->>}
\alias{->>}
\alias{>>}
\title{DuckDB function ->>}
\usage{
`->>`(col0, col1)

`>>`(input, col1)
}
\arguments{
\item{col0}{\code{VARCHAR | JSON}}

\item{col1}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}

\item{input}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{VARCHAR | VARCHAR[]}

\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
DuckDB function \verb{->>()}.

Bitwise shift right.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{8 >> 2
}\if{html}{\out{</div>}}
}

