% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{>>}
\alias{>>}
\title{DuckDB function >>}
\arguments{
\item{input}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}

\item{col1}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise shift right.
}
\examples{
\dontrun{
8 >> 2
}
}
