% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{>>}
\alias{>>}
\title{DuckDB function >>}
\arguments{
\item{input}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}

\item{col1}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise shift right.
}
\section{Overloads}{

\itemize{
\item ```>>`(input = TINYINT, col1 = TINYINT)``
\item ```>>`(input = SMALLINT, col1 = SMALLINT)``
\item ```>>`(input = INTEGER, col1 = INTEGER)``
\item ```>>`(input = BIGINT, col1 = BIGINT)``
\item ```>>`(input = HUGEINT, col1 = HUGEINT)``
\item ```>>`(input = UTINYINT, col1 = UTINYINT)``
\item ```>>`(input = USMALLINT, col1 = USMALLINT)``
\item ```>>`(input = UINTEGER, col1 = UINTEGER)``
\item ```>>`(input = UBIGINT, col1 = UBIGINT)``
\item ```>>`(input = UHUGEINT, col1 = UHUGEINT)``
\item ```>>`(input = BIT, col1 = INTEGER)``
}
}

\examples{
\dontrun{
8 >> 2
}
}
