% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{greatest}
\alias{greatest}
\title{DuckDB function greatest}
\usage{
greatest(arg1 = ANY)
}
\arguments{
\item{arg1}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
Returns the largest value. For strings lexicographical ordering is used. Note that lowercase characters are considered “larger” than uppercase characters and collations are not supported.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{greatest(42, 84)
greatest('abc', 'bcd', 'cde', 'EFG')
}\if{html}{\out{</div>}}
}

