% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{group_concat}
\alias{group_concat}
\title{DuckDB function group_concat}
\arguments{
\item{str}{\code{ANY}}

\item{arg}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Concatenates the column string values with an optional separator.
}
\examples{
\dontrun{
group_concat(A, '-')
}
}
