% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{hamming}
\alias{hamming}
\alias{duckdb_fun_hamming}
\title{DuckDB function hamming}
\arguments{
\item{str1}{\code{VARCHAR}}

\item{str2}{\code{VARCHAR}}
}
\description{
The number of positions with different characters for 2 strings of equal length. Different case is considered different.
}
\examples{
\dontrun{
hamming('duck','luck')
}
}
