% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{hamming}
\alias{hamming}
\title{DuckDB function hamming}
\usage{
hamming(s1 = VARCHAR, s2 = VARCHAR)
}
\arguments{
\item{s1}{\code{VARCHAR}}

\item{s2}{\code{VARCHAR}}
}
\description{
The Hamming distance between to strings, i.e., the number of positions with different characters for two strings of equal length. Strings must be of equal length. Characters of different cases (e.g., \code{a} and \code{A}) are considered different.
}
\examples{
\dontrun{
hamming('duck', 'luck')
}
}
