% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_function_privilege}
\alias{has_function_privilege}
\title{DuckDB function has_function_privilege}
\arguments{
\item{function}{Unspecified.}

\item{privilege}{Unspecified.}

\item{user}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_function_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_function_privilege(`function`, privilege)}
\item \code{has_function_privilege(user, `function`, privilege)}
}
}

