% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_language_privilege}
\alias{has_language_privilege}
\title{DuckDB function has_language_privilege}
\arguments{
\item{language}{Unspecified.}

\item{privilege}{Unspecified.}

\item{user}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_language_privilege()}.
}
