% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{has_sequence_privilege}
\alias{has_sequence_privilege}
\title{DuckDB function has_sequence_privilege}
\arguments{
\item{sequence}{Unspecified.}

\item{privilege}{Unspecified.}

\item{user}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{has_sequence_privilege()}.
}
\section{Overloads}{

\itemize{
\item \code{has_sequence_privilege(sequence, privilege)}
\item \code{has_sequence_privilege(user, sequence, privilege)}
}
}

