% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{hash}
\alias{hash}
\title{DuckDB function hash}
\usage{
hash(value = ANY)
}
\arguments{
\item{value}{\code{ANY}}
}
\value{
\code{UBIGINT}
}
\description{
Returns a \code{UBIGINT} with the hash of the \code{value}. Note that this is not a cryptographic hash.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{hash('🦆')
}\if{html}{\out{</div>}}
}

