% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{hex}
\alias{hex}
\title{DuckDB function hex}
\arguments{
\item{string}{\code{VARCHAR}}

\item{value}{\code{BIGNUM | BIGINT | UBIGINT | HUGEINT | UHUGEINT}}

\item{blob}{\code{BLOB}}
}
\value{
\code{VARCHAR}
}
\description{
Converts the \code{string} to hexadecimal representation.

Converts the \code{value} to \code{VARCHAR} using hexadecimal representation.

Converts \code{blob} to \code{VARCHAR} using hexadecimal encoding.
}
\examples{
\dontrun{
hex('Hello')
hex(42)
hex('\xAA\xBB'::BLOB)
}
}
