% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{histogram}
\alias{histogram}
\title{DuckDB function histogram}
\arguments{
\item{arg}{\code{ANY}}

\item{col1}{\code{ANY[]}}
}
\value{
\code{MAP}
}
\description{
Returns a LIST of STRUCTs with the fields bucket and count.
}
\examples{
\dontrun{
histogram(A)
}
}
