% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ilike_escape}
\alias{ilike_escape}
\alias{duckdb_fun_ilike_escape}
\title{DuckDB function ilike_escape}
\arguments{
\item{string}{\code{VARCHAR}}

\item{like_specifier}{\code{VARCHAR}}

\item{escape_character}{\code{VARCHAR}}
}
\description{
Returns true if the string matches the like_specifier (see Pattern Matching) using case-insensitive matching. escape_character is used to search for wildcard characters in the string.
}
\examples{
\dontrun{
ilike_escape('A\%c', 'a$\%C', '$')
}
}
