% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{is_histogram_other_bin}
\alias{is_histogram_other_bin}
\title{DuckDB function is_histogram_other_bin}
\usage{
is_histogram_other_bin(val = ANY)
}
\arguments{
\item{val}{\code{ANY}}
}
\description{
Whether or not the provided value is the histogram "other" bin (used for values not belonging to any provided bin).
}
\examples{
\dontrun{
is_histogram_other_bin(v)
}
}
