% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{isfinite}
\alias{isfinite}
\title{DuckDB function isfinite}
\arguments{
\item{x}{\verb{FLOAT | DOUBLE | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the floating point value is finite, false otherwise.
}
\section{Overloads}{

\itemize{
\item \code{isfinite(x = FLOAT)}
\item \code{isfinite(x = DOUBLE)}
\item \code{isfinite(x = DATE)}
\item \code{isfinite(x = TIMESTAMP)}
\item \code{isfinite(x = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
isfinite(5.5)
}
}
