% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{isinf}
\alias{isinf}
\title{DuckDB function isinf}
\arguments{
\item{x}{\verb{FLOAT | DOUBLE | DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the floating point value is infinite, false otherwise.
}
\section{Overloads}{

\itemize{
\item \code{isinf(x = FLOAT)}
\item \code{isinf(x = DOUBLE)}
\item \code{isinf(x = DATE)}
\item \code{isinf(x = TIMESTAMP)}
\item \code{isinf(x = `TIMESTAMP WITH TIME ZONE`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{isinf('Infinity'::float)
}\if{html}{\out{</div>}}
}

