% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{isnan}
\alias{isnan}
\title{DuckDB function isnan}
\arguments{
\item{x}{\code{FLOAT | DOUBLE}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the floating point value is not a number, false otherwise.
}
\examples{
\dontrun{
isnan('NaN'::FLOAT)
}
}
