% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{jaccard}
\alias{jaccard}
\title{DuckDB function jaccard}
\usage{
jaccard(s1 = VARCHAR, s2 = VARCHAR)
}
\arguments{
\item{s1}{\code{VARCHAR}}

\item{s2}{\code{VARCHAR}}
}
\description{
The Jaccard similarity between two strings. Characters of different cases (e.g., \code{a} and \code{A}) are considered different. Returns a number between 0 and 1.
}
\examples{
\dontrun{
jaccard('duck', 'luck')
}
}
