% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{jaccard}
\alias{jaccard}
\alias{duckdb_fun_jaccard}
\title{DuckDB function jaccard}
\arguments{
\item{str1}{\code{VARCHAR}}

\item{str2}{\code{VARCHAR}}
}
\description{
The Jaccard similarity between two strings. Different case is considered different. Returns a number between 0 and 1.
}
\examples{
\dontrun{
jaccard('duck','luck')
}
}
