% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{jaro_winkler_similarity}
\alias{jaro_winkler_similarity}
\title{DuckDB function jaro_winkler_similarity}
\arguments{
\item{s1}{\code{VARCHAR}}

\item{s2}{\code{VARCHAR}}

\item{score_cutoff}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
The Jaro-Winkler similarity between two strings. Characters of different cases (e.g., \code{a} and \code{A}) are considered different. Returns a number between 0 and 1. For similarity < \code{score_cutoff}, 0 is returned instead. \code{score_cutoff} defaults to 0.
}
\section{Overloads}{

\itemize{
\item \code{jaro_winkler_similarity(s1 = VARCHAR, s2 = VARCHAR)}
\item \code{jaro_winkler_similarity(s1 = VARCHAR, s2 = VARCHAR, score_cutoff = DOUBLE)}
}
}

\examples{
\dontrun{
jaro_winkler_similarity('duck', 'duckdb')
}
}
