% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{julian}
\alias{julian}
\title{DuckDB function julian}
\arguments{
\item{ts}{\verb{DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{DOUBLE}
}
\description{
Extract the Julian Day number from a date or timestamp.
}
\examples{
\dontrun{
julian(timestamp '2006-01-01 12:00')
}
}
