% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{last}
\alias{last}
\title{DuckDB function last}
\arguments{
\item{arg}{\code{DECIMAL | ANY}}

\item{col0}{\code{T}}
}
\value{
\code{DECIMAL | ANY | T}
}
\description{
Returns the last value of a column. This function is affected by ordering.
}
\section{Overloads}{

\itemize{
\item \code{last(arg = DECIMAL)}: Returns the last value of a column. This function is affected by ordering.
\item \code{last(arg = ANY)}: Returns the last value of a column. This function is affected by ordering.
\item \code{last(col0 = T)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{last(A)
}\if{html}{\out{</div>}}
}

