% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{last}
\alias{last}
\title{DuckDB function last}
\arguments{
\item{arg}{\code{DECIMAL | ANY}}
}
\value{
\code{DECIMAL | ANY}
}
\description{
Returns the last value of a column. This function is affected by ordering.
}
\section{Overloads}{

\itemize{
\item \code{last(arg = DECIMAL)}
\item \code{last(arg = ANY)}
}
}

\examples{
\dontrun{
last(A)
}
}
