% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lcm}
\alias{lcm}
\title{DuckDB function lcm}
\arguments{
\item{x}{\code{BIGINT | HUGEINT}}

\item{y}{\code{BIGINT | HUGEINT}}
}
\value{
\code{BIGINT | HUGEINT}
}
\description{
Computes the least common multiple of x and y.
}
\examples{
\dontrun{
lcm(42, 57)
}
}
