% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{least}
\alias{least}
\title{DuckDB function least}
\usage{
least(arg1 = ANY)
}
\arguments{
\item{arg1}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
Returns the smallest value. For strings lexicographical ordering is used. Note that uppercase characters are considered “smaller” than lowercase characters, and collations are not supported.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{least(42, 84)
least('abc', 'bcd', 'cde', 'EFG')
}\if{html}{\out{</div>}}
}

