% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{least_common_multiple}
\alias{least_common_multiple}
\title{DuckDB function least_common_multiple}
\arguments{
\item{x}{\code{BIGINT | HUGEINT}}

\item{y}{\code{BIGINT | HUGEINT}}
}
\value{
\code{BIGINT | HUGEINT}
}
\description{
Computes the least common multiple of x and y.
}
\examples{
\dontrun{
least_common_multiple(42, 57)
}
}
