% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{left}
\alias{left}
\title{DuckDB function left}
\usage{
left(string = VARCHAR, count = BIGINT)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{count}{\code{BIGINT}}
}
\value{
\code{VARCHAR}
}
\description{
Extracts the left-most count characters.
}
\examples{
\dontrun{
left('Hello🦆', 2)
}
}
