% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{len}
\alias{len}
\title{DuckDB function len}
\arguments{
\item{string}{\code{VARCHAR}}

\item{bit}{\code{BIT}}

\item{list}{\code{ANY[]}}
}
\value{
\code{BIGINT}
}
\description{
Number of characters in \code{string}.

Returns the bit-length of the \code{bit} argument.

Returns the length of the \code{list}.
}
\section{Overloads}{

\itemize{
\item \code{len(string = VARCHAR)}
\item \code{len(bit = BIT)}
\item \code{len(list = `ANY[]`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{length('Hello🦆')
length(42::TINYINT::BIT)
length([1,2,3])
}\if{html}{\out{</div>}}
}

