% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{levenshtein}
\alias{levenshtein}
\alias{duckdb_fun_levenshtein}
\title{DuckDB function levenshtein}
\arguments{
\item{str1}{\code{VARCHAR}}

\item{str2}{\code{VARCHAR}}
}
\description{
The minimum number of single-character edits (insertions, deletions or substitutions) required to change one string to the other. Different case is considered different.
}
\examples{
\dontrun{
levenshtein('duck','db')
}
}
