% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{levenshtein}
\alias{levenshtein}
\title{DuckDB function levenshtein}
\usage{
levenshtein(s1 = VARCHAR, s2 = VARCHAR)
}
\arguments{
\item{s1}{\code{VARCHAR}}

\item{s2}{\code{VARCHAR}}
}
\value{
\code{BIGINT}
}
\description{
The minimum number of single-character edits (insertions, deletions or substitutions) required to change one string to the other. Characters of different cases (e.g., \code{a} and \code{A}) are considered different.
}
\examples{
\dontrun{
levenshtein('duck', 'db')
}
}
