% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lgamma}
\alias{lgamma}
\alias{duckdb_fun_lgamma}
\title{DuckDB function lgamma}
\arguments{
\item{x}{\code{DOUBLE}}
}
\description{
Computes the log of the gamma function.
}
\examples{
\dontrun{
lgamma(2)
}
}
