% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{like_escape}
\alias{like_escape}
\title{DuckDB function like_escape}
\arguments{
\item{string}{\code{VARCHAR}}

\item{like_specifier}{\code{VARCHAR}}

\item{escape_character}{\code{VARCHAR}}
}
\description{
Returns true if the string matches the like_specifier (see Pattern Matching) using case-sensitive matching. escape_character is used to search for wildcard characters in the string.
}
\examples{
\dontrun{
like_escape('a\%c', 'a$\%c', '$')
}
}
