% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{like_escape}
\alias{like_escape}
\title{DuckDB function like_escape}
\usage{
like_escape(string = VARCHAR, like_specifier = VARCHAR, escape_character = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{like_specifier}{\code{VARCHAR}}

\item{escape_character}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if the \code{string} matches the \code{like_specifier} (see Pattern Matching) using case-sensitive matching. \code{escape_character} is used to search for wildcard characters in the \code{string}.
}
\examples{
\dontrun{
like_escape('a\%c', 'a$\%c', '$')
}
}
