% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_cat}
\alias{list_cat}
\alias{duckdb_fun_list_cat}
\title{DuckDB function list_cat}
\arguments{
\item{list1}{\code{ANY[]}}

\item{list2}{\code{ANY[]}}
}
\description{
Concatenates two lists.
}
\examples{
\dontrun{
list_concat([2, 3], [4, 5, 6])
}
}
