% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_concat}
\alias{list_concat}
\title{DuckDB function list_concat}
\usage{
list_concat()
}
\value{
\code{ANY[]}
}
\description{
Concatenates lists. \code{NULL} inputs are skipped. See also operator \code{||}.
}
\examples{
\dontrun{
list_concat([2, 3], [4, 5, 6], [7])
}
}
