% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_contains}
\alias{list_contains}
\title{DuckDB function list_contains}
\arguments{
\item{list}{\code{T[]}}

\item{element}{\code{T}}
}
\description{
Returns true if the list contains the element.
}
\examples{
\dontrun{
list_contains([1, 2, NULL], 1)
}
}
