% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_distance}
\alias{list_distance}
\title{DuckDB function list_distance}
\arguments{
\item{list1}{\code{FLOAT[] | DOUBLE[]}}

\item{list2}{\code{FLOAT[] | DOUBLE[]}}
}
\description{
Calculates the Euclidean distance between two points with coordinates given in two inputs lists of equal length.
}
\examples{
\dontrun{
list_distance([1, 2, 3], [1, 2, 5])
}
}
