% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_distinct}
\alias{list_distinct}
\title{DuckDB function list_distinct}
\usage{
list_distinct(list = `T[]`)
}
\arguments{
\item{list}{\code{T[]}}
}
\description{
Removes all duplicates and \code{NULL} values from a list. Does not preserve the original order.
}
\examples{
\dontrun{
list_distinct([1, 1, NULL, -3, 1, 5])
}
}
