% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_has}
\alias{list_has}
\title{DuckDB function list_has}
\usage{
list_has(list = `T[]`, element = T)
}
\arguments{
\item{list}{\code{T[]}}

\item{element}{\code{T}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if the list contains the element.
}
\examples{
\dontrun{
list_has([1, 2, NULL], 1)
}
}
