% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_has}
\alias{list_has}
\alias{duckdb_fun_list_has}
\title{DuckDB function list_has}
\arguments{
\item{list}{\code{ANY[]}}

\item{element}{\code{ANY}}
}
\description{
Returns true if the list contains the element.
}
\examples{
\dontrun{
list_contains([1, 2, NULL], 1)
}
}
