% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_has_all}
\alias{list_has_all}
\title{DuckDB function list_has_all}
\usage{
list_has_all(list1 = `T[]`, list2 = `T[]`)
}
\arguments{
\item{list1}{\code{T[]}}

\item{list2}{\code{T[]}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns true if all elements of list2 are in list1. NULLs are ignored.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list_has_all([1, 2, 3], [2, 3])
}\if{html}{\out{</div>}}
}

