% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_intersect}
\alias{list_intersect}
\title{DuckDB function list_intersect}
\usage{
list_intersect(l1, l2)
}
\arguments{
\item{l1}{Unspecified.}

\item{l2}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB macro \code{list_intersect()}.
}
