% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_reduce}
\alias{list_reduce}
\title{DuckDB function list_reduce}
\arguments{
\item{list}{\code{ANY[]}}

\item{initial_value}{\code{ANY}}

\item{lambda(x, y)}{\code{LAMBDA}}
}
\description{
Reduces all elements of the input \code{list} into a single scalar value by executing the \code{lambda} function on a running result and the next list element. The \code{lambda} function has an optional \code{initial_value} argument.
}
\examples{
\dontrun{
list_reduce([1, 2, 3], lambda x, y : x + y)
}
}
