% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_reverse_sort}
\alias{list_reverse_sort}
\title{DuckDB function list_reverse_sort}
\arguments{
\item{list}{\code{ANY[]}}

\item{col1}{\code{VARCHAR}}
}
\description{
Sorts the elements of the list in reverse order.
}
\examples{
\dontrun{
list_reverse_sort([3, 6, 1, 2])
}
}
