% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_select}
\alias{list_select}
\title{DuckDB function list_select}
\usage{
list_select(value_list = `T[]`, index_list = `BIGINT[]`)
}
\arguments{
\item{value_list}{\code{T[]}}

\item{index_list}{\code{BIGINT[]}}
}
\description{
Returns a list based on the elements selected by the \code{index_list}.
}
\examples{
\dontrun{
list_select([10, 20, 30, 40], [1, 4])
}
}
