% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_slice}
\alias{list_slice}
\title{DuckDB function list_slice}
\arguments{
\item{list}{\code{ANY}}

\item{begin}{\code{ANY}}

\item{end}{\code{ANY}}

\item{step}{\code{BIGINT}}
}
\value{
\code{ANY}
}
\description{
Extracts a sublist or substring using slice conventions. Negative values are accepted.

list_slice with added step feature.
}
\examples{
\dontrun{
list_slice([4, 5, 6], 2, 3)
list_slice([4, 5, 6], 1, 3, 2)
}
}
