% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_sort}
\alias{list_sort}
\title{DuckDB function list_sort}
\arguments{
\item{list}{\code{ANY[]}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR}}
}
\value{
\code{ANY[]}
}
\description{
Sorts the elements of the list.
}
\section{Overloads}{

\itemize{
\item \code{list_sort(list = `ANY[]`)}
\item \code{list_sort(list = `ANY[]`, col1 = VARCHAR)}
\item \code{list_sort(list = `ANY[]`, col1 = VARCHAR, col2 = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list_sort([3, 6, 1, 2])
}\if{html}{\out{</div>}}
}

