% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{list_unique}
\alias{list_unique}
\title{DuckDB function list_unique}
\usage{
list_unique(list = `ANY[]`)
}
\arguments{
\item{list}{\code{ANY[]}}
}
\value{
\code{UBIGINT}
}
\description{
Counts the unique elements of a \code{list}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{list_unique([1, 1, NULL, -3, 1, 5])
}\if{html}{\out{</div>}}
}

