% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{listagg}
\alias{listagg}
\title{DuckDB function listagg}
\arguments{
\item{str}{\code{ANY}}

\item{arg}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Concatenates the column string values with an optional separator.
}
\section{Overloads}{

\itemize{
\item \code{listagg(str = ANY)}
\item \code{listagg(str = ANY, arg = VARCHAR)}
}
}

\examples{
\dontrun{
listagg(A, '-')
}
}
