% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lower}
\alias{lower}
\title{DuckDB function lower}
\usage{
lower(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Converts \code{string} to lower case.
}
\examples{
\dontrun{
lower('Hello')
}
}
