% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{lpad}
\alias{lpad}
\title{DuckDB function lpad}
\arguments{
\item{string}{\code{VARCHAR}}

\item{count}{\code{INTEGER}}

\item{character}{\code{VARCHAR}}
}
\description{
Pads the \code{string} with the \code{character} on the left until it has \code{count} characters. Truncates the \code{string} on the right if it has more than \code{count} characters.
}
\examples{
\dontrun{
lpad('hello', 8, '>')
}
}
