% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ltrim}
\alias{ltrim}
\title{DuckDB function ltrim}
\arguments{
\item{string}{\code{VARCHAR}}

\item{characters}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Removes any occurrences of any of the \code{characters} from the left side of the \code{string}. \code{characters} defaults to \code{space}.
}
\examples{
\dontrun{
c("ltrim('    test  ')", "ltrim('>>>>test<<', '><')")
}
}
