% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{mad}
\alias{mad}
\title{DuckDB function mad}
\arguments{
\item{x}{\verb{DECIMAL | FLOAT | DOUBLE | DATE | TIMESTAMP | TIME | TIMESTAMP WITH TIME ZONE | TIME WITH TIME ZONE}}
}
\value{
\code{DECIMAL | FLOAT | DOUBLE | INTERVAL}
}
\description{
Returns the median absolute deviation for the values within x. NULL values are ignored. Temporal types return a positive INTERVAL.	.
}
\examples{
\dontrun{
mad(x)
}
}
