% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{make_date}
\alias{make_date}
\title{DuckDB function make_date}
\arguments{
\item{col0}{\code{INTEGER}}

\item{year}{\code{BIGINT}}

\item{month}{\code{BIGINT}}

\item{day}{\code{BIGINT}}

\item{date-struct}{\verb{STRUCT("year" BIGINT, "month" BIGINT, "day" BIGINT)}}
}
\value{
\code{DATE}
}
\description{
The date for the given parts.

The date for the given struct.
}
\examples{
\dontrun{
character(0)
make_date(1992, 9, 20)
make_date({'year': 2024, 'month': 11, 'day': 14})
}
}
