% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{make_time}
\alias{make_time}
\title{DuckDB function make_time}
\usage{
make_time(hour = BIGINT, minute = BIGINT, seconds = DOUBLE)
}
\arguments{
\item{hour}{\code{BIGINT}}

\item{minute}{\code{BIGINT}}

\item{seconds}{\code{DOUBLE}}
}
\value{
\code{TIME}
}
\description{
The time for the given parts.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{make_time(13, 34, 27.123456)
}\if{html}{\out{</div>}}
}

