% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map}
\alias{map}
\title{DuckDB function map}
\arguments{
\item{keys}{\code{K[]}}

\item{values}{\code{V[]}}
}
\value{
\code{MAP("NULL", "NULL") | MAP(K, V)}
}
\description{
Creates a map from a set of keys and values.
}
\section{Overloads}{

\itemize{
\item \code{map()}
\item \code{map(keys = `K[]`, values = `V[]`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{map(['key1', 'key2'], ['val1', 'val2'])
}\if{html}{\out{</div>}}
}

