% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_contains}
\alias{map_contains}
\title{DuckDB function map_contains}
\usage{
map_contains(map = `MAP(K, V)`, key = K)
}
\arguments{
\item{map}{\code{MAP(K, V)}}

\item{key}{\code{K}}
}
\value{
\code{BOOLEAN}
}
\description{
Checks if a map contains a given key.
}
\examples{
\dontrun{
map_contains(MAP {'key1': 10, 'key2': 20, 'key3': 30}, 'key2')
}
}
