% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_from_entries}
\alias{map_from_entries}
\title{DuckDB function map_from_entries}
\usage{
map_from_entries(map = `STRUCT(K, V)[]`)
}
\arguments{
\item{map}{\code{STRUCT(K, V)[]}}
}
\value{
\code{MAP(K, V)}
}
\description{
Returns a map created from the entries of the array.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{map_from_entries([\{k: 5, v: 'val1'\}, \{k: 3, v: 'val2'\}]);
}\if{html}{\out{</div>}}
}

