% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{map_keys}
\alias{map_keys}
\title{DuckDB function map_keys}
\usage{
map_keys(map = `MAP(K, V)`)
}
\arguments{
\item{map}{\code{MAP(K, V)}}
}
\description{
Returns the keys of a map as a list.
}
\examples{
\dontrun{
map_keys(map(['key'], ['val']))
}
}
