% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{max}
\alias{max}
\title{DuckDB function max}
\arguments{
\item{arg}{\code{ANY}}

\item{col1}{\code{BIGINT}}
}
\value{
\code{ANY | ANY[]}
}
\description{
Returns the maximum value present in arg.
}
\examples{
\dontrun{
max(A)
}
}
