% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{md5}
\alias{md5}
\title{DuckDB function md5}
\arguments{
\item{string}{\code{VARCHAR}}

\item{blob}{\code{BLOB}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the MD5 hash of the \code{string} as a \code{VARCHAR}.

Returns the MD5 hash of the \code{blob} as a \code{VARCHAR}.
}
\examples{
\dontrun{
md5('abc')
md5('\xAA\xBB'::BLOB)
}
}
