% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{md5_number}
\alias{md5_number}
\title{DuckDB function md5_number}
\arguments{
\item{string}{\code{VARCHAR}}

\item{blob}{\code{BLOB}}
}
\value{
\code{UHUGEINT}
}
\description{
Returns the MD5 hash of the \code{string} as a \code{HUGEINT}.

Returns the MD5 hash of the \code{blob} as a \code{HUGEINT}.
}
\section{Overloads}{

\itemize{
\item \code{md5_number(string = VARCHAR)}
\item \code{md5_number(blob = BLOB)}
}
}

\examples{
\dontrun{
md5_number('abc')
md5_number('\xAA\xBB'::BLOB)
}
}
