% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{mean}
\alias{mean}
\title{DuckDB function mean}
\arguments{
\item{x}{\verb{DECIMAL | SMALLINT | INTEGER | BIGINT | HUGEINT | INTERVAL | DOUBLE | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TIME | TIME WITH TIME ZONE}}
}
\value{
\verb{DECIMAL | DOUBLE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE | TIME | TIME WITH TIME ZONE}
}
\description{
Calculates the average value for all tuples in x.
}
\section{Overloads}{

\itemize{
\item \code{mean(x = DECIMAL)}
\item \code{mean(x = SMALLINT)}
\item \code{mean(x = INTEGER)}
\item \code{mean(x = BIGINT)}
\item \code{mean(x = HUGEINT)}
\item \code{mean(x = INTERVAL)}
\item \code{mean(x = DOUBLE)}
\item \code{mean(x = TIMESTAMP)}
\item \code{mean(x = `TIMESTAMP WITH TIME ZONE`)}
\item \code{mean(x = TIME)}
\item \code{mean(x = `TIME WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
SUM(x) / COUNT(*)
}
}
