% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{microsecond}
\alias{microsecond}
\title{DuckDB function microsecond}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the microsecond component from a date or timestamp.
}
\examples{
\dontrun{
microsecond(timestamp '2021-08-03 11:59:44.123456')
}
}
