% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{millennium}
\alias{millennium}
\title{DuckDB function millennium}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the millennium component from a date or timestamp.
}
\examples{
\dontrun{
millennium(timestamp '2021-08-03 11:59:44.123456')
}
}
