% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{minute}
\alias{minute}
\title{DuckDB function minute}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the minute component from a date or timestamp.
}
\examples{
\dontrun{
minute(timestamp '2021-08-03 11:59:44.123456')
}
}
