% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{month}
\alias{month}
\title{DuckDB function month}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the month component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{month(ts = DATE)}
\item \code{month(ts = INTERVAL)}
\item \code{month(ts = TIMESTAMP)}
\item \code{month(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
month(timestamp '2021-08-03 11:59:44.123456')
}
}
