% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{monthname}
\alias{monthname}
\title{DuckDB function monthname}
\arguments{
\item{ts}{\verb{DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
The (English) name of the month.
}
\examples{
\dontrun{
monthname(TIMESTAMP '1992-09-20')
}
}
