% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{multiply}
\alias{multiply}
\title{DuckDB function multiply}
\arguments{
\item{col0}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | INTERVAL}}

\item{col1}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | INTERVAL}}
}
\description{
DuckDB function \code{multiply()}.
}
\examples{
\dontrun{
character(0)
}
}
