% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{nanosecond}
\alias{nanosecond}
\title{DuckDB function nanosecond}
\arguments{
\item{tsns}{\verb{DATE | TIMESTAMP | INTERVAL | TIME | TIME_NS | TIME WITH TIME ZONE | TIMESTAMP_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the nanosecond component from a date or timestamp.
}
\examples{
\dontrun{
nanosecond(timestamp_ns '2021-08-03 11:59:44.123456789')
}
}
