% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{nextafter}
\alias{nextafter}
\title{DuckDB function nextafter}
\arguments{
\item{x}{\code{DOUBLE | FLOAT}}

\item{y}{\code{DOUBLE | FLOAT}}
}
\value{
\code{DOUBLE | FLOAT}
}
\description{
Returns the next floating point value after x in the direction of y.
}
\section{Overloads}{

\itemize{
\item \code{nextafter(x = DOUBLE, y = DOUBLE)}
\item \code{nextafter(x = FLOAT, y = FLOAT)}
}
}

\examples{
\dontrun{
nextafter(1::float, 2::float)
}
}
