% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{nextval}
\alias{nextval}
\title{DuckDB function nextval}
\usage{
nextval(`'sequence_name'` = VARCHAR)
}
\arguments{
\item{'sequence_name'}{\code{VARCHAR}}
}
\value{
\code{BIGINT}
}
\description{
Return the following value of the sequence.
}
\examples{
\dontrun{
nextval('my_sequence_name')
}
}
