% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{nfc_normalize}
\alias{nfc_normalize}
\title{DuckDB function nfc_normalize}
\usage{
nfc_normalize(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Converts \code{string} to Unicode NFC normalized string. Useful for comparisons and ordering if text data is mixed between NFC normalized and not.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{nfc_normalize('ardèch')
}\if{html}{\out{</div>}}
}

