% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{nfc_normalize}
\alias{nfc_normalize}
\title{DuckDB function nfc_normalize}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Convert string to Unicode NFC normalized string. Useful for comparisons and ordering if text data is mixed between NFC normalized and not.
}
\examples{
\dontrun{
nfc_normalize('ardèch')
}
}
