% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{octet_length}
\alias{octet_length}
\title{DuckDB function octet_length}
\arguments{
\item{blob}{\code{BLOB}}

\item{bitstring}{\code{BIT}}
}
\value{
\code{BIGINT}
}
\description{
\itemize{
\item \code{octet_length(blob = BLOB)}: Number of bytes in \code{blob}.
\item \code{octet_length(bitstring = BIT)}: Returns the number of bytes in the \code{bitstring}.
}
}
\section{Overloads}{

\itemize{
\item \code{octet_length(blob = BLOB)}
\item \code{octet_length(bitstring = BIT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{octet_length('\\xAA\\xBB'::BLOB)
octet_length('1101011'::BITSTRING)
}\if{html}{\out{</div>}}
}

\seealso{
Other blob:
\code{\link[=decode]{decode()}},
\code{\link[=encode]{encode()}},
\code{\link[=from_base64]{from_base64()}},
\code{\link[=hex]{hex()}},
\code{\link[=md5]{md5()}},
\code{\link[=md5_number]{md5_number()}},
\code{\link{or--or}},
\code{\link[=repeat]{repeat()}},
\code{\link[=sha1]{sha1()}},
\code{\link[=sha256]{sha256()}},
\code{\link[=to_base64]{to_base64()}},
\code{\link[=unbin]{unbin()}},
\code{\link[=unhex]{unhex()}}
}
\concept{bitstring}
\concept{blob}
