% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{octet_length}
\alias{octet_length}
\title{DuckDB function octet_length}
\arguments{
\item{blob}{\code{BLOB}}

\item{bitstring}{\code{BIT}}
}
\value{
\code{BIGINT}
}
\description{
Number of bytes in \code{blob}.

Returns the number of bytes in the \code{bitstring}.
}
\examples{
\dontrun{
octet_length('\xAA\xBB'::BLOB)
octet_length('1101011'::BITSTRING)
}
}
