% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{or--or}
\alias{or--or}
\alias{||}
\title{DuckDB function ||}
\usage{
`||`(arg1 = ANY, arg2 = ANY)
}
\arguments{
\item{arg1}{\code{ANY}}

\item{arg2}{\code{ANY}}
}
\value{
\code{ANY}
}
\description{
Concatenates two strings, lists, or blobs. Any \code{NULL} input results in \code{NULL}. See also \code{concat(arg1, arg2, ...)} and \code{list_concat(list1, list2, ...)}.
}
\examples{
\dontrun{
c("'Duck' || 'DB'", "[1, 2, 3] || [4, 5, 6]", "'\\\\xAA'::BLOB || '\\\\xBB'::BLOB")
}
}
