% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{or-}
\alias{or-}
\alias{|}
\title{DuckDB function |}
\usage{
`|`(left, right)
}
\arguments{
\item{left, right}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise OR.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{32 | 3
}\if{html}{\out{</div>}}
}

