% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parquet_bloom_probe}
\alias{parquet_bloom_probe}
\title{DuckDB function parquet_bloom_probe}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{parquet_bloom_probe()}.
}
\section{Overloads}{

\itemize{
\item \code{parquet_bloom_probe(col0 = VARCHAR, col1 = VARCHAR, col2 = ANY)}
\item \code{parquet_bloom_probe(col0 = `VARCHAR[]`, col1 = VARCHAR, col2 = ANY)}
}
}

