% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parquet_file_metadata}
\alias{parquet_file_metadata}
\title{DuckDB function parquet_file_metadata}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{parquet_file_metadata()}.
}
\section{Overloads}{

\itemize{
\item \code{parquet_file_metadata(col0 = VARCHAR)}
\item \code{parquet_file_metadata(col0 = `VARCHAR[]`)}
}
}

