% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parquet_metadata}
\alias{parquet_metadata}
\title{DuckDB function parquet_metadata}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{parquet_metadata()}.
}
