% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parse_dirname}
\alias{parse_dirname}
\title{DuckDB function parse_dirname}
\arguments{
\item{path}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\description{
Returns the top-level directory name from the given \code{path}. \code{separator} options: \code{system}, \code{both_slash} (default), \code{forward_slash}, \code{backslash}.
}
\examples{
\dontrun{
parse_dirname('path/to/file.csv', 'system')
}
}
