% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parse_dirname}
\alias{parse_dirname}
\title{DuckDB function parse_dirname}
\arguments{
\item{path}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the top-level directory name from the given \code{path}. \code{separator} options: \code{system}, \code{both_slash} (default), \code{forward_slash}, \code{backslash}.
}
\section{Overloads}{

\itemize{
\item \code{parse_dirname(path = VARCHAR)}
\item \code{parse_dirname(path = VARCHAR, separator = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{parse_dirname('path/to/file.csv', 'system')
}\if{html}{\out{</div>}}
}

