% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parse_dirpath}
\alias{parse_dirpath}
\title{DuckDB function parse_dirpath}
\arguments{
\item{path}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the head of the \code{path} (the pathname until the last slash) similarly to Python's \code{os.path.dirname}. \code{separator} options: \code{system}, \code{both_slash} (default), \code{forward_slash}, \code{backslash}.
}
\section{Overloads}{

\itemize{
\item \code{parse_dirpath(path = VARCHAR)}
\item \code{parse_dirpath(path = VARCHAR, separator = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{parse_dirpath('path/to/file.csv', 'forward_slash')
}\if{html}{\out{</div>}}
}

