% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parse_filename}
\alias{parse_filename}
\title{DuckDB function parse_filename}
\arguments{
\item{string}{\code{VARCHAR}}

\item{trim_extension}{\code{VARCHAR | BOOLEAN}}

\item{separator}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the last component of the \code{path} similarly to Python's \code{os.path.basename} function. If \code{trim_extension} is \code{true}, the file extension will be removed (defaults to \code{false}). \code{separator} options: \code{system}, \code{both_slash} (default), \code{forward_slash}, \code{backslash}.
}
\examples{
\dontrun{
parse_filename('path/to/file.csv', true, 'forward_slash')
}
}
