% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parse_formatted_bytes}
\alias{parse_formatted_bytes}
\title{DuckDB function parse_formatted_bytes}
\usage{
parse_formatted_bytes(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{UBIGINT}
}
\description{
Parses a human-readable representation of a size in bytes into an integer.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{parse_formatted_bytes('16 KiB')
}\if{html}{\out{</div>}}
}

