% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{parse_path}
\alias{parse_path}
\title{DuckDB function parse_path}
\arguments{
\item{path}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\value{
\code{VARCHAR[]}
}
\description{
Returns a list of the components (directories and filename) in the \code{path} similarly to Python's \code{pathlib.parts} function. \code{separator} options: \code{system}, \code{both_slash} (default), \code{forward_slash}, \code{backslash}.
}
\section{Overloads}{

\itemize{
\item \code{parse_path(path = VARCHAR)}
\item \code{parse_path(path = VARCHAR, separator = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{parse_path('path/to/file.csv', 'system')
}\if{html}{\out{</div>}}
}

