% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pg_has_role}
\alias{pg_has_role}
\title{DuckDB function pg_has_role}
\arguments{
\item{user}{Unspecified.}

\item{role}{Unspecified.}

\item{privilege}{Unspecified.}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pg_has_role()}.
}
\section{Overloads}{

\itemize{
\item \code{pg_has_role(user, role, privilege)}
\item \code{pg_has_role(role, privilege)}
}
}

