% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{position}
\alias{position}
\title{DuckDB function position}
\usage{
position(string = VARCHAR, search_string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\value{
\code{BIGINT}
}
\description{
Returns location of first occurrence of \code{search_string} in \code{string}, counting from 1. Returns 0 if no match found.
}
\examples{
\dontrun{
position('b' IN 'abc')
}
}
