% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pow}
\alias{pow}
\title{DuckDB function pow}
\usage{
pow(x = DOUBLE, y = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}

\item{y}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Computes x to the power of y.
}
\examples{
\dontrun{
c("pow(2, 3)", "power(2, 3)")
}
}
