% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pow}
\alias{pow}
\alias{duckdb_fun_pow}
\title{DuckDB function pow}
\arguments{
\item{x}{\code{DOUBLE}}

\item{y}{\code{DOUBLE}}
}
\description{
Computes x to the power of y.
}
\examples{
\dontrun{
pow(2, 3)
}
}
