% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{pragma_metadata_info}
\alias{pragma_metadata_info}
\title{DuckDB function pragma_metadata_info}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{pragma_metadata_info()}.
}
\section{Overloads}{

\itemize{
\item \code{pragma_metadata_info()}
\item \code{pragma_metadata_info(col0 = VARCHAR)}
}
}

