% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{prefix}
\alias{prefix}
\title{DuckDB function prefix}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\description{
Returns \code{true} if \code{string} starts with \code{search_string}.
}
\examples{
\dontrun{
prefix('abc', 'ab')
}
}
