% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{prefix}
\alias{prefix}
\title{DuckDB function prefix}
\usage{
prefix(string = VARCHAR, search_string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if \code{string} starts with \code{search_string}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{prefix('abc', 'ab')
}\if{html}{\out{</div>}}
}

