% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{printf}
\alias{printf}
\title{DuckDB function printf}
\usage{
printf(format = VARCHAR)
}
\arguments{
\item{format}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Formats a \code{string} using printf syntax.
}
\examples{
\dontrun{
printf('Benchmark "\%s" took \%d seconds', 'CSV', 42)
}
}
