% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{product}
\alias{product}
\title{DuckDB function product}
\usage{
product(arg = DOUBLE)
}
\arguments{
\item{arg}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Calculates the product of all tuples in arg.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{product(A)
}\if{html}{\out{</div>}}
}

