% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{quantile_disc}
\alias{quantile_disc}
\title{DuckDB function quantile_disc}
\arguments{
\item{x}{\code{ANY}}

\item{pos}{\code{DOUBLE | DOUBLE[]}}
}
\value{
\code{ANY}
}
\description{
Returns the exact quantile number between 0 and 1 . If pos is a LIST of FLOATs, then the result is a LIST of the corresponding exact quantiles.
}
\examples{
\dontrun{
quantile_disc(x, 0.5)
}
}
