% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{query_table}
\alias{query_table}
\title{DuckDB function query_table}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{col1}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{query_table()}.
}
\examples{
\dontrun{
character(0)
}
}
