% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{r_dataframe_scan}
\alias{r_dataframe_scan}
\title{DuckDB function r_dataframe_scan}
\usage{
r_dataframe_scan(col0 = POINTER, experimental = BOOLEAN, integer64 = BOOLEAN)
}
\arguments{
\item{col0}{\code{POINTER}}

\item{experimental}{\code{BOOLEAN}}

\item{integer64}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{r_dataframe_scan()}.
}
