% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{radians}
\alias{radians}
\title{DuckDB function radians}
\usage{
radians(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Converts degrees to radians.
}
\examples{
\dontrun{
radians(90)
}
}
