% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{range}
\alias{range}
\title{DuckDB function range}
\arguments{
\item{col0}{\code{BIGINT | TIMESTAMP}}

\item{col1}{\code{BIGINT | TIMESTAMP}}

\item{col2}{\code{BIGINT | INTERVAL}}

\item{start}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{stop}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{step}{\code{BIGINT | INTERVAL}}
}
\value{
\verb{BIGINT[] | TIMESTAMP[] | TIMESTAMP WITH TIME ZONE[]}
}
\description{
Creates a list of values between \code{start} and \code{stop} - the stop parameter is exclusive.
}
\examples{
\dontrun{
range(2, 5, 3)
}
}
