% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{range}
\alias{range}
\title{DuckDB function range}
\arguments{
\item{col0}{\code{BIGINT | TIMESTAMP}}

\item{col1}{\code{BIGINT | TIMESTAMP}}

\item{col2}{\code{BIGINT | INTERVAL}}

\item{start}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{stop}{\verb{BIGINT | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{step}{\code{BIGINT | INTERVAL}}
}
\value{
\verb{BIGINT[] | TIMESTAMP[] | TIMESTAMP WITH TIME ZONE[]}
}
\description{
Creates a list of values between \code{start} and \code{stop} - the stop parameter is exclusive.
}
\section{Overloads}{

\itemize{
\item \code{range(col0 = BIGINT)}
\item \code{range(col0 = BIGINT, col1 = BIGINT)}
\item \code{range(col0 = BIGINT, col1 = BIGINT, col2 = BIGINT)}
\item \code{range(col0 = TIMESTAMP, col1 = TIMESTAMP, col2 = INTERVAL)}
\item \code{range(start = BIGINT)}
\item \code{range(start = BIGINT, stop = BIGINT)}
\item \code{range(start = BIGINT, stop = BIGINT, step = BIGINT)}
\item \code{range(start = TIMESTAMP, stop = TIMESTAMP, step = INTERVAL)}
\item \code{range(start = `TIMESTAMP WITH TIME ZONE`, stop = `TIMESTAMP WITH TIME ZONE`, step = INTERVAL)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{range(2, 5, 3)
}\if{html}{\out{</div>}}
}

