% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_blob}
\alias{read_blob}
\title{DuckDB function read_blob}
\usage{
read_blob(col0)
}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_blob()}.
}
