% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_blob}
\alias{read_blob}
\title{DuckDB function read_blob}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{union_by_name}{\code{BOOLEAN}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}

\item{filename}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_blob()}.
}
