% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_csv_auto}
\alias{read_csv_auto}
\title{DuckDB function read_csv_auto}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{thousands}{\code{VARCHAR}}

\item{strict_mode}{\code{BOOLEAN}}

\item{dtypes}{\code{ANY}}

\item{column_types}{\code{ANY}}

\item{null_padding}{\code{BOOLEAN}}

\item{column_names}{\code{VARCHAR[]}}

\item{buffer_size}{\code{UBIGINT}}

\item{parallel}{\code{BOOLEAN}}

\item{force_not_null}{\code{VARCHAR[]}}

\item{hive_types}{\code{ANY}}

\item{new_line}{\code{VARCHAR}}

\item{files_to_sniff}{\code{BIGINT}}

\item{dateformat}{\code{VARCHAR}}

\item{delim}{\code{VARCHAR}}

\item{sep}{\code{VARCHAR}}

\item{decimal_separator}{\code{VARCHAR}}

\item{nullstr}{\code{ANY}}

\item{escape}{\code{VARCHAR}}

\item{compression}{\code{VARCHAR}}

\item{encoding}{\code{VARCHAR}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{all_varchar}{\code{BOOLEAN}}

\item{columns}{\code{ANY}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{auto_detect}{\code{BOOLEAN}}

\item{comment}{\code{VARCHAR}}

\item{quote}{\code{VARCHAR}}

\item{max_line_size}{\code{VARCHAR}}

\item{store_rejects}{\code{BOOLEAN}}

\item{union_by_name}{\code{BOOLEAN}}

\item{header}{\code{BOOLEAN}}

\item{types}{\code{ANY}}

\item{skip}{\code{BIGINT}}

\item{filename}{\code{ANY}}

\item{sample_size}{\code{BIGINT}}

\item{timestampformat}{\code{VARCHAR}}

\item{normalize_names}{\code{BOOLEAN}}

\item{ignore_errors}{\code{BOOLEAN}}

\item{names}{\code{VARCHAR[]}}

\item{allow_quoted_nulls}{\code{BOOLEAN}}

\item{maximum_line_size}{\code{VARCHAR}}

\item{rejects_table}{\code{VARCHAR}}

\item{auto_type_candidates}{\code{ANY}}

\item{rejects_scan}{\code{VARCHAR}}

\item{rejects_limit}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_csv_auto()}.
}
\section{Overloads}{

\itemize{
\item \code{read_csv_auto(col0 = VARCHAR, thousands = VARCHAR, strict_mode = BOOLEAN, dtypes = ANY, column_types = ANY, null_padding = BOOLEAN, column_names = `VARCHAR[]`, buffer_size = UBIGINT, parallel = BOOLEAN, force_not_null = `VARCHAR[]`, hive_types = ANY, new_line = VARCHAR, files_to_sniff = BIGINT, dateformat = VARCHAR, delim = VARCHAR, sep = VARCHAR, decimal_separator = VARCHAR, nullstr = ANY, escape = VARCHAR, compression = VARCHAR, encoding = VARCHAR, hive_types_autocast = BOOLEAN, all_varchar = BOOLEAN, columns = ANY, hive_partitioning = BOOLEAN, auto_detect = BOOLEAN, comment = VARCHAR, quote = VARCHAR, max_line_size = VARCHAR, store_rejects = BOOLEAN, union_by_name = BOOLEAN, header = BOOLEAN, types = ANY, skip = BIGINT, filename = ANY, sample_size = BIGINT, timestampformat = VARCHAR, normalize_names = BOOLEAN, ignore_errors = BOOLEAN, names = `VARCHAR[]`, allow_quoted_nulls = BOOLEAN, maximum_line_size = VARCHAR, rejects_table = VARCHAR, auto_type_candidates = ANY, rejects_scan = VARCHAR, rejects_limit = BIGINT)}
\item \code{read_csv_auto(col0 = `VARCHAR[]`, thousands = VARCHAR, strict_mode = BOOLEAN, dtypes = ANY, column_types = ANY, null_padding = BOOLEAN, column_names = `VARCHAR[]`, buffer_size = UBIGINT, parallel = BOOLEAN, force_not_null = `VARCHAR[]`, hive_types = ANY, new_line = VARCHAR, files_to_sniff = BIGINT, dateformat = VARCHAR, delim = VARCHAR, sep = VARCHAR, decimal_separator = VARCHAR, nullstr = ANY, escape = VARCHAR, compression = VARCHAR, encoding = VARCHAR, hive_types_autocast = BOOLEAN, all_varchar = BOOLEAN, columns = ANY, hive_partitioning = BOOLEAN, auto_detect = BOOLEAN, comment = VARCHAR, quote = VARCHAR, max_line_size = VARCHAR, store_rejects = BOOLEAN, union_by_name = BOOLEAN, header = BOOLEAN, types = ANY, skip = BIGINT, filename = ANY, sample_size = BIGINT, timestampformat = VARCHAR, normalize_names = BOOLEAN, ignore_errors = BOOLEAN, names = `VARCHAR[]`, allow_quoted_nulls = BOOLEAN, maximum_line_size = VARCHAR, rejects_table = VARCHAR, auto_type_candidates = ANY, rejects_scan = VARCHAR, rejects_limit = BIGINT)}
}
}

