% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_csv_auto}
\alias{read_csv_auto}
\title{DuckDB function read_csv_auto}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{skip}{\code{BIGINT}}

\item{types}{\code{ANY}}

\item{nullstr}{\code{ANY}}

\item{encoding}{\code{VARCHAR}}

\item{hive_types}{\code{ANY}}

\item{filename}{\code{ANY}}

\item{header}{\code{BOOLEAN}}

\item{delim}{\code{VARCHAR}}

\item{dateformat}{\code{VARCHAR}}

\item{column_names}{\code{VARCHAR[]}}

\item{union_by_name}{\code{BOOLEAN}}

\item{new_line}{\code{VARCHAR}}

\item{escape}{\code{VARCHAR}}

\item{allow_quoted_nulls}{\code{BOOLEAN}}

\item{comment}{\code{VARCHAR}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{sep}{\code{VARCHAR}}

\item{columns}{\code{ANY}}

\item{rejects_limit}{\code{BIGINT}}

\item{force_not_null}{\code{VARCHAR[]}}

\item{auto_type_candidates}{\code{ANY}}

\item{sample_size}{\code{BIGINT}}

\item{timestampformat}{\code{VARCHAR}}

\item{auto_detect}{\code{BOOLEAN}}

\item{all_varchar}{\code{BOOLEAN}}

\item{store_rejects}{\code{BOOLEAN}}

\item{normalize_names}{\code{BOOLEAN}}

\item{rejects_table}{\code{VARCHAR}}

\item{column_types}{\code{ANY}}

\item{compression}{\code{VARCHAR}}

\item{ignore_errors}{\code{BOOLEAN}}

\item{names}{\code{VARCHAR[]}}

\item{max_line_size}{\code{VARCHAR}}

\item{quote}{\code{VARCHAR}}

\item{maximum_line_size}{\code{VARCHAR}}

\item{rejects_scan}{\code{VARCHAR}}

\item{buffer_size}{\code{UBIGINT}}

\item{decimal_separator}{\code{VARCHAR}}

\item{parallel}{\code{BOOLEAN}}

\item{null_padding}{\code{BOOLEAN}}

\item{dtypes}{\code{ANY}}

\item{strict_mode}{\code{BOOLEAN}}

\item{thousands}{\code{VARCHAR}}

\item{files_to_sniff}{\code{BIGINT}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_csv_auto()}.
}
\examples{
\dontrun{
character(0)
}
}
