% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_duckdb}
\alias{read_duckdb}
\title{DuckDB function read_duckdb}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{table_name}{\code{VARCHAR}}

\item{schema_name}{\code{VARCHAR}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{union_by_name}{\code{BOOLEAN}}

\item{filename}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_duckdb()}.
}
\section{Overloads}{

\itemize{
\item \code{read_duckdb(col0 = VARCHAR, table_name = VARCHAR, schema_name = VARCHAR, hive_types_autocast = BOOLEAN, hive_types = ANY, hive_partitioning = BOOLEAN, union_by_name = BOOLEAN, filename = ANY)}
\item \code{read_duckdb(col0 = `VARCHAR[]`, table_name = VARCHAR, schema_name = VARCHAR, hive_types_autocast = BOOLEAN, hive_types = ANY, hive_partitioning = BOOLEAN, union_by_name = BOOLEAN, filename = ANY)}
}
}

