% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{read_text}
\alias{read_text}
\title{DuckDB function read_text}
\arguments{
\item{col0}{\code{VARCHAR | VARCHAR[]}}

\item{union_by_name}{\code{BOOLEAN}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{hive_types_autocast}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}

\item{filename}{\code{ANY}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{read_text()}.
}
\examples{
\dontrun{
character(0)
}
}
