% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{reduce}
\alias{reduce}
\title{DuckDB function reduce}
\arguments{
\item{list}{\code{ANY[]}}

\item{initial_value}{\code{ANY}}

\item{lambda(x, y)}{\code{LAMBDA}}
}
\value{
\code{ANY}
}
\description{
Reduces all elements of the input \code{list} into a single scalar value by executing the \code{lambda} function on a running result and the next list element. The \code{lambda} function has an optional \code{initial_value} argument.
}
\section{Overloads}{

\itemize{
\item \code{reduce(list = `ANY[]`, `lambda(x,y)` = LAMBDA)}
\item \code{reduce(list = `ANY[]`, `lambda(x,y)` = LAMBDA, initial_value = ANY)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{reduce([1, 2, 3], lambda x, y : x + y)
}\if{html}{\out{</div>}}
}

