% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_escape}
\alias{regexp_escape}
\title{DuckDB function regexp_escape}
\usage{
regexp_escape(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Escapes special patterns to turn \code{string} into a regular expression similarly to Python's \code{re.escape} function.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{regexp_escape('https://duckdb.org')
}\if{html}{\out{</div>}}
}

