% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_escape}
\alias{regexp_escape}
\title{DuckDB function regexp_escape}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Escapes special patterns to turn \code{string} into a regular expression similarly to Python's \code{re.escape} function.
}
\examples{
\dontrun{
regexp_escape('https://duckdb.org')
}
}
