% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_escape}
\alias{regexp_escape}
\alias{duckdb_fun_regexp_escape}
\title{DuckDB function regexp_escape}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Escapes all potentially meaningful regexp characters in the input string.
}
\examples{
\dontrun{
regexp_escape('https://duckdb.org')
}
}
