% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_extract_all}
\alias{regexp_extract_all}
\title{DuckDB function regexp_extract_all}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{group}{\code{INTEGER}}

\item{options}{\code{VARCHAR}}
}
\description{
Finds non-overlapping occurrences of the \code{regex} in the \code{string} and returns the corresponding values of the capturing \code{group}. A set of optional regex \code{options} can be set.
}
\examples{
\dontrun{
regexp_extract_all('Peter: 33, Paul:14', '(\w+):\s*(\d+)', 2)
}
}
