% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_full_match}
\alias{regexp_full_match}
\title{DuckDB function regexp_full_match}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{col2}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if the entire \code{string} matches the \code{regex}. A set of optional regex \code{options} can be set.
}
\examples{
\dontrun{
regexp_full_match('anabanana', '(an)*')
}
}
