% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_matches}
\alias{regexp_matches}
\title{DuckDB function regexp_matches}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{options}{\code{VARCHAR}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns \code{true} if \code{string} contains the \code{regex}, \code{false} otherwise. A set of optional regex \code{options} can be set.
}
\section{Overloads}{

\itemize{
\item \code{regexp_matches(string = VARCHAR, regex = VARCHAR)}
\item \code{regexp_matches(string = VARCHAR, regex = VARCHAR, options = VARCHAR)}
}
}

\examples{
\dontrun{
regexp_matches('anabanana', '(an)*')
}
}
