% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regexp_replace}
\alias{regexp_replace}
\title{DuckDB function regexp_replace}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{replacement}{\code{VARCHAR}}

\item{options}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
If \code{string} contains the \code{regex}, replaces the matching part with \code{replacement}. A set of optional regex \code{options} can be set.
}
\section{Overloads}{

\itemize{
\item \code{regexp_replace(string = VARCHAR, regex = VARCHAR, replacement = VARCHAR)}
\item \code{regexp_replace(string = VARCHAR, regex = VARCHAR, replacement = VARCHAR, options = VARCHAR)}
}
}

\examples{
\dontrun{
regexp_replace('hello', '[lo]', '-')
}
}
