% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{regr_slope}
\alias{regr_slope}
\title{DuckDB function regr_slope}
\usage{
regr_slope(y = DOUBLE, x = DOUBLE)
}
\arguments{
\item{y}{\code{DOUBLE}}

\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the slope of the linear regression line for non-NULL pairs in a group.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{COVAR_POP(x, y) / VAR_POP(x)
}\if{html}{\out{</div>}}
}

