% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{remap_struct}
\alias{remap_struct}
\title{DuckDB function remap_struct}
\arguments{
\item{input}{\code{ANY}}

\item{target_type}{\code{ANY}}

\item{mapping}{\code{ANY}}

\item{defaults}{\code{ANY}}
}
\description{
Map a struct to another struct type, potentially re-ordering, renaming and casting members and filling in defaults for missing values.
}
\examples{
\dontrun{
remap_struct(
  {'i': 1, 'j': 2},
  NULL::ROW(v1 INT, v2 INT, v3 INT),i
  {'v1': 'j', 'v3': 'i'},i
  {'v2': NULL::INTEGER}i
)
}
}
