% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{repeat}
\alias{repeat}
\title{DuckDB function repeat}
\arguments{
\item{string}{\code{VARCHAR}}

\item{count}{\code{BIGINT}}

\item{blob}{\code{BLOB}}

\item{col0}{\code{T[]}}

\item{col1}{\code{BIGINT}}
}
\value{
\code{VARCHAR | BLOB | T[]}
}
\description{
Repeats the \code{string} \code{count} number of times.

Repeats the \code{blob} \code{count} number of times.
}
\examples{
\dontrun{
repeat('A', 5)
repeat('\xAA\xBB'::BLOB, 5)
character(0)
}
}
