% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{repeat}
\alias{repeat}
\title{DuckDB function repeat}
\arguments{
\item{col0}{\code{ANY | T[]}}

\item{col1}{\code{BIGINT}}

\item{string}{\code{VARCHAR}}

\item{count}{\code{BIGINT}}

\item{blob}{\code{BLOB}}
}
\value{
\code{VARCHAR | BLOB | T[]}
}
\description{
Repeats the \code{string} \code{count} number of times.

Repeats the \code{blob} \code{count} number of times.
}
\section{Overloads}{

\itemize{
\item \code{`repeat`(col0 = ANY, col1 = BIGINT)}
\item \code{`repeat`(string = VARCHAR, count = BIGINT)}
\item \code{`repeat`(blob = BLOB, count = BIGINT)}
\item \code{`repeat`(col0 = `T[]`, col1 = BIGINT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{repeat('A', 5)
repeat('\\xAA\\xBB'::BLOB, 5)
}\if{html}{\out{</div>}}
}

