% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{replace}
\alias{replace}
\title{DuckDB function replace}
\arguments{
\item{string}{\code{VARCHAR}}

\item{source}{\code{VARCHAR}}

\item{target}{\code{VARCHAR}}
}
\description{
Replaces any occurrences of the \code{source} with \code{target} in \code{string}.
}
\examples{
\dontrun{
replace('hello', 'l', '-')
}
}
