% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{replace_type}
\alias{replace_type}
\title{DuckDB function replace_type}
\usage{
replace_type(param = ANY, type1 = ANY, type2 = ANY)
}
\arguments{
\item{param}{\code{ANY}}

\item{type1}{\code{ANY}}

\item{type2}{\code{ANY}}
}
\description{
Casts all fields of type1 to type2.
}
\examples{
\dontrun{
replace_type({duck: 3.141592653589793::DOUBLE}, NULL::DOUBLE, NULL::DECIMAL(15,2))
}
}
