% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{reverse}
\alias{reverse}
\title{DuckDB function reverse}
\usage{
reverse(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Reverses the \code{string}.
}
\examples{
\dontrun{
reverse('hello')
}
}
