% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{right}
\alias{right}
\title{DuckDB function right}
\arguments{
\item{string}{\code{VARCHAR}}

\item{count}{\code{BIGINT}}
}
\description{
Extract the right-most \code{count} characters.
}
\examples{
\dontrun{
right('Hello🦆', 3)
}
}
