% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{right}
\alias{right}
\title{DuckDB function right}
\usage{
right(string = VARCHAR, count = BIGINT)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{count}{\code{BIGINT}}
}
\value{
\code{VARCHAR}
}
\description{
Extract the right-most \code{count} characters.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{right('Hello🦆', 3)
}\if{html}{\out{</div>}}
}

