% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{round}
\alias{round}
\title{DuckDB function round}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL}}

\item{precision}{\code{INTEGER}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL}
}
\description{
Rounds x to s decimal places.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{round(42.4332, 2)
}\if{html}{\out{</div>}}
}

