% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{round}
\alias{round}
\title{DuckDB function round}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL}}

\item{precision}{\code{INTEGER}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL}
}
\description{
Rounds x to s decimal places.
}
\section{Overloads}{

\itemize{
\item \code{round(x = TINYINT)}
\item \code{round(x = TINYINT, precision = INTEGER)}
\item \code{round(x = SMALLINT)}
\item \code{round(x = SMALLINT, precision = INTEGER)}
\item \code{round(x = INTEGER)}
\item \code{round(x = INTEGER, precision = INTEGER)}
\item \code{round(x = BIGINT)}
\item \code{round(x = BIGINT, precision = INTEGER)}
\item \code{round(x = HUGEINT)}
\item \code{round(x = HUGEINT, precision = INTEGER)}
\item \code{round(x = FLOAT)}
\item \code{round(x = FLOAT, precision = INTEGER)}
\item \code{round(x = DOUBLE)}
\item \code{round(x = DOUBLE, precision = INTEGER)}
\item \code{round(x = DECIMAL)}
\item \code{round(x = DECIMAL, precision = INTEGER)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{round(42.4332, 2)
}\if{html}{\out{</div>}}
}

