% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{round}
\alias{round}
\title{DuckDB function round}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL}}

\item{precision}{\code{INTEGER}}
}
\description{
Rounds x to s decimal places.
}
\examples{
\dontrun{
round(42.4332, 2)
}
}
