% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{rpad}
\alias{rpad}
\title{DuckDB function rpad}
\usage{
rpad(string = VARCHAR, count = INTEGER, character = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{count}{\code{INTEGER}}

\item{character}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Pads the \code{string} with the \code{character} on the right until it has \code{count} characters. Truncates the \code{string} on the right if it has more than \code{count} characters.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{rpad('hello', 10, '<')
}\if{html}{\out{</div>}}
}

