% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{rpad}
\alias{rpad}
\title{DuckDB function rpad}
\arguments{
\item{string}{\code{VARCHAR}}

\item{count}{\code{INTEGER}}

\item{character}{\code{VARCHAR}}
}
\description{
Pads the string with the character from the right until it has count characters.
}
\examples{
\dontrun{
rpad('hello', 10, '<')
}
}
