% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{rtrim}
\alias{rtrim}
\title{DuckDB function rtrim}
\arguments{
\item{string}{\code{VARCHAR}}

\item{characters}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Removes any occurrences of any of the \code{characters} from the right side of the \code{string}. \code{characters} defaults to \code{space}.
}
\section{Overloads}{

\itemize{
\item \code{rtrim(string = VARCHAR)}
\item \code{rtrim(string = VARCHAR, characters = VARCHAR)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{rtrim('    test  ')
rtrim('>>>>test<<', '><')
}\if{html}{\out{</div>}}
}

