% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{rtrim}
\alias{rtrim}
\title{DuckDB function rtrim}
\arguments{
\item{string}{\code{VARCHAR}}

\item{characters}{\code{VARCHAR}}
}
\description{
Removes any occurrences of any of the \code{characters} from the right side of the \code{string}. \code{characters} defaults to \code{space}.
}
\examples{
\dontrun{
c("rtrim('    test  ')", "rtrim('>>>>test<<', '><')")
}
}
