% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{second}
\alias{second}
\title{DuckDB function second}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIME | TIMESTAMP | TIME WITH TIME ZONE | TIME_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the second component from a date or timestamp.
}
\section{Overloads}{

\itemize{
\item \code{second(ts = DATE)}
\item \code{second(ts = INTERVAL)}
\item \code{second(ts = TIME)}
\item \code{second(ts = TIMESTAMP)}
\item \code{second(ts = `TIME WITH TIME ZONE`)}
\item \code{second(ts = TIME_NS)}
\item \code{second(ts = `TIMESTAMP WITH TIME ZONE`)}
}
}

\examples{
\dontrun{
second(timestamp '2021-08-03 11:59:44.123456')
}
}
