% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{setseed}
\alias{setseed}
\title{DuckDB function setseed}
\usage{
setseed(col0 = DOUBLE)
}
\arguments{
\item{col0}{\code{DOUBLE}}
}
\value{
\code{"NULL"}
}
\description{
Sets the seed to be used for the random function.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{setseed(0.42)
}\if{html}{\out{</div>}}
}

