% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sha1}
\alias{sha1}
\title{DuckDB function sha1}
\arguments{
\item{value}{\code{VARCHAR}}

\item{blob}{\code{BLOB}}
}
\description{
Returns a \code{VARCHAR} with the SHA-1 hash of the \code{value}.

Returns a \code{VARCHAR} with the SHA-1 hash of the \code{blob}.
}
\examples{
\dontrun{
sha1('🦆')
sha1('\xAA\xBB'::BLOB)
}
}
