% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sha256}
\alias{sha256}
\title{DuckDB function sha256}
\arguments{
\item{value}{\code{VARCHAR}}

\item{blob}{\code{BLOB}}
}
\value{
\code{VARCHAR}
}
\description{
Returns a \code{VARCHAR} with the SHA-256 hash of the \code{value}.

Returns a \code{VARCHAR} with the SHA-256 hash of the \code{blob}.
}
\examples{
\dontrun{
sha256('🦆')
sha256('\xAA\xBB'::BLOB)
}
}
