% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sign}
\alias{sign}
\title{DuckDB function sign}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\description{
Returns the sign of x as -1, 0 or 1.
}
\examples{
\dontrun{
sign(-349)
}
}
