% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sign}
\alias{sign}
\title{DuckDB function sign}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}
}
\value{
\code{TINYINT}
}
\description{
Returns the sign of x as -1, 0 or 1.
}
\section{Overloads}{

\itemize{
\item \code{sign(x = TINYINT)}
\item \code{sign(x = SMALLINT)}
\item \code{sign(x = INTEGER)}
\item \code{sign(x = BIGINT)}
\item \code{sign(x = HUGEINT)}
\item \code{sign(x = FLOAT)}
\item \code{sign(x = DOUBLE)}
\item \code{sign(x = UTINYINT)}
\item \code{sign(x = USMALLINT)}
\item \code{sign(x = UINTEGER)}
\item \code{sign(x = UBIGINT)}
\item \code{sign(x = UHUGEINT)}
}
}

\examples{
\dontrun{
sign(-349)
}
}
