% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{signbit}
\alias{signbit}
\title{DuckDB function signbit}
\arguments{
\item{x}{\code{FLOAT | DOUBLE}}
}
\value{
\code{BOOLEAN}
}
\description{
Returns whether the signbit is set or not.
}
\section{Overloads}{

\itemize{
\item \code{signbit(x = FLOAT)}
\item \code{signbit(x = DOUBLE)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{signbit(-0.0)
}\if{html}{\out{</div>}}
}

