% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{skewness}
\alias{skewness}
\title{DuckDB function skewness}
\usage{
skewness(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the skewness of all input values.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{skewness(A)
}\if{html}{\out{</div>}}
}

