% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{/}
\alias{/}
\title{DuckDB function /}
\arguments{
\item{col0}{\code{FLOAT | DOUBLE | INTERVAL}}

\item{col1}{\code{FLOAT | DOUBLE}}
}
\value{
\code{FLOAT | DOUBLE | INTERVAL}
}
\description{
DuckDB function \verb{/()}.
}
\section{Overloads}{

\itemize{
\item ```/`(col0 = FLOAT, col1 = FLOAT)``
\item ```/`(col0 = DOUBLE, col1 = DOUBLE)``
\item ```/`(col0 = INTERVAL, col1 = DOUBLE)``
}
}

