% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sniff_csv}
\alias{sniff_csv}
\title{DuckDB function sniff_csv}
\usage{
sniff_csv(col0 = VARCHAR, delim = VARCHAR, dateformat = VARCHAR, column_names = `VARCHAR[]`, sep = VARCHAR, hive_partitioning = BOOLEAN, header = BOOLEAN, escape = VARCHAR, allow_quoted_nulls = BOOLEAN, comment = VARCHAR, maximum_line_size = VARCHAR, new_line = VARCHAR, columns = ANY, rejects_limit = BIGINT, force_not_null = `VARCHAR[]`, timestampformat = VARCHAR, auto_detect = BOOLEAN, sample_size = BIGINT, auto_type_candidates = ANY, nullstr = ANY, encoding = VARCHAR, normalize_names = BOOLEAN, rejects_table = VARCHAR, column_types = ANY, skip = BIGINT, types = ANY, max_line_size = VARCHAR, quote = VARCHAR, rejects_scan = VARCHAR, ignore_errors = BOOLEAN, compression = VARCHAR, names = `VARCHAR[]`, force_match = BOOLEAN, store_rejects = BOOLEAN, all_varchar = BOOLEAN, buffer_size = UBIGINT, decimal_separator = VARCHAR, parallel = BOOLEAN, null_padding = BOOLEAN, dtypes = ANY, strict_mode = BOOLEAN, thousands = VARCHAR, files_to_sniff = BIGINT, filename = ANY, union_by_name = BOOLEAN, hive_types = ANY, hive_types_autocast = BOOLEAN)
}
\arguments{
\item{col0}{\code{VARCHAR}}

\item{delim}{\code{VARCHAR}}

\item{dateformat}{\code{VARCHAR}}

\item{column_names}{\code{VARCHAR[]}}

\item{sep}{\code{VARCHAR}}

\item{hive_partitioning}{\code{BOOLEAN}}

\item{header}{\code{BOOLEAN}}

\item{escape}{\code{VARCHAR}}

\item{allow_quoted_nulls}{\code{BOOLEAN}}

\item{comment}{\code{VARCHAR}}

\item{maximum_line_size}{\code{VARCHAR}}

\item{new_line}{\code{VARCHAR}}

\item{columns}{\code{ANY}}

\item{rejects_limit}{\code{BIGINT}}

\item{force_not_null}{\code{VARCHAR[]}}

\item{timestampformat}{\code{VARCHAR}}

\item{auto_detect}{\code{BOOLEAN}}

\item{sample_size}{\code{BIGINT}}

\item{auto_type_candidates}{\code{ANY}}

\item{nullstr}{\code{ANY}}

\item{encoding}{\code{VARCHAR}}

\item{normalize_names}{\code{BOOLEAN}}

\item{rejects_table}{\code{VARCHAR}}

\item{column_types}{\code{ANY}}

\item{skip}{\code{BIGINT}}

\item{types}{\code{ANY}}

\item{max_line_size}{\code{VARCHAR}}

\item{quote}{\code{VARCHAR}}

\item{rejects_scan}{\code{VARCHAR}}

\item{ignore_errors}{\code{BOOLEAN}}

\item{compression}{\code{VARCHAR}}

\item{names}{\code{VARCHAR[]}}

\item{force_match}{\code{BOOLEAN}}

\item{store_rejects}{\code{BOOLEAN}}

\item{all_varchar}{\code{BOOLEAN}}

\item{buffer_size}{\code{UBIGINT}}

\item{decimal_separator}{\code{VARCHAR}}

\item{parallel}{\code{BOOLEAN}}

\item{null_padding}{\code{BOOLEAN}}

\item{dtypes}{\code{ANY}}

\item{strict_mode}{\code{BOOLEAN}}

\item{thousands}{\code{VARCHAR}}

\item{files_to_sniff}{\code{BIGINT}}

\item{filename}{\code{ANY}}

\item{union_by_name}{\code{BOOLEAN}}

\item{hive_types}{\code{ANY}}

\item{hive_types_autocast}{\code{BOOLEAN}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{sniff_csv()}.
}
\examples{
\dontrun{
character(0)
}
}
