% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{split}
\alias{split}
\title{DuckDB function split}
\usage{
split(string = VARCHAR, separator = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\value{
\code{VARCHAR[]}
}
\description{
Splits the \code{string} along the \code{separator}.
}
\examples{
\dontrun{
split('hello-world', '-')
}
}
