% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{split}
\alias{split}
\alias{duckdb_fun_split}
\title{DuckDB function split}
\arguments{
\item{string}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\description{
Splits the string along the separator.
}
\examples{
\dontrun{
string_split('hello-world', '-')
}
}
