% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_AsGeoJSON}
\alias{ST_AsGeoJSON}
\title{DuckDB function ST_AsGeoJSON}
\usage{
ST_AsGeoJSON(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{JSON}
}
\description{
Returns the geometry as a GeoJSON fragment

This does not return a complete GeoJSON document, only the geometry fragment.
To construct a complete GeoJSON document or feature, look into using the DuckDB JSON extension in conjunction with this function.
This function supports geometries with Z values, but not M values. M values are ignored.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_AsGeoJSON('POLYGON((0 0, 0 1, 1 1, 1 0, 0 0))'::GEOMETRY);
}\if{html}{\out{</div>}}
}

