% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_AsMVT}
\alias{ST_AsMVT}
\title{DuckDB function ST_AsMVT}
\usage{
ST_AsMVT(col0, col1, col2, col3, col4)
}
\arguments{
\item{col0}{\code{ANY}}

\item{col1}{\code{VARCHAR}}

\item{col2}{\code{INTEGER}}

\item{col3}{\code{VARCHAR}}

\item{col4}{\code{VARCHAR}}
}
\value{
\code{BLOB}
}
\description{
Make a Mapbox Vector Tile from a set of geometries and properties
The function takes as input a row type (STRUCT) containing a geometry column and any number of property columns.
It returns a single binary BLOB containing the Mapbox Vector Tile.

The function has the following signature:

\verb{ST_AsMVT(row STRUCT, layer_name VARCHAR DEFAULT 'layer', extent INTEGER DEFAULT 4096, geom_column_name VARCHAR DEFAULT NULL, feature_id_column_name VARCHAR DEFAULT NULL) -> BLOB}
\itemize{
\item The first argument is a struct containing the geometry and properties.
\item The second argument is the name of the layer in the vector tile. This argument is optional and defaults to 'layer'.
\item The third argument is the extent of the tile. This argument is optional and defaults to 4096.
\item The fourth argument is the name of the geometry column in the input row. This argument is optional. If not provided, the first geometry column in the input row will be used. If multiple geometry columns are present, an error will be raised.
\item The fifth argument is the name of the feature id column in the input row. This argument is optional. If provided, the values in this column will be used as feature ids in the vector tile. The column must be of type INTEGER or BIGINT. If set to negative or NULL, a feature id will not be assigned to the corresponding feature.
}

The input struct must contain exactly one geometry column of type GEOMETRY. It can contain any number of property columns of types VARCHAR, FLOAT, DOUBLE, INTEGER, BIGINT, or BOOLEAN.

Example:

\if{html}{\out{<div class="sourceCode sql">}}\preformatted{SELECT ST_AsMVT(\{'geom': geom, 'id': id, 'name': name\}, 'cities', 4096, 'geom', 'id') AS tile
FROM cities;
}\if{html}{\out{</div>}}

This example creates a vector tile named 'cities' with an extent of 4096 from the 'cities' table, using 'geom' as the geometry column and 'id' as the feature id column.

However, you probably want to use the ST_AsMVTGeom function to first transform and clip your geometries to the tile extent.
The following example assumes the geometry is in WebMercator ("EPSG:3857") coordinates.
Replace \code{{z}}, \code{{x}}, and \code{{y}} with the appropriate tile coordinates, \verb{\{your table\}} with your table name, and \code{{tile_path}} with the path to write the tile to.

\if{html}{\out{<div class="sourceCode sql">}}\preformatted{COPY (
    SELECT ST_AsMVT(\{\{
        "geometry": ST_AsMVTGeom(
            geometry,
            ST_Extent(ST_TileEnvelope(\{z\}, \{x\}, \{y\})),
            4096,
            256,
            false
        )
    \}\})
    FROM \{your table\} WHERE ST_Intersects(geometry, ST_TileEnvelope(\{z\}, \{x\}, \{y\}))
) to \{tile_path\} (FORMAT 'BLOB');
```.
}\if{html}{\out{</div>}}
}
\section{Overloads}{

\itemize{
\item \code{ST_AsMVT(col0 = ANY)}
\item \code{ST_AsMVT(col0 = ANY, col1 = VARCHAR)}
\item \code{ST_AsMVT(col0 = ANY, col1 = VARCHAR, col2 = INTEGER)}
\item \code{ST_AsMVT(col0 = ANY, col1 = VARCHAR, col2 = INTEGER, col3 = VARCHAR)}
\item \code{ST_AsMVT(col0 = ANY, col1 = VARCHAR, col2 = INTEGER, col3 = VARCHAR, col4 = VARCHAR)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

