% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_AsSVG}
\alias{ST_AsSVG}
\title{DuckDB function ST_AsSVG}
\usage{
ST_AsSVG(geom, relative, precision)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{relative}{\code{BOOLEAN}}

\item{precision}{\code{INTEGER}}
}
\value{
\code{VARCHAR}
}
\description{
Convert the geometry into a SVG fragment or path

The SVG fragment is returned as a string. The fragment is a path element that can be used in an SVG document.
The second boolean argument specifies whether the path should be relative or absolute.
The third argument specifies the maximum number of digits to use for the coordinates.

Points are formatted as cx/cy using absolute coordinates or x/y using relative coordinates.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_AsSVG('POLYGON((0 0, 0 1, 1 1, 1 0, 0 0))'::GEOMETRY, false, 15);
-- M 0 0 L 0 -1 1 -1 1 0 Z
}\if{html}{\out{</div>}}
}

