% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{st_aswkb}
\alias{st_aswkb}
\title{DuckDB function st_aswkb}
\usage{
st_aswkb(geom = GEOMETRY)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BLOB}
}
\description{
Returns the Well-Known Binary (WKB) representation of the geometry.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{st_aswkb(ST_GeomFromWKB(X'01010000000000000000000000000000000000000000000000000'))
}\if{html}{\out{</div>}}
}

