% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Collect}
\alias{ST_Collect}
\title{DuckDB function ST_Collect}
\usage{
ST_Collect(geoms)
}
\arguments{
\item{geoms}{\code{GEOMETRY[]}}
}
\value{
\code{GEOMETRY}
}
\description{
Collects a list of geometries into a collection geometry.
\itemize{
\item If all geometries are \code{POINT}'s, a \code{MULTIPOINT} is returned.
\item If all geometries are \code{LINESTRING}'s, a \code{MULTILINESTRING} is returned.
\item If all geometries are \code{POLYGON}'s, a \code{MULTIPOLYGON} is returned.
\item Otherwise if the input collection contains a mix of geometry types, a \code{GEOMETRYCOLLECTION} is returned.
}

Empty and \code{NULL} geometries are ignored. If all geometries are empty or \code{NULL}, a \verb{GEOMETRYCOLLECTION EMPTY} is returned.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- With all POINT's, a MULTIPOINT is returned
SELECT ST_Collect([ST_Point(1, 2), ST_Point(3, 4)]);
}\if{html}{\out{</div>}}
}

