% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CoverageInvalidEdges}
\alias{ST_CoverageInvalidEdges}
\title{DuckDB function ST_CoverageInvalidEdges}
\usage{
ST_CoverageInvalidEdges(geoms, tolerance)
}
\arguments{
\item{geoms}{\code{GEOMETRY[]}}

\item{tolerance}{\code{DOUBLE}}
}
\value{
\code{GEOMETRY}
}
\description{
Returns the invalid edges in a polygonal coverage, which are edges that are not shared by two polygons.
Returns NULL if the input is not a polygonal coverage, or if the input is valid.
Tolerance is 0 by default.
}
\section{Overloads}{

\itemize{
\item \code{ST_CoverageInvalidEdges(geoms = `GEOMETRY[]`, tolerance = DOUBLE)}
\item \code{ST_CoverageInvalidEdges(geoms = `GEOMETRY[]`)}
}
}

