% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_CoverageUnion}
\alias{ST_CoverageUnion}
\title{DuckDB function ST_CoverageUnion}
\usage{
ST_CoverageUnion(geoms)
}
\arguments{
\item{geoms}{\code{GEOMETRY[]}}
}
\value{
\code{GEOMETRY}
}
\description{
Union all geometries in a polygonal coverage into a single geometry.
This may be faster than using \code{ST_Union}, but may use more memory.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

