% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{st_crs}
\alias{st_crs}
\title{DuckDB function st_crs}
\usage{
st_crs(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the Coordinate Reference System (CRS) identifier of the geometry.
}
\seealso{
Other geometry:
\code{\link[=st_astext]{st_astext()}},
\code{\link[=st_aswkb]{st_aswkb()}},
\code{\link[=st_geomfromwkb]{st_geomfromwkb()}}
}
\concept{geometry}
