% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Dimension}
\alias{ST_Dimension}
\title{DuckDB function ST_Dimension}
\usage{
ST_Dimension(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{INTEGER}
}
\description{
Returns the "topological dimension" of a geometry.
\itemize{
\item For POINT and MULTIPOINT geometries, returns \code{0}
\item For LINESTRING and MULTILINESTRING, returns \code{1}
\item For POLYGON and MULTIPOLYGON, returns \code{2}
\item For GEOMETRYCOLLECTION, returns the maximum dimension of the contained geometries, or 0 if the collection is empty.
}
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Dimension('POLYGON((0 0, 0 1, 1 1, 1 0, 0 0))'::GEOMETRY);
}\if{html}{\out{</div>}}
}

