% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Distance_Spheroid}
\alias{ST_Distance_Spheroid}
\title{DuckDB function ST_Distance_Spheroid}
\usage{
ST_Distance_Spheroid(p1, p2)
}
\arguments{
\item{p1, p2}{\code{POINT_2D}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the distance between two geometries in meters using an ellipsoidal model of the earths surface

The input geometry is assumed to be in the \href{https://en.wikipedia.org/wiki/World_Geodetic_System}{EPSG:4326} coordinate system (WGS84), with [latitude, longitude] axis order and the distance limit is expected to be in meters. This function uses the \href{https://geographiclib.sourceforge.io/}{GeographicLib} library to solve the \href{https://en.wikipedia.org/wiki/Geodesics_on_an_ellipsoid#Solution_of_the_direct_and_inverse_problems}{inverse geodesic problem}, calculating the distance between two points using an ellipsoidal model of the earth. This is a highly accurate method for calculating the distance between two arbitrary points taking the curvature of the earths surface into account, but is also the slowest.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Note: the coordinates are in WGS84 and [latitude, longitude] axis order
-- Whats the distance between New York and Amsterdam (JFK and AMS airport)?
SELECT st_distance_spheroid(
st_point(40.6446, -73.7797),
st_point(52.3130, 4.7725)
);
}\if{html}{\out{</div>}}
}

