% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Extent_Approx}
\alias{ST_Extent_Approx}
\title{DuckDB function ST_Extent_Approx}
\usage{
ST_Extent_Approx(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOX_2DF}
}
\description{
Returns the approximate bounding box of a geometry, if available.

This function is only really used internally, and returns the cached bounding box of the geometry if it exists.
This function may be removed or renamed in the future.
}
