% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_HasZ}
\alias{ST_HasZ}
\title{DuckDB function ST_HasZ}
\usage{
ST_HasZ(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{BOOLEAN}
}
\description{
Check if the input geometry has Z values.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- HasZ for a 2D geometry
SELECT ST_HasZ(ST_GeomFromText('POINT(1 1)'));
}\if{html}{\out{</div>}}
}

