% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_InterpolatePoint}
\alias{ST_InterpolatePoint}
\title{DuckDB function ST_InterpolatePoint}
\usage{
ST_InterpolatePoint(line, point)
}
\arguments{
\item{line, point}{\code{GEOMETRY}}
}
\value{
\code{DOUBLE}
}
\description{
Computes the closest point on a LINESTRING to a given POINT and returns the interpolated M value of that point.

First argument must be a linestring and must have a M dimension. The second argument must be a point.
Neither argument can be empty.
}
