% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MakeBox2D}
\alias{ST_MakeBox2D}
\title{DuckDB function ST_MakeBox2D}
\usage{
ST_MakeBox2D(point1, point2)
}
\arguments{
\item{point1, point2}{\code{GEOMETRY}}
}
\value{
\code{BOX_2D}
}
\description{
Create a BOX2D from two POINT geometries.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_MakeBox2D(ST_Point(0, 0), ST_Point(1, 1));
}\if{html}{\out{</div>}}
}

