% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MakeLine}
\alias{ST_MakeLine}
\title{DuckDB function ST_MakeLine}
\usage{
ST_MakeLine(start, end)
}
\arguments{
\item{start, end}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
\itemize{
\item \code{ST_MakeLine(geoms = `GEOMETRY[]`)}: Create a LINESTRING from a list of POINT geometries.
\item \code{ST_MakeLine(start = GEOMETRY, end = GEOMETRY)}: Create a LINESTRING from two POINT geometries.
}
}
\section{Overloads}{

\itemize{
\item \code{ST_MakeLine(geoms = `GEOMETRY[]`)}
\item \code{ST_MakeLine(start = GEOMETRY, end = GEOMETRY)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_MakeLine([ST_Point(0, 0), ST_Point(1, 1)]);
SELECT ST_MakeLine(ST_Point(0, 0), ST_Point(1, 1));
}\if{html}{\out{</div>}}
}

