% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MakePoint}
\alias{ST_MakePoint}
\title{DuckDB function ST_MakePoint}
\usage{
ST_MakePoint(x, y, z, m)
}
\arguments{
\item{x, y, z, m}{\code{DOUBLE}}
}
\value{
\code{POINT_2D | POINT_3D | POINT_4D}
}
\description{
\itemize{
\item \code{ST_MakePoint(x = DOUBLE, y = DOUBLE)}: Creates a GEOMETRY point from an pair of floating point numbers.

For geodetic coordinate systems, x is typically the longitude value and y is the latitude value.

Note that ST_Point is equivalent. ST_MakePoint is provided for PostGIS compatibility.
\item \code{ST_MakePoint(x = DOUBLE, y = DOUBLE, z = DOUBLE)}: Creates a POINT_3D.
\item \code{ST_MakePoint(x = DOUBLE, y = DOUBLE, z = DOUBLE, m = DOUBLE)}: Creates a POINT_4D.
}
}
\section{Overloads}{

\itemize{
\item \code{ST_MakePoint(x = DOUBLE, y = DOUBLE)}
\item \code{ST_MakePoint(x = DOUBLE, y = DOUBLE, z = DOUBLE)}
\item \code{ST_MakePoint(x = DOUBLE, y = DOUBLE, z = DOUBLE, m = DOUBLE)}
}
}

\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_AsText(ST_MakePoint(143.3, -24.2));
-- POINT (143.3 -24.2)
}\if{html}{\out{</div>}}
}

