% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MaximumInscribedCircle}
\alias{ST_MaximumInscribedCircle}
\title{DuckDB function ST_MaximumInscribedCircle}
\usage{
ST_MaximumInscribedCircle(geom, tolerance)
}
\arguments{
\item{geom}{\code{GEOMETRY}}

\item{tolerance}{\code{DOUBLE}}
}
\value{
\verb{STRUCT(center GEOMETRY, nearest GEOMETRY, radius DOUBLE)}
}
\description{
Returns the maximum inscribed circle of the input geometry, optionally with a tolerance.

By default, the tolerance is computed as \code{max(width, height) / 1000}.
The return value is a struct with the center of the circle, the nearest point to the center on the boundary of the geometry, and the radius of the circle.
}
\section{Overloads}{

\itemize{
\item \code{ST_MaximumInscribedCircle(geom = GEOMETRY)}
\item \code{ST_MaximumInscribedCircle(geom = GEOMETRY, tolerance = DOUBLE)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Find the maximum inscribed circle of a square
SELECT ST_MaximumInscribedCircle(
    ST_GeomFromText('POLYGON((0 0, 10 0, 10 10, 0 10, 0 0))')
);
}\if{html}{\out{</div>}}
}

