% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_MemUnion_Agg}
\alias{ST_MemUnion_Agg}
\title{DuckDB function ST_MemUnion_Agg}
\usage{
ST_MemUnion_Agg(col0)
}
\arguments{
\item{col0}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Computes the union of a set of input geometries.
"Slower, but might be more memory efficient than ST_UnionAgg as each geometry is merged into the union individually rather than all at once.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

