% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Multi}
\alias{ST_Multi}
\title{DuckDB function ST_Multi}
\usage{
ST_Multi(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Turns a single geometry into a multi geometry.

If the geometry is already a multi geometry, it is returned as is.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Multi(ST_GeomFromText('POINT(1 2)'));
}\if{html}{\out{</div>}}
}

