% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_NumGeometries}
\alias{ST_NumGeometries}
\title{DuckDB function ST_NumGeometries}
\usage{
ST_NumGeometries(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{INTEGER}
}
\description{
Returns the number of component geometries in a collection geometry.
If the input geometry is not a collection, this function returns 0 or 1 depending on if the geometry is empty or not.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

