% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Points}
\alias{ST_Points}
\title{DuckDB function ST_Points}
\usage{
ST_Points(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{GEOMETRY}
}
\description{
Collects all the vertices in the geometry into a MULTIPOINT.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{SELECT ST_Points('LINESTRING(1 1, 2 2)'::GEOMETRY);
}\if{html}{\out{</div>}}
}

