% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_RemoveRepeatedPoints}
\alias{ST_RemoveRepeatedPoints}
\title{DuckDB function ST_RemoveRepeatedPoints}
\usage{
ST_RemoveRepeatedPoints(line, tolerance)
}
\arguments{
\item{line}{\code{LINESTRING_2D}}

\item{tolerance}{\code{DOUBLE}}
}
\value{
\code{LINESTRING_2D | GEOMETRY}
}
\description{
\itemize{
\item \code{ST_RemoveRepeatedPoints(line = LINESTRING_2D)}, \code{ST_RemoveRepeatedPoints(line = LINESTRING_2D, tolerance = DOUBLE)}: Remove repeated points from a LINESTRING.
\item \code{ST_RemoveRepeatedPoints(geom = GEOMETRY)}, \code{ST_RemoveRepeatedPoints(geom = GEOMETRY, tolerance = DOUBLE)}: Returns the geometry with repeated points removed.
}
}
\section{Overloads}{

\itemize{
\item \code{ST_RemoveRepeatedPoints(line = LINESTRING_2D)}
\item \code{ST_RemoveRepeatedPoints(line = LINESTRING_2D, tolerance = DOUBLE)}
\item \code{ST_RemoveRepeatedPoints(geom = GEOMETRY)}
\item \code{ST_RemoveRepeatedPoints(geom = GEOMETRY, tolerance = DOUBLE)}
}
}

