% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_RotateY}
\alias{ST_RotateY}
\title{DuckDB function ST_RotateY}
\usage{
ST_RotateY(geom, radians)
}
\arguments{
\item{geom, radians}{Unspecified.}
}
\value{
Unspecified.
}
\description{
Rotates a geometry around the Y axis. This is a shorthand macro for calling ST_Affine.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Rotate a 3D point 90 degrees (π/2 radians) around the Y-axis
SELECT ST_RotateY(ST_GeomFromText('POINT Z(1 0 0)'), pi()/2);
}\if{html}{\out{</div>}}
}

