% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_Transform}
\alias{ST_Transform}
\title{DuckDB function ST_Transform}
\usage{
ST_Transform(box, source_crs, target_crs, always_xy)
}
\arguments{
\item{box}{\code{BOX_2D}}

\item{source_crs}{\code{VARCHAR}}

\item{target_crs}{\code{VARCHAR}}

\item{always_xy}{\code{BOOLEAN}}
}
\value{
\code{BOX_2D | POINT_2D | GEOMETRY}
}
\description{
Transforms a geometry between two coordinate systems

The source and target coordinate systems can be specified using any format that the \href{https://proj.org}{PROJ library} supports.

The third optional \code{always_xy} parameter can be used to force the input and output geometries to be interpreted as having a [easting, northing] coordinate axis order regardless of what the source and target coordinate system definition says. This is particularly useful when transforming to/from the \href{https://en.wikipedia.org/wiki/World_Geodetic_System}{WGS84/EPSG:4326} coordinate system (what most people think of when they hear "longitude"/"latitude" or "GPS coordinates"), which is defined as having a [latitude, longitude] axis order even though [longitude, latitude] is commonly used in practice (e.g. in \href{https://tools.ietf.org/html/rfc7946}{GeoJSON}). More details available in the \href{https://proj.org/en/9.3/faq.html#why-is-the-axis-ordering-in-proj-not-consistent}{PROJ documentation}.

DuckDB spatial vendors its own static copy of the PROJ database of coordinate systems, so if you have your own installation of PROJ on your system the available coordinate systems may differ to what's available in other GIS software.
}
\section{Overloads}{

\itemize{
\item \code{ST_Transform(box = BOX_2D, source_crs = VARCHAR, target_crs = VARCHAR)}
\item \code{ST_Transform(box = BOX_2D, source_crs = VARCHAR, target_crs = VARCHAR, always_xy = BOOLEAN)}
\item \code{ST_Transform(point = POINT_2D, source_crs = VARCHAR, target_crs = VARCHAR)}
\item \code{ST_Transform(point = POINT_2D, source_crs = VARCHAR, target_crs = VARCHAR, always_xy = BOOLEAN)}
\item \code{ST_Transform(geom = GEOMETRY, source_crs = VARCHAR, target_crs = VARCHAR)}
\item \code{ST_Transform(geom = GEOMETRY, source_crs = VARCHAR, target_crs = VARCHAR, always_xy = BOOLEAN)}
\item \code{ST_Transform(geom = GEOMETRY, target_crs = VARCHAR)}
\item \code{ST_Transform(geom = GEOMETRY, target_crs = VARCHAR, always_xy = BOOLEAN)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Transform a geometry from EPSG:4326 to EPSG:3857 (WGS84 to WebMercator)
-- Note that since WGS84 is defined as having a [latitude, longitude] axis order
-- we follow the standard and provide the input geometry using that axis order,
-- but the output will be [easting, northing] because that is what's defined by
-- WebMercator.

SELECT
    ST_Transform(
        st_point(52.373123, 4.892360),
        'EPSG:4326',
        'EPSG:3857'
    );
}\if{html}{\out{</div>}}
}

