% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_TransScale}
\alias{ST_TransScale}
\title{DuckDB function ST_TransScale}
\usage{
ST_TransScale(geom, dx, dy, xs, ys)
}
\arguments{
\item{geom, dx, dy, xs, ys}{Unspecified.}
}
\value{
Unspecified.
}
\description{
Translates and then scales a geometry in X and Y direction. This is a shorthand macro for calling ST_Affine.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- Translate by (1, 2) then scale by (2, 3)
SELECT ST_TransScale(ST_Point(1, 1), 1, 2, 2, 3);
}\if{html}{\out{</div>}}
}

