% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{ST_ZMFlag}
\alias{ST_ZMFlag}
\title{DuckDB function ST_ZMFlag}
\usage{
ST_ZMFlag(geom)
}
\arguments{
\item{geom}{\code{GEOMETRY}}
}
\value{
\code{UTINYINT}
}
\description{
Returns a flag indicating the presence of Z and M values in the input geometry.
0 = No Z or M values
1 = M values only
2 = Z values only
3 = Z and M values.
}
\section{Provided by}{

The \code{spatial} extension (\code{LOAD spatial;}).
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{-- ZMFlag for a 2D geometry
SELECT ST_ZMFlag(ST_GeomFromText('POINT(1 1)'));
}\if{html}{\out{</div>}}
}

