% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{starts_with}
\alias{starts_with}
\title{DuckDB function starts_with}
\usage{
starts_with(string = VARCHAR, search_string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\description{
Returns \code{true} if \code{string} begins with \code{search_string}.
}
\examples{
\dontrun{
starts_with('abc', 'a')
}
}
