% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{starts_with}
\alias{starts_with}
\title{DuckDB function starts_with}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\description{
Returns true if string begins with search_string.
}
\examples{
\dontrun{
starts_with('abc','a')
}
}
