% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{stats}
\alias{stats}
\title{DuckDB function stats}
\usage{
stats(expression = ANY)
}
\arguments{
\item{expression}{\code{ANY}}
}
\value{
\code{VARCHAR}
}
\description{
Returns a string with statistics about the expression. Expression can be a column, constant, or SQL expression.
}
\examples{
\dontrun{
stats(5)
}
}
