% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{stddev}
\alias{stddev}
\title{DuckDB function stddev}
\usage{
stddev(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the sample standard deviation.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sqrt(var_samp(x))
}\if{html}{\out{</div>}}
}

