% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{stddev_pop}
\alias{stddev_pop}
\title{DuckDB function stddev_pop}
\usage{
stddev_pop(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the population standard deviation.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sqrt(var_pop(x))
}\if{html}{\out{</div>}}
}

