% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{stddev_samp}
\alias{stddev_samp}
\title{DuckDB function stddev_samp}
\usage{
stddev_samp(x = DOUBLE)
}
\arguments{
\item{x}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Returns the sample standard deviation.
}
\examples{
\dontrun{
sqrt(var_samp(x))
}
}
