% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{storage_info}
\alias{storage_info}
\title{DuckDB function storage_info}
\usage{
storage_info(col0 = VARCHAR)
}
\arguments{
\item{col0}{\code{VARCHAR}}
}
\value{
Unspecified.
}
\description{
DuckDB function \code{storage_info()}.
}
