% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{str_split}
\alias{str_split}
\title{DuckDB function str_split}
\arguments{
\item{string}{\code{VARCHAR}}

\item{separator}{\code{VARCHAR}}
}
\description{
Splits the \code{string} along the \code{separator}.
}
\examples{
\dontrun{
str_split('hello-world', '-')
}
}
