% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{str_split_regex}
\alias{str_split_regex}
\title{DuckDB function str_split_regex}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{options}{\code{VARCHAR}}
}
\description{
Splits the \code{string} along the \code{regex}. A set of optional regex \code{options} can be set.
}
\examples{
\dontrun{
str_split_regex('hello world; 42', ';? ')
}
}
