% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strftime}
\alias{strftime}
\title{DuckDB function strftime}
\arguments{
\item{data}{\verb{DATE | TIMESTAMP | TIMESTAMP_NS | VARCHAR | TIMESTAMP WITH TIME ZONE}}

\item{format}{\verb{VARCHAR | DATE | TIMESTAMP | TIMESTAMP_NS | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{VARCHAR}
}
\description{
Converts a \code{date} to a string according to the format string.
}
\section{Overloads}{

\itemize{
\item \code{strftime(data = DATE, format = VARCHAR)}: Converts a \code{date} to a string according to the format string.
\item \code{strftime(data = TIMESTAMP, format = VARCHAR)}: Converts a \code{date} to a string according to the format string.
\item \code{strftime(data = TIMESTAMP_NS, format = VARCHAR)}: Converts a \code{date} to a string according to the format string.
\item \code{strftime(data = VARCHAR, format = DATE)}: Converts a \code{date} to a string according to the format string.
\item \code{strftime(data = VARCHAR, format = TIMESTAMP)}: Converts a \code{date} to a string according to the format string.
\item \code{strftime(data = VARCHAR, format = TIMESTAMP_NS)}: Converts a \code{date} to a string according to the format string.
\item \code{strftime(data = `TIMESTAMP WITH TIME ZONE`, format = VARCHAR)}: Converts a \code{date} to a string according to the format string.
\item \code{strftime(data = VARCHAR, format = `TIMESTAMP WITH TIME ZONE`)}: Converts a \code{date} to a string according to the format string.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{strftime(date '1992-01-01', '\%a, \%-d \%B \%Y')
}\if{html}{\out{</div>}}
}

