% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strftime}
\alias{strftime}
\title{DuckDB function strftime}
\arguments{
\item{data}{\code{DATE | TIMESTAMP | TIMESTAMP_NS | VARCHAR}}

\item{format}{\code{VARCHAR | DATE | TIMESTAMP | TIMESTAMP_NS}}
}
\value{
\code{VARCHAR}
}
\description{
Converts a \code{date} to a string according to the format string.
}
\examples{
\dontrun{
strftime(date '1992-01-01', '\%a, \%-d \%B \%Y')
}
}
