% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strftime}
\alias{strftime}
\title{DuckDB function strftime}
\arguments{
\item{data}{\code{DATE | TIMESTAMP | TIMESTAMP_NS | VARCHAR}}

\item{format}{\code{VARCHAR | DATE | TIMESTAMP | TIMESTAMP_NS}}
}
\value{
\code{VARCHAR}
}
\description{
Converts a \code{date} to a string according to the format string.
}
\section{Overloads}{

\itemize{
\item \code{strftime(data = DATE, format = VARCHAR)}
\item \code{strftime(data = TIMESTAMP, format = VARCHAR)}
\item \code{strftime(data = TIMESTAMP_NS, format = VARCHAR)}
\item \code{strftime(data = VARCHAR, format = DATE)}
\item \code{strftime(data = VARCHAR, format = TIMESTAMP)}
\item \code{strftime(data = VARCHAR, format = TIMESTAMP_NS)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{strftime(date '1992-01-01', '\%a, \%-d \%B \%Y')
}\if{html}{\out{</div>}}
}

