% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{string_split_regex}
\alias{string_split_regex}
\title{DuckDB function string_split_regex}
\arguments{
\item{string}{\code{VARCHAR}}

\item{regex}{\code{VARCHAR}}

\item{options}{\code{VARCHAR}}
}
\value{
\code{VARCHAR[]}
}
\description{
Splits the \code{string} along the \code{regex}. A set of optional regex \code{options} can be set.
}
\section{Overloads}{

\itemize{
\item \code{string_split_regex(string = VARCHAR, regex = VARCHAR)}
\item \code{string_split_regex(string = VARCHAR, regex = VARCHAR, options = VARCHAR)}
}
}

\examples{
\dontrun{
string_split_regex('hello world; 42', ';? ')
}
}
