% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strip_accents}
\alias{strip_accents}
\title{DuckDB function strip_accents}
\usage{
strip_accents(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Strips accents from \code{string}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{strip_accents('mühleisen')
}\if{html}{\out{</div>}}
}

