% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strip_accents}
\alias{strip_accents}
\title{DuckDB function strip_accents}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Strips accents from \code{string}.
}
\examples{
\dontrun{
strip_accents('mühleisen')
}
}
