% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strlen}
\alias{strlen}
\title{DuckDB function strlen}
\usage{
strlen(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{BIGINT}
}
\description{
Number of bytes in \code{string}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{strlen('🦆')
}\if{html}{\out{</div>}}
}

