% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strlen}
\alias{strlen}
\alias{duckdb_fun_strlen}
\title{DuckDB function strlen}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Number of bytes in string.
}
\examples{
\dontrun{
strlen('🦆')
}
}
