% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{strptime}
\alias{strptime}
\title{DuckDB function strptime}
\arguments{
\item{text}{\code{VARCHAR}}

\item{format}{\code{VARCHAR}}

\item{format-list}{\code{VARCHAR[]}}
}
\value{
\code{TIMESTAMP}
}
\description{
Converts the \code{string} text to timestamp according to the format string. Throws an error on failure. To return \code{NULL} on failure, use try_strptime.

Converts the \code{string} text to timestamp applying the format strings in the list until one succeeds. Throws an error on failure. To return \code{NULL} on failure, use try_strptime.
}
\section{Overloads}{

\itemize{
\item \code{strptime(text = VARCHAR, format = VARCHAR)}
\item \code{strptime(text = VARCHAR, `format-list` = `VARCHAR[]`)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{strptime('Wed, 1 January 1992 - 08:38:40 PM', '\%a, \%-d \%B \%Y - \%I:\%M:\%S \%p')
strptime('4/15/2023 10:56:00', ['\%d/\%m/\%Y \%H:\%M:\%S', '\%m/\%d/\%Y \%H:\%M:\%S'])
}\if{html}{\out{</div>}}
}

