% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_extract}
\alias{struct_extract}
\title{DuckDB function struct_extract}
\arguments{
\item{struct}{\code{STRUCT}}

\item{'entry'}{\code{VARCHAR | BIGINT}}
}
\value{
\code{ANY}
}
\description{
Extract the named entry from the STRUCT.
}
\section{Overloads}{

\itemize{
\item \code{struct_extract(struct = STRUCT, `'entry'` = VARCHAR)}
\item \code{struct_extract(struct = STRUCT, `'entry'` = BIGINT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{struct_extract(\{'i': 3, 'v2': 3, 'v3': 0\}, 'i')
}\if{html}{\out{</div>}}
}

