% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_has}
\alias{struct_has}
\title{DuckDB function struct_has}
\arguments{
\item{struct}{\code{STRUCT}}

\item{'entry'}{\code{ANY}}
}
\description{
Check if an unnamed STRUCT contains the value.
}
\examples{
\dontrun{
struct_has(ROW(3, 3, 0), 3)
}
}
