% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_indexof}
\alias{struct_indexof}
\title{DuckDB function struct_indexof}
\usage{
struct_indexof(struct = STRUCT, `'entry'` = ANY)
}
\arguments{
\item{struct}{\code{STRUCT}}

\item{'entry'}{\code{ANY}}
}
\value{
\code{INTEGER}
}
\description{
Get the position of the entry in an unnamed STRUCT, starting at 1.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{struct_indexof(ROW(3, 3, 0), 3)
}\if{html}{\out{</div>}}
}

