% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{struct_insert}
\alias{struct_insert}
\title{DuckDB function struct_insert}
\usage{
struct_insert()
}
\value{
\code{STRUCT}
}
\description{
Adds field(s)/value(s) to an existing STRUCT with the argument values. The entry name(s) will be the bound variable name(s).
}
\examples{
\dontrun{
struct_insert({'a': 1}, b := 2)
}
}
