% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{substring}
\alias{substring}
\title{DuckDB function substring}
\arguments{
\item{string}{\code{VARCHAR}}

\item{start}{\code{BIGINT}}

\item{length}{\code{BIGINT}}
}
\value{
\code{VARCHAR}
}
\description{
Extracts substring starting from character \code{start} up to the end of the string. If optional argument \code{length} is set, extracts a substring of \code{length} characters instead. Note that a \code{start} value of \code{1} refers to the first character of the \code{string}.
}
\section{Overloads}{

\itemize{
\item \code{substring(string = VARCHAR, start = BIGINT, length = BIGINT)}
\item \code{substring(string = VARCHAR, start = BIGINT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{substring('Hello', 2)
substring('Hello', 2, 2)
}\if{html}{\out{</div>}}
}

