% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{suffix}
\alias{suffix}
\title{DuckDB function suffix}
\usage{
suffix(string = VARCHAR, search_string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{search_string}{\code{VARCHAR}}
}
\description{
Returns \code{true} if \code{string} ends with \code{search_string}.
}
\examples{
\dontrun{
suffix('abc', 'bc')
}
}
