% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sum}
\alias{sum}
\title{DuckDB function sum}
\arguments{
\item{arg}{\code{DECIMAL | BOOLEAN | SMALLINT | INTEGER | BIGINT | HUGEINT | DOUBLE | BIGNUM}}
}
\value{
\code{DECIMAL | HUGEINT | DOUBLE | BIGNUM}
}
\description{
Calculates the sum value for all tuples in arg.
}
\section{Overloads}{

\itemize{
\item \code{sum(arg = DECIMAL)}: Calculates the sum value for all tuples in arg.
\item \code{sum(arg = BOOLEAN)}: Calculates the sum value for all tuples in arg.
\item \code{sum(arg = SMALLINT)}: Calculates the sum value for all tuples in arg.
\item \code{sum(arg = INTEGER)}: Calculates the sum value for all tuples in arg.
\item \code{sum(arg = BIGINT)}: Calculates the sum value for all tuples in arg.
\item \code{sum(arg = HUGEINT)}: Calculates the sum value for all tuples in arg.
\item \code{sum(arg = DOUBLE)}: Calculates the sum value for all tuples in arg.
\item \code{sum(arg = BIGNUM)}: Calculates the sum value for all tuples in arg.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sum(A)
}\if{html}{\out{</div>}}
}

