% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sum}
\alias{sum}
\title{DuckDB function sum}
\arguments{
\item{arg}{\code{DECIMAL | BOOLEAN | SMALLINT | INTEGER | BIGINT | HUGEINT | DOUBLE | BIGNUM}}
}
\value{
\code{DECIMAL | HUGEINT | DOUBLE | BIGNUM}
}
\description{
Calculates the sum value for all tuples in arg.
}
\section{Overloads}{

\itemize{
\item \code{sum(arg = DECIMAL)}
\item \code{sum(arg = BOOLEAN)}
\item \code{sum(arg = SMALLINT)}
\item \code{sum(arg = INTEGER)}
\item \code{sum(arg = BIGINT)}
\item \code{sum(arg = HUGEINT)}
\item \code{sum(arg = DOUBLE)}
\item \code{sum(arg = BIGNUM)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sum(A)
}\if{html}{\out{</div>}}
}

