% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sum}
\alias{sum}
\title{DuckDB function sum}
\arguments{
\item{arg}{\code{DECIMAL | BOOLEAN | SMALLINT | INTEGER | BIGINT | HUGEINT | DOUBLE | BIGNUM}}
}
\value{
\code{DECIMAL | HUGEINT | DOUBLE | BIGNUM}
}
\description{
Calculates the sum value for all tuples in arg.
}
\examples{
\dontrun{
sum(A)
}
}
