% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sum_no_overflow}
\alias{sum_no_overflow}
\title{DuckDB function sum_no_overflow}
\arguments{
\item{arg}{\code{INTEGER | BIGINT | DECIMAL}}
}
\value{
\code{HUGEINT | DECIMAL}
}
\description{
Internal only. Calculates the sum value for all tuples in arg without overflow checks.
}
\section{Overloads}{

\itemize{
\item \code{sum_no_overflow(arg = INTEGER)}
\item \code{sum_no_overflow(arg = BIGINT)}
\item \code{sum_no_overflow(arg = DECIMAL)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sum_no_overflow(A)
}\if{html}{\out{</div>}}
}

