% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{sumkahan}
\alias{sumkahan}
\title{DuckDB function sumkahan}
\usage{
sumkahan(arg = DOUBLE)
}
\arguments{
\item{arg}{\code{DOUBLE}}
}
\value{
\code{DOUBLE}
}
\description{
Calculates the sum using a more accurate floating point summation (Kahan Sum).
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{sumkahan(A)
}\if{html}{\out{</div>}}
}

