% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{switch}
\alias{switch}
\title{DuckDB function switch}
\arguments{
\item{key}{\code{K | MAP(K, V)}}

\item{map}{\code{MAP(K, V) | V}}

\item{value}{\code{V}}
}
\value{
\code{V}
}
\description{
Creates a switch statement similar to CASE WHEN/THEN.
}
\section{Overloads}{

\itemize{
\item \code{switch(key = K, map = `MAP(K, V)`)}: Creates a switch statement similar to CASE WHEN/THEN.
\item \code{switch(key = K, map = `MAP(K, V)`, value = V)}: Creates a switch statement similar to CASE WHEN/THEN.
\item \code{switch(key = `MAP(K, V)`, map = V)}: Creates a switch statement similar to CASE WHEN/THEN.
\item \code{switch(key = `MAP(K, V)`)}: Creates a switch statement similar to CASE WHEN/THEN.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{switch(x, map(\{1 : 1\}, default)
}\if{html}{\out{</div>}}
}

