% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{time_bucket}
\alias{time_bucket}
\title{DuckDB function time_bucket}
\arguments{
\item{bucket_width}{\code{INTERVAL}}

\item{timestamp}{\verb{DATE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}

\item{origin}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE | VARCHAR}}
}
\description{
Truncate TIMESTAMPTZ by the specified interval bucket_width. Buckets are aligned relative to origin TIMESTAMPTZ. The origin defaults to 2000-01-03 00:00:00+00 for buckets that do not include a month or year interval, and to 2000-01-01 00:00:00+00 for month and year buckets.
}
\examples{
\dontrun{
time_bucket(INTERVAL '2 weeks', TIMESTAMP '1992-04-20 15:26:00-07', TIMESTAMP '1992-04-01 00:00:00-07')
}
}
