% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{*}
\alias{*}
\title{DuckDB function *}
\arguments{
\item{col0}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | INTERVAL}}

\item{col1}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | INTERVAL}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | INTERVAL}
}
\description{
DuckDB function \verb{*()}.
}
\section{Overloads}{

\itemize{
\item ```*`(col0 = TINYINT, col1 = TINYINT)``
\item ```*`(col0 = SMALLINT, col1 = SMALLINT)``
\item ```*`(col0 = INTEGER, col1 = INTEGER)``
\item ```*`(col0 = BIGINT, col1 = BIGINT)``
\item ```*`(col0 = HUGEINT, col1 = HUGEINT)``
\item ```*`(col0 = FLOAT, col1 = FLOAT)``
\item ```*`(col0 = DOUBLE, col1 = DOUBLE)``
\item ```*`(col0 = DECIMAL, col1 = DECIMAL)``
\item ```*`(col0 = UTINYINT, col1 = UTINYINT)``
\item ```*`(col0 = USMALLINT, col1 = USMALLINT)``
\item ```*`(col0 = UINTEGER, col1 = UINTEGER)``
\item ```*`(col0 = UBIGINT, col1 = UBIGINT)``
\item ```*`(col0 = UHUGEINT, col1 = UHUGEINT)``
\item ```*`(col0 = INTERVAL, col1 = DOUBLE)``
\item ```*`(col0 = DOUBLE, col1 = INTERVAL)``
\item ```*`(col0 = BIGINT, col1 = INTERVAL)``
\item ```*`(col0 = INTERVAL, col1 = BIGINT)``
}
}

