% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{timezone}
\alias{timezone}
\title{DuckDB function timezone}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE | VARCHAR}}

\item{col1}{\verb{TIME WITH TIME ZONE | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the timezone component from a date or timestamp.
}
\examples{
\dontrun{
timezone(timestamp '2021-08-03 11:59:44.123456')
}
}
