% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_base}
\alias{to_base}
\title{DuckDB function to_base}
\arguments{
\item{number}{\code{BIGINT}}

\item{radix}{\code{INTEGER}}

\item{min_length}{\code{INTEGER}}
}
\value{
\code{VARCHAR}
}
\description{
Converts \code{number} to a string in the given base \code{radix}, optionally padding with leading zeros to \code{min_length}.
}
\examples{
\dontrun{
to_base(42, 16, 5)
}
}
