% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_base64}
\alias{to_base64}
\title{DuckDB function to_base64}
\usage{
to_base64(blob = BLOB)
}
\arguments{
\item{blob}{\code{BLOB}}
}
\value{
\code{VARCHAR}
}
\description{
Converts a \code{blob} to a base64 encoded string.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{to_base64('A'::BLOB)
}\if{html}{\out{</div>}}
}

