% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_binary}
\alias{to_binary}
\title{DuckDB function to_binary}
\arguments{
\item{string}{\code{VARCHAR}}

\item{value}{\code{BIGNUM | UBIGINT | BIGINT | HUGEINT | UHUGEINT}}
}
\description{
Converts the \code{string} to binary representation.

Converts the \code{value} to binary representation.
}
\examples{
\dontrun{
to_binary('Aa')
to_binary(42)
}
}
