% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_centuries}
\alias{to_centuries}
\title{DuckDB function to_centuries}
\arguments{
\item{integer}{\code{INTEGER | BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a century interval.
}
\section{Overloads}{

\itemize{
\item \code{to_centuries(integer = INTEGER)}
\item \code{to_centuries(integer = BIGINT)}
}
}

\examples{
\dontrun{
to_centuries(5)
}
}
