% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_centuries}
\alias{to_centuries}
\title{DuckDB function to_centuries}
\arguments{
\item{integer}{\code{INTEGER | BIGINT}}
}
\description{
Construct a century interval.
}
\examples{
\dontrun{
to_centuries(5)
}
}
