% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_days}
\alias{to_days}
\title{DuckDB function to_days}
\arguments{
\item{integer}{\code{INTEGER | BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a day interval.
}
\examples{
\dontrun{
to_days(5)
}
}
