% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_decades}
\alias{to_decades}
\alias{duckdb_fun_to_decades}
\title{DuckDB function to_decades}
\arguments{
\item{integer}{\code{INTEGER}}
}
\description{
Construct a decade interval.
}
\examples{
\dontrun{
to_decades(5)
}
}
