% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_hex}
\alias{to_hex}
\title{DuckDB function to_hex}
\arguments{
\item{string}{\code{VARCHAR}}

\item{value}{\code{BIGNUM | BIGINT | UBIGINT | HUGEINT | UHUGEINT}}

\item{blob}{\code{BLOB}}
}
\value{
\code{VARCHAR}
}
\description{
Converts the \code{string} to hexadecimal representation.

Converts the \code{value} to \code{VARCHAR} using hexadecimal representation.

Converts \code{blob} to \code{VARCHAR} using hexadecimal encoding.
}
\section{Overloads}{

\itemize{
\item \code{to_hex(string = VARCHAR)}
\item \code{to_hex(value = BIGNUM)}
\item \code{to_hex(blob = BLOB)}
\item \code{to_hex(value = BIGINT)}
\item \code{to_hex(value = UBIGINT)}
\item \code{to_hex(value = HUGEINT)}
\item \code{to_hex(value = UHUGEINT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{to_hex('Hello')
to_hex(42)
to_hex('\\xAA\\xBB'::BLOB)
}\if{html}{\out{</div>}}
}

