% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_hours}
\alias{to_hours}
\alias{duckdb_fun_to_hours}
\title{DuckDB function to_hours}
\arguments{
\item{integer}{\code{BIGINT}}
}
\description{
Construct a hour interval.
}
\examples{
\dontrun{
to_hours(5)
}
}
