% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_microseconds}
\alias{to_microseconds}
\title{DuckDB function to_microseconds}
\usage{
to_microseconds(integer = BIGINT)
}
\arguments{
\item{integer}{\code{BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a microsecond interval.
}
\examples{
\dontrun{
to_microseconds(5)
}
}
