% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_millennia}
\alias{to_millennia}
\title{DuckDB function to_millennia}
\arguments{
\item{integer}{\code{INTEGER | BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a millenium interval.
}
\section{Overloads}{

\itemize{
\item \code{to_millennia(integer = INTEGER)}
\item \code{to_millennia(integer = BIGINT)}
}
}

\examples{
\dontrun{
to_millennia(1)
}
}
