% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_milliseconds}
\alias{to_milliseconds}
\title{DuckDB function to_milliseconds}
\usage{
to_milliseconds(double = DOUBLE)
}
\arguments{
\item{double}{\code{DOUBLE}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a millisecond interval.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{to_milliseconds(5.5)
}\if{html}{\out{</div>}}
}

