% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_months}
\alias{to_months}
\title{DuckDB function to_months}
\arguments{
\item{integer}{\code{INTEGER | BIGINT}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a month interval.
}
\examples{
\dontrun{
to_months(5)
}
}
