% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_seconds}
\alias{to_seconds}
\title{DuckDB function to_seconds}
\usage{
to_seconds(double = DOUBLE)
}
\arguments{
\item{double}{\code{DOUBLE}}
}
\value{
\code{INTERVAL}
}
\description{
Construct a second interval.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{to_seconds(5.5)
}\if{html}{\out{</div>}}
}

