% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_timestamp}
\alias{to_timestamp}
\title{DuckDB function to_timestamp}
\usage{
to_timestamp(sec = DOUBLE)
}
\arguments{
\item{sec}{\code{DOUBLE}}
}
\description{
Converts secs since epoch to a timestamp with time zone.
}
\examples{
\dontrun{
to_timestamp(1284352323.5)
}
}
