% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{to_weeks}
\alias{to_weeks}
\alias{duckdb_fun_to_weeks}
\title{DuckDB function to_weeks}
\arguments{
\item{integer}{\code{INTEGER}}
}
\description{
Construct a week interval.
}
\examples{
\dontrun{
to_weeks(5)
}
}
