% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{transaction_timestamp}
\alias{transaction_timestamp}
\title{DuckDB function transaction_timestamp}
\usage{
transaction_timestamp()
}
\value{
\verb{TIMESTAMP WITH TIME ZONE}
}
\description{
Returns the current timestamp.
}
\examples{
\dontrun{
transaction_timestamp()
}
}
