% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{translate}
\alias{translate}
\title{DuckDB function translate}
\usage{
translate(string = VARCHAR, from = VARCHAR, to = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}

\item{from}{\code{VARCHAR}}

\item{to}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Replaces each character in \code{string} that matches a character in the \code{from} set with the corresponding character in the \code{to} set. If \code{from} is longer than \code{to}, occurrences of the extra characters in \code{from} are deleted.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{translate('12345', '143', 'ax')
}\if{html}{\out{</div>}}
}

