% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{translate}
\alias{translate}
\title{DuckDB function translate}
\arguments{
\item{string}{\code{VARCHAR}}

\item{from}{\code{VARCHAR}}

\item{to}{\code{VARCHAR}}
}
\description{
Replaces each character in string that matches a character in the from set with the corresponding character in the to set. If from is longer than to, occurrences of the extra characters in from are deleted.
}
\examples{
\dontrun{
translate('12345', '143', 'ax')
}
}
