% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{trim}
\alias{trim}
\title{DuckDB function trim}
\arguments{
\item{string}{\code{VARCHAR}}

\item{characters}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Removes any occurrences of any of the \code{characters} from either side of the \code{string}. \code{characters} defaults to \code{space}.
}
\examples{
\dontrun{
trim('    test  ')
trim('>>>>test<<', '><')
}
}
