% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{trunc}
\alias{trunc}
\title{DuckDB function trunc}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}

\item{col1}{\code{INTEGER}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
Truncates the number.
}
\examples{
\dontrun{
trunc(17.4)
}
}
