% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{trunc}
\alias{trunc}
\title{DuckDB function trunc}
\arguments{
\item{x}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}}

\item{col1}{\code{INTEGER}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | FLOAT | DOUBLE | DECIMAL | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT}
}
\description{
Truncates the number.
}
\section{Overloads}{

\itemize{
\item \code{trunc(x = TINYINT)}
\item \code{trunc(x = TINYINT, col1 = INTEGER)}
\item \code{trunc(x = SMALLINT)}
\item \code{trunc(x = SMALLINT, col1 = INTEGER)}
\item \code{trunc(x = INTEGER)}
\item \code{trunc(x = INTEGER, col1 = INTEGER)}
\item \code{trunc(x = BIGINT)}
\item \code{trunc(x = BIGINT, col1 = INTEGER)}
\item \code{trunc(x = HUGEINT)}
\item \code{trunc(x = HUGEINT, col1 = INTEGER)}
\item \code{trunc(x = FLOAT)}
\item \code{trunc(x = FLOAT, col1 = INTEGER)}
\item \code{trunc(x = DOUBLE)}
\item \code{trunc(x = DOUBLE, col1 = INTEGER)}
\item \code{trunc(x = DECIMAL)}
\item \code{trunc(x = DECIMAL, col1 = INTEGER)}
\item \code{trunc(x = UTINYINT)}
\item \code{trunc(x = UTINYINT, col1 = INTEGER)}
\item \code{trunc(x = USMALLINT)}
\item \code{trunc(x = USMALLINT, col1 = INTEGER)}
\item \code{trunc(x = UINTEGER)}
\item \code{trunc(x = UINTEGER, col1 = INTEGER)}
\item \code{trunc(x = UBIGINT)}
\item \code{trunc(x = UBIGINT, col1 = INTEGER)}
\item \code{trunc(x = UHUGEINT)}
\item \code{trunc(x = UHUGEINT, col1 = INTEGER)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{trunc(17.4)
}\if{html}{\out{</div>}}
}

