% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{~}
\alias{~}
\title{DuckDB function ~}
\arguments{
\item{input}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise NOT.
}
\section{Overloads}{

\itemize{
\item ```~`(input = TINYINT)``
\item ```~`(input = SMALLINT)``
\item ```~`(input = INTEGER)``
\item ```~`(input = BIGINT)``
\item ```~`(input = HUGEINT)``
\item ```~`(input = UTINYINT)``
\item ```~`(input = USMALLINT)``
\item ```~`(input = UINTEGER)``
\item ```~`(input = UBIGINT)``
\item ```~`(input = UHUGEINT)``
\item ```~`(input = BIT)``
}
}

\examples{
\dontrun{
~15
}
}
