% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{~}
\alias{~}
\title{DuckDB function ~}
\arguments{
\item{input}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise NOT.
}
\section{Overloads}{

\itemize{
\item \code{`~`(input = TINYINT)}
\item \code{`~`(input = SMALLINT)}
\item \code{`~`(input = INTEGER)}
\item \code{`~`(input = BIGINT)}
\item \code{`~`(input = HUGEINT)}
\item \code{`~`(input = UTINYINT)}
\item \code{`~`(input = USMALLINT)}
\item \code{`~`(input = UINTEGER)}
\item \code{`~`(input = UBIGINT)}
\item \code{`~`(input = UHUGEINT)}
\item \code{`~`(input = BIT)}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{~15
}\if{html}{\out{</div>}}
}

