% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{unbin}
\alias{unbin}
\title{DuckDB function unbin}
\usage{
unbin(value = VARCHAR)
}
\arguments{
\item{value}{\code{VARCHAR}}
}
\value{
\code{BLOB}
}
\description{
Converts a \code{value} from binary representation to a blob.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{unbin('0110')
}\if{html}{\out{</div>}}
}

