% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{unbin}
\alias{unbin}
\title{DuckDB function unbin}
\usage{
unbin(value = VARCHAR)
}
\arguments{
\item{value}{\code{VARCHAR}}
}
\description{
Converts a \code{value} from binary representation to a blob.
}
\examples{
\dontrun{
unbin('0110')
}
}
