% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{unhex}
\alias{unhex}
\title{DuckDB function unhex}
\usage{
unhex(value = VARCHAR)
}
\arguments{
\item{value}{\code{VARCHAR}}
}
\value{
\code{BLOB}
}
\description{
Converts a \code{value} from hexadecimal representation to a blob.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{unhex('2A')
}\if{html}{\out{</div>}}
}

