% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{unhex}
\alias{unhex}
\title{DuckDB function unhex}
\arguments{
\item{value}{\code{VARCHAR}}
}
\description{
Converts a \code{value} from hexadecimal representation to a blob.
}
\examples{
\dontrun{
unhex('2A')
}
}
