% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{unicode}
\alias{unicode}
\title{DuckDB function unicode}
\arguments{
\item{str}{\code{VARCHAR}}
}
\description{
Returns the unicode codepoint of the first character of the string.
}
\examples{
\dontrun{
unicode('ü')
}
}
