% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{union_extract}
\alias{union_extract}
\title{DuckDB function union_extract}
\usage{
union_extract(union = UNION, tag = VARCHAR)
}
\arguments{
\item{union}{\code{UNION}}

\item{tag}{\code{VARCHAR}}
}
\value{
\code{ANY}
}
\description{
Extract the value with the named tags from the union. NULL if the tag is not currently selected.
}
\examples{
\dontrun{
union_extract(s, 'k')
}
}
