% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{union_tag}
\alias{union_tag}
\alias{duckdb_fun_union_tag}
\title{DuckDB function union_tag}
\arguments{
\item{union}{\code{UNION}}
}
\description{
Retrieve the currently selected tag of the union as an ENUM.
}
\examples{
\dontrun{
union_tag(union_value(k := 'foo'))
}
}
