% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{union_value}
\alias{union_value}
\title{DuckDB function union_value}
\usage{
union_value()
}
\value{
\code{UNION}
}
\description{
Create a single member UNION containing the argument value. The tag of the value will be the bound variable name.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{union_value(k := 'hello')
}\if{html}{\out{</div>}}
}

