% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{unpivot_list}
\alias{unpivot_list}
\title{DuckDB function unpivot_list}
\usage{
unpivot_list()
}
\value{
\code{LIST}
}
\description{
Identical to list_value, but generated as part of unpivot for better error messages.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{unpivot_list(4, 5, 6)
}\if{html}{\out{</div>}}
}

