% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{upper}
\alias{upper}
\alias{duckdb_fun_upper}
\title{DuckDB function upper}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Convert string to upper case.
}
\examples{
\dontrun{
upper('Hello')
}
}
