% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{url_decode}
\alias{url_decode}
\title{DuckDB function url_decode}
\arguments{
\item{input}{\code{VARCHAR}}
}
\description{
Unescapes the URL encoded input.
}
\examples{
\dontrun{
url_decode('this\%20string\%20is\%2BFencoded')
}
}
