% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{url_decode}
\alias{url_decode}
\title{DuckDB function url_decode}
\usage{
url_decode(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Decodes a URL from a representation using Percent-Encoding.
}
\examples{
\dontrun{
url_decode('https\%3A\%2F\%2Fduckdb.org\%2Fwhy_duckdb\%23portable')
}
}
