% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{url_encode}
\alias{url_encode}
\title{DuckDB function url_encode}
\usage{
url_encode(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\value{
\code{VARCHAR}
}
\description{
Encodes a URL to a representation using Percent-Encoding.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{url_encode('this string has/ special+ characters>')
}\if{html}{\out{</div>}}
}

