% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{url_encode}
\alias{url_encode}
\alias{duckdb_fun_url_encode}
\title{DuckDB function url_encode}
\arguments{
\item{input}{\code{VARCHAR}}
}
\description{
Escapes the input string by encoding it so that it can be included in a URL query parameter.
}
\examples{
\dontrun{
url_encode('this string has/ special+ characters>')
}
}
