% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{url_encode}
\alias{url_encode}
\title{DuckDB function url_encode}
\usage{
url_encode(string = VARCHAR)
}
\arguments{
\item{string}{\code{VARCHAR}}
}
\description{
Encodes a URL to a representation using Percent-Encoding.
}
\examples{
\dontrun{
url_encode('this string has/ special+ characters>')
}
}
