% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{uuid_extract_timestamp}
\alias{uuid_extract_timestamp}
\title{DuckDB function uuid_extract_timestamp}
\usage{
uuid_extract_timestamp(uuid = UUID)
}
\arguments{
\item{uuid}{\code{UUID}}
}
\value{
\verb{TIMESTAMP WITH TIME ZONE}
}
\description{
Extract the timestamp for the given UUID v7.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{uuid_extract_timestamp('019482e4-1441-7aad-8127-eec99573b0a0')
}\if{html}{\out{</div>}}
}

