% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{variant_extract}
\alias{variant_extract}
\title{DuckDB function variant_extract}
\arguments{
\item{col0}{\code{VARIANT}}

\item{col1}{\code{VARCHAR | UINTEGER}}
}
\value{
\code{VARIANT}
}
\description{
DuckDB function \code{variant_extract()}.
}
\examples{
\dontrun{
character(0)
}
}
