% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{variant_typeof}
\alias{variant_typeof}
\title{DuckDB function variant_typeof}
\usage{
variant_typeof(input_variant = VARIANT)
}
\arguments{
\item{input_variant}{\code{VARIANT}}
}
\value{
\code{VARCHAR}
}
\description{
Returns the internal type of the \code{input_variant}.
}
\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{variant_typeof(\{'a': 42, 'b': [1,2,3]\})::VARIANT)
}\if{html}{\out{</div>}}
}

