% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{version}
\alias{version}
\title{DuckDB function version}
\usage{
version()
}
\value{
\code{VARCHAR}
}
\description{
Returns the currently active version of DuckDB in this format: v0.3.2	.
}
\examples{
\dontrun{
version()
}
}
