% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{version}
\alias{version}
\title{DuckDB function version}
\value{
\code{VARCHAR}
}
\description{
Returns the currently active version of DuckDB in this format: v0.3.2	.
}
\section{Overloads}{

\itemize{
\item \code{version()}
\item \code{version()}
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{version()
}\if{html}{\out{</div>}}
}

