% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{weekday}
\alias{weekday}
\title{DuckDB function weekday}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the weekday component from a date or timestamp.
}
\examples{
\dontrun{
weekday(timestamp '2021-08-03 11:59:44.123456')
}
}
