% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{weekofyear}
\alias{weekofyear}
\title{DuckDB function weekofyear}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the weekofyear component from a date or timestamp.
}
\examples{
\dontrun{
weekofyear(timestamp '2021-08-03 11:59:44.123456')
}
}
