% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{xor}
\alias{xor}
\title{DuckDB function xor}
\arguments{
\item{left}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}

\item{right}{\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}}
}
\value{
\code{TINYINT | SMALLINT | INTEGER | BIGINT | HUGEINT | UTINYINT | USMALLINT | UINTEGER | UBIGINT | UHUGEINT | BIT}
}
\description{
Bitwise XOR.
}
\section{Overloads}{

\itemize{
\item \code{xor(left = TINYINT, right = TINYINT)}: Bitwise XOR.
\item \code{xor(left = SMALLINT, right = SMALLINT)}: Bitwise XOR.
\item \code{xor(left = INTEGER, right = INTEGER)}: Bitwise XOR.
\item \code{xor(left = BIGINT, right = BIGINT)}: Bitwise XOR.
\item \code{xor(left = HUGEINT, right = HUGEINT)}: Bitwise XOR.
\item \code{xor(left = UTINYINT, right = UTINYINT)}: Bitwise XOR.
\item \code{xor(left = USMALLINT, right = USMALLINT)}: Bitwise XOR.
\item \code{xor(left = UINTEGER, right = UINTEGER)}: Bitwise XOR.
\item \code{xor(left = UBIGINT, right = UBIGINT)}: Bitwise XOR.
\item \code{xor(left = UHUGEINT, right = UHUGEINT)}: Bitwise XOR.
\item \code{xor(left = BIT, right = BIT)}: Bitwise XOR.
}
}

\section{SQL examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{xor(17, 5)
}\if{html}{\out{</div>}}
}

