% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{year}
\alias{year}
\title{DuckDB function year}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\value{
\code{BIGINT}
}
\description{
Extract the year component from a date or timestamp.
}
\examples{
\dontrun{
year(timestamp '2021-08-03 11:59:44.123456')
}
}
