% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-funs.R
\name{yearweek}
\alias{yearweek}
\title{DuckDB function yearweek}
\arguments{
\item{ts}{\verb{DATE | INTERVAL | TIMESTAMP | TIMESTAMP WITH TIME ZONE}}
}
\description{
Extract the yearweek component from a date or timestamp.
}
\examples{
\dontrun{
yearweek(timestamp '2021-08-03 11:59:44.123456')
}
}
